Implement support for menus in fragments.
This uses a little trick to be able to take care of rebuilding
the menu when the fragments change, when running on pre-HC
platforms.  When running on HC or better, it uses the new API
to invalidate the menu.

Also adds a little bit of a helper for using new menu APIs and
fixes the implementation of dump() to call up to the super class
dump() when running on HC.

And fix a few problems that came up when running on Donut.

Change-Id: I2420545632796b3199d5eec90aa3f3372ceedb90