Tweak view save/restore state so it will play well with list views.
We need to restore the state *after* the adapter has been set;
setting the adapter clears the state.  To do this, we move the
state restore from immediately after the view is created to after
we call Fragment.onActivityAttached().

Also introduced a new Fragment.onViewCreated() callback which is
nice for fragments that want to do some setup after onCreateView()
but allow for subclasses to override that method.  (ListFragment
I'm looking at you.)

Change-Id: Ib06d7b0208306f2e4a807917a5c1d99415d31203