Fix issue #5238823, #5100340 and #5249654.
5238823: java.lang.NullPointerException at...
...android.support.v4.view.ViewPager.pageRight(ViewPager.java:1335)

5100340	Leaking reallyStopped in FragmentActivity

5249654 ViewPager resets focus on layout when descendant of...
...fragment.getView() is focusable (FragmentPagerAdapter bug)

Also there was a bug in the "really stopped" thing that was
previously introduced -- fragments wouldn't stop their loaders
until hitting that state, but they wouldn't go into that state
if the fragment was deactivated or put on the back stack without
the containing activity really stopping.

To fix this, we now implement the really stop facility as
an actual state, using the ACTIVITY_CREATED state, so that we
can know when to do the really stop for both cases.

Change-Id: Ia9bd55f41ad30c3ce07f99b6ac6d9113b15fe46d