AccessibilityServiceInfo does not report properly the feedback type.
Since the feedback type is an integer the bits of which correspond to a given
feedback type, the AccessibilityNodeInfoCompat#feedbackTypeToString should return a
string representation of all feedback types. The current implementation
incorrectly assumes that the argument had only one flag set. Note that this
is a new ICS API and we would like to have this fixed.

bug:5486933

Change-Id: Id6299968d82c8f9916a52938c0d5bda1f57d3f96