Change the "start deferred" fragment API to "user visible hint" - Support lib edition
Allow a fragment to set a hint of whether or not it is currently user
visible. This will be used implicitly to defer the start of fragments
that are not user visible until the loaders for visible fragments have
run. This hint defaults to true.

Modify FragmentPagerAdapter to use the new APIs both in
frameworks/base and here.

Change-Id: Idb5bce8a9ddf484186942555cd3c581d833eba7c