Initial work on ActionBar support library
This library contains a compatibility ActionBar implementation and
a number of other resources and helpers for implementing apps that
support multiple versions of the Android platform.

Imported most, if not all, dependent classes and resources for Eclair
implementation.

There's still some work that needs to be done on resources (both
for organization, and getting the build working), as well as getting
Menus working properly.

ICS and JB libraries not implemented yet.

Base framework design by Adam Powell <adamp@google.com>.

Change-Id: I64cc3b5cabcbd638890af51d93e0c4fd62879225