Cache HTTP headers on disk.
This change makes the DiskBasedCache to also store HTTP response
headers on disk. When parsing a cache entry, these are restored and
forwarded to the response parser. This is needed as soon as the
response parser depends on any HTTP response header. If we don't do
this, the parsed response hitting the disk cache might be different
from the parsed response coming from the network.

Bug: 7273175
Change-Id: I414ce6309b05fe427153fd2c1c16a8122c0005f4