SDK only: improve SearchViewCompat
There is an issue with SearchView where if you use back to dismiss
it, the search text does not get cleared.  This change fixes it in
the support library by overriding the callback about the action
view being collapsed to also clear the search text.  (Note that this
will only be fixed in API 14 or later since this callback doesn't
exist on earlier versions.)

Also add a bunch of additional methods for performing API calls
on the SearchView.

Change-Id: I97241f165eb6076329ff7108c4596da72d02cb6c