Move support Menu interfaces + callbacks to v4
This moves the Menu* interfaces used for the ActionBar
into android-support-v4. This allows us to remove
ActionBarFragment and related hackery.

This commit just keeps us at the same level of functionality,
but is a cleaner implementation. A future feature would be
to enable the use of the support Menu when not using
ActionBarActivity.

Change-Id: I3de7695d4db0b9626f2bc9ed49ad0528146dc093
Signed-off-by: Chris Banes <chrisbanes@google.com>