Add support for media route provider services.
Media route provider services are services that publish media routes
to applications and to the system.  The registered media route
provider enumerates the installed services, connects to them,
and listens for route updates.

As part of this change, extended media route to use a set of
media route control intent filters to describe its capabilities
instead of just one.  This is necessary for matching uri-based
media control intents such as ACTION_PLAY.

Added support for routes with icons.

Change-Id: I8f00fac27914662bec3df0e5a212e81a9d306081