Add support for media route provider services.
Media route provider services are services that publish media routes
to applications and to the system.  The registered media route
provider enumerates the installed services, connects to them,
and listens for route updates.

As part of this change, extended media route to use a set of
media route control intent filters to describe its capabilities
instead of just one.  This is necessary for matching uri-based
media control intents such as ACTION_PLAY.

Added support for routes with icons.

Added support for querying information about the provider of
each route and the list of all available providers.

Fixed a bug in the system media route provider that resulted
in duplicate routes being created.

Ensure that routes are notified as being unselected before
being notifies as being removed.  The invariant here is that
an application should only perceive a route to be selected
between the time it was added / enabled and the time it
was removed / disabled.

Change-Id: If1a81ace4ef17643f219f7fed06a80855dc9a7e1