Support layout_margin* on DrawerLayout children
Child views of DrawerLayout now support margins. This allows
applications to offset both content or drawer views as needed.

Left and right margins are not permitted on drawer views and will be
ignored.

Vertical layout_gravity settings on drawer views are now supported as
well as drawer views with a height less than the full height of the
DrawerLayout.

Apps should note that introducing a top/bottom margin for a drawer
view will be less efficient in terms of fill rate than a drawer that
spans the full height.

Change-Id: Ia1a595ec50fa853039fa0ca287e4e3cf1e287047