Add a short delay for drawer peeking in DrawerLayout
Don't have a drawer peek right away, have a short delay to account for
other UI affordances that may be placed a bit too close to the
drawer's edge.

Add support methods to ViewDragHelper to check if any pointer has
crossed the helper's touch slop.

Add support methods to ViewDragHelper to check if an edge was touched.

Change-Id: I872612376c091deb21914c8aa43ce39a31aa4210