Flesh out the transport controller...  now transport mediator.
This introduces a bunch of formal interfaces to represent the
player and controller sides.  The TransportController, now called
TransportMediator, is now the central place responsible for
propagating the interactions between the player and on-screen
controls; as it does this it can also reflect this information
through the remote control client and back from the media button
receiver.

This simplifies code using the class, getting rid of various
connections they would otherwise need to supply.

Change-Id: I36121d96e98267554f044712f3b3b018cdc445bb