Fix ViewPager touch handling in scrolling parent
ViewPager at the moment does not handle being in a
vertically scrolling parent well. It's onTouchEvent
sets mIsDragging on every ACTION_DOWN.
This commit fixes that so that mIsDragging is only
set if the ACTION_MOVE is checked against the touch
slop and is in the correct direction.
I also added in a call to
requestDisallowInterceptTouchEvent().

Fixes 8636395.

Change-Id: I3e4d26209708b67b84da1ba8fe460eb45e11a3e2
Signed-off-by: Chris Banes <chrisbanes@google.com>