Add a way for apps to hint at the types of routes they want.
Some route providers may need additional information about the kinds
of routes to be discovered.  We define a concept of a Selector
to allow an application to specify these capabilities.
This concept is a bit new and not plumbed all the way through yet.
We will refine it as we build more of the UI bits on top.

Added APIs for starting active route scans.

Bug: 8175766
Change-Id: Ib2d2f7bd6cd0610edc8c1f68d879f386c340f682