ViewPager was not repositioning the PagerTitleStrip properly
Gmail had an artifact where animating the ViewPager in from the right caused
the title strip at the bottom to disappear at the end of the animation.
The problem was that the view pager was being redrawn at the end of the
animation (due to disabling the layer that was being animated), but the
title strip was mis-positioned off the right of the screen.

The root cause was ViewPager's layout process, which was not properly
repositioning the title strip during layout, when the scrollX field
changed.

Issue #9336914 ViewPager title strip not positioned correctly during layout

Change-Id: Idfcb70e2f28b98233fe6cb0e9b20bc2b6582ac1b