Fix Tab FragmentTransaction on ICS
At the moment the transaction given to onTabUnselected
is never committed. This is because each ft is created
within the TabWrapper, rather than globally within
the Action Bar.

Moving the ft handling up to the Action Bar means that
the correct behaviour happens: each Tab refers back to
a single ft.

https://code.google.com/p/android/issues/detail?id=58602

Change-Id: I71b6ef83d46cb72450400e5abf80f216d6811b43
Signed-off-by: Chris Banes <chrisbanes@google.com>