Use correct open/close content description strings.
The JD for the costructor says that the openDrawerContentDescRes
describes the "open action". This means that it should be set
when the drawer is closed. Same for the closeDrawerContentDescRes
that should be set when the drawer is opened.

However, the code in syncState, setDrawerIndicationEnabled, onDrawerClosed
and onDrawerOpened swapped the meaning of the two strings.

Bug: 11062863
Change-Id: I9df100f5520f992190d7c479a99ba5471b4a3822