Use a temp file during extract.
This is to avoid a problem when multiple start at the same time
on the same apk, the first was creating the file while the second
was trying to dexopt the empty new file. With this change, when
facing the race condition, the extraction will be made twice but
it should be of no consequence but loss of time.

Change-Id: I20d8d7b790de6debdba0c3a19f55e362f088ea81