Add color mode to PrintHelper.
Most of the time people pringing a bitmap would like the output
to be in color. Hence, when calling the platform print method
this class has to provode the corresponding color mode hint. This
change adds a color mode option to the PrintHelper class and
defaults to color. The client can override this behavior.

bug:11216402

Change-Id: Ie554e79c4b776a1582aa2b9915672bbf84519f49