recompute mSlideOffset in onLayout to avoid rounding errors
b/11432787 Opened conversation with David, got the account picker

We get a call to continueSettling() which correctly sets the left of the slide view to 1.
However, when the subsequent call to continueSettling() comes in the left position of the
slide view is 0.  The scroller tells us we should go to zero (i.e. mScroller.getCurrX()
returns zero).

Since both are zero dx is zero and because dx is zero we don't do mCallback.onViewPositionChanged.
Because we don't do that callback SlidingPaneLayout doesn't set mSlideOffset to zero.
So when the callback fires mSlideOffset is wrong (close to zero but not zero).

Change-Id: I558defba9d48ee428860d1e3e25fd918f9413857