Remove obsolete assumption in RecyclerView
Don't assume position is a proxy for identity unless we have another
positive signal that data hasn't changed. (This will continue to
evolve as we start adding more internal features.) For now, serve up a
scrap view that is either marked dirty or has the same position as
long as the types are correct.

Change-Id: Ida371f109a0c4fcc5898a567388ed0864fd4ea3c