RecyclerView API: create/onCreateViewHolder, bind/onBindViewHolder
Split the create/bind process for RecyclerView adapters into on*
methods to be implemented by the subclass and keep the 'raw' methods
final to perform internal bookkeeping. This ensures that the resulting
ViewHolders are consistent; an app may choose to preemptively create
views from an adapter that may be expensive to inflate and dump them
into a shared pool for later use.

Add a structureChanged bit to LayoutManager#layoutChildren. This gives
a coarse-grained hint to the LayoutManager that can be used to invoke
various fast paths if no new items have been added or removed and only
potential rebinds or relayouts of existing item views may be required.

Change-Id: Ia704bd22a465bc2f0b22e4de43ddfae956d51623