List Layout Manager
Layout Manager that can be attached to a RecyclerView and provide similar
functionality to android.widget.ListView.

This is work in progress but stable enough to start using. Keep in mind,
API may change.

So far, it supports:
 * Horizontal and vertical layouts
 * RTL layouts
 * Reverse layout (similar to RTL but also works vertical)
 * Stack from end (similar to AbsListView's stackFromBottom but in both
   orientations).
 * View recycling

It does NOT add dividers by default but you can provide them as ItemDecorators.

Change-Id: Ic13922bac938b40fd74fa2ed910cdd1c3ea735b7