fix undesired horizontal scroll when focus from fastlane to row.
We have code in addFocusables() to avoid focus to views
in padding area.  But when we switch from fast lane to
row, addFocusables() is not called,  instead hroziontalGridView.requestFocus()
is called, which will by default find the first item in horizontalgridview.
Overrides onRequestFocusInDescendant() so that we only returns items
not overlapping with padding area.

enforce "focus after descendant".

Renamed some internal methods to avoid possible conflicts
if recyclerview add methods later.

Change-Id: I24fd175d7fe48924793f2b12fc8ea146280dcb3b