fix undesired horizontal scroll when focus from fastlane to row.
We have code in addFocusables() to avoid focus to views
in padding area.  But when we switch from fast lane to
row, addFocusables() is not called,  instead hroziontalGridView.requestFocus()
is called, which will by default find the first item in horizontalgridview.
Overrides onRequestFocus() so that we only returns items
not overlapping with padding area.

remove the descendant focusability code since gridmanager should be always
focusing after descendant.

Renamed some internal methods to avoid possible conflicts
if recyclerview add methods later.

Change-Id: Iefd24a7f2a09341daeebe94736cfd4a5c62cd530