BaseCardView and ImageCardView for leanback library.
A BaseCardView defines regions that can be displayed based on view state
changes on the layout. Three regions are defined: main, info, and extra,
and info and extra regions may become visible when the layout becomes
either activated or selected. Info must appear before or at the same
time as extra.

ImageCardView is a widget based on BaseCardView, defining a simple card
widget with a main image, a title, and a content string. Optionally the
card may also contain a small image badge. The text and badge are
contained in the info region, and there is no extra section.

Change-Id: I0cc497a7eaf434da3633ff307ab10d730731bbed