DO NOT MERGE DrawerLayout emits inconsistent accessibility tree.
DrawerLayout is hiding the content and closed drawer if one
is opened. However, the connection between the ignored child
and the drawer via the parent relation was not broken, thus
an inconsistent node three may be produced. An example of this
is an accessibility service holding a node to the content,
followed by opening a drawer which hides the content, and now
the service refreshing the node. In this case following the
parent relation one can get from the hidden content to the
drawer but the latter will not have the content as a child.

Change-Id: Ia0ffbd390fb5d073ec9907a948248eb288ee8480