GridLayoutManager: avoid redundant dispatchChildSelected
If it's fastRelayout,  we fire dispatchChildSelected only on mFocusPosition
changes,  this can avoid mutliple events fired by view size change animation.
If it's not fastRelayout,  it can be a structure change,  so we
always dispatchChild regardless mFocusPosition changes or not.

In scrollToView triggered not in layout,  there is no structure change
so it will dispatchChildSelected only if mFocusPosiiton changes.

b/14807080

Change-Id: I5ac770181542955360c4d0531d5110694db04b89