GridLayoutManager: Fix WindowAlignment algorithm
The current algorithm makes assumption that if the scrollCenter of a view
is before the middle point of window,  we will align the first item to
start edge.   This does not work if the first item is too large,  e.g. in
Search, the item's center position is after the middle point of window.

The fix is to add more flags for algorithm so if it's first item,  it should
always be aligned to start edge of window.

b/15547766

Change-Id: I81f8b68c1739b7df1fa08d94d3a23b6c9e495747