Remove hidden attribute.
The support lib doesn't support hidden attributes that are put into
themes. Apparently theme attributes can't be hidden and try to link
against the hidden attribute (which fails).

This isn't different anywhere anyways, so I just removed the
attribute and inlined the value in the only layout it's used in order
to fix the SDK build.

Change-Id: I4c06dba1746f4d50f7bd48c474bd7001ad54ff77