Refactor appcompat themes
The old structure was created when we had to have
aliases for v14+ themes to use the framework action
bar. As we're now using the unbundled toolbar everywhere
we need a new structure which allows theme inheritance,
rather than a single parent.

Added a THEMES.txt which explains the new layout.
This new structure should make it loads easier
to theme widgets which use new attributes (i.e.
editTextBackground in v11).

Change-Id: Iafc588610256ba6feff4327e6dfbf75bf1466e39