Revert "Suspected cause of regression affecting Timely "
This reverts commit 634291e533ebdbf683b4349eb551a83213140cc3.

Thanks to Sblitz, xav, adamp, vsiva, deepanshu for 
helping getting to the bottom of this. It looks like 
this was just resource id skew caused by a app build 
problem which seems to have been caused by prebuilts 
for the v7 support libraries not being updated 
correctly. 

At this point we don't know of any changes that 
are needed to the original CL or to the 
weight-enabled version of GridLayout in L. 

Reverting the revert.

Change-Id: I0e052a3707207e15f6dbf8b94225cb64dd2321da