Use unique lists to hold new animations spawned
Previously all new move/change/add animations would be placed on a global
list to avoid starting. That meant that a later request for unrelated animations
would populate the same lists and would thus potentially start at the wrong
time (coincident with the original animations instead of with the new
animation requirements).

This change changes the list instances to be lists-of-lists. When animation runnables
are first created, new lists are created and populated and then these new lists are
added to the list-of-list instances. Then the later runnables will act only on the
unique list associated with the proper animation event. Meanwhile, there is still global
access to all animations in case any of the animations need to get canceled.

Issue #16550053 DefaultItemAnimator should run unique animation lists

Change-Id: I87206da608eb4205405d9b99879c38d018ad6feb