DetailRow: Dont focus to button during transition
fix No.2 of b/16900335.

This is not stricly following the ux spec because running
the fading animation of parent viewgroup and a background fading
animation at the same time is really slow.  The button fading
is started after activity transition is done.

Also clean up and removed ChangeBounds workaround, (reparent) is no longer
used.  Moved classes to transition package, it is used by both widget
and app package.

Change-Id: I00463642843b3f3ef9c9845de451712d65745a89