GridLayoutManager: fix selection offset problem
When user calls "setSelection" *before* add items
to the adapter.  The mFocusPosition is already set to
the position,  then adding more items will incorrectly
add offsets to not yet created item.  We should not
apply pending offset in such situtation.

Ideal fix is that: user should add all items to the adapter
*before* set adapter on the view.

b/17331101

Change-Id: Ic07b68874c3742e6d29275a1c5921fa8afc08011