Fix for b/16400888 - API Review: GridLayout attributes
Column weights specify the relative proportions of space distribution
within a *row*. With that in mind, I flipped the vertical and horizontal
axes in the docs - but that was a mistake. This CL flips the docs back.
The code works the right way around so doesn't need to be modified.

Change-Id: I129788edc8099eca05c524683aac0de9a08cfdd7
(cherry picked from commit 9877a1ad5c6b1705afe76bf61ad875553474a6a8)