Fix List Menu for Toolbar Action Bars
The problem is that PhoneWindow keeps hold of
the first menu returned from onCreatePanelView(). The update
of that is controlled by PanelStateFeature.refreshDecorView flag
is rarely set, so the new view returned by onCreatePanelView()
is never actually used.

This CL works around it by keeping one ListMenuPresenter (and thus
one MenuView) and passing it to each ToolbarActionBar. It can
then update the view as needed.

BUG: 17049242

Change-Id: Iedaa93c1dfc0068e4ce588aee28548949f9c4d15