Add kill-switch for AppCompat's compatible widget styles
Added a new theme attribute and ActionBarActivity method
to enabled/disable the default widget styling. This does
not completely disable the tinting functionality, just disables
the default styles (which are used by the tinting functionality).

This means that you can still manually set a
Widget.AppCompat.Blah style on a compatible widget and
still get tinting, which is needed for the internal views.

BUG: 18050902
Change-Id: Iaee8ff5eb2e82106a07e02fbb78e477b0535d49e