Workaround bug with EventLog.writeEvent + Spannables DO NOT MERGE
Activity.onMenuItemSelected() calls EventLog.writeEvent()
with the menu item's condensed title. If the title is a
non-String CharSequence then writeEvent will throw an exception.

This was fixed in Activity for JB-MR2 so we need to workaround it
for devices before that.

BUG: 18072335
Change-Id: I6787170eabb2f840caa5e4e97663d835e06688b3
(cherry picked from commit 937df84589168e159706463f25780f35ec63b750)