Compute scroll offset bugfixes
This CL fixes two related scroll offset calculation bugs in LayoutManagers.

LinearLayoutManager was not checking if View is visible when finding start and end
children. This creates a problem if extra space is not 0. This CL changes that code
to check item visibility and also prefer fully visible item if smooth scrolling
is disabled. Not doing so will create problems when top of the list is reached.

StaggeredGridLayoutManagerw was already checking item visibility but it was not
falling back to a partially visible child if none of the children is fully visible.
This would jump offset to 0 as no children can be found. Now, it accepts a partially
visible child if none of the children is fully visible.

Bug: 18314430
Bug: 18344415

Change-Id: Ie2710cb10289cad91af1901e7d15a6a1a96040c4