Improve GridLayout's weight calculations
This CL improves how excess space is distributed in GridLayout. Previously,
GridLayout would find unused space in each axis and distribute it among
the Views by weights. It would re-run the constraint solver after this step.

This was performing properly when Views are in different
columns/rows because total remaining space would not be consumed, leaving the last
span in each axis with more space. (in BF solver, remaining space
goes to the last span of the axis).

This CL changes the weight distribution to use Newton method to find the max
distributed space that will not invalidate the constraints. It runs the solver
log(<axis size> * <number of Views>) times by different distributed space
guesses until the MAX is found. Because it is logarithmic,
it does not bring any significant performance cost if the number of views is
too many.

As a side effect, thhis CL also fixes a bug in GridLayout where remaining space
was distributed only in primary direction.

Bug: 17485996
Change-Id: Ie549e8ae4fd93a2f417db22f9a087f46a129f6e0