Add generic scroll support to RecyclerView
This CL adds support for scrolling RecyclerView with a separate input
device (e.g. mouse).

AbsListView handles generic motion scroll as an instant scroll event,
so does RecyclerView (instead of smooth scrolling).

Since most of the necessary APIs are added on Froyo or Honeycomb, this
CL adds InputDeviceCompat to access constants in InputDevice.

It also extends the scope of MotionEventCompat to suport axis and source.

Bug: 18796745
Change-Id: I8f12d452d39b27bf2b2da425d57e87b09ac1bf26