Disable predictive animations when span count changes
When span count is updated, we cannot query SpanSizeLookup
in pre-layout because it will assume it has the latest
span size. With this CL, we disable predictive animations
until span size change is applied in a layout pass.

This can be further improved by passing span size to
SpanSizeLookup#getSpanSize to support predictive
animations

Change-Id: I51373343a6fbd24442641b142d376afed28c28d9
Bug: 19356551