Notify listeners when resetting scroll in ViewPager.
This fixes a bug in ViewPager on small (wearable) screens, where if
you swipe very gently, it doesn't retract fully. It retracts almost to
the end, but stops a pixel (or 2 or 3) before reaching the end.

When retracting if a touch is detected, the scroll state is immediatele
set to 0. The listeners, however, are not informed about this event and
they state in the state delivered to them by the last scrolling update.

Change-Id: Ief0d40ba04a73c71ab6767d4201cabdadf8d1fb8