Add the support lib for VD and AVD
By using auto namespace, we have all the functionality we want for support lib
of VD and AVD.

All the java files in the support lib are based upon the framework ones, with
slight modification.

The test cases are also from framework test, with namespace changed.
Note that the animator part are still using android namespace since they are
supported in v11.

TODO:
1 integrate the AAPT support to use android namespace.
2 find the best way to integrate the framework VD AVD for L+.

b/19200714

Change-Id: I10a80a89d26237ea780368e19929d83cd7678fff