Fix missing dispatchChildSelected() for staggered grid
This bug does not happen to single row grid where we
change mFocusPosition before createItem.

For staggered grid,  when scroll up to item 1: we first
create item1, then consumePendingMovesAfterLayout() changes
mFocusPosition to 1; the event is not fired in createItem
and it's not fired in smoothscroller onStop() if mPendingMoves
is 0.

Bug 19620864

Change-Id: If90fa1b11f7ba689015ecd5462482b83254f345b