Update annotation retention for resource type annotations
The resource type annotations in the support library should use
class file retention. That's more compact than having to use external
annotations for these. (IntDef and StringDef on the other hand should
just use source retention, since the class file format for annotations
only allows storing the values of the constants they refer to, not
references to the constants.)

Also allows the resource type annotations to be applied to local
variables as requested in
150475: Update Res support annotations to permit LOCAL_VARIABLE
        annotations

Change-Id: I29bdfe3c258b349bd6c2a89c41b2bb76454d0ed4