Inflate RV attr to create the layout manager.
Make RecyclerView more toolable. Add a custom attribute (layoutManager)
which accepts the class name of the LayoutManager to be used. The
LayoutManager specified should have a default constructor or a
constructor with the signature: (android.content.Context,
android.util.AttributeSet, int, int).

Example usage:
<android.support.v7.widget.RecyclerView
  android:layout_width="match_parent"
  android:layout_height="match_parent"
  res_auto:layoutManager="LinearLayoutManager"/>

<android.support.v7.widget.RecyclerView
  android:layout_width="match_parent"
  android:layout_height="match_parent"
  res_auto:layoutManager=".CustomLayoutManager"/>

(res_auto is is the namespace prefix for
http://schemas.android.com/apk/res-auto)

Bug: 19780401
Change-Id: I46b26fce28c5b79b3a1e4321cf5c1e54ad089a5e