Fix find first/last child logic for big views
StaggeredGridLayoutManager was assuming all children all smaller
than RecyclerView when finding first/last visible children.
This is a bad assumption and breaks compatibility with swipe refresh
layout.

This CL fixes that bug and also extends unit test to cover that
case.

Previous tests was not calling the global method which is why
the issue is missed.

Bug: 19496362
Change-Id: I7fe021216af8d544318c06b80ebb51379a019102