Fix SGLM child position bug
Looks like SGLM had a bug where it would put children with gaps
when filling above and below anchor line. This would mostly happen
if clip to padding is false because it was not properly handling
clip to padding.

This CL changes how SGLM handles recycling in a long with fill state
preparetion. This also makes it 100% compatible with clip to padding
false.

Bug: 20565830
Change-Id: Iae9a5bd50865e552120f85c8f511986c8e9f4395