Theme.AppCompat.DayNight working back to API v14
This CL enables the use of -night and -notnight resource
qualifiers, and thus allowing Theme.AppCompat.DayNight
to work on those API levels.

TwilightCalculator and TwilightManager are simplified versions
of those internally in the framework. Since we don't want to
force the use of location permissions, we have a fallback of
hardcoded 'night' time.

BUG: 21119724

Change-Id: I48ff5f0ea0c02ba60d3c2979aa9fa2997080de37