Allow FragmentActivity to hook in to Factory2 method
Since Activity will automatically install itself as a
private factory on HC+, we need to ensure that FragmentActivity
also hooks in as a Factory2. Currently, if FragmentActivity's
Factory is *not* called, the framework FragmentManager will
attempt to inflate a support Fragment and crash.

Also removed some AppCompat code which is now uneeded.

BUG: 21287184
Change-Id: Id163caa3ec46654e25706e58415d4aa7b0213dd1