Add previewsdk support library
This tiny library ties together the build-time constants used to
populate build.prop for devices and checks for the
Build.VERSION.PREVIEW_SDK_INT value to ensure binary compatibility
across prerelease builds with breaking API changes. This allows apps
to easily check if their assumptions about the SDK they built with
hold on the current device. If not, the app can fall back to the
functionality available in the previously released public API level.

Change-Id: I3d1975fe9483ca6c4f4e24e08844052be314459b