Hiding removed child views from accessibility.
Accessibility code won't know that hidden child views of RecyclerView
exist. Such views exist during animations that collapse deleted child
views.

Implemented with setImportantForAccessibility. There was an attempt
to implement this using AccessibilityNodeProvider, but after some
soul-searching with accessibility pundits, it was decided that
AccessibilityNodeProvider doesn't quite match this scenario, so I
fell back to the 'mark as important' approach.

Bug: 20055027
Change-Id: Ibb8478ed85990248821ccb1d772b4f2f9c02aa74