Pass generic motion event to scrollInternal
We had a crash where if RecyclerView is used inside a
nested scrolling and event is a generic motion event;
we would get a NPE while trying to access a non-existing
event.
This CL makes a change to check for null events to avoid
the NPE and also starts passing motion event to scroll
internal so that generic motion events work as expected
as well.

Bug: 21561903
Change-Id: Ic48a75d83f083691295e59910fa14de1c6781fd8