RecyclerView: Add payload for efficient bind
Add an optional payload in onBindViewHolder and notifyItemChanged.
The payload can be used to indicate the partial change
of an item so that onBindViewHolder can do an efficient rebinding.

We are going to add in-place Change animation in next CL. The in-place
Change animation happens on the existing holder, Adapter will receive
a payload in onBindViewHolder.

The existing crossfading CHANGE involves two viewHolders,  the oldViewHolder
is faded out without rebinding, the newViewHolder is binding to a
scrap view and faded in.

Change-Id: I117e5fe24d757dc89e68bb5860b5f4fb71bd9cf5