Prioritize delete animation over move animation
RecyclerView was running a move animation when an item
is deleted but also moved. We should run delete animation
instead to make sure proper callbacks arrive.

In the future, this can be changed but for now, it is not
LayoutManager's responsibility to move a deleted item so
occurrences of a "deleted item being moved" can be ignored.

Bug: 21639668
Change-Id: I5b1cfa37a7e5e7a2ef34d7cc76fbebc41907cf9a