Remove package private method access from LinearLayoutManager
This CL removes some package private access to RecyclerView from
LinearLayoutManager. This is useful when people try to customize
it.
We should deprecate getScrapList : ViewHolder API to avoid further
similar problems. LayoutManager should not be accessing the view
holder and this API is creating problems.

I also fixed a layout test which was going into recursion.

Bug: 21847010
Change-Id: Ibfbd5b817affd7c93aafc0ca6cd99f62313be556