Fixing a crash introduced by my previous CL.
I didn't notice that method deleteItem(View) is used via reflection.
Restoring its (unused) parameter, which is needed for reflection,
but keeping a better name deleteSelectedItems.

Actually, we have a problem that click handlers like this one are also
used as helper methods for other handlers, where View parameter is
unnecessary, but I don't think it's worth fixing, so let's pass View and
not use it.

Bug: 21643257
Change-Id: Ibc0cd919706156ddc00890f322e51cb0774bc26b