Fix AppBarLayout scrolling when scrolling view is small
Currently AppBarLayout will only start scrolling if the
the View starting the nested scroll is large enough to be
scrolled. This works fine when that view is a direct sibling
of AppBarLayout, but when it is not a direct sibling we should
use it's parent instead of the view directly when checking the
height.

BUG: 22789052
Change-Id: Ibedbcb29d2a5ee917ad0b6140ad4ce7ff6fd04f2