Cherrypick from mnc-preview-docs docs: Fixing javadoc formatting. Original Change-Id: I844d9b03985138b2cb7733038bf4042ddc3479bc
There was a major formatting problem in
ShareActionProvider.setShareHistoryFileName() (because of badly-formed
closing tags) and a minor one in setShareIntent() (an extra space in
the code-snippet-background color).

Chris also caught some substantive errors that I fixed (I'm guessing
because this Javadoc was copied from the Framework version of the
method and some changes needed to be made).

See first comment for doc stage location.

bug: 22857881
Change-Id: I844d9b03985138b2cb7733038bf4042ddc3479bc