Add a way of generating a CustomTabsSessionToken from an intent
Currently there is no way of obtaining a session token from an
intent on the browser side. We add  the session's ICustomTabsCallback
to the intent with CustomTabsIntent.Builder constructor but since
CustomTabsSessionToken don't have any public constructors and
ICustomTabsCallback is hidden we can't generate one using the IBinder
we got from the intent.

This change add a support library convenience call to do all this
behind the scenes.

b/23389835

Change-Id: Iff0b28c2f85d0400646517dacacfa947a6e556f7