Fix animation listener being replaced by HWLayer animation listener
Previously, as an optimization, a custom animation listener is created in
FragmentManager to put alpha animations on HW layer. Due to the
limitation of the View.Animation, only one listener can be set
on an Animation. As a result, this optimization caused user-defined
animation listeners to be replaced by the custom animation listener.
To fix this issue, the custom animation listener can now wrap around
any user-defined listener, so that listener events can be sent first
to our custom listener, and then to user-defined listener.

Bug: 23404179
Change-Id: Id939178f24e900f8be8de088438c77978963ce25