Cherrypick from mnc-docs docs: Fixed bad links in LinearLayoutManager Original Change-Id: Ie4bef05857a2a29fd4f5409b33eff04c2a7eaf84
(Cherrypicking per JF's instructions to get this into the live docs.)

One bad link was a link to a non-public field (which turns into
a bad link in the public Javadocs); the other was just a badly-coded
link.

Also slightly improved the @returns doc for getOrientation().

See first comment for doc stage location.

bug: 23689635
Change-Id: I2973d1cb3a6f723f2b1a8c8d811cb2775e3c1843