Fix TabLayout's flickering for good
The problem is due us not handling ViewPager's
page change listeners correctly. onPageSelected() can
be called while the ViewPager is still mostly in the
previous page. This means that the next onPageScrolled()
int will not match the selected page.

The fix is to skip updating the selected text, like we already
do with the indicator.

BUG: 22709728
Change-Id: Ic17a136b3352ae5c620f4295f852d5f9e5a53f92