[Custom Tabs] Add updateVisuals() in CustomTabsService
In ICustomTabsService.aidl, updateVisuals() is a generic API that
takes a bundle and an IBinder. The bundle needs to have the same structure as
a custom tabs VIEW INTENT, in terms of UI customizations.

In CustomTabsSession, a public method is created for updating the action
button for the current session with a new Bitmap icon and new content
description.

Change-Id: I6f30da7c524599342997987081838c73a54e4880