Hide custom offset circle when transferring nested scroll events
If an app has a custom offset for the start of their circle, and are not scaling it up
into existance, its because they are hiding it behind something.
That something might move when nested scrolling events are dispatched.
Hide the circle view when it gets back to its original hidden position.
Fixes b/24333860 Pull to refresh circle shows up unexpectedly in SDK 23

Change-Id: Ia87c1c650048ab1a6bc35641e875a2ea57c8b740