Respect a FloatingActionButton's visibility when anchored
Currently when a FAB is anchored to an AppBarLayout, it
controls the FAB's visibility (for the automatic seam
functionality). This meant that any user defined visibility
was ignored since we had no way of distinguishing what was
user defined or not.

This CL fixes that by recording what the user defined visibility
is, and only updates the FAB's visibility from the ABL if the user
has set it to be visible.

BUG: 24973851

Change-Id: Iee9e95a6eac551934844fbfdbd9ca8fe68bcb28b