MediaRouter: Check a route's supporting control category
If a media route doesn't support remote playback, it means that the
route is for live audio / video mirroring such as A2DP speaker, headset,
and mirroring display and the media route provider does not provide
playback information such as media metadata and playing status.
See the details : android.support.v7.media.MediaControlIntent

The previous code checked if metadata or playback state is not null
to see a route's playback control availability. It caused an unwanted
bug when having a mirroring route and (non-null) empty metadata or
playback state.

Bug: 24888113
Change-Id: I7bcbbbf0139e1c653705f8054e1a3264c387fba2