Add convenient GuidedStepFragment finish methods
finishGuidedStepFragments() would either finish activity
or pops all GuidedStepFragments if was launched on top
of other content.  This made it possible that fragment
does not care about whether it's launched in separate
activity or on top of other content when it handles "finish"
action.

popBackStackToGuidedStepFragment() would pop several
GuidedStepFragments from stack.

Also fixed a bug of losing focus to content below
GuidedStepFragment when switch to next action.

Change-Id: I4345bfe9e4dac73b915f0f6318ce19d90098b45a