Fix swipe to refresh arrow trimming and scaling issues for quick scrolls.
For fast scrolls refresh circle arrow would get stuck in a bad state because
we only called setStartEndTrim and setArrowScale only when
overscrollTop < mTotalDragDistance. So it was possible that the last update
for adjusted percentage was less than one.

This CL calls setStartEndTrim and setArrowScale for all values of overscrollTop.

Bug: 25311564
Change-Id: Iae7182b42c8f2c61a64e78c36e82f20ff5ea7403