Leanback: Force text alignment to start
This is annoying that if a TextView match_parent in width and is
detected as LTR language, it will by default align text to left
in RTL mode.   This doesn't affect BrowseRow header in RTL because
the BrowseRow Header is wrap_content and layout align the header
view to right in parent.

However the ImageCardView and GuidedAction is using match_parent
for TextView width,  aligns English to left in RTL.

The fix forces every TextView to align text to viewStart except Guidance
Text are aligned to viewEnd.

Change-Id: I0c155a5daf61c31e7b940686aaad9526e99c8317