Re-add DayNight functionality to AppCompat
There are now a number of themes with the prefix
Theme.AppCompat.DayNight which support automatic
light/dark switching based on the time.

There are now two main APIs to enable these themes
to work. AppCompatDelegate.setDefaultNightMode() allows
setting of the night mode globally within an app.
setLocalNightMode() allows overriding of the default
for the specfic host component instance.

Change-Id: If0349165fc9163deb2bab974ea63939edc00c6dd