Fix test case for PercentFrameLayout.
This is the same as b/26002719 that was fixed by ag/825775.
On GB and older devices FrameLayout has a bug that makes it
incorrectly treat margins when gravity is not specified.

Since the implementation of percent-based layouts is to
"convert" custom percent-based attributes to pixel-based
margins on the MarginLayoutParams and let the base classes
handle the rest of the logic, we need to work around bugs
in those base classes.

Change-Id: I072c0f9557b00759e7fec715ea50f7dc22f0345c