Ensure MediaDescriptionCompat retains Media URI on API 21-22
Media URI was added to the framework MediaDescription in API 23, but
to support it as part of MediaDescriptionCompat on all API levels, it
is important that the data remains even when the object is parceled or
converted to and from a framework object on API levels 21 and 22.

Using a custom key in the Bundle associated with MediaDescriptionCompat
allows the data to be saved and restored automatically, including cases
such as with MediaControllerCompat where the MediaDescription framework
objects are being passed as part of framework QueueItems.

BUG: 26058433
Change-Id: I5b019c41ef2716bf3b7ccc04b17d8953d3c499c1