Fix CollapsingToolbarLayout not drawing a title
Happens when used in a parent which is initially GONE. An example
is a fragment which is used as a drawer in a DrawerLayout.

CollapsingToolbarLayout uses isShown() to work out if it has
enough space to draw a title. isShown() doesn't work in this
situation since it crawls up the view hierarchy to check that
all ancestors are visible. This doesn't matter to us, so
we now just check if the dummy view is attached and visible.

BUG: 26554039
Change-Id: I77541023d2079cfb7d4f16602b30db3ba707ba67