Fix window insets for AppBarLayouts in ViewPager
Caused by the standard dispatchApplyWindowInsets()
not being appropriate for a horizontally scrolling
widget with full-height children. This is because
if child 0 consumes the insets, the rest of the children
will not receive the dispatch. To workaround this
we manually dispatch the original insets, not allowing
children to consume them from each other.

Also tidied up the window insets dispatching in CoL,
AppBarLayout and CollapsingToolbarLayout.

BUG: 22869860

Change-Id: I26544267f672b3b4d041d4ef8b29dd0d5f706455