mediarouter: Fix api-check rule.
We're asking doclava to generate stubs for v7.app, so we
must point it at the corresponding sources. This also includes
an API update because the previous API definition was incorrectly
generated. It had no access to source so it would lose all
@deprecated information.

It's probably an error to overlap API definitions like this, and
that should be cleaned up in a follow up change.

Change-Id: I71ee1aafd6aa5ad501651a5204a1d40ddad676d4