mediarouter: Fix api-check rule.
We're asking doclava to generate stubs for v7.app, so we
must point it at the corresponding sources. This also includes
an API update because the previous API definition was incorrectly
generated. It had no access to source so it would lose all
@deprecated information.

It's probably an error to overlap API definitions like this, and
that should be cleaned up in a follow up change.

(cherry picked from commit 65469b1fd471cba714505cacbda99ec82e27c0a4)

Change-Id: I8421fbfd1436606427d6317e4ae7016ed2f4de04