Ignore views that appear and disappear
This CL fixes a bug where if a view is added in a layout pass
and removed in a following layout pass, we would try to run
animate disappearance although it never appeared. RV already
handles the case if the view is added in pre-layout. This CL
fixes the case where it happens between different layout
passes, which might happen when the LM uses auto-measure.

Bug: 27457120
Change-Id: I4c044c2eb8ff23883e1522491c1e053645f96131