Leanback Support: Fixed the constant date range updates in DatePicker which also fixed animation jumps
When scrolling vertically across a datepicker field, the date ranges kept
being changed resulting in the scroller adapter to be constantly updated.
This in turn resulted in undesirable animation jumps when hitting the top
or bottom range values.
Now, the range is always kept up-to-date and is updated only when
needed.

Bug: 27527022
Change-Id: I8f2799a05c1432839ec8a034f4b85da30e2124fb