Special handle match parent children in GridLayoutManager
GridLayoutManager measures children first with their measure spec
then re-measures them to have same height. This was creating
larger than expected if the child is MATCH_PARENT and also has
decor offsets / margins.
This CL fixes it by adding insets to the very first measure step
of the child if it has MATCH_PARENT spec in the scroll direction.

This is not an issue in LinearLayoutManager or
StaggeredGridLayoutManager because they don't have any inter-child
alignment and always take insets into account.

Bug: 27479792
Change-Id: Ib5f897f67fe9dd569cc6fbf477bb2c1141bb4a54