Fix TextInputLayout clearing the EditText tint
When a TIL's error is cleared, it clears the ColorFilter
on the EditText, which should reset it tint back
the original color. On < API 21 this isn't currently
happening since we added a drawable state workaround,
which sets a new background drawable on the EditText.

This then clears out AppCompat's internal tint, and thus
the tint is lost. Fixed by only using the workaround on
the platforms it is needed: API 21 + 22.

BUG: 27611753
Change-Id: Ia91a8bfa1295bd66c926541f98cfe7e148971f62