Fix issues around setting text color
In AppCompat, we now reset the text color if we detect
a String reference. Since most platform styles declare
a string referenced color in their text appearance, this
means that the text appearance can override an explicit
android:textColor on the view.

Fixed by always resetting the text color on API < 23.

BUG: 27784645
Change-Id: Ic5a2c7e1fb49da419bbfd3c6176c7a28c6eb4ac1