Preserve focus after layout
LayoutManagers try to keep focused view visible during a layout
calculation. This does not work well if the View is being replaced
with another view due to type change and/or animation. It is
beyond LayoutManager's intended control to care for this case so
RecyclerView will track it and recover focus to the new View
if necessary.

Bug: 25510761
Change-Id: I65680577587839ecfe84a931998da8b9926b5566