Adding a test case to reproduce crash in recyclerview.
When we detach a recycler view, it sets
mFirstLayoutComplete = false in onDetachedFromWindow method.
Then if the user adds back the view and runs a smoothScrollBy
animation, we skip consumePendingUpdateOperations(). This results
in an exception if user removes data in the middle of an animation
(because we skipped processing changes to the adapter).

One way to fix it might be to not reset mFirstLayoutComplete = false
in onAttachedToWindow/onDetachedFromWindow call.

Bug: 27926253
Change-Id: I5436072b6a7039878222a74c0f9213184b29a4a1