Do not recycle visible views
This CL adds a new API to the LayoutManager where the LayoutManager
can get the bounds of a View while taking into account its
transformation matrix.
This API is mainly useful for recycle decision. I've also changed
built in layout managers to respect this value in addition to layout
position of the view.

Bug: 26402811
Change-Id: If75072c73d3374bfac0439d7714714c82684096a