Adds support for MessagingStyle in the Compat lib
This change adds support for the MessagingStyle in NotificationCompat
v4 (see ag/904266 for framework implementation).

This change also puts MessagingStyle extras into the Notification to
allow a NotificationListener compiled against N on a pre-N phone to
read a MessagingStyle set by an app compiled against N on a pre-N phone.

Fallback to BigText on API<24 tracked by b/28132121.

Bug:28044251
Change-Id: I38ef9a9fd59b929b644cd9bcca9941f0119f46b4