RecyclerView focus search improvements
This CL fixes a couple of bugs in RecyclerView's focus search
handling.
* Built in layout managers were handling focus FWD/BCK wrong if
the layout is RTL.

* RecyclerView was not letting the focus leave RecyclerView if
focus FWD/BCK is called and RV does not have more children in
that direction. It was going back to the top of the list instead
of letting the view system give focus to the next view outside RV.

* RecyclerView was losing focus when focus FWD/BCK is called and
RecyclerView has more children to layout but they are not laid
out. In this case, focus was escaping the RecyclerView because
it was not laying out new children in the focus direction because
it already has some valid candidates. In this case, RecyclerView's
candidates would be overridden by other views outside the
RecyclerView.

Bug: 27222294
Change-Id: Ia5583c21df0a354d53c013aa76c4f9b475240c1f