Fix CSLs with alpha attribute in AppCompat again
The original fix was to double set android:alpha
and app:alpha. This has unforeseen issues on API 21 + 22
since it reads android:alpha but expects it to a literal
color value (so it blows up with an attribute value).

This CL fixes it by adding a v23 override for each color,
setting app:alpha in color/ and android:alpha in color-v23.

BUG: 28334410
Change-Id: I65ff37b5ef94d0887b809f09e7330f3a76b94855