Add a custom tabs API to use RemoteViews in the secondary toolbar
This introduces three new extras and a single call that uses all
of them inside the Builder. The client can now add a remote view
hierarchy to be inserted into the secondary toolbar. With the
RemoteViews, the client can give a list of View IDs that will
have their onClick behavior overridden to get back the current
url of the custom tab as data.

BUG=27970505

Change-Id: Ia2d912e44f873619ce48c340699d888111d709df