BrowseFragment: allow page fragment to delay loading data
When first row of BrowseFragment is a page fragment,
startEntranceTransition() is called when BrowseFragment's adapter is
ready.   It is also when page fragment is just created from first
Row of BrowseFragment.   But page fragment can take another second
to load its own data,  so startEntranceTransition() cannot start
transition immediately,  it needs wait the current fragment's data
to be ready.

Added new API for page fragment to notify data is ready,  this doesn't
apply to fragment implemented RowsFragmentAdapter because in that case
the fragment is sharing the same adapter of BrowseFragment.

Bug 28471536

Change-Id: Ie9fd475868e0f7cc530b38769d31c0f71d1a2c6c