PlaybackOverlayFragment: fixed the wrong focus behavior after fade-in & fade-out
When a song starts playing, PlaybackOverlayFragment fades out
all the components. After a user pressed a DPAD button however,
the focus shifted to the media player
while the highlight selector still remained on that row.

Now after a fade-out, the focus remains at the same
location as it was if a media item row is clicked.
If the focus was on the media player before fade-out, the focus
resets to the play button after fade-in happens.

Bug: 28916752
Change-Id: I1584a8a96687ed36b740aa033f253f23c89364a9