Add autoHide to FloatingActionButton
Add a new attribute "behavior_autoHide" to
FloatingActionButton.Behavior. The FAB automatically hides when there is
not enough space for it if behavior_autoHide is set to true. It defaults
to true.  FAB.Behavior currently supports AppBarLayout and
BottomSheetBehavior.

This also fixes the issue that CoordinatorLayout stops sending
onDependentViewChanged after the child view reaches at an edge of the
CoordinatorLayout.

Bug: 28962378
Change-Id: Ic17bda6fc448f0c483752f1a3ed0df50fe60a458