Make some Compat classes have more coverage
Add missing methods and public fields to AccessibilityNodeInfoCompat and
AccessibilityWindowInfoCompat so that their interfaces are almost fully
compatible with the corresponding framework base classes.
The statement above says almost because these support lib classes do not
implement the Parcelable interface (this issue also applies to
PointerIconCompat) and so cannot serve as drop-in replacements in *ALL*
cases (in particular when parceling is used). Please see the discussion
on the bug linked below for a detailed explanation.

Bug: 29148805
Change-Id: I5b9406de1d26adc6f11a7ac24152ff3171c97e15