Support7Demos: Build metadata using the stored title in the playlist
Currently the media session retrieves the title of current media from
the provider by sending a private control intent. However, because the
title is already known, it is unnecessary to get this information from
the provider.

This CL stores title in the playlist in order to use it for building the
media metadata. Therefore it can show the title in the controller
dialog, even when the current route provider is not a sample provider.

Bug: 25732463
Change-Id: Ia9f859f02f45e5fed0bb6ede06c79b5bc98712cb