Change layout span to match SpanLookup span
There was an inconsistency in GridLayoutManager's API where the
getSpanIndex method in GLM.LayoutParams would always assign
leftmost span as span 0 (even if it RTL) while GLM.SpanLookup
would assign rightmost span as 0 if it is RTL (and leftmost if
it is LTR).

SpanLookup's API makes a lot more sense since it avoids checking RTL
parameter and we only need to check it when positioning views.

Changing the API docs and also fixing a bug that was
caused by this confusion.

Bug: 30143272
Change-Id: I8653c0ab390bfe28dd36e2c5560d656b0c658fe4