Improve GridLayoutTest
A GridLayoutTest was relying on the fact that a certain text
won't fit horizontally which breaks in landscape.

This CL limits the width of the container to ensure it does
not fit. Even this is flaky since it is using a TextView but
enough to fix the test in landscape for now.

Bug: 30206950
Change-Id: I88387a7ee58ff4585d5a0c16a9f6aff05e1c74a8