MediaRouter: Fix the unmatched color issue in MRCD
The theme of MediaRouteControllerDialog(MRCD) should be determined based
on the 'colorPrimary' value of the current dialog theme because the
value will be used as a background color. However, current
implementation retrieves 'colorPrimary' value from the current context
but not the dialog theme, causing the unmatched color issue between text
color and background color. This change pre-evaluates dialog theme to
make the MRCD have the right theme.

Bug: 25731047
Change-Id: I68b968fe0b03c2602f4c776d6c5682bc0c1b7b89