Invalidate decor offsets after rebound
RV invalidates decor offsets eagerly which creates a problem
if predictive animations are enabled. In that case, we will
be updating decor offsets before updating the item and
will not update after updating the item.

This CL changes the behavior such that we will use dirty
offsets in pre-layout if item is not re-bound yet and
will invalidate the decor offsets when it is bound which
will trigger a call to the item decoration to update the
offsets.

Bug: 29989333
Change-Id: Idc23f83be19164cf137fa81242e6c90c4e85fbae