Add a demo for InputConnectionCompat#commitContent
This CL demonstrates what changes need to be done in application to
support IMEs that can commit contents such as GIF image via
InputConnectionCompat#commitContent() support library API (or
InputConnection#commitContent() API added in API 25).

Although this sample uses WebView to show the image sent from the IME to
make the sample code as simple as possible, the key part here is that
app developers will receive content URIs and it is up to them (and their
applications) how receved contents need to be handled.

Bug: 29833914
Change-Id: Ib0f927b553a1f557f4cf81747ae1cd72e371a287