Make expanded title bounds start from top of Toolbar
Currently they're from the top of the CTL, which isn't
quite correct when the CTL is consuming insets.
Developers expect these bounds to be similar to the
collapsed bounds, therefore we will start them
from the top of the Toolbar.

This allows consistency of setting a top margin,
regardless of whether the view is using a transparent
status bar or not.

BUG: 31392796

Change-Id: I54dddfd038bce90acb2299d8b85958ac736ae16d