Support custom views on snackbars
Split the existing Snackbar class into two - the base class
with the common functionality for the lifecycle management
and action, and the extending class that "adds" a text message.

The extending class maintains the existing API surface.

This CL has sample code for a custom snackbar in SupportDesignDemo,
as well as unit tests in CustomSnackbarTest class.

Bug: 31658766
Test: ./gradlew support-design:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=android.support.design.widget.SnackbarTest
Test: ./gradlew support-design:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=android.support.design.widget.CustomSnackbarTest
Test: ./gradlew support-design-demos:assembleDebug; adb install -d -r ../../out/host/gradle/frameworks/support/support-design-demos/build/outputs/apk/support-design-demos-debug.apk
Change-Id: Ic042ad84f56645b48ad6a4d46e22a8e9352f5268