MediaRouter: Synchronize route selecting information
If the selected route of an application was a system media route such
as the device speaker, a wired headset, or a bluetooth speaker, and
the selected route is changed by other processes, no callback could
be called in API 24 and before, and once it happens, getSelectedRoute()
could return a wrong value and selectRoute() could be no-op.

This CL ensures that getSelectedRoute() always returns the currently
selected route and selectRoute() works.

Bug: 31921702
Test: Support7Demos on API 24
Change-Id: I0edcbe4d9f2fe42e0fc6ba78802320eaa24fe24d