Added focus recovery mechanism to RecyclerView
Fixed RV's focus issue in two cases: If RV received focus before or
during layout, the focus was not passed down to the children. The seocnd
case was when RV's prior focused child became unfocusable and animation
was disabled, RV didn't redeliver the focus to a child after layout is
complete.
Also, made a best-effort redelivery of focus to a child when the
previously focused item is no longer valid.
Added several tests for reproducing the bugs and verifying focus
recovery logic across different adapter change scenarios as well as adding
animation parameter for tests.

Fixes: 32280153
Fixes: 32486474
Fixes: 32227010
Test: ./gradlew support-recyclerview-v7:connectedCheck
-Pandroid.testInstrumentationRunnerArguments.class=android.support.v7.widget.RecyclerViewFocusRecoveryTest

Change-Id: Ibaff9ee6699efe4a6800faf0f34c701a24baafdc