Add RestrictTo scope for the same library
This will allow a single library to have public implementation classes
that can be referenced from other packages within the same library,
but without exposing them to other libraries in the same group.

Also added an alias for GROUP_ID called LIBRARY_GROUP since I think it
reads better than GROUP_ID in code, and pairs better with
LIBRARY. (The alternative was to use ARTIFACT_ID for the new library
scope, but that didn't seem right (and we really want to match *both*
group id and artifact id, not just the latter.)

Test: The new annotation scope and the aliasing of GROUP_ID and
LIBRARY_GROUP is checked by lint unit tests in lint's codebase.

Change-Id: I909106ceba8712e0923dc9ae865d36609fb270e2