Surpress test on kitkat
Current test fails on kitkat because mockito wraps some
method that it should not. Usually we override the method and
make it pulic but I cannot override this package private method
on View class.
The test is not really about API versions so I'm just disabling
it on older devices.

Bug: 33092812
Test: RecyclerViewAccessibilityLifecycleTest passes on api 21.
Change-Id: I3821ec0ef9187b43da4ffade11a9e1263b6de106