Supress a test on kitkat.
Current test fails on kitkat because mockito wraps some
method that it should not. Usually we override the method and
make it pulic but I cannot override this package private method
on View class.
The test is not really about API versions so I'm just disabling
it on older devices.

Follow up on ag/1664657 that restricted the wrong test.

Bug: 33092812
Test: Only one test now runs on JB devices
Change-Id: If3c6a071198e167f23919b5827af8b7466446bcd