Tweaking the sample music player app
* Add two new states: INITIAL and PREPARING
* INITIAL state is used to explicitly identify the part of the
lifecycle before the first playback. It is also used to explicitly
start playback of first track from the FAB
* PREPARING state is used to identify that part of the lifecycle
in which our media player is preparing the next track. Move the
transition to PREPARING state to be right before the call to
prepare(), instead of in repository's track index update as it
was before
* The service is now explicitly started from only one place
- activity's onCreate(). The rest of the communication from the UI
is done by the activity / adapter updating the LiveData and the
service observing those changes.

Test: build and run the app
Change-Id: Ia0ade70991941002d26888d253fbe5f26779f2bd