Fix BoxInsetLayout tests for phone devices.
This Cl ensures that the tests are run on
a relevan Activity - without a status bar,
navigation bar or an ActionBar. BoxInsetLayout
is not compatible with the above as it is meant
to be used as the top-level hierarchy element
on Wear devices which do not support the above
elements.

Test: Performed manual testing on a phone and a round
and square watch.

Bug: 34115480

Change-Id: Icc4dbefed8ea2efebba9ae17f0fe5569386daa03