Bind Cursor queries in a type safe API
The rawQuery API in SQLiteDatabase class expects all bind arguments
as Strings. Even though this usually works fine for other types due
to SQLite's type affinity logic, it fails in some cases.

To avoid that, this CL replaces the usage of rawQuery with a new
class called SupportSQLQuery which has a callback to bind items
to a query using the prepared statement.

There is no public API to achieve this so this CL adds a new method
to SupportSQLiteDatabase that accepts SupportSQLiteQuery.
In the framework version, this method is implemented using a
CursorFactory.

This callback is not necessarily instantly invoked so to avoid
the risk of binding to wrong values (because method parameters
may change after the call), the SupportSQLiteStatement is prepared
instantly. For this reason, we have a pooled RoomSQLiteStatement
class that implements the SupportSQLiteProgram API and can rebound
its parameters into another SupportSQLiteProgram. In queries, we
instantly populate it and it can be re-run any later time.
This implementation also avoids creating inner classes for each
query.

Bug: 32342709
Bug: 34503410
Test: RoomSQLiteQueryTest, QueryWriterTest, DaoWriterTest
Change-Id: Id74ba3e02d8de1f7df839249327b3baf933f5f84