Database Verification
This CL adds logic to verify @Query sql's at compile time.
This is done by creating an in-memory version of the database
during compile time and preparing queries against it.

This is an optional process that developers can disable via
SkipQueryVerification since there is always a chance that we
won't be able to find the sqlite on the host machine. To minimize
the risk, we are shipping xerial with the compiler.
SkipQueryVerification might be necessary if the developer modifies
the database directly.

Right now, we don't use this information besides making sure
query compiles. In followup CLs, we'll use this information
to validate return types and also allow arbitrary java classes
as return types (fi their fields match the fields in the response).

Bug: 33463891
Test: DatabaseVerifierTest
Change-Id: I9667a3bba15a54f25bbe7795de02eaeb435900f9