Prefer public fields in entities / pojos
This CL changes the Pojo parsing heuristic to prefer most
open method over less open ones. This allows us to generate
code w/o checking visibility constraints. Because we are
always doing the best possible, if it does not compile,
nothing else we can do. The only downside is that we
cannot provide a nice error message.

Bug: 33463891
Test: EntityProcessorTest
Change-Id: I24affe344d88423c9870492cffd3b4db6ea6aeb7