Refactor the flow of search query in the main repo list
* Move the auth token lifecycle into the auth token model so
that it can be used if necessary by fragments that wish to
reload their data on auth token re-acquiring.
* Introduce a new activity-level model that only keeps track
of the current top-level search query.
* Scope the repo list model to its fragment (instead of the
activity as before).
* Repo list fragment observes changes to that search query model
and updates its repo list model + adapter accordingly.
* The last three points effectively allow executing any number
of search queries with the top-level search box. Moving the repo
list model into the fragment scope also reduces data coupling
between the activity and the fragment.

Test: build and run the app
Change-Id: I8cef576024bc9fc27220e8039342f90bd580ab82