Move type converters to be one way
Previously, TypeConverters were enforced to be 2 way which does not make
sense in all cases because it might be necessary just for reading or just
for writing.

This CL refactors the TypeAdapterStore to make them one way. This opens up
the path for potentially serializing / deserializing a value via different
paths.

This should usually be OK but to avoid possible issues, for adapters, we
try to reverse the forward path (the path that takes user's data into
database). If that works, we use it. If not, we try to find a path from the
last node of forward path to the user's value.

For query parameter adapters, we don't really need the logic to read from
cursor. To support then, I've divided ColumnTypeAdapter into 2 interfaces.
StatementValueBinder, CursorValueReader.

TypeAdapterStore now has a specific function to get StatementValueBinder
which does not search for the path to read a cursor value. A similar logic
will be introduced for fields of pojos.

Bug: 32342709
Bug: 34610784
Test: existing tests pass + TypeAdapterStoreTest
Change-Id: I09789e4c0e97debaaa1d9ce422437f401d8fd310