Move FragmentState out of Fragments.java
Fragments.java contained 2 top level classes FragmentState and Fragments.
This CL moves FragmentState into its own java file. On the plus side we
will have one top level class per file as mandated by Google style, on the
downside per line history will be a bit harder to do for FragmentState class,
however it does not seem to have a lot of logic so it should be an OK loss.

Test: None
Change-Id: I6b83c9b2478af66bc68f9d91cf12484bb4a23532