Fixed visibility of inactive columns in Leanback Picker widget
When the picker widget is activated and the column size grows, the
picker items were not rebound. Thus the views that were cached by the
RecyclerView and just get laid out could hold an outdated alpha.
This is because views' alphas get updated in either during binding
or a focus change.
This for example created a problem when setting dates for an
inactive picker, and later activating it. The inactive columns
could still display unselected rows.

Bug: 35179122
Test: ./gradlew support-leanback-v17:connectedCheck
-Pandroid.testInstrumentationRunnerArguments.class=android.support.v17.leanback.widget.DatePickerTest#testInvisibleColumnsAlpha

Change-Id: Ib343f1a4ebcbb9e6a8caa81cc673a1026d873225