Disabled unfocusable scrolling when RV has no focused child
When no focusable items are found in a given layout direction, RV
scrolls to the next partially invisible child while keeping the
currently focused child visible. However, it's possible that the
focused child is cleared and before RV gains focus, focusSearch
is called on RV.
Since no focused child exists in this case, scrolling to the next
unfocusable view is not meaningful since there's no focused child
to keep visible on the screen. Returned the focusSearch to the
parent view of RV in this case.

Bug: 35182817
Test: ./gradlew support-recyclerview-v7:connectedCheck
-Pandroid.testInstrumentationRunnerArguments.class=android.support.v7.widget.LinearLayoutManagerTest#unfocusableScrollingWhenFocusCleared

Change-Id: I2d07a1664f0721cfc62cbdcf9dc11b31e521f50e