Fix flaky LoaderTest.testLeak()
Bug 36044087

There's a little challenge when it comes to forcing garbage collection.
There's no guarantee about things being collected. It appears that
the problem is that the previous activity hasn't been destroyed and
has no reference when the GC happens. This CL waits for the old activity
to be destroyed and for a few cycles of the UI thread before continuing
with the GC. It also changes the forceGC() for O+.

Test: ran testLeak
Change-Id: I9c94c16e76990b9c15b776356fb32e5fe1a982f7