Initial commit of Navigation into ApiReviewDemo
Add first draft of Navigation code into the ApiReviewDemo,
allowing users to see a list of items and click on one to
go to a detail screen of that item.

Also creates a series of navigation XML files that can be
used to step through the functionality available to developers.

Test: Run ApiReviewDemo's News activity
Change-Id: I2758718814ce673d29e15d0ee2f72a4fe484b18c