Create Menus from a NavController
Allow destinations within a NavGraph to be marked
as "primary" or "secondary". With the help of the addition
of title and icon support, you can now generate a complete
Menu from a NavGraph.

This is used in the testapp to populate:
1) The Action Bar (title and Up button)
2) A NavigationView
3) A BottomNavigationView

And allow all three to stay in sync with the current destination.

Test: testapp's BottomNavActivity and DrawerActivity work.
Change-Id: Icace71c471e5fb24e4003c31d06bd624efe24354