Fixed unfocusableScrollingWhenFocusCleared failure on API 15 and below
This test examines the scrolling behavior when focusSearch is called
while no child holds focus. This happens on API 16 and above in
onFocusChanged callback because when clearFocus is called and before
any child receives focus next, there's a transient state when no child
holds focus. This transient state does not exist on API 15- because
ViewRootImpl calls requestFocus before onFocusChange callback is called
thus this test is not applicable to those APIs.

Bug: 35639134
Test: ./gradlew support-recyclerview-v7:connectedCheck
-Pandroid.testInstrumentationRunnerArguments.class=android.support.v7.widget.LinearLayoutManagerTest
Change-Id: I1f96cccbfcd02aba30eb1f2d0009806dadecf586