Fix saving target fragment when Fragment not created.
Bug 35928527

During optimized transactions, a fragment may be removed without
being created. That leaves the state of the fragment in INITIALIZING
and previously, that state wasn't ever saved. This CL allows a
fragment that is being removed to be brought up to the CREATED state
so that it can be saved during saveAllState().

Chrry-picked from: Iea922369ab9bd91a7ee7a4e9e5a638e58dd97430

Matches Framework changes:
* I649f1931745be43087ec3578e9195624e80821dc
* Ie7207cc647312d38b377405bc5ec8721db757d2e

Test: ran fragment tests
Change-Id: Iea922369ab9bd91a7ee7a4e9e5a638e58dd97430