Standardize Navigator names and attributes
Navigators now have a Navigator.Name annotation for
the name they want to use in navigation graph XML
files and attribute names are standardized to be the
class name of the Navigator.

This ensure that tooling can cross reference a
Navigator, its name in navigation graph XML files,
and the custom attributes unique to that destination.

Test: testapp still runs

BUG: 37252062
Change-Id: Icc53c53e966a9feba340394b7abc9070a74afa4b