Hide TypefaceCompat and introduce FontsContractCompat instead.
This change brings the supportlib implementation of Downloadable
Fonts in sync with the recent changes in the framework.

TypefaceCompat.create is kept but hidden to deal with xml layout font
loading, as is in the framework atm.

The API26 implementation still points to API24, as framework APIs are
not final yet.

Bug: 35381428
Test: support-compat:connectedDebugAndroidTest
Change-Id: I4d73d7a547a0d9b79c4b7158d431a573b80417f3