API change in the constructor of SpringAnimation and FlingAnimation
This CL adds a new class FloatValueHolder that wraps a float value
and exposes the access to it through getter and setter (i.e.
setValue(float) and getValue()).

The previous single argument constructor in SpringAnimation
and FlingAnimation have been changed to use a different type of
argument - FloatValueHolder. This change makes the constructor API
look much more straight forward.

Note this CL changes the API that hasn't been released yet.

Bug: 30999909
Test: Included in this CL

Change-Id: I1cf2793cc4594d7b3ff35b31bb242ea9bfa0c894