Support setting threshold for custom properties
This CL removes the (unreleased) APIs set/getVelocityThreshold,
as the velocity threshold is not a straight forward idea, and
it requires develoeprs to convert a per frame threshold to
per second.

Instead, set/getMinimumVisibleChange(float) is added. This expects
a minimum value change that is visible to end users. This value
is then used to calculate a reasonable velocity and value threshold
for animations.

BUG: 38001368
Test: Included
Change-Id: I22e46d9650be74fbcad638ab2e5c1cacde357149