AutoSize TextView - fallback to OS version for O and above
Re-enable delegating to the framework version of the auto-size
implementation for OS versions O and above. Although the clean
preferred way of switching implementations is built into
TextViewCompat (this is what's public) had to check and switch
implementations in AppCompatTextView because internally we read
the auto-size attrs @ AppCompatTextHelper#loadFromAttributes and
after we need to make a call to the framework implementation
to populate with data. This stackoverflowed without the OS
check because TextViewCompat always saw the TextView as an
instance of AutoSizeableTextView and thus was always calling
the support implementation (which in turn called the helper).
Also introduces a new test to verify that setTextSize is
working when auto-size is disabled and make it clear in all
support tests that we are testing AppCompatTextView.

Bug: 37958118
Test: existing tests pass
Test: ./gradlew support-appcompat-v7:connectedCheck --info --daemon
-Pandroid.testInstrumentationRunnerArguments.class=android.support.v7.widget.AppCompatTextViewAutoSizeTest

Change-Id: Ie00a0f69842ae291ea5cd721b506363457859289