Adjust layout width to fix broken test
To ensure the espresso tests work on non-watch devices, the layout
height and width were constrained to 213dp. This works okay for e.g. a
phone in portrait mode, however it doesn't work for large devices in
landscape. This is because the action drawer has margins which are
specified as a percentage of the entire screen. As a result, the action
drawer list elements were off screen and thus "isCompletelyDisplayed"
was always false.

To resolve this, layout height remains constrained at 213dp, but layout
width is now match_parent even for large devices. While debugging, I
found it useful to make the drawers have distinct colors (red and blue)
so that it's easy to tell when the drawer is open. I have left these
changes since it should help with future debugging.

Test: Espresso tests on a Ryu, Angler, Swordfish, Tetra
Bug: 38041618
Change-Id: I910ec418949a91e37407917e0e3ee2e59eb4cf1e