Add Set<String> persist capabilities to Preference
Since the minimal SDK is 14 we can add support for Set<String> that we
already have in the framework version of preferences. SharedPreferences
started support Set<String> from API 11. Set<String> API was only at
MultiSelectListPreference in v14 which will be now removed.

There were no tests for the "persist" logic so to be able to test the
additions I have added a new test class that tests all the persist APIs.

Bug: b/38153879
Test: PreferencePersistTest

Change-Id: I9efc977d2a17a659427196661cda1c15f735e125