Simpler SQLite interface
This CL reduces the API surface of the SupportSQLiteDatabase not to include
factory methods or methods. It also changes query methods that receive
String[] to Object[] to workaround the binding bug.

This is still not enough as it depends on android classes. We may later
further reduce this to use no-android APIs but that api will look very
limited unless we can use default methods (which requres android
gradle plugin 3.0 that is not stable right now).

I've also added a simple QueryBuilder to replace methods in the SQLiteDatabase
that receive a million arguments.

Bug: 38481640
Test: existing tests pass
Change-Id: I37dc666235fa31877725d85cc98fb01542d4638a