MediaSessionCompat: Fix NoSuchMethodError in API 19
Whenever RemoteControlClient.MetadataEditor#putObject() is called
in API 19, it throws NoSuchMethodError when compileSdkVersion >= 21.
The method was added in API 21. Before API 21, only the parent class
MediaMetadataEditor had the method putObject().

This seems as a bug of android compiler because calling the method
throws NoSuchMethodError instead of calling the parent class's method.

Upcasting the RemoteControlClient.MetadataEditor object to
MediaMetadataEditor before calling the method makes the error disappear.
This CL makes MediaSessionCompat avoid the crash by this way.

Bug: 62335842
Test: Run MediaSessionCompatTest.testSetMetadata() in API 19-26
Change-Id: I016c171d9d745a6332ef22d17ea7a95bbecf32a2