Stop copying files as much as possible.
Due to lack of API before API 25 or before, need to prepare the real
file for creating Typeface. The previous implementation creates
temporary file and create Typeface from it. However, with that
implementation, the mapped memory won't be shared between applications.
To share the memory as much as possible, introduced following approach,
- API 26 or newer:
  Use framework implementation.
- API 21 - API 25
  Try to find the file path from file descriptor. If found, use it for
  creating Typeface. If not found, use fallback file copying
  implementation.
- API 20 or before
  Still use file copying approach.

This CL also fixes font style resoltion from FontFamily on API 23 or
before.

Bug: 37749510
Bug: 62199090
Test: ./gradlew support-compat:connectedDebugAndroidTest
      on API 15, 16, 17, 18, 19, 21, 22, 23, 24, 25, 26
Test: ./gradlew support-emoji:connectedDebugAndroidTest
      on API 15, 16, 17, 18, 19, 21, 22, 23, 24, 25, 26
Test: ./gradlew support-appcompat-v7:connectedDebugAndroidTest
      on API 15, 16, 17, 18, 19, 21, 22, 23, 24, 25, 26

Change-Id: Ie2dc9594b99598292c10a5402b1b903e27390dab