Fix Toolbar title appears to lose bold
The issue here was not bold style, but sans-serif being used
instead of sans-serif-medium.

The code that handles setTextAppearance was not clearing the existing
value for the Typeface when a new value is present. Because of this,
if TypedArray#getFont threw an exception, the typeface remained in its
old value, and the following code relies on null as a signal that a
system font might need loading.
By resetting the value before re-calculating, the null
check works as expected.

Bug: 62449517
Test: AppCompatTextViewTest#testSetTextAppearance_resetTypeface
Change-Id: I4aaa7fb7f3d3a7f615a03fe6baeca7b5fce2d3d2