Introduce auto retry to FontRequestEmojiCompatConfig.
This CL introduces auto retry mechanism to FontRequestEmojiCompatConfig.
To implement this features, provides setter functions for Handler and
RetryCallback.

The Handler is used for waiting the content observer notification and
asynchronous repeated retrying. If developers don't give this, metadata
loader implicitly creates the thread and use it for asynchronous
retrying.

The RetryCallback is used for giving control to the developers of retry
interval. Developers can use the default retry callback
DEFAULT_RETRY_CALLBACK.

Test: ./gradlew support-emoji:connectedDebugAndroidTest
      on API 15, 16, 17, 18, 19, 21, 22, 23, 24, 25, 26
Test: Verified by using test app
      on API 15, 16, 17, 18, 19, 21, 22, 23, 24, 25, 26
Bug: 62227949

Change-Id: I17e941caeae7b9c0651072b83e40fd99bf93d475