Make NotificationCompat.Style apply themselves
Instead of the NotificationCompatImpl using
instanceof checks to handle each Style, create a
generic apply() method that each Style can implement
to do their own encapsulated version checking and
changes that are specific to that style.

Updates BigTextStyle, InboxStyle, BigPictureStyle,
and the API 24+ route for MessagingStyle to use this
new functionality.

Test: Manual inspection
Bug: 33042051
Change-Id: I304dacee71bf25a9d8f2c5a24e319671465e8d2b