Added a workaround for ValueAnimator issue
ValueAnimator calls onAnimationRepeat before it calls onAnimationUpdate
for last frame (interpolated time 1.0). This causes spinner to jump back
and forth for that frame, as parameters of drawing change for the next
repetition of the animation.

ValueAnimator usually skips the last frame if animation is repeating,
so this was happening only occasionally.

Test: Done manually
Change-Id: Ifd4adcfb2bd75fffa9a40fc7aa201f23127c4b24