Do not update currently open tooltip.
Currently if View.setTooltipText is called while
the tooltip is being shown for that view, it will
update the displayed text. The tooltip then will
resize to wrap around the new text, but not change
its position. This looks confusing if the new text
is significantly shorter or longer.

Removing this functionality until proper
re-positioning is implemented.

This also fixes a linked bug which is triggered
by menu item view reuse

Bug: 38491655
Test: manual
Change-Id: I1fb3ff2f66d146530d2de850a67631243c9492cb