DO NOT MERGE: Fix Fragment.onCreate() not being called in some cases.
Bug 62798751

When an application process has been killed with a retained instance
fragment on its back stack, and a configuration change causes
state save, the retained instance fragment was not calling onCreate()
when it was popped back into existence.

The cause of the problem is that the mRetaining flag was being set,
and when it was supposed to go through onCreate(), it was skipped.
This CL uses a new flag to judge whether or not a Fragment has
called onCreate() or not.

Test: ran fragment tests
Change-Id: Ia66a1f474d815f657b6962097139c796f7fa198b