leanback: fix selected position after remove all
The bug is when remove all items and then add few items, the focus
position was set to last item of newly added items.

The current code tries to be smart to select the next item when
focused item is removed.  But it's impossible to know if there is
item after the removed one. So stop tracking the offset change
when focused item is removed.

Bug: 64235615
Test: GridWidgetTest

Change-Id: I654e906bccff04ce59ec840f9356a17717eaba2d