ExifInterfaceTest: Fix failing tests
After ag/2634960 is merged, some of the ExifInterface tests failed.
It is because the tag TAG_ISO_SPEED_RATINGS was deprecated and was
replaced with TAG_PHOTOGRAPHIC_SENSITIVITY in that CL, but the test
was unchanged at that time so that it is still using the old tag.

This CL fixes the failing tests, as well as making ExifInterface
does not break the backward compatibility.

Bug: 64467277
Test: Passed ExifInterfaceTest
Change-Id: Ic5dcd4d355e2935c7773384bf9e2576d4b49e46f