Fixed time format for API below 18
For API below 18, SimpleDateFormat.getTimeInstance is used to extract
time format. It was using full style that made it pick time zone
field (zzzz); Changed it to short so that it only includes hours, minute
and am/pm fields.
Also, removed "a" for 24 hours format. For API 18 and above, these were
already handled using the capital 'H' in the skeleton argument.

Bug: 64355457
Test: TimePickerTest on Nexus 4 API 17
Change-Id: Ic7390c4e783701cfc29df50ee06341a205577ba5