Ensure action drawer updates when menu modified
Currently, when the action drawer has a title and the menu is modified,
we notify the RecyclerView about the wrong index. Because of that, the
item's associated view will not be updated.

This change takes into account whether or not the drawer has a title
when calculating the view's adapter position.

Test: Added tests and ran them on swordfish & angler
Bug: 64609474
Change-Id: I2cdf4d8daeb65d7cc2716d5387305b9064f1c2e3