Allow preferences to be hidden from the preference group.
When a preference group with many children is shown, we would like to
add the capability to hide part of the children, and provide an expand
button for user to show all the hidden children.

The number of initially expanded children is configured through the new
attribute of the preference group. This count will only apply to
end-children, that is, if a child contains any nested children, the
nested children will count towards the limit, but their parent will not
be counted.

The hidden preferences' titles will be listed as the summary of the
expand buttoni, so as to give a hint to the user what preferences are
being hidden.

Bug: 63985174
Test: ./gradlew preference-v7:connectedCheck --info --daemon
Change-Id: I30b1ea4578793de6ff43b46644e9145734d49aa2
(cherry picked from commit c88dea08fa07d811eb7cfe77c7a523fd2b0ce339)