Implement Font callback for asynchronous font fetching
When a font needs to be fetched asynchronously, previously we were
keeping a reference to the TextView that needed it and set it
when available from FontsContractCompat. As this is needed by
callers to ResourcesCompat.getFont as well when their font resource
might be async, this change introduces a new callback.

Ideally all font requests should be asynchronous, however we already
released a synchronous API so we will live with both. The developer
can't know if a font will require async fetching beforehand easily,
therefore should always favor the async API.

Bug: 65280123

Test: ./gradlew support-compat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=android.support.v4.content.res.ResourcesCompatTest
./gradlew support-compat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=android.support.v4.provider.FontsContractCompatTest
./gradlew support-compat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=android.support.v4.graphics.TypefaceCompatTest

Change-Id: I1c4c7501df6503c2d758ff8ad2a1504e07496e3c