Remove animated views when canceled.
Bug: 64771466

When removed fragments have early-terminated animations,
the fragment views were not being removed from their
parents. This CL uses the startViewTransition() and
endViewTransition() to force the disappearing views
to be removed.

I wasn't able to reproduce this on a test app, but
the manual test on the Project Fi app showed success.

Test: manual on Project Fi app
Change-Id: I2c69d49b1147ac4a3cdb8e74cb698ade64a14807