Generate ProGuard rules to keep generated adapters
Once an adapter class is generated, also generate a ProGuard
configuration file that only marks the generated class as a seed (entry
point to be kept) if the peer class is ever instantiated. If the
generated classes are always kept (since they are referenced through
reflection) then the peer class is always kept. This can cause large
amounts of code to remain in the final binary that is unreferenced.

This change depends on ProGuard being able to pull the configuration
files from inside the jar file.

A new Unused lifecycle observer was added as a canary to ensure that
ProGuard can be taught to use the new keep rules.

Test: Rebuilt the integration test APK and verified the keep rules were
      generated.
Test: :lifecycle:compiler unit tests

Bug: 65293954

Change-Id: I56c3d8a228e74048eac057ac2da265be046c45fc