Fix failing AppCompatTextViewTest#testTypefaceAttribute_fontFamilyHierarchy
There was a race condition, where if we request the font in the first pass
(TextAppearance) and it takes time to come back, we go through the
second pass and find the Typeface attr (in style), but after that the
async font comes back and overrides.

Fix for now by making the font call blocking as it is done in the framework.

Bug: 65357330
Test: ./gradlew appcompat-v7:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=android.support.v7.widget.AppCompatTextViewTest
Change-Id: I3353726d71716a404f180c5de576fbb933780e99