Fix stack overflows and invalid deep link intents
NavControllerTest became flaky due to a
StackOverflowError caused by including the
NavDeepLinkBuilder's internal Intent as an extra
in itself. Avoid this problem by creating a
copy of the Intent before passing it to
TaskStackBuilder.

This has the added benefit of ensuring that the
internal Intent does not leak out to external
developers.

NavDeepLinkBuilder now also throws an
IllegalStateException in cases where the graph
or destination has not yet been set.

Test: NavControllerTest now consistently passes
BUG: 36817937
Change-Id: I01495062fe44f4ed01c3b015f23d7380054b15e2