Add drawer activities.
The drawer component allows car applications to have an action bar that
is styled appopriately according to AAE's UX guidelines.

This drawer components is moved over from the car-stream-ui-lib with
some updates to naming.

In addition, add some attributes that are needed by the drawer but have
not been supported yet in PagedListView.

Test: built a demo app that has PagedListView and Drawer and ensure that
everything works correctly
Bug: 67383920
Change-Id: Ib0fbc14a57708012b8486e18108f62a366872c61