Allow multiple jobs to be enqueued together.
This allows us to create a tree or graph of dependencies, essentially allowing us
to do doAllOf(...).then(...) (note that there is no "doAllOf"; we use the same
primitives we already have).

Added a much more robust testapp impl that initializes a db, reads and processes
several text files in separate (potentially parallel) work, and then collates the
results back into the db (which can be then displayed on-screen).

Test: Updated tests, testapp.
Change-Id: Iabea473cdf0186e27535c9952ad85e39e837bc7e