Add preprocessor to jetifier to generate types map
The rewrite rules we have been using until now are easy to write but at
the end we will need a map of every single change that happened in the
support library due to the refactoring. Because of that I'm introducing
a preprocessor tool that will firstly run all the rules over the real
support libraries to build the complete map. This map is then added to
a newly generated config. This generated config will be shipped with
jetifier.

The libraries bytecode and XML resources rewriting was updated to use
the generated types map instead of the rules itself. That makes it
simpler and also prevents rewriting of anything that is not expected
to be rewritten.

The tool has also few new extra checks. If the preprocessor will find
any java type from support library that is not covered by any of
the given rules it will fail. That will ensure that our mapping is not
missing anything. Also the jetifier itself will fail if it finds
a type that is supposed to be rewritten but there is no rule in
the map. This will allow us to detect that we are rewriting older
versions of support libraries for instance.

Bug: b/64833662
Test: MapGenerationTest
Change-Id: I5d326717b5c224cad363835d2c65eb2673e4927e