Remove ConstraintsState and replace with ConstraintListener(s).
We have a basic ConstraintListener interface that is extended by specific listeners
(StorageNotLowListener, etc.).  ConstraintTracker now expects a ConstraintListener
parameter; each of the concrete implementations provide an appropriate listener.
ConstraintController also expects this in its generic form. The only implementation
of this is ConstraintsTracker for the moment, but we can use this for (1) better
testing and (2) a Builder pattern for ConstraintsTracker so we can customize the
listeners we care about.

This does NOT add any code to the listener callbacks.  That's coming in the next
CL.

Test: Updated tests; this is more or less removing ConstraintsState and replacing
it with various ConstraintListeners.

Change-Id: I48a5287c17567e7b6e2420e79e8b580b80e12bf6