Inform ConstraintsTracker about potentially blocked/unblocked WorkSpecs.
This is nearing the final architecture - in this CL, I have created concrete
implementations of the various ConstraintControllers, each with their own
ConstraintListener. When the listener gets triggered, we know that the
matching WorkSpecs for that constraint are either blocked or unblocked. We
pass this information upstream (to the ConstraintsTracker), which can
determine fully if each WorkSpec is constrained by checking each
ConstraintController.

The next step is hooking up the final piece of wiring and informing the
processor.  I didn't do it in this CL to keep it constrained (har har).

Test: Incoming!
Change-Id: I60122be2207a68cb3353d1eca1519b609d6805ec