leanback: move childrenStates invalidation to AdapterObserver
ag/2577510 causes an extra LayoutManager.onItemsChange() event
fired when adapter is attached. Leanback implementation was relying
on that event in order to invalidate saved children state. The extra
event unexpectedly clears children states so that leanback lost
its children state when restoring view hiearchy.
(Unfortunately leanback did not have a TestCase and failed catch the
regression.)

The fix is to move away from the LayoutManager.onItemsChange() event
and relies on Adapter's DataObserver in addition to onAdapterChange.

Bug: 65381722
Test: testRestoreStateAfterAdapterChange
Change-Id: I3b7b9e3f5c85c3b6e19974477b0da706dd99ccb1