Don't have talkback read action item twice
When navigating through the action drawer with talk back, each row is
focused. If the action's icon and text both have content descriptions,
then the text is read twice. To avoid that, I've ensured the text will
be read aloud, but I've marked the icon's ImageView has not important
for accessibility and removed its content description.

Test: Verified with test app
Bug: 67369424
Change-Id: I76dd61aaa8779e5df3a076513b47570752b97cea