Add 2 extra preferences to SupportPreferenceDemos
For this bug we're moving some styles from SettingsLib into the support
library to be better able to share styles in regular and Google
settings. This CL adds a preference with an icon and a SwitchPreferece
to the SupportPreferenceDemos, to help show the effects of the style
changes.

Bug: 64689839
Test: manual (build and run support-preference-demos app, and open
SupportPreferenceFragment section)

Change-Id: I9a42bcf35030bd48161d9b6ab1390c6aac26093b