Initialize WorkManager on app startup.
This is based on a piece of feedback from API walkthroughs.  People were
getting confused that the context/activity being passed to
WorkManager.getInstance was actually returning an instance bound to that
context or its lifecycle.

This will also allow us to add a hook for WorkManager configuration if
and when we decide to provide that.

Test: N/A, ran demo app.
Change-Id: Idc25c7e06bfd530a58a371184e7b9ddde3a9d846