Switch @IntDef from long to int, and add @LongDef
The @IntDef annotation actually uses a *long* to store the set
of expected constants. This was to avoid having to have
separate annotations for int- and long-constants.

This works well in Java since int's are silently upconverted
to longs.

However, in retrospect, that was a very bad idea. In Kotlin,
there's no such silent upconversion; referencing an int constant
in an @IntDef does not compile. It insists you explicitly call
toLong() on each constant -- which not only is ugly, but doesn't
even compile in an annotation.

This means @IntDef is completely broken for Kotlin.

This CL fixes this by switching the value type back to int, and
introducing a @LongDef annotation for the rare cases where the
constants being listed are actually longs.

Note that this is an incompatible change: code which was
referencing long constants from an @IntDef will no longer
compile. However, this is really rare:
- frameworks/base has nearly 400 IntDef annotations and only a
  handful of them longs.
- frameworks/support has ~125 IntDef annotations and only 2 of
  them were using longs.

The developer experience after upgrading to a new version if
using long constants is an error message from the compiler like
this:

PlaybackStateCompat.java:48: error: incompatible types: possible lossy
conversion from long to int
    @IntDef(flag=true, value={ACTION_STOP, ACTION_PAUSE, ACTION_...

Test: These annotations have source retention so there is no
      runtime impact. The test is the compiler.

Change-Id: If888e4f5e629655a095dfbfb77e2e92c5c31b382
(cherry picked from commit 77795c65aa61622895e12f3266051d6e5df12440)