Use a modified LinearSnapHelper to replace the PagedLayoutManager.
Create a new PagedSnapHelper that extends LinearSnapHelper. This
PagedSnapHelper duplicates the functionality of the
PagedLayoutManager. This allows users of this PagedListView to use
whatever LayoutManager they want.

Also, remove the extension of RecyclerView. This removes the ability to
fade out the last item. This functionality will be added back in a later
change.

Lastly, the setting of max pages will not work properly at the moment.
This change is left here to be modified in a followup change.

Test: Re-enable espresso tests and run tests (./gradlew
car:connectedCheck)
Test: Test code on sample app.
Bug: 69479394
Change-Id: If3258741976800779df76e5261f752fece3b2447