Jetifier gradle plugin improvements.
We have now 2 tasks in jetifier:

1) jetifyLibsTask
Transforms individual libraries using jetifier by directly inlining it
into dependency configuration. This is what we already had.
Example:
compile jetifier.process('test:myDependency:1.0')

2) jetifyGlobalTask
Automatically processes a whole given configuration so the developer does
not have to explicitely state any specific dependencies that should
be transformed. This is good for automatic upgrades. Note that this task
does not cache and it will require extra work to make it cache manually.
So this is more like a prototype.
Example:
jetifier.addConfigurationToProcess(configurations.implementation)

Bug: b/64833662
Test: Local tests (will be added later)

Change-Id: If2931e5a42b2a33322608567df41d3799dc6086a