Implements proper handling of EdgeEffect for clipToPadding on NestedScrollView on API 21 and above.
This CL makes makes NestedScrollView act exactly the same
as ScrollView in relation to how the EdgeEffect reacts to clipToPadding for
API >= 21.  As for API < 21, NestedScrollView already acts like ScrollView
in that they both ignore clipToPadding for their EdgeEffects, always
respecting left and right padding, and always ignoring top and bottom padding.

Bug: 65469708
Test: Manually verified fix. It's very difficult to write a test for, so not writing one.

Change-Id: I3804f35cb2f5b8ef9e505c9037ff898d43fb03ff