Add PageKeyedDataSource
Fixes: 68705650

Add PageKeyedDataSource to support request-based keys, common in
loading from the network.

This DataSource isn't currently capable of consuming an initial key
(since we can't extract an initial key from an item, only from a page).

This initial version differentiates between loadAfter and loadBefore
with a boolean in the load param, but we can revisit this if it doesn't
fit well with other data sources.

Test: ./gradlew runBuildServerCompilationTasks
Test: tests in paging-common, paging-runtime, room-integration-tests-testapp

Change-Id: Idf26a640bf573eb8ac1e357b7e3293ab65ba7983