Restructure ConstraintsTracker to combine individual LiveData for each ConstraintController
Removes individual LiveData from ConstraintControllers. There is now one
LiveData in ConstraintsTracker which passes WorkSpecs to all
ConstraintControllers. This is to avoid changes to ForegroundProcessor
for the time being and will be removed in a future CL.

The goal is to allow ForegroundProcessor or SystemAlarmService to pass
WorkSpecs that ConstraintsTracker should monitor constraints for.

- Add WorkSpecDao#getConstraintsTrackerEligibleWorkSpecs
- Add ConstraintsTracker#update(List<WorkSpec>) and #reset() as the
  interface to manage WorkSpecs to track constraints for.
- Changed ConstraintController#isWorkSpecConstrained(WorkSpec) logic. A
  null current value indicates that the controller is not being used
  and that there are no WorkSpecs with its constraint.

Test: Added tests in ConstraintControllerTest and ConstraintsTrackerTest
Change-Id: I5043b80b29c0131f37741187ee76f48cbf7ff9b8