Refactor logic to extract template content from a Slice
Previously all of the logic to identify what bits of
content in a slice go into which parts of a template were
within the related view.

This CL moves that logic into separate classes which
process the slice and provide getters for the different
bits of info.

This only does list / row content, grid / messaging will
come in follow up.

Test: manual, use slice browser make sure list-based
      slices look normal
Change-Id: Iae4876c23b9331f961ba1fd23e53e5228b70ce94