handle large key sets in relationships
If a relationship result has 1000 >= items, sqlite rejects the query.
We cannot blame the developer because query is generated by room so
we rather divide it into sub maps when this happens.

It is not the most efficient way to divide the map but since this is
unlikely to happen, seems fair.

Bug: 64643894
Test: PojoWithRelationTest#largeRelation_*
Change-Id: Ie0f2f4520e1e26681dfb3ed7239fe3c6f5e31586