Create static method for getting a LoaderManager
Allow developers to create a LoaderManager from any
LifecycleOwner+ViewModelStoreOwner class, rather
than only from FragmentActivity and Fragment.

Remove LoaderManager instance held by FragmentActivity
and Fragment (as they are not the sole creators of a
LoaderManager), instead calling
LoaderManager.getInstance() to get a new instance. As
all of the state is stored in the LoaderViewModel
singleton, this is a cheap operation.

Test: ./gradlew checkApi
BUG: 70897911
Change-Id: Ib62384804c4443314e63db1c3e47ec56f4962073