Implement a speed bump view.
This view is meant as a wrapper for any content that a developer wishes
to monitor taps over. It will monitor for a set number of taps within a
given frame of time. If that value is passed over, then further
interaction is locked out.

By default, speed bump wraps the drawer and content views.

The values for maximum number of taps is also hardcoded right now. This
will be changed in the future once a system wide configuration mechanism
is implemented.

Test: Test on sample app and unit tests.
Fixes: 70338632
Change-Id: Ib2b78ed2d72ef081792acdc91b595c658d47c48e