Safeguard against empty type adapters
Somehow, room ends up in a state where there is a converter path to a type
but the column adapter is empty. This results at a compile time crash which
hides the real problem. This safeguard will allow a logical error to populate
to the user (or even, room might find the correct one).

Bug: 71855433
Bug: 71387107
Test: existing tests pass
Change-Id: I7804328c4ae5d47a859e38c1e0aa5b3817ea5bbb