Add Action#getShowsUserInterface()
We don't know whether or not an Action's PendingIntent will show a user
interface when triggered. This new flag within
NotificationCompat.Action will let developers signal that this Action
won't display UI elements when fired. This is useful for applications
that can consume Notifications and provide a hands-free ui-less
experience.

Test: ./gradlew support-compat:connectedCheck
Bug: 67717177
Change-Id: If71583f9270999e421f40376dbb56b84f2e25ebe