Port ListAdapter from Paging Library to RecyclerView
Bug:67420319
Bug:70411933
Bug:70678044
Bug:72116084

Test: ListAdapterHelperTest

Api / behavior changes as part of the port:

1) Hid ListAdapterConfig.Builder#setMainThreadExecutor, since it shouldn't be necessary for external
usage

2) Put DiffCallback in ListAdapterConfig.Builder constructor, since it's required, and enables
diamond operator (or type inference, in Kotlin)

3) Removed Arch executors, and in their place use a) a main thread handler, and b) a static
background thread pair. Neither is ideal, but can be cleaned up in the future once supportlib has
these sorts of Executors.

4) Make DiffUtil.ItemCallback no longer @NonNull, since arbitrary lists in ListAdapter contain
arbitrary data. Once we have type annotations in support lib, we can fix this.

5) Move AdapterCallback to top level public method, since it's usage makes ListAdapterHelper more
convenient.

Making the Paging lib depend on this version of ListAdapter in supportlib will come in a follow up.

Change-Id: I0123ce4a9319434208b6228dee1c6563c774f145