Return boolean from Processor#process(String)
- Update ForegroundProcessor#process(String) to return boolean
- Remove outdated test, ProcessorTest#testProcess_noWorkInitialDelay

The boolean indicates whether or not the work was successfully enqueued
for processing. If process(workId) was already called before for that
workId, it will return false.

Currently, Processor#process() silently checks if a work was already
set to be processed, but does not notify the caller. This will be used
to determine whether or not a Work was actually enqueued for processing.
Otherwise, the caller would need to keep its own set of work IDs that
it called Processor#process().

ProcessorTest#testProcess_noWorkInitialDelay took 2 seconds to check
that a work was put into running state. That functionality is tested in
WorkerWrapperTest#testRunning.

Test: Update and run ProcessorTest and ForegroundProcessorTest on
      API 16, 18, 23, 26
Change-Id: I2b82cde2a36498f31ff0b0050f4c27cfdd090d0f