leanback: make sure header view gets focus in touch mode
If user taps on a card view and press back, fast lane item
does not gets focus as it's currently in touch mode and
focusableInTouchMode is false. This also fix problem if the
UiBench test puts Activity in touch mode initially the header
wont be selected. generatev4.py is executed to sync source
code of support/framework fragment tests.

Bug: 71904218
Test: BrowseSupportFragmentTest#testTouchMode on master build
      of phone and TV.

Change-Id: I493c5aeef475b9137131fceb1c4db7566694e98d