Change WorkStatus to an enum.
Lots of files were affected, but the change is pretty
straightforward. The only thing worth noting is that
Room SQL queries don't automatically typeconvert the
enum, so I made some int constants to back the enums
that I used in the SQL queries. You can find that in
EnumTypeConverters.java (which I will use for other
conversions in the future, like BackoffPolicy).

Also renamed getStatusById -> getStatus.

Test: Fixed and ran tests.
Change-Id: I57d293eafea07f9e661dfbdb28fdc82cbc301273