Fix parent not being null after animating removed fragment.
Bug: 71889623

When animating a fragment out with Animations, the view that was
removed was not clearing its parent field. This was caused by
both startAnimation() and startViewTransition() being called
and then endViewTransition() and finishAnimation() being called
in the wrong order.

This CL wraps the Animation in an AnimationSet that delays ending
the animation until endViewTransition() is called.

Test: parentNullAfterAnimation
Change-Id: I3c239613ab8f208ee008db1f7e3d89caff7c73c5