Ignore some states when checking KEEP_EXISTING_WORK policy.
KEEP_EXISTING_WORK should only bail on the new work if there
is something with the given tag that is ENQUEUED or RUNNING.
Completed or cancelled statuses should not count since the
work chain can be restarted. BLOCKED work should not factor
into things since they are either orphaned due to a failure
or waiting for an ENQUEUED/RUNNING item.

Bug: 72525224
Test: added new test, ran tests.
Change-Id: Id59a6189da507f7fcd21ab1104c3bd29b0cb3f76