Implemented java.util.Optional as an Adapter type in Room.
This permits it
to be used as the naked ORM return type (Optional<T>) or as part of the
inner type of complex return types (Flowable<Optional<T>>).

This is implemented differently to the Guava Optional. This delegates to
a SingleEntityQueryResultAdapter instead of reimplementing its body.
This can be fixed in that implementation too.

Bug:72121980
Change-Id: I60f8f65d2d24f3694a75680aed2c0ba10c00dfa7
Fixes:
Test: Unit test covering presence/absence of raw Optional.