Fix DialogFragment Dialog visible lifecycle
When calling DialogFragment#dismiss(), the
dialog should be immediately dismissed to
allow developers who override onDismiss to
check for isAdded(), etc. and inform other
components of dismiss events.

Going forward, the recommendation for checking
for manual dismissal will be to check
getDialog().isShowing() in onStop() - if it is
not showing, it was manually dismissed.

Test: new DialogFragmenTest cases pass
BUG: 72808587
Change-Id: If4b19b21d8abc01197ed5171cf83b11efac58e56