Add NotificationCompat#addInvisibleAction.
Invisible actions allow apps to add actions that won't show in the
Android OS notification, while providing applications that listen for
notifications extended functionality. An example of this usage is for
hands-free applications which listen for notifications containing
messages. An app developer may not want to add visible "mark as read" or
"reply" actions, but may want to allow for voice commands read and "mark
as read" and reply.

Invisible actions are stored in the CarExtender extras in a
backwards-compatible way, but is not a field within CarExtender. We do
this to ease off of the use of CarExtender from the app developer
perspective.

Test: ./gradlew support-compat:connectedCheck
Bug: 67717177
Change-Id: I15da349eff6a85666df808404e89a8bd0e022ec1