Change getInstance() to getInstance(Context).
This removes the ContentProvider-initialized singleton in the
manifest.  In general, we are prioritizing faster startup over
the marginal readability improvement.  There was some concern
in previous Googler studies about "which context", "will it
be cached and scoped to this lifecycle", etc.  We believe that
the documentation is sufficiently clear on this.  In addition,
this follows a common pattern in Android apps.

Test: Updated and ran tests.

Change-Id: I5f24724934ed1e16bddb0d9906326ad2e54e398d