Double-schedule jobs on API 23 to overcome bugs.
API 23 JobScheduler only kicked off jobs if there
were at least two jobs in the queue, even if the
job constraints were met.  This behavior was
considered undesirable and later changed in
Marshmallow MR1.  To match the new behavior, we
double-schedule jobs on API 23 and dedupe them
in SystemJobService as needed.

The JobScheduler behavior was fixed in ag/794033.

Change-Id: I5dc3049f65e5716258ab6de02b53d502e670f389
Fixes: 73383970
Test: Added and ran tests.