Remove unused resource overrides
AFAICT these resource overrides are not used, but (if accessed
by library clients) can lead to crashes.

R.dimen.abc_action_bar_progress_bar_size doesn't appear to be
used anywhere.

R.bool.abc_action_bar_embed_tabs is only accessed in appcompat,
where it also has definitions both in values/ and values-port/.
I'm removing the lone declaration in compat/values-port/. Perhaps
this is an accidental leftover from the code moving?

These were found by a new lint check in 3.2; a generalized version
of the ExtraTranslations check which flags resources that appear
only in resource overrides, but not in the base folder (with some
special cases.)

This is the issue explanation:

   Explanation for issues of type "MissingDefaultResource":
   If a resource is only defined in folders with qualifiers like -land or -en,
   and there is no default declaration in the base folder (values or layout
   etc), then the app will crash if that resource is accessed on a device
   where the device is in a configuration missing the given qualifier.

   As a special case, drawables do not have to be specified in the base
   folder; if there is a match in a density folder (such as drawable-mdpi)
   that image will be used and scaled. Note however that if you  only specify
   a drawable in a folder like drawable-en-hdpi, the app will crash in
   non-English locales.

   There may be scenarios where you have a resource, such as a -fr drawable,
   which is only referenced from some other resource with the same qualifiers
   (such as a -fr style), which itself has safe fallbacks. However, this still
   makes it possible for somebody to accidentally reference the drawable and
   crash, so it is safer to create a default dummy fallback in the base
   folder. Alternatively, you can suppress the issue by adding
   tools:ignore="MissingDefaultResource" on the element.

   (This scenario frequently happens with string translations, where you might
   delete code and the corresponding resources, but forget to delete a
   translation. There is a dedicated issue id for that scenario, with the id
   ExtraTranslation.)

Change-Id: I52b01819998671fec019b148d9536ec0b80c39a4