Allow embedded and relation in raw query observed tables
RawQuery expects developer to list accessed tables in the
annotation. Many people got confused w/ the fact that it is
limited to the @Entity annotated classes. It becomes very
inconvenient to list all entities when you are fetching a
composite object (like an object w/ relations or embedded
fields).

This CL allows putting any java class there as long as it
has a relation or embedded field.

Bug: 74041772
Test: RawQueryProcessorTest, RawQueryTest (integration)
Change-Id: Ibf21ccb566a155a9ee9cee8075dd956c83056fe3