Supports expandable preferences in every PreferenceGroup.
Previously only the last preference in a PreferenceScreen could become
expandable. With this change multiple PreferenceGroups can have their
own expandable preferences and be displayed in the same
PreferenceScreen.

Nesting expandable preferences is not supported and will throw an exception.

PreferenceGroups containing an expandable preference must have a key defined
to properly save state between transitions.

Nested groups inside an expandable preference will only add the title of
the group to the summary of the expandable preference, and not the title
of their children.

Bug: b/68811148
Test: added
Change-Id: I176cee0edd6fad3382b5f2fb873496c3bf04591a