[WebView Support Library] Support incompatible WebView APKs.
Most functionality in the WebView Support Library requires an up-to-date
WebView APK.
We need a way of dealing with WebView APKs that are not up-to-date.
In this CL we ensure that if the current WebView APK is not up-to-date
(i.e. the support library cannot bypass the framework to call into the
 WebView APK) the support library declares that the APK supports no
features at all.

Bug: 76202025
Test: run androidx.webkit
Change-Id: I30ecf2459fc793c10f1817a431b18804344e2b8e