Convert MessagingStyle to use Person.
This change replaces the MessagingStyle.Message#mSender field with
 #mPerson as the backing field. This change also deprecates the
references to mSender (#getSender). This change takes into account the
possibility of an app using an older version of support, encoding the
sender into the Message bundle instead of Person and a newer versioned
app reading that bundle. We handle this by watching for the legacy
field and handle accordingly.

Test: ./gradlew core:connectedCheck
Bug: 72057404
Change-Id: I3cf1ecd57c59954bc775b63457110293adaed92b