Use setCustomSelectionActionModeCallback
Change Ib2348aef79a5153a40ed3ff39712879316cecab7 adds a possibility for
developers to fix 66939160 in support library. The bug causes the menu
created on text selection for ACTION_PROCESS_TEXT not to work properly
for O and O_MR1 devices.

This CL updates support library TextView extension widgets to use the
bug fix introduced in the previous CL.

Note that the first version of this CL
(I56e4a33d8849b46dce6a4e8c657490f4683b549e) had to be reverted as it was
causing an infinite loop between TextViewCompat and AppCompatTextView
(actually any extension of TextView). Now we are always calling the
super implementation of setCustomSelectionActionModeCallback to avoid
this situation.

Bug: 68991560
Test: none
Change-Id: Ic7e7411c3e1fc5735290e9707cdcf6ac2dbb3746