Rename two icon dimens to make clashes less likely
Apps will sometimes have their own idea of what a notification icon is
and how large it should be. We shouldn't be making that decision for
them if they happen to declare a dimen with the same name.

Follow-up to Ib1228cee0d368472a97c210b6d90007c953e2fbc

Bug: 77963298
Test: ./gradlew build
Change-Id: I7ddfe39efce8d5bfc64bc09dec441b2c3342a838