[WebView Support Lib] Add non-framework impl for WebMessage* APIs
Add support for the chromium (support library glue) path for WebMessage
related APIs.

WebMessagePortCompat is implemented in a way that's consistent with
per-method feature detection (each method call can either go through the
Android framework or through the support library glue).

WebMessageCompat does not call into both the framework and the support
library - because it's a POD class so the data it needs is created
before its ctor is run.

Bug: 73150657
Test: run androidx.webkit tests for an L-device using an up-to-date
build of Chromium (including the change this CL depends on).
Test: run a hacked version of androidx.webkit.PostMessageTest that
switches between using the frameworks and support library versions of
WebMessagePortImpl - to ensure the switching logic works.

Change-Id: Ie44f0f7af9de97423f01372a239bdbf523ebb40e