[WebView Support Library] split testOnSafeBrowsingHit into two tests
This splits testOnSafeBrowsingHit into testOnSafeBrowsingHitBackToSafety
and testOnSafeBrowsingHitProceed in WebViewClientCompatTest in the
support library. This is intended to be the support lib analog to
http://ag/3933408.

This also changes SafeBrowsingProceedClient and
SafeBrowsingBackToSafetyClient to inherit from a common
SafeBrowsingClient class.

Also, because not all devices support Safe Browsing, this changes an
'if' check into an assumeTrue statement, to make the meaning clear.

Bug: 77979250
Test: ./gradlew :webkit:connectedAndroidTest
Change-Id: I5e2fe877f41e71eb5dcf29d0cb6f03b6dc6ca804