[WebView Support Library] fix test for APK features
This fixes a test case designed to assert the WebView APK contains the
expected set of Features.

The previous design for the test was to assert that the WebView APK
supports at least all of the Features supported by the current version
of the webkit support library module. This design posed several
problems:

 * If we add new Features to the webkit module, old APKs will fail the
   test.
 * We can't intentionally remove features from the APK.
 * We can't support Features in the webkit module until we drop an APK
   (restricts development)

This CL changes the test to instead build a Set of the expected
Features, based on the APK's version. This also serves to document
precisely when we add or remove Features in the APK. This has the
consequence that we need to update this test before each time we drop an
APK (which adds or removes Features). However, this should at least
unblock automated tests.

This also prints slightly more helpful debugging information (it prints
the full Sets of expected Features and actual Features).

Bug: 79908426
Bug: 79907442
Test: ./gradlew :webkit:connectedAndroidTest
Change-Id: I6bc1ed03eead7fb11dc2ef734abf80e52251a4bd