Moved UnhandledKeyEvent support into ComponentActivity
Introduces a KeyEventDispatcher utility in core which can be
used by any window to enable full KeyEvent compat support. In
this case, it specializes for Dialog and Activity to handle
UnhandledKeyEvents at the correct times during dispatch and
ComponentActivity hooks up to it.

Also, synced up behavior with framework UnhandledKeyEvents

Bug: 79739636
Test: expanded tests in KeyEventsTestCaseWithWindowDecor
Change-Id: I8fb58f8a8d716afee7e13f194944695f2a2d76e7