Make PagedScrollBarView a custom view.
This allows the scroll bar thumb to be measured and laid out without
having the entire view be remeasured.

The measurement of the scroll bar thumb has to happen after the
containing RecyclerView has laid itself out.

Fixes: 73290584
Test: Boot up sample app and inflate a PagedListView. Ensure
requestLayout() improperly called error is not present.
Test: test performance of PagedScrollBarView before and after change to
ensure there is a performance improvement.
Change-Id: Id33ec4f14b14b4b46780b75489b7431d4fd191f9