Improve Jetifier's prefix configuration.
1) The reversed prefixes are no longer hardcoded in the Processor.kt but
are part of the config file.
2) Originally the prexifes were used as the first line of defense when
filtering what should be rewritten. This stopped to be intuitive since we
introduced quite a lot of types of rules. And it is not immideately
obvious when someone adds a new rule which is not covered by any prefix it
gets ignored. Because of that from now we will use the prefixes as a
check if we should report error with missing mapping or not or in cases
where checking the rules is expensive. So in case of package map we
directly check if package is there and if yes there is no need to look
at prefixes. If the package is not there but prefix is satisfied we
treat it as an error.
3) Fix bug with not dejetifying material package properly. First of all
the changes from 2) will make sure this won't happen again. But also add
suffix of "/" to make sure the prefix regex is matched.

Bug: 80208210
Test: Added, updated and manually checked that material
AndroidManifest.xml gets dejetified properly.

Change-Id: I0e899cf1f80e8393f1593c372f66609527d80610