Fix positional prepend
When new PagedList latches in, copy the mLastLoad position from the
previous PagedList and transform it to the item's new position in the
PagedList. This position is passed (our best guess for) the most
recently accessed load position. This ensures that the next PagedList
to be created will be initialized with a position relevant to the
viewport.

This fixes a bug where the mLastLoad position would not be initialized
on a fresh PagedList if no in-viewport re-binds occur.

Fixes: 80149146
Test: ./gradlew :paging:runtime:cC

Change-Id: I3d84f8728e39171d899d7135eca7a97ad479a19b