Allow Fragment navigate() calls in Fragment lifecycle calls
Previously, FragmentNavigator relied entirely on
FragmentManager for the current state, necessitating
the use of executePendingTransactions after each
navigate() call. By switching FragmentNavigator to
its own storage of the current state, it is possible
to switch to traditional queued fragment transactions.

Besides enabling the ability to call navigate() in
lifecycle methods (which didn't work before because
we were executing a pending transaction), this should
also enable use of transitions.

Test: ImmediateNavigationTest doesn't crash
Test: Manual testing in the test-app
BUG: 79632233
Change-Id: I7f823cb7e068e13845312065a0eafc938f099852