Apply pop first in OnNavigatorNavigatedListener
Avoid IllegalArgumentExceptions when popping from
a destination back to a nested destination when
the Navigator initiates the pop operation (i.e.,
when you hit the system back button when using
Fragment destinations).

By popping destinations off the NavController's
back stack first, we ensure that the subsequent
findDestination() call runs with the correct
current destination, allowing nested destinations
to be found.

Test: new NavControllerTest
BUG: 80453447
Change-Id: I71297c60d4a794cb43f5c55ffd1fb98f81a4ca9f