Fix NavigationUI when using nested start destinations
Updates NavigationUI to handle cases where
the start destination of the NavController's
graph is a nested graph. This ensures that the
correct element is selected in a NavigationView
or BottomNavigationView as well as ensures
that the up button only appears as an arrow when
there's actually a destination on the back stack.

Test: Ran the updated integration test app
BUG: 109675998
Change-Id: Iece8d71c177915baba058765ca04a909fd0b019a