Version-compat-test: Fix flaky test(Un)Subscribe
The reason for flakiness was that previous tests used unstable way to
check whether the callback is called. Depending on the timing,
isOnSubscribeCalled() could be called before the onSubscribe()
is called.

This CL rewrites the two flaky MediaBrowser2Test methods into the new
test class MediaLibrarySessionCallbackTest so that they do not flake:
 - testSubscribe -> testOnSubscribe()
 - testUnsubscribe -> testOnUnsubscribe()

Bug: 110053551
Test: ./runtest.sh 1
Change-Id: Ic424b2f8d014d8d5f6a1976bba5e17b844053d35