Use proper class loader for unparceling
MediaMetadataCompat and MediaDescriptionCompat unparcels the given
bundle without setting proper class loader in some points, and
MediaMetadataCompat.getBundle() returns the reference of bundle
which could cause crashes when developers added app private
parcelables into it directly.

This CL sets a class loader to avoid exceptions while unparceling,
and returns a copy of bundle when MediaMetadataCompat.getBundle()
is called. Also it removes unused code in MediaUtiles2 which has
codes somewhat related with this fix.

Bug: 77141224
Test: passed VersionCompatTests
Change-Id: Ide439114eedabb5a3919e65e4dc0240a566d0a11