Support singleTop for NavGraphNavigator
NavGraphNavigator shouldn't use
BACK_STACK_DESTINATION_ADDED when the singleTop
NavOptions flag is used and the NavGraph is already
on the top of the stack.

Special care is needed for nested graphs - singleTop
should also handle cases where NavGraph A starts
NavGraph B starts NavGraph C. If either A, B, or C
is navigated to with singleTop, that shouldn't affect
the back stack (since otherwise we'd have two copies
of the same stack - exactly what singleTop tries to
avoid).

Test: NavGraphNavigatorTest
Test: Sample project attached to the bug report also works
BUG: 109916080
Change-Id: Ie5269891c420dbb74fbe08dbb6fddb6598a7f618