NavUI: Add CollapsingToolbarLayout setup method
CollapsingToolbarLayout is a special layout that
you can wrap your Toolbar in that takes over the
display duty of the title, automatically resizing/
moving it based on how collapsed the layout is.

Therefore we need a separate setup method for
specifically the CollapsingToolbarLayout that
mirrors the Toolbar version, but calls setTitle
on the CollapsingToolbarLayout.

BUG: 110887183
Test: The code flow is exactly like the Toolbar version
Change-Id: If46d583f18670b0cbc50e9de4cc87e0731b5a44b