Fix bug where RV.smoothScrollToPosition fails sometimes.
Preivously, if RecyclerView.smoothScrollToPosition(...)
was called in RecyclerView.OnScrollListener.onScrolled(...)
in the final scroll update, it would not be honored.
Now it is.

Also updated subtle bug where if smoothScrollBy was called
with a custom Interpolator, future calls to fling would
also use that custom interpolator (despite the fact that
you can't provide a custom interpolator to fling. This is
also fixed.

Fixes: 112421248
Test: RecyclerViewBasicTest
	#smoothScrollBy_withoutCustomInterpolator_resetsToDefaultInterpolator
	#fling_resetsInterpolatorToDefault
      RecyclerViewLayoutTest
        #smoothScrollToPosition_calledOnPreviousFinished_scrollsToPosition
Change-Id: Idcb5bccec15fa5061f168ecfe95c7a516b886fd0