Pass default style values to parent constructors.
Pass the default attribute style attribute and style resource to the
parent class. This allows any attribute that the custom views do not
explicitly parse to be parsed by the parent class. For example, an
android:background specified in a style will now be applied correctly.

Fixes: 112423815
Test: ./gradlew car:connectedCheck
Test: Verify that a style with an attribute not explicitly parsed by
either view is applied.

Change-Id: I5acc7b5978833f23ab14c62dac7dde2465d754cf