Implement passing lifecycle events from FlutterActivity and FlutterView and subscribing to it in SchedulerBinding.
Important change:
1) Added :core as a dependency to have access to jetpack's lifecycle implementation
2) Currently forcing FlutterActivity to extend ComponentActivity from core module to have access to an LifecycleOwner. Having it we can skip writing again all the lifecycle events passing through all the classes. I believe later we will have ability to inject LifecycleOwner into FlutterView, otherwise we would just workaround it.

Also I decided to try BroadcastChannel from Coroutines for subscribing for out mapped lifecycle states. It is an analogue of PublishSubject from rx and can simplify our code: without it we would need to create an interface for a listeners, store list of listeners somewhere and then call all of them manually. Please let me know if you think the approach with listeners is better for us. An example of subscribing is in SchedulerBinding.

Test: n/a
Change-Id: Icb55bf87a75998be75249cdfe8c3dee03802ea4b