Fixed flaking test in RecyclerViewLayoutTest.
The test was confusing, so I first cleaned it
up a bit.  Then made the test more determinsitic
by triggering an operation synchronously on the
UI thread.  (Previously the operation was being
triggered from the test thread, to start on the
UI thread.  My suspicion is that the the animated
scroll was sometimes completing before the other
operation actually started.

Bug: 112157426
Test: androidx.recyclerview.widget.RecyclerViewLayoutTest
	#smoothScroll_destinationDeletedDuring_newDestinationReached

Change-Id: Ie17ba8480eff789298ba116d14c8ff3410070ab2