Fix OnBackStackChangeListener race condition.
The current logic blocks until the last Fragment is removed from the
FragmentManager but before the OnBackStackChangeListener is invoked.
This leads to a race condition between the teardown method of the
unit test that removes the OnBackStackChangeListener and the
reportBackStackChange method within FragmentManager that is invoked
shortly after the Fragment is actually removed.

Change-Id: I2375ee70abd3a40d79adf8351481b0694ee23e15
Fixes: 70727622
Test: this