Fix periodic work so it executes repeatedly.
Processor.onExecuted is the wrong place for cancelling things in
other schedulers.  It doesn't have the proper information about
what needs to be cancelled.  For example, PeriodicWork should NOT
be cancelled.  This should be done right before Schedulers.schedule
at the end of WorkerWrapper.

Bug: 112859683
Test: Added new tests for API 22-, manually tested using
integration app for 23+ since JobScheduler enforces a 15 minute
limit between executions

Change-Id: Ia0e6e17367fbc2493339b9b1275453267c571580