Complete matrix utils
The CL adds all the methods existent in matrix_utils.dart.
It also changes Matrix and Vector classes to use Doubles instead of
Floats everywhere (dart seems to rely on Doubles, Romain's
implementation was using Floats).
It also fixes the inc() and dec() operators to comply with kotlin's
practice that the operation should be immutable.
Tests will come in a further CL.

Test: not yet
Change-Id: I2fe0b4a3c49cbf619a3e3e30b07f1875b235bb84