Convert Window and Bindings into classes instead of singleton objects
With the original Flutter approach it is not possible to have more then one FlutterView as Window and all the Binding classes are singletons but bound to a specific view. For example Window.phisicalSyze it is actually a size of the view, not Window. And Window.render() is a command for a specific view, now Window to draw a widget. The same with bindings, all they are helping with some part of logic to render or schedule drawing of widgets in the specific view.
I made them just a classes instead of Kotlin objects and creating instances during the initialization of SimpleFlutterView.
Later we need to think about proper destroying of Bindings when the view is destroyed(they have some subscriptions) and maybe rename Window and split it into smaller parts.

Test: Tested manually that now we can have two SimpleFlutterViews in the hierarchy and both are rendered properly (didn't commit it to not add an additional complexity to hello world)
Change-Id: I0db9c7edb101be47e59ff86f2507c2b4168bfe2e