DynamicRoute: Revise sample routes of SampleDynamicGroupMrp
Since the name of an ungroupable route "DG Not Groupable Route" makes
confusion that it is totally ungroupable(exclusive route) which doesn't
match with API implementation.

So, this CL revises sample routes for possible test scenario that TVs
are ungroupable and speakers are groupable. Below are routes published
by SampleDynamicGroupMrp.

  - DG Dynamic Route 1 - TV
  - DG Dynamic Route 2 - TV
  - DG Dynamic Route 3 - Speaker
  - DG Dynamic Route 4 - Speaker

Bug: 114664342
Test: Manual test with support-v7-demos
Change-Id: I93e2037e6b515a2754799cb9553b743657170472