Prevent potential foreign key constraints.
Scheduling happens in a separate DB transaction from making things
as enqueued (for a variety of historical reasons that will soon be
changing).  This allows for a race condition where a synchronous
cancel/enqueue/enqueueUniqueWork can change the database and lead
to a foreign key constraint violation because the original WorkSpec
no longer exists.  This code handles that gracefully; we can
remove it down the road when additional infrastructure and changes
come online.

Bug: 114705286
Test: Added and ran tests.
Change-Id: Id1c6cf48b305a3526e2913a38ee0b5398c244ba4