Fix resources rewrite bug in Jetifier.
Previous CL for annotations processing (aosp/765522) relaxed the
selector for tokens in XML files. However the original filter was
preventing symbols like "?android:attr/textAppearanceMedium" to be
picked up for processing. After such token got into processing it
uncovered that we did not assert properly that we should not handle
cases with mixed "/" and "." as separators. For instance the factory
of JavaType from dot notation was supposed to assert case like this
for long time already as it is non-reversible operation.

Due to the new assert I have also discovered that we had this issue in
String processing already. However we were quite lucky as we apparently
do not have a strings that would mix "." and "/" separators.

Test: Added the xml snippet that uncovered the breakage
Change-Id: I224e274293f0c9a648892dfbb9aa6b98bb6d0881