[WebView Support Library] initial demo app
This adds an initial demo app. This is a ListView which points to other
Activities (each of which may similarly point to other Activities, or
may actually exercise a feature).

So far, none of these activities exercise any WebView features, other
than the MainActivity which adds the WebView version to its title (using
WebViewCompat#getCurrentWebViewPackage).

This creates an empty SafeBrowsingActivity for the sake of populating
the top-level ListView with something. We'll later populate
SafeBrowsingActivity with a similar ListView for each specific Safe
Browsing edge case.

Bug: 78154299
Test: ./gradlew :webkit-demos:assembleDebug :webkit-demos:installDebug \
    && adb shell monkey -p "com.example.androidx.webkit" \
    -c android.intent.category.LAUNCHER 1

Change-Id: I67724ff344d96b2576366f039591a22dd4c1eb3b