MediaSession2: Merge test timeout values
Previously timeout for expected failure was less than the timeout for
expected success not to waste time by spending time on expected timeout.
However, if we do, we cannot catch unexpected calls after the smaller
timeout. So we'd better to have only one timeout for both tests.

This indirectly fixes a flaky test
MediaBrowserCompatTestWithMediaLibraryService2#testGetRoot by increasing
the connection wait time from 500ms to 1000ms. Note that binding to a
service takes around 400ms, so we'd better to have relaxed timeout.

Test: $ ./gradlew media2:connectedCheck
      $ media/version-compat-tests/runtest.sh 1
Change-Id: I7c0773cd01df475bc9ef9aae5c0fcd6ba5b1f47e