Avoid using a non-final getLifecycle() on back pressed
Always use mLifecycleRegistry for the Lifecycle of
ComponentActivity when it comes to
addOnBackPressedCallback() since that method can be
called from the constructor (for example, in
FragmentActivity).

Ideally, getLifecycle() would be final and something
we could rely on always returning mLifecycleRegistry,
but that's a breaking change for current users of
FragmentActivity / AppCompatActivity.

Test: Activity tests still pass
BUG: 118690391
Change-Id: I7cd4d1477ff16bdeca41bdf794227fef9af8f83e