Fix crash in RTL environment
This fixes the crash in PrecomputedText with RTL context.
The root cause of this crashing is that the text direction is computed
lazily in measure time since the view might be detatched from
RootViewImpl.

This CL removing direction check from setText for PrecomputText and
start using Compat version in Android P since we cannot fix the
PrecomputedText in framework.

This doesn't contain the test cases since it is hard to set up the test
case in a reliable way.

Bug:113635044
Test: TreeHugger
Test: ./gradlew core:connectedAndroidTest on
API 15, 16 17, 18, 19, 21, 22, 23, 24, 25, 26, 27, 28, Q
Test: Locally verified the crash has gone.

Change-Id: I1aac8ba7583ec898b3d2017b8ac2058e15ca8403