Force external AndroidX dependencies to transitive
We use @aar when specifying prebuilt dependencies to generate
the correct pom but that also makes them non-transitive by
default. This requires declaring all of transitive dependencies
and is very manual to keep up to date.

This CL adds a dependency hook to force all of them to be
transitive without us needing to declare every dependency transitive.

Bug: 119426437
Test: local
Change-Id: I613b2ee72d1443989ddd0262081207da2b31a16a