Accessibility actions in viewCompat
The developer can now add an accessibilityAction using the support
library, by overriding perfom in AccessibilityActionCompat and passing
that into addAccessibilityAction in viewCOmpat.
A few parts to it,
1) Have accessibilityActions be associated with views using setTag.
2) On creation of the node info combine the ones on the view and the
delegate there.
3) On perform action, if the normal delegate call fails, try to perfom
an action using the ones tagged on the views.
4) AddAccessibilityAction takes a lambda paramter
(accessibilityViewCOmmand), which takes a CommandArguments parameter
(as opposed to a bundle), allows for stronger type checking and it makes
it cleaner
5) We ahve to use resource ids as actionIds because of a check in frameworks that doesn't allow actions outside of resourceId range, so make 32 of them beforehand
(no view should have anywhere near 32 actions, those people would be bad people).

Change-Id: I705742619c2ce0343ad97965efe4c3b393b9cc88
Fix:31493845
Test: Unit tests.