Fix flaky MediaPlayer2.reset() tests
reset() could be called just after a task was posted to the task handler
but before it started executing. At the same time as posting the task
mCurrentTask was set to reference it. In this case, the pending task was
cleared from the task handler, but we would wait for it to complete (it
would never run due to having been cleared).

This condition can be reproduced reliably by posting a task that does a
sleep to the Handler just before the actual task in processPendingTask,
which delays the Handler from processing the newly-posted task.

Fix the issue by clearing tasks only after the current task has run to
completion.

Bug: 80232248
Bug: 119207664
Test: Ran testReset repeatedly with Exo/MP implementations.
Change-Id: I057ac2d777c9487769c34b294ad2256721dc398e