Don't use ignoredColumns fields when finding POJO getters and setters.
When using ignoredColumns from the @Entity annotation, the processor
was still attempting to find getters and setters for such columns.
This change makes it so that ignored fields via the ignoredColumns
behave the same as @Ignore, in which no getter nor setter are required
for such fields, the fields are also not needed as a constructor params.

Bug: 119830714
Test: ./gradlew :room:room-compiler:test \
                :room:integration-tests:testapp:cC

Change-Id: Ie148b502403b73e09ac3a3328270aed0b13a3109