Media2: Fix timing issue in MediaSessionLegacyStub
TL;DR: Timing issue happens because ControllerCb#equals() wasn't defined

Longer version:
Due to the timing issue, MediaSessionLegacyStub may create multiple
ControllerInfos for a RemoteUserInfo.

In that case, those multiple ControllerInfos are considered as all
different because ControllerInfo#equals() only checks the object
equality of the ControllerCb, which is the wrapper of the
RemoteUserInfo.

This CL fixes the timing issue, defines ControllerCb#equals(), and
removes previously workaround in ConnectedControllersManager.

Test: $ gradlew media2:connectedCheck
      $ media2/version-compat-tests/runtest.sh 1
Change-Id: I19a3ee5e7328987f3531e91d4fa8f77f03eaa5f1