Remove unnecessary call to getAnimationListener
It is not necessary to wrap the existing AnimationListener, because the
animation is wrapped in an AnimationSet to begin with. Therefore the new
listener can just be added to that AnimationSet, and both the new and
the original listener will be called.

Also renamed EndViewTransitionAnimator to EndViewTransitionAnimation
because it is an Animation, not an Animator.

Bug: b/117519981
Test: ./gradlew fragment:connectedMaxDepVersionsDebugAndroidTest
Change-Id: Ib8e6510dac09adaff0c35cf4caeb6c7326aeb5cd