[WebView Support Library] fix API 1.0.0 txt file
This removes 2 APIs which were incorrectly added during the switch to
metalava (http://aosp/814037).

That CL was based on the state of src directories at
3478ef7cbc6b27d8c1497d76c3ffff688771380e (http://ag/4400052). However,
this is not a correct algorithm, since that commit is on the master
branch, and 1.0.0 APIs were finalized on the pi-androidx-dev branch. The
correct commit should've been a97ea0302322a843ac4d229244c27f46ce358c4e
(http://ag/4369068).

This manually edits the file (verified by diffing before
http://aosp/814037), since I'm not sure how to replicate the process of
the original metalava CL with the correct commit.

Bug: 120172279
Test: N/A
Change-Id: I92728e05a8bfb8481c12522f9a00a24ae640c684