Support deferred types on @Query with INSERT/UPDATE/DELETE statements.
Created separate value objects for a QueryMethod that is a SELECT
statement versus a INSERT/UPDATE/DELETE. This helps us diverge on the
binding path since non-SELECT statement are prepared and shared.

To support the various deferred types, Rx, Guava and Couroutines a new
set of binders and providers where created for the 'prepared queries'.

Bug: 120227284
Test: ./gradlew room:room-compiler:test \
                room:integration-tests:room-testapp-kotlin:cC \
                room:integration-tests:room-testapp:cC

Change-Id: Iad6c3f8bfaf58d6a5ac80bc2461f9290cea0c65a