Call getApplicationIdTextResource via reflection in Safe Args.
Safe Args depends on AGP 3.2 which doesn't declare
getApplicationIdTextResource. On 3.3 getApplicationIdTextResource is
recommended and on 3.4 getApplicationId is completely depreciated and
will throw. This changes makes it so that we call the recommended method
via reflection to temporarily support newer AGP versions while Safe
Args still depends on a stable one.

Bug: 119662045
Test: ./gradlew navigation:navigation-safe-args-gradle-plugin:test
Change-Id: I8539ce40d3f45ea0268db9b76dc6e9f14edda7f7