Add more FontFamily test font files
Each font contains [a,b,c,d,e,f,g,h,i,j,k,l,m,n,o,p,q,r] characters 
and [wide,narrow] glyph types. Each font file includes one character
that is represented with wide glyph, and others with narrow. This is
used in the tests in order to differentiate the font that is loaded for 
a specific weight/style in the FontFamily.
- 200 italic has [a] with wide glyph
- 200 regular has [b] with wide glyph
- ...
- 900 italic has [q] with wide glyph
- 900 regular has [r] with wide glyph

Test: None

Bug: 113815795
Change-Id: I6925c1b3e973c52d08d0ee31db619a4dce3e4459