[R4A] Add lifecycle hooks to the composer
The composer will now notify objects that implement ComposerLifecycleObserver
when the object enters and leaves the composition.

Technically a value enters composition when it is recorded in the slot table
for the first time. It leaves when the instance is removed from the slot table
the last time. This is difficult to describe in more practical terms, however.
More understandable lifecycles can be built using this primitive and are,
once they are created, recommended over using this primitive directly.

Also contains a fix to the frame manager where the invalidations could be sent
unnecessarily.

Fixes: b/122323961
Change-Id: I555e98f4c952553afc536eea9b3a66a23e0cfa66