Move useTextureView checkbox to VideoPlayerActivity in support-media-demos
Before this CL, there are two ways to change viewType of the VV in demo:
1. Checking the useTextureView box from VideoSelectorActivity
2. Changing screen orientation by tilting the device

It is confusing because there is also an attribute called viewType in
activity_video_player.xml and it is set to textureView. Even if tester
does NOT check the useTextureView box from the selector activity the
viewType of the VV will be textureView.

Also, switching the viewType when screen orientation is changed makes
testing the layout hard. When the viewType is changed, requestLayout
will be called one more time, and some testers want to test the
textureView in both screen orientation, portrait and landscape.

This CL resolves the issues.

Test: Run support-media-demos and manually test the checkbox
Change-Id: I715bf413002c09c502374ae6e3df7aa813fe4818