[wv] Fix loadUrlAndWaitForCompletion behaviour on the UI thread.
Currently the code calls mWebView.getHandler() in order to get a
Handler to assist with pumping the main event loop. However it's called
on a WebView that's not attached to the view hierarchy, and getHandler()
returns null in that case. The only reason that TracingControllerTest is
passing is because if loadUrlAndWaitForCompletion is called on the UI
thread, it doesn't actually wait for completion.

Bug: 122530218
Test: ./gradlew webkit:connectedCheck
Change-Id: Idf4c1e3479789801eec99e8698294ca157477d7d