Core adapter module with R4a Composable workarounds
I applied suggested by Chuck Jazdzewski approach of adding an adapter module without IR for ui-core components so they works better in a multimodule world. See b/120971484
Originally Chuck suggested this approach to make composable functions discoverable between the modules but apparently it is also helping to workaround some weird issues with Component classes as well.
I created adapters for MeasureBox and Draw. Now we will need to import this functions from android.ui.core.adapter instead of android.ui.core.
It helped to fix recomposing when we dynamically change some parameter for measuring or drawing composable.
For example you have a padding widget and you want to change its padding provided as a parameter after 1 sec. It will work when all the code is in one module, but to make it work between modules we need this adapters. More details here b/120971484

Also I added a new RecomposeTestActivity to verify recomposing is working for both update the parameter and adding a new widget cases. It turned out we also forgot to make Dimension class a data class so equals checks for it didn't work properly.

Test: added RecomposeTest
Change-Id: I9fd6b81477cb345f08730bb3bcc918b4ba261118