Hide ResolvableFuture and AbstractResolvableFuture from public API
We expose minimalistic safer CallbackToFutureAdapter, that allows wrap
simple callbacks as ListenableFuture. It is not targeted for any
complicated async pipelines, for such cases you should use coroutines,
guava, rxjava.

Unfortunately internal libraries already depend on that so it is
downgrade only to LIBRARY_GROUP_PREFIX. Going forward we need to migrate
libraries to CallbackToFutureAdapter

Test: ./gradlew lintMinDepVersionsDebug
Change-Id: I4d9999de56770c913995ef4544f510cbc93d4467