Fix popBackStack when nesting NavHostFragments
Turns out that FragmentManager.popBackStack()
automatically forwards popBackStack() to its
childFragmentManager. In FragmentNavigator's
case, we actually want it to pop a Fragment off
of the current FragmentManager, so make it explicit
on what we want to pop back to.

This fixes the common case where a Fragment
has its own child Fragments and back stack and
ensures that the FragmentNavigator's idea of what
is on the back stack matches the FragmentManager's
idea of what is on the back stack.

Test: new FragmentNavigatorTest test
BUG: 122770335
Change-Id: I8cd74b9a8a7a10c81f63473cf6a17b4ee91074ec