Fixes crash on dialog configuration change
Previously we tried to get a reference to the underlying preference
after a dialog had been dismissed due to rotation. Due to a bug with
Fragment, the target fragment containing this preference was not
available - so this crashed after rotating another time. This will be
fixed in Fragment anyway, but there's no reason for us to get the
Preference unless we have a positive result from the user anyway.

Bug: b/122167543
Test: Manually verified
Change-Id: Ibec8e88701a8a4c7dacf22a9f316db96ff73c59c