Adds preferenceCategoryTitleTextAppearance attr and cleans up layout files
There is no safe framework attribute for text appearance we can use for
the title of a PreferenceCategory, so we need to add our own attribute
that allows developers to set the textAppearance used by the default
PreferenceCategory layout.

Also extracts API specific styles out of layout files and defines all
padding attributes so we can remove a bunch of duplicated layouts.

Bug: b/123234058
Bug: b/123223813
Test: Manually verified
Change-Id: I2c470cb4ce0f54651836ebe46b1ff54f0a96bad1