Change names of conflicting classes.
ComponentNodes and their cousins were introduced in
the ui-port module and then copied to ui-core. The
copy makes it painful to find the right classes/files
because completion could take you to the wrong class.

This CL renames the old classes and files so that they
don't have the same conflicts.

Test: ran ui-port android tests
Change-Id: Iab18b16fe5d20ccf44f95f4e317490bea33b2776