Update Constraints class and add ConstrainedBox
The CL updates the Constraints class to reflect the current state of the
layout protocol. The ConstrainedBox widget is added; it can be used to
enforce additional sets of Constraints in layout hierarchies. The CL
also adds InfiniteDimension as a constant, replacing the previous usages
of Float.POSITIVE_INFINITY.dp.

Bug: 115463945
Test: none yet
Change-Id: I0a3d5eab3b0dbc82cb0616ba3f8a3612e261ead7