Remove a bunch of reflection usages from Transitions library
We made public a bunch of apis in Q which were annotated as @hide in platform before. Now we can use it not only after Q but as it is the same method which were available before we can link it directly with just wrapping the call into try catch for safety.
Methods usage updated:
1) View.animateTransform(Matrix)
2) ViewGroup.suppressLayout(boolean)
3) View.setTransitionAlpha(float)
4) View.getTransitionAlpha()
5) View.transformMatrixToGlobal(Matrix)
6) View.transformMatrixToLocal(Matrix)
7) View.setAnimationMatrix(Matrix)
8) View.setLeftTopRightBottom(int, int, int, int)

Test: run tests on Q, 28, 21, 19, run sample app
Bug: 117519704
Bug: 117521477
Bug: 117521009
Bug: 117520180
Bug: 117521189
Bug: 117520374
Bug: 117521349
Bug: 117521191
Change-Id: I44f4e388b13d1cb495613c88e1bd979d94ef72a2