item delegate is not being re-added on bind.
When an item view is added to the RV pool, its accessibility delegate is
set to null. ViewCompat adds a default delegate when the delegate is
set to null to maintain certain AccessibiltyDelegateCompat
functionality, so the item delegate isn't re-added when the view is
reused.

We now only add the item delegate when the view is created versus on
bind and the RV doesn't remove the item delegate itself.

Now, if a developer first adds their own item delegate, and then removes
it, the item delegate isn't re-added by RV. This seems like the expected
behavior.

Test: Tested with RecyclerViewAccessibilityTest and
RecyclerViewAccessibilityLifecycleTest. Tested with ViewPager-2 demos to
make sure the CollectionInfo was being updated correctly.

Change-Id: I365ff9f1792a1a81895bfd90b813602c79f82998