Few small improvements in VP2's tests and comments
1. Simplify SwipeTest.createTestSet
Modify the (adapterProvider to orientation) pattern to our common
pattern which is listOf(options).flatMap { option -> ... }

2. Improve comment on mTarget == NO_POSITION
Comment now reads ".. when mTarget == NO_POSITION .." instead of
".. when there is no target ..", which makes it easier to relate to the
actual boolean expression.

3. Assert that finalEvents has events before querying [0]
Before verifying assertions on finalEvents[0] in PageChangeCallbackTest,
verify the size of the list first to prevent out of bounds errors.

Test: ./gradlew viewpager2:connectedCheck
Change-Id: I31d4680f693f2fd588a1d14e7d9a1cdeebc03d3d