Add TextLayoutSpanTest class for span test
This CL creates a TextLayoutSpanTest to check the correctness
of spans in the bottom layer. This class is needed because
previously we only checked if spans are set as expected, but
there isn't any test to check if TextLayout is processing
spans correctly.
By adding those span tests, we can make sure that in each
"span transition" the TextPaint are set with correct value.

Bug: 121157451
Test: ./gradlew ui-android-text: test
Test: ./gradlew ui-android-text: connectedAndroidTest
Change-Id: I14742348605ebf132b2936aac4d5531f8d143c9a