item delegate not being re-added on bind.
When an item view is added to the RV pool, its accessibility delegate is
set to null. ViewCompat adds a default delegate when the delegate is
set to null to maintain certain AccessibiltyDelegateCompat
functionality, so the item delegate isn't re-added when the view is
reused.

Now during binding we check if the delegateCompat on an itemView is a default
delegateCompat by matching classname instead of checking for null. If
so, the itemDelegate is added.

Test: Tested with RecyclerViewAccessibilityTest and
RecyclerViewAccessibilityLifecycleTest. Tested with ViewPager-2 demos to
make sure the CollectionInfo was being updated correctly.

Change-Id: I4549a577cfc3c3aa45c569b0673833c8eabb505c