Only run small & medium tests when building all
When all modules must be built (because either no file change was
detected, or because a file change in root was detected), we no longer
want to return allProjects for the case of detecting CHANGED_PROJECTS,
since the test runner for that set runs @large tests. This resulted in
extremely long runs. In these cases, we still want allProjects returned,
but to the DEPENDENT_PROJECTS test runner instead. This one runs only
@small and @medium tests, making the runtime much more reasonable.

Bug: 127686528, 127794700
Test: Updated/ran unit tests.
Change-Id: I8f760146e1b25900e85446d214d9243684099dd2