Fix VideoCaptureUseCase crash due to surface being released.
DeferrableSurface.setOnSurfaceDetachedListener being called right after ImmediateSurface is created, however the listener is called immediately because currently surface is not attached yet. As a result, surface is released and cause the crash.

The solution is to make sure mCameraSurface.release() is always called at OnSurfaceDetachedListner callback, and we only call setOnSurfaceDetachedListener when we want to release that surface. This listener will be called right away if the surface is already detached or not even attached yet.  so we don't
have to worry that the listener is set after the surface is detached.

When the video recording is finished, VideoCaptureUseCase calls setupEncoder() again and also change it to release surface when surface is attached to fix the video recordig issue at legacy devices (J2 prime will halt the capture session after recording)

Test: manual test core test app / cameraview test with Video capture.
Change-Id: I8e78fa3038e5fcb75b81fc13905907b083522469