[wv] Terminate preexisting renderer before starting renderer tests.
We make sure to destroy WebView instances after each test, and this
causes the WebView renderer process to exit cleanly. However this does
not happen synchronously, and it appears to be possible for a newly
created WebView to revive the exiting renderer.

This CL mitigates this by creating a dummy WebView in the renderer test
setup, and terminating its renderer.

Test: ./gradlew webkit:connectedCheck
Fixes: 128551624
Change-Id: I6609592b21c06bd31a840c0da42e3fb8ff75bc42