Fix method ordering when using @ContentView+<fragment>
When calling through to a @ContentView annotated
constructor of FragmentActivity (or AppCompatActivity),
Fragments created via <fragment> are inflated before
the FragmentManager is moved to the CREATED state.

By adding a check around ensureInflatedFragmentView(),
we can avoid cases where onCreateView() and onViewCreated()
are called before the Fragment is attached and created.

Test: ContentViewTest and updated Navigation testapp
Fixes: 129907905
Change-Id: I278beedacf8633f84bfc918246ac8caef894ab1e