Simple and probably temporary solution to disable transition animations in tests
Introduced a DisableTransition test rule. When applied all the Transition tags will have no animations and will switch to the final step immediately. It will help us to write reliable screenshot tests.
The solution of disabling ripples completely when transitions are disabled is temporary. Probably we will need something better, but I don't want to add a lot of code for this and ripples are not using the Transition tag to reuse the logic.

Bug: 129678459
Test: Manually tested, CheckboxUiTest now works without animations.
Change-Id: I0168c19102650cc66214a5fe3a5e68fe0952e69f