Update ConstrainedBox and Container to use dp
Before this CL, ConstrainedBox and Container were taking Constraints
instances as arguments, inconsistently with the rest of the API surface
for layouts - Constraints has IntPx values, while layouts use dp. The CL
updates the APIs of ConstrainedBox and Container to use DpConstraints,
constraints working with dp, added to the ui-layout module.

Bug: 129758859
Test: ran all ui-layout tests and ran all demos
Change-Id: Iac03617ff39a6d9cf74803014e9d5ada1804d228