Fix DayNight updates when in background
We currently unregister the BroadcastReceivers
for battery saver mode in onStop(), meaning that
the Activity will NOT be recreated automatically
when in the background. This means that apps get the
normal system behavior, which can be broken on older
platforms. We now just let the existing cleanup happen
in onDestroy().

Also tidied up exception logging in
ActivityCompat.recreate().

BUG: 130307919
Test: manual

Change-Id: I4e090ed95c1644319feee21f3faae6663ba31c51