Restrict usage of animateLayoutChanges by documentation
When any ViewGroup in the hierarchy of a page has
animateLayoutChanges="true" (i.e., has a LayoutTransition), and a child
of that ViewGroup is added/removed during a scroll, the animation will
interfere with the layout calculations of RecyclerView, resulting in
gaps between pages. These gaps trigger an IllegalStateException in
ScrollEventAdapter’s position calculations.

When a child is added/removed, the LayoutTransition sets up a change
animation to that child's siblings and its ancestors. It waits until the
layout is updated (OnLayoutChangeListener) and for all relevant values
that have changed an animation will be started to animate that change
over time. Because RecyclerView moves the pages for the scroll that is
in progress, the position of the page (t/r/b/l) is incorrectly
identified as in need of animation.

In order to prevent this from happening, all LayoutTransitions present
in a page's view hierarchy need to have mAnimateParentHierarchy set to
false. It is not possible to fix this from within ViewPager2, because
LayoutTransitions can be added, removed and modified at any moment.
We can detect it though when the damage has already been done, and we
can address it in documentation.

Bug: 129530305
Test: None
Change-Id: I9311b6293fa9e91de2ab9f7bed159a7165fe87a8