Send broadcast to close system dialog before espresso test.
For BasicUITest, it might failed because some app crash and system alert
dialog popup.

Send a broadcast to close system dialog, this might mitigate these kind
problems.

The exception looks like following:
java.lang.RuntimeException: Waited for the root of the view hierarchy to have window focus and not request layout for 10 seconds. If you specified a non default root matcher, it may be picking a root that never takes focus.

Test: manual test with emulator API 26, API28.
Change-Id: I62daf092afdd2ba514be24f8f1978b3f4cd742eb