Remove jdiff and legacy doclava api tracking.
This change removes jdiff usage that was used for generating diffs
between library versions that no longer works due to independent
versioning.

It also removes tasks that use doclava for API tracking. We have moved
to using metalava for all the libraries and we have no plans to revert
back to doclava for api tracking.

Test: ./gradlew buildOnServer
Change-Id: I4f64e17ac5ecd70822181beefb369a9ef3f0c53e