Add repaint boundary
Repaint boundaries allow us to segment where invalidations
propagate in the hierarchy. On older platforms, Views are
repaint boundaries. On newer platforms, RenderNodes are
used for repaint boundaries.

This CL also allows for a difference between remeasure
and relayout calls. When a Model is read in the layout()
lambda, the Layout only needs to call the layout lambda
and not the whole lambda of the Layout when that Model
changes. This saves measurement of most of the children.
This difference is especially useful for something like
VerticalScroller, where the layout lambda only moves
its child, which is in a RepaintBoundary.

Test: gradlew connectedCheck
Test: ran ui-demos

Change-Id: Ibbf8c8f9de90984e2113a4e1f13fa8fa3904e66c