Remove back stack from NavGraphNavigator
When using navigate via deeplink across graphs, we add the appropriate
parents to the navController back stack, but we do not add any parents
to the NavGraphNavigator back stack. This causes an issue when we pop
a destination that was navigated to via deep link, where an extra
navGraph is removed from the NavGraphNavigator back stack.
Eventually, the NavGraphNavigator gets to the root and it is
impossible to return back from the deep link destination.

We no longer need the NavGraphNavigator back stack and can allow the
DestinationNavigator to manage pop operations. Removing the back stack
also means we can remove save state methods as well.

Test: New test, navigation tests
BUG: 132509387
Change-Id: I811282e359c0bc2ee04fd8aba4dfc6c9d50855f6