Workaround for jittery test execution in FakeDragTest
FakeDragTest makes a swipe gesture on a view, which translates the
motion events to a fake drag on ViewPager2. If the motion events are
delivered with jitter, the fakeDragBy invocations follow a different
timing pattern than the motion events. This can cause the velocity
tracked by the fakeDragBy invocations to be incorrect. This only happens
rarely, so it's good enough to retry a few times.

Bug: 132863303
Test: 1000x ./gradlew viewpager2:integration-tests:testapp:cC \
        -Pandroid.testInstrumentationRunnerArguments.class=\
        androidx.viewpager2.integration.testapp.test.FakeDragTest

Change-Id: I9f46cfe161d3de5b14ad2ac58434114144bebb60