Fix unit test ImageCaptureTest.camera2InteropCaptureSessionCallbacks() fail
The Interop callback will get preview and image capture results. If it verifies
that the last capture was an image capture, it might fail because the last result
might be a preview.

Therefore, modify the validation criteria to ensure that there is only one image
capture result from all captured results, not just the last one.

Bug: 133917045
Test: ./gradlew camera:camera-camera2:connectedAndroidTest -Pandroid.testInstrumentationRunnerArguments.class=androidx.camera.camera2.ImageCaptureTest#camera2InteropCaptureSessionCallbacks
Change-Id: I5b3949ecead80776144b795c62170297b2546820