Add Identity Credential Jetpack.
The Identity Credential APIs provides an interface to a secure store for
user identity documents.  These APIs are deliberately fairly general and
abstract. To the extent possible, specification of the message formats
and semantics of communication with credential verification devices and
Issuing Authorities (IAs) is out of scope for these APIs.

These APIs are modeled on the android.security.identity.* APIs (and the
Identity Credential HAL) which is part of Android 11 and Android API
level 30. Consistency between this new Jetpack API and the API in the
platform has been an explicit goal.

If running on Android 11 and the device has hardware-backed Identity
Credential support then this Jetpack simply forwards calls to the
platform API. Otherwise, an Android Keystore-backed implementation
will be used. While the Android Keystore-backed implementation does not
provide the same level of security and privacy it is perfectly adequate
for both holders and issuers in cases where all data is issuer-signed.

Our open-source mDL app at https://github.com/google/mdl-ref-apps
has been ported to use this code and works as intended both with and
without hardware-backed support.

This Jetpack is designed to run on any Android device with API level
24 or later.

Bug: 130658702
Test: Many new unit tests (both with and without hw-backed support)
Test: Tested with https://github.com/google/mdl-ref-apps
Relnote: Added Identity Credential Jetpack.
Change-Id: Icf90bf39c485c00acb4485c1b402c894cde3e317