Fix stop button (Ctrl + F2) in studio.
Running ./studiow launches studio as a background process using a
trailing & in bash, which causes bash to run the process with a signal
mask that blocks SIGINT due to POSIX-compliance reasons. This change
applies set -m shell option, which allows the backgrounded process to
run with job control enabled, which lifts this signal mask.

Test: ./studiow -y && A=$(ps | grep studio | cat /proc/`awk '{print $1}'`/status | grep SigBlk) && if [ "${A2:-1}" = "1" ] ; then exit 1 ; fi
Change-Id: I291c5ff21b3b4d2fc3581cedaba5c3d11ea0c47f