Remove border param from the Shape interface.
After further investigations with usages of Shapes API I decided it will not work nicely if we try to combine borders into shapes. Let's split this entities completely. We will later have separate DrawShape and DrawBorder components. The main reason is previously I wanted to combine all the drawing and clipping into the same Draw component, and it will work if we use Canvas clipping. But to achieve the clipping with antialiasing based on View/RenderNode outline clipping we should split our components a bit further. So our simplified surface will work like this:
DrawShape(shape)
Clip(shape) {
   children()
}
DrawBorder(shape, border)
We don't want to put DrawShape and DrawBorder inside Clip as we want to draw with antialiasing, and clip will support antialiasing not for every shape.

Test: none, just a structural change
Bug: 129278276
Change-Id: I50438762f1365023cdab083b2ab70f7d57dfad04