Call onCurrentMIChanged when setMediaItem is called twice
Calling setMediaItem twice is a unique case, and unlike skipping
to different items in a playlist, it executes the commands
setNextMediaItemInternal and skipToNextInternal, but none of these
commands call onCurrentMediaItemChanged after they are completed.

This CL also fixes the bug where skipping to a different media item
in a playlist does not call onCurrentMediaItemChanged.

Bug: 136526695, Bug: 136715413
Test: ./gradlew :media2:media2-player:connectedCheck
Change-Id: I34036eca6519397bb7b4f404d19b0fc8824b5973