Allow add() or replace() that take a Class
Rather than manually construct a Fragment to
pass it into a FragmentTransaction's add() and
replace() methods, leverage the FragmentFactory
already set on the FragmentManager to instantiate
the Fragment.

Besides allowing us to build reified versions of
these methods for Kotlin users, it also ensures
that both the intial creation and recreation
go through the same path.

Test: updated tests
BUG: 126124987
Change-Id: I726aee9a82d22c8803cf5e7aec858539c12d40bc