Fix IllegalArgumentException when RV used with Transitions on APIs below 18
When we introduced an override for suppressLayout we also added IllegalArgumentException into setLayoutTransition as this method is not supported anymore. But this is breaking a compat logic implemented as a backport for suppressLayout for APIs below 18 in Transitions. There we create some fake empty LayoutTransition just to eat requestLayout() calls. Currently this logic is causing IllegalArgumentException because of calling setLayoutTransition().
We can detect this case and instead of really applying this empty LayoutTransition just apply suppressLayout as it is the real intended behaviour. Later when LayoutTransitions will be cleared we set suppressLayout back.

Bug: 136475957
Test: new tests in RecyclerViewWithTransitionsTest
Change-Id: I45a2f6bed5adddf626c5fd0367d089f968d848a1