Use minHeight with vertical paddings for Button
Previously if you use the Button with some large font size the text would be clipped as we had hardcoded height 36.dp according to the spec. Now we treat this as minimum height and add vertical paddings for the text. Current constant(8.dp) is not yet finalized, but the logic will stay the same.

Bug: 135588104
Test: new test added
Change-Id: I4ddc3fc5acb12f4a244c351bb121352e981a196b