Ensure Transitions on exiting Fragments end before views are destroyed
Currently the FragmentManager starts Fragment transitions, and does not
ever consider how those transitions are executed when moving Fragments
through the proper state. This can cause an issue were Fragment views
are destroyed before transitions are completed.

This change allows FragmentManager to delay the call to onDestroyView by
considering whether a Fragment currently has exiting transitions. If it
does have exiting transitions, the call to onDestroyView is executed at
the end of the transition. The transitions are all merged so the call is
made only after the last transition has finished. EXIT transitions
(exit, return, sharedElementReturn) are the only transitions considered.

There are two exceptions:
One is that if the FragmentMangager is being destroyed (likely
because the activity is being destroyed), the Fragment view is destroyed
immediately because the transition is no longer needed.

The other is that we decided not to destroy the Fragment and it is being
given another view, so we need to destroy the current one immediately so
the new one can be created.

Test: Added test and added a check to all old tests
BUG: 136110528
Change-Id: I40f4244c0f7d7207cac32508f15c32affcff6142