Add BiometricPrompt device credential support for P
Currently, BiometricPrompt only supports authentication with device
credentials (e.g., PIN, pattern, or password) on Android Q. This is
because prior versions implemented device credential authentication by
passing an activity result. Because BiometricPrompt is hosted in a
client activity, this would cause the client activity to receive the
result, leading to incorrect behavior.

This commit fixes that issue and introduces support for device
credential authentication to P (which uses
androidx.biometric.BiometricFragment internally). It does this by
launching a transparent handler activity which intercepts the result
from the device credential authentication activity and calls the
appropriate authentication callback for the client.

Notably, this commit does NOT add support for device credential
authentication for versions O and below. This is because these versions
use androidx.biometric.FingerprintDialogFragment internally, which
requires additional work before it can support device credential auth.

Test: Biometric support demo app on Android O, P, and Q devices

Bug: 123595980
Change-Id: I8c9fb1431e9777cd5ff273d686bdbc5cb9f2409c