Replace framework constant with hardcoded value
Replaced android.R.attr.fillType with it's hardcoded value as it
is introduced in a newer API level. Keep the constant around
so we can parse and support the same functionality supported
API levels or backport it. This follows a similar pattern to
AndroidResources.java that includes hardcoded values of attributes
introduced in newer API levels.

Bug: 138922076
Test: Re-ran XmlVectorParserTest on Nexus 5 API level 22 emulator
Change-Id: Ia1d13ad128f6fc6ee5695a8538dd8309b06f60f0