Make TrustedWebActivityIntentBuilder build TrustedWebActivityIntent instead of raw intent
In an upcoming feature we will need to grant uri permissions to certain files prior to
launching a Trusted Web Activity. We need a TrustedWebActivityIntent class built by the Builder,
that holds the uris and grants the permission, similarly to how CustomTabIntent holds and applies
startAnimationBundle.

This CL does the non-backwards-compatible part of that - changes the signature of
TrustedWebActivityIntentBuilder#build.

Test: TrustedWebActivityIntentBuilderTest is updated

Change-Id: I03fb621220dd24cd59095ded8c55bba70305f6d3