Deal with overshoot of RecyclerView
When frames are skipped, RecyclerView can overshoot its target during a
smooth scroll. LinearLayoutManagerExtraLayoutSpaceTest relies on
reaching the target from a predictable direction, which is violated when
it overshoots.

This should be fixed in RecyclerView, but for now fix this test by
detecting the overshoot and scrolling 1px in the opposite direction and
back in the correct direction.

Bug: 131356269
Test: ./gradlew recyclerview:recyclerview:connectedCheck \
        -Pandroid.testInstrumentationRunnerArguments.class=\
        androidx.recyclerview.widget.LinearLayoutManagerExtraLayoutSpaceTest

Change-Id: I52c9baa5ddc7c851447289af316a9871c9260e57