Allow getBackStackEntry() for any destination type
Instead of only allowing getBackStackEntry() for
navigation graphs, open the behavior surface up
so that developers can access the NavBackStackEntry
for any destination on the back stack.

It is still recommended that within a Fragment, you
use the Fragment as the LifecycleOwner,
ViewModelStoreOwner, SavedStateRegistryOwner, etc.
as that is the one tied to the Fragment itself.

However, for custom Navigators that use destination
implementations that don't have access to these
interfaces, this provides an alternative to
implementing them separately.

Test: updated tests pass
BUG: 36818088
Change-Id: I834447cb563c18b48bc1c5c8e5a4c08b44539948