Adds inset FAB (cutout / notch) support to BottomAppBar
Moves away from the fixed 'center / end / center cut' fab positions,
so that now a fab can either be center or end aligned, and can have a
cutout in both configurations.

If cutoutShape in FabConfiguration is not null, the shape
will be enlarged on all sides and drawn directly around the FAB,
carving a notch into the BottomAppBar.

Also cleans up AppBarActivity a bit to work with the larger number of
app bar samples we now have.

Bug: b/139866383
Test: AppBarActivity, AppBarTest
Change-Id: Ia91488e12ffd536160991dbd30e7594638c639df