Add test for BiometricManager and improve existing test
Adds a unit test for the androidx.biometric.BiometricManager class,
making use of the androidx truth extension for assertions. Also updates
the previously added DeviceCredentialHandlerBridge test to use truth
assertions as well.

Note that the new test only exercises the case of using the framework
BiometricManager on API 29 and higher. Test cases exercising
FingerprintManagerCompat on API 28 and below are still forthcoming.

Test: ./gradlew biometric:connectedAndroidTest

Bug: 139874508
Change-Id: Ib1ba1e8f63ed7d6a426b42af99298c7909de29b8