Revert "Populate the correct ScrollDeltaX/Y values of AccessibilityEvent when"
This reverts commit c5e76dabc5a0f49c48e232e1895d521ba8a3dcff.

Reason for revert: This method won't be of much use. In P and Q the framework makes a copy of the event and sends this to a11y services, so the scroll values won't be populated in the first place.

cs/android/frameworks/base/services/accessibility/java/com/android/server/accessibility/AbstractAccessibilityServiceConnection.java?q=abstractaccessibilityserviceconnection&dr=CSs&l=1051

Change-Id: I8843f082b413f9739fbec1f60dd133007c03a753