Make provided alignment lines override inherited
Before this CL, if both a parent and a child were providing values for
an alignment line, the parent would merge their value with the one of
their child. This CL updates the logic to use the value of the parent,
without merging it with the one of the child.

Bug: 140415643
Test: ran AndroidLayoutDrawTest
Change-Id: I99538ab16e6fba8b6defae17b7bc284629226a41