Do not validate session info bundles on sender side
This CL makes MediaSessionCompat don't throw IAE even if the given
session info bundle couldn't be unparcelled on receiver side. In other
words, it keeps the protection on receiver side only while removes the
validation on sender side to avoid extra time and memory consumption
when writing bundle to a throw-away parcel.

Bug: 137796430
Test: ./media/version-compat-tests/runtest.sh 4 (on API 21 and 29)
Change-Id: I32ec1bee38bdf4435896f5d6be08149ec1cc7875