Deprecate Fragment getFragmentManager and requireFragmentManager
GetFragmentManager() does not clearly define which FragmentManager is
being retrieved. Deprecating it (and requireFragmentManager()) in favor
of getParentFragmentManager() creates an explicit API so it is clear
that this method returns the FragmentManager that this Fragment is added to.

GetParentFragmentManager() must return a FragmentManager or an
IllegalStateException will be thrown, exactly like
requireFragmentManager().

Test: All tests pass
BUG: 140574496
Change-Id: I0981a7319feccd0bd951b8e37b2906be0fd91345