Add dropdown alignment for the Popup
Add the feature of aligning the Popup below the parent, similar to
how the showAsDropdown() method does for the PopupWindow.
This however does not have the same visual effect, because it does
not include an animation for the moment.
Added examples inside the PopupActivity.kt file.
Modified some of the exaples to better ilustrate the Popup's behaviour.

Test: Added
Bug: b/140853156
Change-Id: I185954c5ef5b6a5f38c9c1aa3173a0dc938d89a1