Move CutCornerShape from ui-material to ui-foundation
All other shapes are in ui-foundation. Originally I thought it is good idea to keep CutCornerShape separately in ui-material as it feels really material design specific, but the previous user studies proved this makes hard to discover the existence of this shape as all other shapes are located in the same package of the same module. To improve discoverability I am moving this class near the other shapes.

Bug: 141172569
Test: CutCornerShapeTest passing
Change-Id: I29f05ada91ec02dd7c5cc2b20b2b862992dc5dc9