GreedyScheduler no longer ignores foreground work.
* Previously we would always trampoline off a background scheduler to
start working on a WorkRequest which needed to run in the context of a
foreground service. This was because we were considering enforcing the 10 minute
execution windows for such Workers. After some back and forth we decided that
this restriction was going to make Workers that need to run in the context of a
foreground service - far less useful.

Test: Added unit tests.
Change-Id: I91d7b94f9c902e7af9a35fcb21f2eac3220b9470