Use Log.isLoggable() for LoaderManager debug logging
Instead of requiring developers to rebuilt their app
with LoaderManager.enableDebugLogging(true), we
should also respect Log.isLoggable() to determine
what VERBOSE and DEBUG logging to output.

This also updates logging tied to the creation
and destruction of LoaderInfo to DEBUG level,
ensuring that this critical logging is more
noticable when logging is enabled.

Test: ./gradlew checkApi
Change-Id: Iac989721d2f6244fcbd316f7930ef5e9bdb3ac34