More declarative Draggable
This CL indended to make Draggable "more declarative", by making it
controlled, which means now the user is the owner of the value that is
being dragged. This leads to a very good improvements, such as:
-- Consumers of this value are not bounded by recomposition anymore and
can choose when they want to read value, therefore cause only
relayout/redraw
-- There's a choice for users to NOT to update value if they don't want
to by simply to nothing in `onDragValueChange` callback to controll flow
100%, which is thedirection we want to drive all our APIs to.
-- There's no intricate classes anymore invloved in Draggable API. No
controllers or some magic-effects, only @Composable and Value.

To achieve this, ValueHolder was introduced in ui-foundation, which anyone can
use in order to pass not the value itself, but the ability to read this
value in the future when it suits clients better.

Bug: 142385777
Test: All tests passed

Change-Id: Id84fdeb4c127e761c659a22bfc6cceae0632bc49