Avoid redundant notification of TrackInfo and VideoSize
This CL removes additional calls of onTracksChanged and
onVideoSizeChanged when current media item changes because they could
be called with wrong values (i.e. the previous tracks and video size).
The correct values will be notified when they are actually updated by
player (typically right after onCurrentMediaItemChanged).

It also modifies a flaky test in VideoView_WithPlayerTest in order to
get more information when the test fails.

Bug: 140270775
Test: ./gradlew	:media2:media2-player:connectedCheck
  :media2:media2-widget:connectedCheck
Change-Id: I0cb4f003dd0db07886ba91e1d5ec3912406b211d