ItemDelegate should not call view implmentation of onInitializeAccessibilityNodeInfo twice.
This is problematic for ViewGroup, which will add the children twice.
I may change that on the platform.
The problem is if an AccessibilityDelegateCompat is associated with an item view when it is not bound,
The new item delegate will take a reference to it, and call it also for accessibility functions.
Both end up calling into the view implmentation of onInitializeAccessibilityNodeInfo.

Change-Id: Id70111fade49e81b02c0e2571984bb158c0c3aa2
Fix: 141754356
Test: RecyclerViewAccessibility*Tests, added a test in RecyclerViewLifeCycler too. Added accessibility action switch to support-v7 in the nested recycler view, which allows us to quickly verify that complex hierarchies work with the delegate net of Accessibility.