Smarter Ripple color selection
In current implementation we just use primary on light backgrounds which is not comply with the specs. Instead the color of the ripple should be taken from the text or iconography in the component. To make it automatically work for majority of cases I updated the logic to apply the same logic we use for defining text color based on the background. For example if the background is `surface` we will use `onSurface` for Ripple. Then if the color is custom we will use White if the background is dark, and Black in the opposite case. For the cases when automatic logic is not working I provide ability to apply any custom color as param of Ripple. For example I now use this logic in Buttons: when ButtonStyle was configured to apply `primary` as a text color I would also apply this color as the ripple color.
Additionally I implemented ripple color opacity requirement from specs: Opacity is 12% on light theme and 24% on dark theme. To make this work I added an additional callback to RippleTheme - opacityCallback

Bug: 124500407
Test: tested manually, new tests in RippleEffectTest and TextColorsTest
Change-Id: Id82263ff83622915c7524bbd9e4480d8ecabf24b