Make sure we call into custom item delegates.
This is quite annoying because getItemDelegate() doesn't ahve to be an ItemDelegate.
That means if someone is using AccessibilityDelegateCompat here instead, they will miss out on layout manager, and original delegate functionality.
That being said we should do our best here and accomodate for all the people who are using getItemDelegate() with an ItemDelegate.
In theory we can push out the original delegate code from ItemDelegate into AccessibilityDelegateCompat. But I decided not to do that because
a) It's too late to make a change that risky.
b) It will create a dependency between RV and Core.
c) We are deprecating all of this item delegate, recycler view delegate stuff soon.

Bug: 141895738
Test: Added a custom delegate unit tests to make sure we don't lost this again.

Change-Id: I8cc9eea8c82252dd40963f66cd97ac76007e895c