Remove view destroyed check in FragentTransitionAnimTest
This change makes broken Animation tests more closely reflect the
passing Animator tests. That includes explicitly declaring the animation
duration time on the fragment at the beginning of the test, and removing
the view destroyed check on Animation End.

This change is mostly to check what issues could be causing the
Animation tests to fail.

Test: all tests pass locally
BUG: 140580302
Change-Id: Idc0cd5b3310d21b469613c16f73465cc908aed83