Temporarily disable Compose IDE plugin for androidx
The version of Kotlin we are using with androidx is incompatible with
the version of Kotlin we are using with studio, which is causing the IDE
to crash in unfortunate ways.  Since we switched to FCS, the IDE support
is less nessary.  Let's disable until we get studio updated, and then we
can re-enable and rethink our circular dependencies.

Change-Id: I5c5897c5f9afe6144399bd4c097c13ac35540863