Unifies dependencies across Compose and main project
Previously we split dependencies (such as Kotlin versions) in both
build_dependencies and Dependencies, and there were some dependencies
(Lint) that we didn't split, when we should. Havig two places where we
set these becomes really confusing, and hard to maintain. This CL splits
lint dependencies to match AGP version, and ensures that versions that
*can* change in Dependencies.kt always come from build_dependencies, as
the source of truth.

Test: ./gradlew bOS
Change-Id: I99f62beb99bfcda84076f2066765c1298fc416f1