Prevent jetifier from rewriting unmodified libs.
In case copyUnmodifiedLibsAlso is true, jetifier also copies libraries
that were not changed. However it wasn't doing just pure copy but it
was dumping content from the memory. Such content could be changed due
to formatting differences. This was breaking signatures. Current fix
makes sure that we copy the files directly and preserving checksums.

Bug: b/142580430
Test: Added
Change-Id: I533fd823495032cd40123ba4bf2461e1d98f1e2d