introducing raw image modes in slices
Include an additional image modes that allows developer using custom
sized image in slices at their own discretion.

Developers can now use either RAW_IMAGE_LARGE or RAW_IMAGE_SMALL for
images in slices.

In a RowView, the image will be rendered at its original size, as
opposed to getting resized to 48x48. Note that if the image is too
large for the row, it will be cropped to fit into the row.

In a GridRowView, the image will be scaled wtih CENTER_INSIDE if
its too large for the grid row, as oppose to CENTER_CROP.

The difference between RAW_IMAGE_LARGE and RAW_IMAGE_SMALL is backward
compatibilities. If the call-site is using a previous version of slice
view library, RAW_IMAGE_LARGE will be considered equivalent to
LARGE_IMAGE, and RAW_IMAGE_SMALL will be SMALL_IMAGE.

Test: manually verify in SliceBrowser following slices
/notification -> verify image in the first row is 36x36
/cat -> verify the image scale mode is CENTER_INSIDE
for backward-compatibility test, remove logic related to hint_raw
and check the slices
/notification -> verify image in the first row is 48x48
/cat -> verify the image scale mode is CENTER_CROP

Bug: 144180711
Change-Id: I4357dd0dcca5405433fb902dd8287e008b123e47