Slider material component
This CL adds slider as a part of material components.

Alongside with other input-based components, Slider designed in the way
that itself it is trying to hold as less state as possible, while
providing flexibility for developers to work with the state if they want
to own it.

Supports descrete values via step count and via granular support of
values that concrete slider could have.

Bug: 113856614
Test: included in this CL
Change-Id: I002e004f6ad7309eb9809f4c8886496f03fbaeb8