Add ANDROID_HOME env variable to gradlew & studio
This sets the ANDROID_HOME environment variable by default for studio
and for running gradlew tasks from the command line. Tests for custom
lint checks need an sdk defined, but will use the sdk defined by the
ANDROID_HOME variable if the sdk is not explicitly defined. This change
is primarily to remove defining the sdk in each lint detector test.

Test: ./gradlew bOS
Change-Id: Id34fe5648df1792595932eb915cf1995b90fd765