Expose HandlerExecutor publicly
Since API guidance is to standardize on Executor only APIs moving
forward, apps which have not converted everything to Executors will
require a trampoline to to use the new APIs. In addition, this class can
be used to substantially simplify implementation of APIs that need to
support both Handlers and Executors (simply wrap Handlers with this
class, and then treat everything as an Executor). Finally, while this
class is relatively simple, trivial implementations usually contain bugs
(see other HandlerExecutor implementation which do not properly throw
RejectedExecutionException or NPE).

Bug: 145015834
Test: HandlerExecutorTest
Change-Id: Iea942e14ab9f610bdec86faefa54c971d0f0a26f