Multi-dimensional animation
This CL made a fundamental change to the animation infrasture to support
multi-dimensional animation. The previous implementation assumes all
animation to be 1D. As a result, positional velocity cannot be tracked
in one animation. And instead multiple animations, one for each dimension
of the position, would need to be created.

The new change requires the dimensional info to be explicitly specified in
AnimationBuilders. This enables multi-dimensional velocity tracking in
the form of Vectors, which is also introduced in this CL. Also included
in this CL is support for creating a multi-dimensional animation from
a 1D animation by using the same animation on all dimensions.

Test: included in this CL

Change-Id: I6e3d1a5674072bde28b64486552e942fc85ac5d0