Implement LoadHint -> PageEvent stream
Paging logic to convert a stream of LoadHints from RecyclerView bind
calls to PageEvents, which are incremental updates as pages are fetched.
This is implemented the new Pager, PagerState and PageFetcher.

A Pager is a snapshot / single generation representation of the backing
PagedSource. Converts a stream of LoadHints / bind calls from
RecyclerView into PageEvents by making calls to PagedSource.load. The
stream of PageEvents are exposed via PagedData. A Pager lives within a
job that is cancelled whenever the PagedSource is invalidated.

PagerState holds state that must persist across individual resolution of
LoadHints. It must be accessed with a lock within Pager as it is shared
between async loads (start / end can happen at the same time).

PageFetcher converts a stream of refresh tokens, which are sent whenever
a PagedSource is invalidated, into a stream of PagedData. PageFetcher is
multi-generation across generations of PagedData, created using Pager,
PagerState pairs.

Test: ./gradlew paging:paging-common:test
Change-Id: Iff0c232746592ad0ac8c2c5f908085663a957f72