Fix flaky FragmentTransitionAnimTest
FragmentTransitionAnimTest using animations are failing on API 21
devices because the tests were requiring animations to get an
onAnimationStart callback.

Window animations should be disabled for all androidx tests so this
should not be an issue, but it seems that animations still attempt to
run in API 21 tests.

This change removes the requirement for animations to start and only
starts the endAnimationLatch if an animation is actually run.

Test: change in test
BUG: 140580302
Change-Id: Ia6beb5b392c6d17c01d44e44c750657a03668968