Fix position of PointerInputWrapper in test layout
The test layout of SendSwipeTest places the PointerInputWrapper at the
highest level in the composition, meaning it catches input on the entire
screen instead of just on the widget where we inject the swipes. To be
sure that we verify pointer input on that widget, we need to only listen
to that widget.

Bug: 143480427
Test: ./gradlew ui:ui-test:cC
Change-Id: I627da97253a3c8c10094b1e6cc3842e2c8c5609b