Fix incorrect size for parent data modifiers.
Before this CL, the size returned by the ModifiedParentDataNode wrapper is always zero(Unmeasured) since this modifier doesn't set the size.
After the change the ModifiedParentDataNode will return the size of the wrapped node instead.
But because ModifiedLayoutNode extends ModifiedParentDataNode, there's a check that fallback to wrapped size only happens if the size is Unmeasured. This will ensure the correctness of ModifiedLayoutNode size.

Test: new modifiers_validateCorrectSizes() test; ui-layout, ui-material, ui-framework
Change-Id: Idc88669e70d937f64fcfd4f78db6357d9dc5659f