Indicate null behavior in findFragmentByTag docs
FragmentManager's findFragmentByTag() accepts a null tag, but if a
Fragment with a null tag exists in the FragmentManager it will never
actually be found.

FindFragmentByTag is public API and changing the behavior could be a
breaking change, so we should adjust the documentation so that this is
clear to users.

BUG: 146936114
Test: ./gradlew bOS
Change-Id: I42b81e0cb5fa9816550f9aa795a6db0936455109