Buffer / conflate last hint sent to hintChannel across generations
Provides an initial value to new generations of Pager based on the last
hint, to allow new generations to fulfill prefetchDistance after REFRESH

Improves the use case where a developer implements getRefreshKey
in a way that returns the same list after invalidate, which would cause
a user who has already scrolled to an end of a list to be unable to scroll
further in that direction.

Conflation is handled within AsyncPagedDataDiffer, which has access to
DiffResult returned by DiffUtil, so it can map ViewportHints across
generations.

Fixes: 147155784
Test: ./gradlew paging:paging-runtime:test
Change-Id: Ic16116a64cbee1667b5fbc54ce74e37949628edf