Ensure childFragmentManager gets moved to attached
If you add a child fragment to a fragment's childFragmentManager before
the fragment is in the CREATED state, the childFragmentManager never
moves to the ATTACHED state. Since ATTACH is never dispatched to the
childFragmentManager, it does not get a host or a viewModel. This causes
a NPE when attempting to move the fragment to the CREATED state.

This change dispatches ATTACH to childFragmentManagers when its parent
fragment is attached, therefore it is attached when the fragment
attempts to move to CREATED.

FIXES: 146441558
Test: Added FragmentLifecycleTest
Change-Id: I78a0784f63648b2f869eb788cf3e205a87f6d76f