ImageCaptureTest.canSupportGuaranteedSize fail by NullPointerException
[Root cause]
The test use mockito API to do "answer" and "verify" on the same API
onCapturedSuccess() but in different thread. It will cause a race
condition and fail when the actions after "verify" is running before
"answer" task is complete.

[Solution]
Wait the "answer" task to be complete and then do actions after "verify".

Bug: 147728563
Test: ./gradlew camera:camera-camera2:connectedAndroidTest
Change-Id: If171ea49c53cee5dc96baef657f4865c0d996b21