Fix for Fragment menu visibility issue in FSA
The fragment's menu visibility is directly related to the fragment's max
lifecycle. When the max lifecycle is STARTED, the menu should not be
visible. When the max lifecycle is RESUMED, the menu should be visible.

This CL fixes the bug where a fragment that is added to the fragment
manager by FragmentStateAdapter did set the max lifecycle to STARTED,
but did not set the menu visibility to false.

Relnote: "Fixed `FragmentStateAdapter` issue with initial fragment menu
visibility when adding a fragment to the `FragmentManager`."

Bug: 144442240
Test: ./gradlew viewpager2:viewpager2:cC \
        -Pandroid.testInstrumentationRunnerArguments.class=\
        androidx.viewpager2.widget.FragmentLifecycleTest

Change-Id: I9d2ffa0efcb03b63eeca25f2d4e74c81446a5774