Add support to find components in popups in tests
Finders like findByTag weren't able to find nodes inside popups, because
we only searched in the main application's window, and popups are in
another window. This CL searches through all windows (roots) and
searches for all SemanticsTreeProviders (AndroidComposeViews) in all of
them.

In order to find the root view in each window, PopupLayout is now
assigned the id android.R.id.content.

Bug: 139861182
Test: ./gradlew ui:ui-test:cC && ./gradlew ui:ui-foundation:cC
Change-Id: I0b71ddfb900e89dee696a11bd1c0991888c61349