Ensure findFragment works with <fragment> tag
When calling FragmentManager.findFragment() with the view of a fragment
that was added using the <fragment> tag, that fragment was being ignored
(either the parentFragment was found or it was crashing saying the view
did not have a fragment set).

This change sets the proper view tag on <fragment> tag inflated
fragments.

Test: Added test
Bug: 147784323
Change-Id: I4346b694ba2af1fab1f68a879df96beaf129cd0c