Introduce TextState and TextFieldState to store less objects in the slot table
This change allows us to use only 2 slots instead of 17 in runtime slot table in Text, as we have only one remember and onCommit with one param instead of a few remember/onCommit calls with a lot of params.
We not use just an object fields instead of state for storing TextLayoutResults and LayoutCoordinates which optimizes reads and writes for this properties as we didn't really rely on the state mechanism for this fields. We also don't recreate TextSelectionDelegate us we can reuse the object with the lambda delegates. Similar optimizations were applied for TextField

Test: manually, unit tests
Bug: 148767197
Change-Id: If45213c8fad27860c7f33cf81534671cce94c062