Re-enable animations for transition tests
Recently we disabled all the animations for all our tests. This change re-enables them (changes animator duration scale back to 1) for the tests which requires it. The tests need it because of their asynchronous nature. For example we switch to ui thread and start animation. Then we switch back to the test thread. wait here for the animation start callback and assert the state is correct while animation is running. With disable animations when we switched to the another thread the animation is always already not only started, but finished as well.

Fixes: 146997833
Fixes: 146169628
Fixes: 146199009
Fixes: 146778652
Test: all the tests are passing locally
Change-Id: I04b738849afb19573f8d22257981db6208e43eb2