Moved UseCaseSurfaceOccupancyManager from camera2 to core
Moved the UseCaseSurfaceOccupancyManager class from camera2 to core and
renamed it to UseCaseSurfaceOccupancy to better reflect the purpose.

The UseCaseSurfaceOccupancy check is now done in CameraX.java before
resolution calculation instead of as part of the resolution calculation.

Relnote: "Enforce that only a single VideoCapture and single
ImageCapture can be used at one time as part of core instead of camera2."

Bug: b/147698557

Test: ./gradlew camera:camera-core:connectedCheck
      manual test of core test app

Change-Id: I48e1c31df058b2237da6c4a447f8b4465f31b498