Fix RotationTransformTest.
RotationTransformTest was setting the screen orientation by calling setRequestedOrientation() on FakeActivity. The reversePortrait() test was failing, setRequestedOrientation(REVERSE_PORTRAIT) wasn’t rotating the display by 180 degrees. I read that this might be because the activity needs a certain screenOrientation flag to be set to sensorFull in the manifest to support all orientations.

This CL tests RotationTransform in an other way, and makes sure the degrees it returns are correct.

Bug: b/149937115
Test: RotationTransformTest.java
Change-Id: I4d231217ed00bb1db21f4540aece0c809264dc32