Preview parameter source API
Adds a new @ParameterProvider annotation and the
PreviewParameterProvider interface that allow create data sources that
can be fed to a @Preview parameter. The PreviewParameterProvider acts as
a generator that can create a potentially infinite number of parameters.
This allows easily having multiple previews that will be rendered using
different input information.

Also add an initial sample data source that can generate Lorem Ipsum.

Bug: 139476601
Test: New tests pass
Change-Id: I4e030d7bb6f0fe090afcc92a43ed8ee0d2afcc02