Don't export hidden FTS columns in schema bundles.
FTS tables always have a hidden rowid table and an optional
languageid. When exporting the schema to a bundle that gets
serialized to a JSON, we shouldn't include such columns
because during verification they will not be available when
executing PRAGMA table_info.

This fixes users of the testing library that had migrations
to an FTS table with either a named rowid column or a languageid
column.

Bug: 145858914
Test: FtsMigrationTest
Change-Id: I3ff09337c861fdca6056c2bce29f2873dd8166c1