Allow calling getLayoutInflater in onCreateDialog
Calling getLayoutInflater in a custom DialogFragment's overridden
onCreateDialog method causes a StackOverflowError because
DialogFragment's onGetLayoutInflater relied on calling onCreateDialog.

In onCreateDialog, we should return the layout inflater from the
Fragment onGetLayoutInflater instead, which has no idea about the
context or theme of the Dialog. The layout inflater returned in
onCreateView will remain the same, but the two will be different.

This also fixes an issue where FragmentContainerView and the fragment
tag cannot inflate fragments within a DialogFragment. The inflated
fragments are properly added to the DialogFragment's
childFragmentManager.

Test: Added DialogFragmentInflatedChildTest test
Bug: 117894767
RelNote: "Fixed a bug in DialogFragment that caused a StackOverflowError
when calling getLayoutInflater from within onCreateDialog."

Change-Id: Ib55b53de9ccd5ccb69111d519fd7ec02fbf9726e