Enfore rect format for HINGE and FOLD feaures.
Fold display features are expected to have a zero-area rect
that spans the entire length of the window (width for horizontal
folds and height for vertical folds).

Hinge display features are expected to have bounds that span the
enture length of the window (width for horizontal hinges and
height for vertical hinges).

Test: DisplayFeatureTest
Test: ExtensionCompatTest
Test: SidecarCompatTest

Fixes: 150899271
Change-Id: If4768905c5d7e32e17dc60fe16898490b1693edd