Ensure DialogFragment call setContentView before shown
The lifecycleObserver in dialog fragment does not get an onChanged call
until the lifecyle moves to STARTED, which means the call to
setContentView happens after the dialog is already shown.

The viewLifecycleOwnerLiveData should use observeForever, instead of
observe, to ensure that setContentView is called in time for it to have
an affect when the dialog is shown. It is okay for DialogFragment to use
this since is manually removes the observer in onDetach.

Test: add DialogFragmentTest test
Bug: 151652127
Relnote: "Fixed a regression from Fragment 1.3.0-alpha02 that caused
DialogFragment set the content view after the dialog was shown."

Change-Id: Ia27adfcd90d7a9420fcec0309c9cba774f600311