Ripple converted to Modifier
Apart from it:
Semantics emitting components like Clickable or Toggleable now have a modifier param as we need to pass ripple modifier before the gesture detector modifier emitted by this components. This will be simplified further when we convert semantics into modifiers.
True is now a default param for "bounded" param. Seems like we can say this is more popular usecase(for example there are more usages like this in Jetnews, and the most of usages of bounded = false will be replaced by Icon component)
I wrote more tests for Ripples using the bitmap comparision api.

Fixes: 151331852
Fixes: 136804289
Fixes: 150060763
Test: manually on material demos, new tests
Relnote: Ripple is now a Modifier. While Clickable is not yet converted the recommended usage is Clickable(onClick = { ... }, modifier = ripple())
Change-Id: Ie52007f6948838a64fb25dba4dfbb7853d0e442f