Reset capture session surface if it is using an invalid surface.
When using Preview with a SurfaceViewImplementation, there's a race condition where the capture session might be opened before SurfaceView's onSurfaceCreated callback is invoked, which causes a crash. This CL makes it so that when the surface the capture session is using is invalid, a new surface needs to be provided.

Relnote: "Fixes a previous known issue where PreviewView's surfaceView implementation wasn't working well on certain devices, and would cause the app to crash after resuming preview."
Bug: b/152729941
Test: PreviewViewAppPreviewUpdateTest.java
Change-Id: I5ed6b7d6ee6a4d9aeb6643e06440597f9674191c