Avoid destroying the target fragment's view too early
When using setReorderingAllowed(true) with a replace
operation that has the incoming fragment with a target
fragment of the outgoing fragment, the new fragment
goes through attach() before transitions are started.

Unconditionally calling moveToExpectedState() when
USE_STATE_MANAGER is false will move the target (i.e.,
the first fragment) all the way down to CREATED since
it is no longer added.

By checking the target's fragment state, we maintain
the guarantee that the target fragment is at least
created when the other fragment is attached without
moving the state downward.

This doesn't affect the USE_STATE_MANAGER case since
the expected state will be AWAITING_EXIT_EFFECTS and
the view won't be destroyed.

Test: new FragmentTransitionTest test
BUG: 152023196
Change-Id: I41f2f8cf3f0a33e846be91f0356adb197be9a386