Only call onRestoreInstanceState on a dialog once
With previous changes to DialogFragment (aosp/1263328), dialog fragments
with views call onRestoreInstanceState twice on their dialogs. Once when
the dialog is prepared (which is required for dialog fragments without a
view) another in onViewStateRestored (which is required for dialog
fragments with views to be restored after dialog.setContentView()). This
causes problems in BottomSheetDialogFragment where it only accepts the
first call to onRestoreInstanceState and ignores the second, which is
the one it actually needs.

We should make sure that we only call onRestoreInstanceState once so the
state restored is always the proper one.

Test: tested in testApp
Fixes: 152225734
Change-Id: I50237ea5daf36b59ae165946f0b09098c47c0a06