Implement HiltViewModelProcessor
Add code generation to the processor. Two sources are generated per
ViewModel whose constructor is annotated with @ViewModelInject. One
class is a ViewModelAssistedFactory implementation for doing assisted
construction of a ViewModel who might have requested a SavedStateHandle.
The second class is a Dagger module with Hilt's InstallIn for adding
the generated ViewModelAssistedFactory to the multibinding map used by
the runtime Hilt module ViewModelFactoryModule.

Test: Manual via integration app
Change-Id: I1582f9d9d651a5a6a17bb457c60ee4fe301ec729