MaterialTheme updates / refactoring
Moves emphasisLevels outside of MaterialTheme, as it's not really related to the 'themable' parts of a Material application.
This should instead be queried from the ambient directly, using EmphasisAmbient.current

Additionally, changes `colors` and `typography` parameters on MaterialTheme to inherit from the parent by default, allowing overriding
a specific part of a MaterialTheme within a section of the hierarchy, such as a particular screen. This makes it consistent with `shapes`.

Bug: b/151727018
Test: n/a
Change-Id: Ib5e40c2d26737829a80ec16b390b125b570e5cbb
Relnote: "Removes MaterialTheme.emphasisLevels, instead use EmphasisAmbient.current to retrieve the emphasis levels"