Rename all "children" named composable params in ui-material.
We want to be more explicit about the intent. For example, by naming a param on FloatingActionButton as `icon` or `text` on Button if makes it clear what we expect to be passed, even if by the nature of being a slot you can pass in whatever you want. When the slot is indeed un-opinionated we call it content. This also matches with our `contentColor` concept.

Relnote: "children" trailing parameter was renamed for a lot of Material components, please use the new name if you used it via named params (or just use the trailing lambda syntax)
Fixes: 150695302
Test: material tests
Change-Id: Ia6d19fb822f29f13d774ca84777713e3b45b0cbd