LifecycleOwner integration
ComponentActivity is now mandatory. We just pass LifecycleOwner directly inside activity.setContent{} to not wait until the View will be attached. When ViewGroup.setContent{} is used we get the value using ViewTreeLifecycleOwner once the view is attached. If it is not available(we are used in just regular Activity or inside AppCompatActivity which currently has a bug and not populating the ViewTreeLifecycleOwner) we use our fake LifecycleRegistry for now(we move it to resumed when view is attached and to stopped when detached). This behavior is to be reconsidered once we fix the bug in AppCompatActivity.

Relnote: "New LifecycleOwnerAmbient is now available. Now an Activity you are using with Compose UI should extend androidx.activity.ComponentActivity (or AppCompatActivity). setContent on android.app.Activity is now deprecated"
Test: manually, plus new tests to ensure the ambient is populated
Fixes: 153141084
Bug: 151835993
Change-Id: Idb25a736332b17aebbfb96e919b932c2cc284d56