Add -ktx mirrored dependencies to preference-ktx
When other library groups (such as fragment) force
the same version code for all artifacts, it is
important for users that upgrade to a newer version
of preference-ktx not only get the updated fragment
dependency, but the upgraded fragment-ktx
dependency as well.

By adding the -ktx dependencies mirroring the library's
dependencies, we can ensure that both are updated
at the same time. This aligns androidx.preference with
other AndroidX libraries which do the same thing.

Test: ./gradlew bOS
Change-Id: I3c33a6866f7d215e58af61d29e8478837681f069