Support nullable types used with savedInstanceState()
We restrict usage of rememberSavedInstanceState to only non-nullable types which makes sense, as there is nothing to store. But savedInstanceState is different as it creates a holder object, and the holder object with a nullable inner value is a meaningful state we should save and restore. The regular state{} function works fine with the nullable types.

Change-Id: I6847f1a78afc14061082993583a6fcfc374e3277
Relnote: savedInstanceState() can now be used with nullable types
Bug: 153532190
Test: new tests