Fix bug where Image data not rewound
In ImageCapture the ImageProxy's buffer is used to retrieve the exif
orientation, however the buffer was not rewound. This change fixes it so
that buffer is rewound.

Also added test in ImageCaptureTest which accesses the buffer so that it
verifies the buffer has been rewound.

Relnote: "Fix bug where buffer in ImageProxy from ImageCapture was not
rewound"

Bug: b/153249512
Test: androidx.camera.camera2.ImageCaptureTest
Change-Id: I0c45524cdf5cc44e2b0441e82c4cc467d1e611e3