Support TransitionAnimation tracking in ui-tooling
Make sure that TransitionAnimations are parsed into ComposeAnimations and
notified to Android Studio when subscribing to the PreviewAnimationClock.
This way, the Animation Inspector will be able to track and display these
animations and their corresponding states.

From the Animation Inspector in Android Studio, users will be able to see
and play around with all the TransitionAnimation in the same @Composable
tree, setting the from and to states of each one individually and analyzing
how they animate simultaneously by manipulating a timeline that will control
the PreviewAnimationClock time. In addition to that, users will be able to
inspect the PropKeys being animated as well as their value at a given time.

The main changes made to PreviewAnimationClock to achieve that were:

* notifySubscribe and notifyUnsubscribe now receive a ComposeAnimation
  object instead of a AnimationClockObserver, making it much easier to
  handle in Android Studio.
* Parse TransitionAnimationClockObserver into ComposeAnimation in
  notifySubscribe to notify Android Studio when TransitionAnimations are
  subscribed.

In addition to that, the following new APIs had to be added to make it
possible to track and manipulate TransitionAnimations:

* updateSeekableAnimation: updates the cached SeekableAnimation associated
  with the given ComposeAnimation. The SeekableAnimation is used to get the
  animation properties in setClockTime.
* updateAnimationStates: updates the from and to states of the actual
  TransitionAnimations using the values from the cached SeekableAnimations,
  and resets the clock so all the animations go to the initial state.
* getMaxDuration: returns the time of the longest TransitionAnimation being
  tracked, taking into account the current from and to states.
* getAnimatedProperties: receives a ComposeAnimation and returns a list of
  label -> value pairs corresponding to an animated property (e.g. PropKey)
  label and its value at the current time.

Relnote: N/A
Test: PreviewAnimationClockTest
Bug: 160267340
Bug: 157896171
Change-Id: Ie8e05b5996fc541868a9fbf8aacdca361b667ea0