Move RectangleShape from ui-foundation to ui-graphic
RectangleShape is such a generic shape which is the only reasonable default for a lot of API so it makes sense to keep it near the Shape interface. This move will allow us to use this shape in ui-core and ui-framework.

Relnote: RectangleShape moved from androidx.ui.foundation.shape.* to androidx.ui.graphics.*
Bug: 154507984
Test: tests are passing
Change-Id: Ia74d5a3bbe2ee3a28bbddb57a2aef2607679d4ac