Add test for multiple fragment  startActivityForResult calls
It is possible to call startActivityForResult back to back. In a
Fragment this starts both activities, and once the user returns back to
the original activity, it disptaches onActivityResult in the order that
the activities were started.

This change adds a test in FragmentReceiveResultTest to ensure there are
no regressions.

Test: FragmentReceiveResultTest
Change-Id: I7c1e01970caacababc22db493aa9b75d0188ca58