Redesign SemanticsMatcher and nodes collection.
This removes the list of SemanticNodeInteraction elements as the return
type when working with multiple nodes. From now on we return
SemanticsNodeInteractionCollection where we have better control over
what is happening. This removes the forEeach as out of the box solution
since we find it too tricky to use.

As part of that I have implemented the suggestion to make selectors
lazy. It turned out it bringed lots of simplfiications in the process,
especially when transitioning between single & multiple items and
implementing lazy assertDoesNotExist().

When doing that I realized that I overloaded SemanticsMatcher with way
too much responsibility. Suddently matcher was not just matching but
was doing projections. Example is that in theory it was valid to type
findAll(isClickable().children()) which is not something we wanted.

I realized this is mixing two concepts. So I introduced
SemanticsSelector. The way it works is that find() and assert() work
with SemanticsMatcher that is a boolean function. However the
SemanticsInteracitonNode transforms it into SemanticsSelector.
Operations like children() return SemanticsSelector instead. The nice
benefit is that SemanticsMatcher is simple again and SemanticsSelector
can provide flow for error reporting in cases where selector expects
one element but received multiple.

Fixes: 154809726
Bug: 150293477
Test: Added
Change-Id: I32fc93c67a50fbd3b7b441c9543301243c120fa9