Small code reorganization in ScrollerLayout
There is no need to calculate the size inside the layout block as the size will not change and this happens on every scroll. Plus removed conditional logic for the cases when there is no placeables - it is impossible as we are manually adding Box and control the count of children.

Test: manually, scroller tests
Change-Id: I9594f7a3f411b928f651f2d3528af4bbe5786113