Determine direction based on FragmentTransactions
Rather than have DefaultSpecialEffectsController
try to determine whether the set of Operations
should be considered a pop or not, pass that information
through from FragmentManager which determines
direction by looking at the the FragmentTransactions
themselves.

By using only the direction of the last
FragmentTransaction, we can ensure that all operations
being executed are going in the same direction - this is
important when using non symmetric animations (such as
sliding to the left vs right).

Test: no regressions when running tests with USE_STATE_MANAGER
BUG: 147749580
Change-Id: I458ad8e2e7f728881b76fb2bcb25c4bdecf276d4