Add @ChecksSdkIntAtLeast annotation
This lets us and users annotate methods and fields that
represent an SDK_INT check. This lets lint stop hardcoding
knowledge about these methods inside of lint (such as
BuildCompat.isAtLeastR), and more importantly lets users
designate their own API utility methods (this is pretty
common in Kotlin). Lint has been using a lot of method
name logic to figure out if something is a version check
method (e.g. it understands a number of prefixes such as
isAtLeast and has, suffixes like OrLater, and variations
of codenames such as O and Oreo. However, this is not
perfect and frequently users file bugs that things aren't
working for their custom checks. In the IDE, lint will
resolve method calls and look inside version utility
checks, but this does not work from Gradle when the
utility calls are in separate libraries, since the method
resolve will point to a compiled method instead.

The annotation's javadocs shows some of the various
use cases for this.

Bug: 120255046
Test: Tested in lint along with the code to process these
  annotations
Change-Id: I89a5407693458796a6f67446ad107a82dbe47329