Get the IntentSenderRequest from the intent
Currently, we use the IntentSenderRequest that is passed into invoke()
directly to retrieve all of the variables need to properly make the call
to startIntentSenderForResult. Although this is a great fit for the
design of the ActivityResult APIs (i.e. creating and passing in the
input at launch time) and bypasses the need to place the
IntentSenderRequest into the intent, it makes the data flow slightly
more complex.

To attempt to clear this up, we will handle this similar to the
RequestPermissions contract and place the IntentSenderRequest in the
intent produced by the createIntent method. Then have invoke() retrieve
it from there instead of using the input directly.

Test: ./gradlew checkApi
Bug: 153433387
Change-Id: I7347713e1268283bfc9e34f90fdff887db8ca5ac