Remove CollectedOwners and restructure bitmap capturing
CollectedOwners was only used for two purposes: retrieving a list of all
SemanticsNodes and getting a reference to an Activity. The Activity
reference was only needed for displayMetrics and a window reference.

Both the displayMetrics and the window reference can be extracted on
demand from a SemanticsNode and doesn't require the CollectedOwners.

The list of all SemanticsNodes is now the only thing left, so
SynchronizedTreeCollector.collectOwners() can now completely bypass the
CollectedOwners and just return the list of SemanticsNodes immediately.

Incidentally, this removes two unnecessary calls to fetch all
SemanticsNodes when capturing a bitmap.

Bug: 155959983
Test: ./gradlew ui:ui-test:cC
Change-Id: I9b66ac72ecef331923e3dd10888ac3e312d8553f