Require non-null ViewTreeLifecycleOwner
We will now crash if the parent doesn't provide a correct owner.
Renamed setOnLifecycleOwnerAvailable to setOnInitialized to use the same callback for all owners, in the follow up cl ViewModel and SavedState related owners will also became available the this callback triggered.
Also added dependencies to the alpha versions of the libraries which correctly propagate ViewTreeLifecycleOwner for AppCompatActivity and Fragment.
Plus we will now provide a fake lifecycle in ComposeViewAdapter used by the preview in Studio.
Added ViewAmbient and removed deprecated OwnerAmbient. We need to have a reference to the view in Popups and Dialogs to be able to propagate owners from the main View to the new one we manually create.

Fixes: 158200117
Test: manually, new tests
Change-Id: Ic8f31f377c3d3d448596a130718efb436d42aa5d