Clear navHostFragments from container when destroyed
When a NavHostFragment's view is created, if that view has a parent, we
set a tag for the navController on the parent. This means if we have
nested NavHostFragments and attempt to return from the inner NavHost to
the outer NavHost, the tag is never cleared, which ends up creating a
leak.

If the view of a NavHost is being destroyed and the tag on its parent is
still set to its navController, we need to clear the tag.

Because there can only be one navController set on a view (instead of an
add/remove design), if developers are still using the fragment tag as
their container, this can create a state where there are no navControllers
set in the view hierarchy since the original NavHost will not have a parent.
So developers need to use FragmentContainerView as the container to ensure there is always a navController set in the hierarchy.

Test: MultiNavHostFragmentTest
Bug: 154441073
Change-Id: Ife3fb66fcc9ac924b4078001a7f153425199627a