Extract measure and layout logic from AndroidComposeView to MeasureAndLayoutDelegate
This allows to properly spilt responsibilities and keep AndroidComposeView simple. I am going to introduce further optimizations for the logic in MeasureAndLayoutDelegate but wanted to first split it.
I also change the behaviour so now we will requestLayout() only when the remeasure is scheduled for the root node and post the message to the Handler otherwise instead of causing unnecessary redrawing as it was before(we were calling invalidate(). If it happens that drawing will start before the handler execute its callback then it is also fine as we call measureAndLayout() from dispatchDraw as well.

Fixes: 156264832
Test: ui-core tests
Change-Id: I20de9254de31a6b06c0b4be883facf0ed98190df