Setting PreviewChannel#description as empty for null cases
Bug : 119800858

This is an improvement over the previous CL which added the null check inside the description setter.
https://android-review.googlesource.com/c/platform/frameworks/support/+/1304173

Use Cases : If a developer is using the same builder to construct different PreviewChannel Objects by overriding different fields, then setting a null description might result in setting the description of the previous builder object's data in the current scenario.

Test: Unit-test imitating the above use case
./gradlew buildOnServer : Successful

Change-Id: I82e0b1250015b95b405180e7fd95f91d6e1e83c4