Add Modifier.zoomable
This modifier aims to match the Modifier.scrollable API, but provide
standard pinch-to-zoom functionality.

Right now the two interactions this provides are:

1. Zooming via pinch
2. Programatic zooming (i.e. for interactions like zoom-on-double-tap)

Note that this does not currently provide a Maps-like "fling"
effect when zooming, and instead matches the Photos-like behavior of
stopping the scaling when the gesture stops.

Bug: 156763679
Test: `./gradlew test` (new test added)
Test: Added a ZoomableSample to the demo app
Relnote: "Added Modifier.zoomable for pinch-to-zoom functionality"
Change-Id: Id5d63ff7d29f9dedd6d9d28df2ce74081a892cda