Fixed Material Checkbox
During a rebase the change to
DrawScope.drawRoundRect lost the default
parameter behavior of radiusY matching
that of radiusX if it was not provided.
As a result material checkbox was being
drawn as a square as the previous default
value of radiusY was 0.0f. Updated Checkbox
to explicitly pass a radius parameter.

Also reverted logic of CheckboxScreenshotTest
to have the original threshold specified.

Fixes: 157244624
Test: re-ran compose tests

Change-Id: I0957f1ba78db955895a1aed9b56b21505bc59f34