Relayout optimization
Previously when we needed to remeasure the node we were finding the first parent of it which was measuring the child in the layout block, not the measure block. Instead now we will do it smarter: we will only remeasure the target node and then if as a result the measured size(or alignments) has been changed we will also remeasure the parent. This allows us to significantly decrease the amount of nodes we need to remeasure/relayout with every change.
The real performance gain highly depends on the layout hierarchy and the amount of work we do in every measure/layout block, this CL just allows us to re-execute them less often. For example the performance of RadioGroupBenchmark.toggleRadio_layout / RadioGroupBenchmark.toggleRadio_measure was improved by almost 40 percent.

Bug: 155078844
Test: a lot of new tests, ui-core tests are passing
Change-Id: If992dd4ff8b4ba19e5cfaff741f5ecfb4a963880