Fix issue with mixing get and require APIs
When mixing the get_() and require_() fragment APIs with a kotlin non-null
assertion operator, if the operator comes after the require_() the lint
rule gets confused and produces a fix that is exactly the same.

This change does a piecewise like conversion that places the non-null
operator with the get_() call, which will then allow the rule to work
appropriately.

Test: Added test to UseRequireInsteadOfGetTest
Bug: 157677616
Change-Id: Idbae59654471067269426e787938e885af6880e2