Fix crash during the AdapterList measuring
Inside the measure block we are calling consumePendingScroll which calls composeAndMeasureNextItem which was running this code:
Constraints(maxWidth = rootNode.width, maxHeight = IntPx.Infinity)
It is incorrect to call rootNode.width during the measuring as this rootNode is currently being measuring and has no measured width yet. Instead of doing it we need to take the width from the parent constraints.

Fixes: 154653504
Test: manually
Change-Id: Icedc2d6af3b82d530a606584321bff283ef7f157