Add Lifecycle.withStateAtLeast APIs
Add suspending APIs to await a lifecycle state and run a non-suspending
block of code synchronously at the point of state change, then resume
with the result. These APIs differ from the existing when* methods as
they do not permit running suspending code and do not employ a custom
dispatcher.

Test: WithLifecycleStateTest
Change-Id: I67b7dc9c4cd824ccb410091999ed67dc671c3e5d