Rename AdapterList to LazyColumnItems
It was also moved into a `lazy` sub-package and split into two files. Plus I renamed params:
1) data -> items. this seems to be more meaningful then just raw `data`
2) itemCallback -> itemContent. this is more meaningful and we generally don't use word callback in the lambda names, especially for composable lambdas
Sidenote: I named the file with a deprecated function `Temporary`, as if I name it `AdapterList` git will not understand that most of the content has been moved to the new file `LazyColumnItems` and we will lose more git history for lines.

Fixes: 158833694
Relnode: AdapterList was renamed to LazyColumnItems
Test: manually
Change-Id: I847c174939f28eab71aad934d35945b584e01edf