Notify animation (un)subscribtions to PreviewAnimationClock
When an AnimationClockObserver subscribes or unsubscribes to a
PreviewAnimationClock, we should notify Android Studio, to the tools can
react accordingly. We do that by calling notifySubscribe from subscribe and
notifyUnsubscribe from unsubscribe. Android Studio will use bytecode
manipulation to intercept calls to these methods and redirect them to the
Animation inspector, which will handle them properly.

Test: PreviewAnimationClockTest
Bug: 158752769
Change-Id: Ie5622f6ed04adff764239343cf327f2e4553c7ff