Fix touch focus fail on Samsung S7
when starting a focusAndMetering action , if there is a active
action,  it will try to cancel it first.  and that will trigger
a CONTROL_AF_TRIGGER_CANCEL which causes the issue because it will
cancel the upcoming AF_TRIGGER as well.

The solution is not to call CONTROL_AF_TRIGGER_CANCEL when we are
about to start another AF_TRIGGER action.

Relnote: "Fixed the issue where startFocusAndMetering fails to
focus successfully on Samsung Galaxy S7."
Bug: 159039233
Test: CameraControlDeviceTest, FocusMeteringControlTest

Change-Id: If3be7c3359687a47a61fbd0f9c7f018d5d8fd323