Suppress various Room test app warnings.
* Missing index in junction warning.
This warning can be suppressed but turns out we never added an
entry for it in the public API of suppressible warnings. We don't
fix them because we want to test the path where Room does the right
thing when the index is missing.

* Transaction warnings
These are easy fix, just add @Transaction to the DAO method. We
already had coverage on those without it.

* Mismatched getters and setters
These warning have fixes but instead we suppress them because the
semi-incorrect code is left on purpose for test coverage.

Bug: 158839766
Test: No warning on ./gradlew room:integration-tests:room-testapp:compileDebugAndroidTestSources
Change-Id: I544eae88257eb08a5dbfa079a45e417ea10d80e9