Throw error for synchronization on main thread
When synchronization is attempted on the main thread, a deadlock is
caused. Rather than letting it happen and timing out, detect the issue
and throw an error.

Fixes: 159224666
Test: ./gradlew cC
Relnote: "An IllegalStateException is now thrown if you attempt to
trigger a synchronization on the main thread, e.g. when you do:
```
runOnIdleCompose {
    findByTag(\"tag\").assertExists()
}
```
Remove the call to runOnIdleCompose in such occurrences:
```
findByTag(\"tag\").assertExists()
```
"

Change-Id: Idced7d74747addff0653a645f2ccc059156430f2