Add Request objects for CameraPipe.
This CL adds the primary Request object that will be submitted to a
CameraGraph instance. Requests are designed so that they do not need
to be constructed from a CameraDevice but contain the information that
will be needed to build a CaptureRequest when the CaptureSession is
ready. This allows them to be queued, reused, and/or resubmitted (when
used as a repeating request).

Test: Added unit tests for Request extension methods.
Test: Updated fakes for RequestMetadata and ResponseMetadata
Test: All existing tests continue to pass.

Change-Id: I87d98e1dc5dbd47010b8d836b2fb258493b0bea7