Fix for performance regression in CheckboxesInRowsBenchmark.first_compose
aosp/1316677 introduced some performance regression in *first_compose benchmarks as we started to do a bit more work in requestRemeasure() method. This change improved the performance of this code.
First we will not call methods on Handler when it is not necessary as these a native calls and can take some time, plus we will not schedule the callback when the view layout is requested as it will trigger remeasure.
Second we will not add nodes in relayoutNodes treemap when we can skip it as adding into this data structure is not fast. We can skip adding the node into the list if the parent is already added there as measuring the parent will include measuring the child.

Bug: 158461153
Test: manually
Change-Id: Ia45ae73d4e42d94ee241d6c102f30a72cab3814b