Add CameraDevice and CaptureSession wrappers
These wrapper interfaces and classes allow interactions with the camera2
CameraDevice and CameraCaptureSession to be tested via fake versions of
these types. The implementations (Prefixed with Android*) pass calls to
the actual class and do a small ammount of exception handling and API
compatibility checking.

These interfaces and classes are an implementation detail of CameraPipe
and are not intended to be used directly.

Test: All existing tests continue to pass
Test: Lint (and API compatibility checks) continue to pass.

Change-Id: Ic6d43890b9b58a152db5797906e5e29dcb57b405