Prevent focus and metering from being triggered when tap outside preview
When the preview in PreviewView has padding either to the sides, or at the top and bottom (for example when the scale type is FIT_*), tapping on the padding close enough to the preview may trigger focus and metering. This is because converting the preview tap position (previewX, previewY) may be a valid point in the camera sensor surface. This depends on the aspect ratios of the camera sensor and the preview.

This issue is reproducible on Pixel 3 where the camera sensor has an aspect ratio of 4:3, and setting the preview aspect ratio to 16:9, the preview scale type to FIT_CENTER  and making the PreviewView wide enough so that there are black paddings to the right and left.

Bug: 158780652
Test: FocusMeteringControlTest.java
Change-Id: I69b7f1f1d54ab8f366cdc67d6e316ee7bc7ba35b