Final migration of material size tests.
Also removes the old sizes testing infra.

It turned out that having a utility that sets a content and returns it
for measuring in a safe way (by using parent) is quite less error prone
and requires less code so I have kept that patten but hooked it to the
new APIs.

I couldn't migrate for instance text position tests as that would
require access to unmerged semantics. But I still at least fixed the
positions used, so it won't cause issues on cutouts anymore.

Bug: b/157657482
Test: N/A
Relnote: Removed obsolete size testing APIs.
Change-Id: Iba0a086e8c88cf44684cba56766792614201ba30