Fix interop redrawing bug in specific scenario
The View would stop redrawing because the dirtyLayers list was getting
out of sync with the isDirty flag on RenderNodeLayers. The bug was only
happening on >= API 29. This could happen when, as part of the
updateDisplayList call, the AndroidViewsHandler was receiving
onDescendantInvalidated corresponding to the View, which was calling
invalidate on the corresponding LayoutNode. This was setting the isDirty
flag to true, but the LayoutNode was removed from the dirtyLayers list
at the end of updateDisplayList. Therefore, at subsequent invalidations
requested, the LayoutNode was not added back because it was already
marked isDirty, preventing further redrawing. This CL delays setting
isDirty to false until redrawing is finished, and this is consistent
with when the flag is set for ViewLayers.

Bug: 160219011
Test: AndroidViewCompatTest
Change-Id: I709976c24a36794592e2543263efbd41c80ecaf6