Make getTypeElement calls null-safe
We use elementUtils.getTypeElement in many places and sometimes it can be null,
sometimes it shouldn't.

This CL moves all of those calls into extension functions on the
processing environment where the expected nullability is explicit
in the method call.

There was also many usages of getTypeElement().asType() and
getTypeElement(typeName.toString) so I've created find/requreTypeMirror
and overloads that receive TypeName.

Bug: 160323720
Test: existing test suite
Change-Id: I3f31a38c86abd1b43a8a93d9b457c61bac802167