Isolate access to ElementFilters
This CL attempts to isolate access to ElementFilters into extension
methods.
As part of it, I've also removed all access to ElementUtils except for
the extension files.

Unfortunately, the difference between declared methods / methods
including supers / instance methods is not super clear in our codebase.
I tried a best effort to group them in 3 sections but I'm not 100% sure
if this is correct.

Added an additional test for element extensions (probably needs to be
further extended).

Bug: 160323720
Test: existing test suite, ElementExtTest
Change-Id: I962758128d3123b4c120fe48630ad54202047833