Add Executable element support
This CL adds executable element support to xprocessing.
Unlike the current implementation in room, most of the kotlin
handling is abstracted in XProcessing. We still leak it
via the API to check things for suspend functions or being
able to see suspend parameter names in class vs in kotlin metadata
but they are fairly limited.

Bug: 160322705
Bug: 160323720
Test: compiler-xprocessing tests
Change-Id: I375d7c5e5f1616000d231900a4c89796a6e09332