MediaRouter: Add a Callback#onRouteSelected
When dynamic grouping is used (or MR2 is used), the selected route
may be different from the requested route.

To support apps that requires originally requested route,
a new variant of onRouteSelected is added.

The default implementation is calling onRouteSelected with the selected
route.

Before this CL, when route A is selected and a dynamic group A' is created,
 onSelect(A) -(after a while)-> onUnselect(A) -> onSelect(A') are called
After this CL,
 (after a while) onRouteSelected(A, A') is called

Relnote: "Add a new `MediaRouter.Callback#onRouteSelected`
to get notified when the selected route are different from the requested
route"

Bug: 159733199
Test: Manually with support v7 demos enabling/disabling
MR2, confirmed onRouteSelected is called only once.

Change-Id: Ieee163d3487d28be5088311316b881885c5d6f08