Use reflection name for Worker and ViewModel multibinding keys.
Worker factories search for an assisted factory based on the
reflection name of the class and not its canonical name,
e.g. TopClass$NestedWorker vs TopClass.NestedWorker. This
changes fixes the generated multibinding key (@StringKey) so that
the reflection name is used and nested Worker classes get
instantiated by Hilt. Furthermore to keep the keys consistent,
this change updates HiltViewModelFactory to also use the
reflection name.

Bug: 160524718
Test: Hilt Worker and ViewModel test apps.
Change-Id: I3f9255c2e8601fbf7cb2c0160c4e6aae2fe6c58f