CompositionLifecycleObserver docs and defaults
Expand the documentation for CompositionLifecycleObserver and give the
onEnter and onLeave methods empty default implementations. Document the
potential side effects of exposing CompositionLifecycleObserver
references as API.

Change the return value of rememberCoroutineScope and
compositionReference to not directly implement
CompositionLifecycleObserver to avoid the documented side effects.

Test: existing tests
Relnote: "Added default method implementations for
CompositionLifecycleObserver"

Change-Id: I2428951907b8256d698fc11291dc7d8c3b756d4c