Fix lifecycle-livedata-core-ktx-lint ObsoleteLintCustomCheck
Using the latest version of lint-core (27.1.0) as a dependency for the
livedata-core lint rule causes an ObsoleteLintCustomCheck for studio 4.0
and 4.1 (along with the subsequent gradle versions).

Instead of using the latest, using the minimum version seems to work
appropriately.

Added a lintMinVersion variable to the dependcies file and used that to
make updating easier.

Test: tested in studio 4.0 and 4.1-Canary 8
Bug: 158699265
Change-Id: If1636b3e08a368cf25fb268f42754c4c6ff2f463