New Focus Modifier Implementation
This is the first CL in a series of CLs which implement the new focus API. The new implementation consists of multiple modifiers each implementing a specific use case.
This CL adds FocusModifier2 and ModifiedFocusNode2 (Which will be renamed to FocusModifier and ModifiedFocusNode after the existing classes are deleted).

Bug: 160924778
Relnote: Added Modifier.focus which replaces FocusModifier.
Test: Built and launched the demo app
Change-Id: Ib852a056a0f3c76757f0fdef07e75e82bf178b8d