Mark some of the AppCompat code-paths to be R8-friendly
The two methods - setSupportBackgroundTintMode and
setSupportButtonTintList are already marked as @RestrictTo
and should only be called by the matching ViewCompat APIs.

The implementation in ViewCompat only calls the "fallback" /
backport methods on pre-21 versions of the platform. Adding
explicit build checks in all the AppCompat classes that
implement the TintableBackgroundView interface allows R8 to
strip out the passthrough to AppCompatBackgroundHelper when
min SDK level is at least 21 and make the resulting minified
binary smaller.

Change-Id: I611542e05414ea9670b586ae67f823de05a97bc3
Test: ./gradlew appcompat:appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.package=androidx.appcompat.widget
Bug: 126209808
Bug: 126204182