Start tracking androidx.hilt APIs.
Add generates API files for androidx.hilt. To do this
make the constructor of the worker and viewmodel factories
restricted since they take a parameter that reference the
assisted factory which themselves are restricted. These
factories are not meant to be constructed by the user but
instead are available to be injected by Hilt so are only
needed to be referenced via their type. A more proper API
for building them will be available in a follow-up CL
to address https://github.com/google/dagger/issues/1938.

Bug: 161380664
Test: n/a
Relnote: n/a
Change-Id: Ic757283bbce4b82ddfcedf494a8c4373b2bc12a0