Ensure ActivityResultLauncher can be launched in onCreate
When attempting to launch an ActivityResultLauncher in onCreate() of a
Fragment, it fails since the the Fragment's Lifecycle does not move to
CREATED until after the onCreate() callback in executed. We should
register the callback as soon as the fragment is attached since that is
the point at which we have all of the needed information.

This change also adds tests to verify that ComponentActivity does not
suffer from the same issue.

Test: Added Activity and Fragment tests
Bug: 161464278
RelNote: "You can now call launch() on an ActivityResultLauncher in the
onCreate() lifecycle method of a fragment."

Change-Id: Ifd98f308fdfbd7c9ce3ae93d032b140d6f3ec35e