Fix flaky postsubmit imageCapture rotation test
When an activity has a locked orientation, OrientationEventListener is used to update the image capture's target rotation, which ensures the captured images have the correct rotation and orientation. OrientationEventListener is called whenever the orientation sensor detects the slightest change in the device's orientation, which causes flakiness in the test, since a new orientation can be set between the moment the rotation is done and when the test assertion is run (which verifies the the captured image is rotated correctly, and may use the target rotation of the image capture use case). To work around this, the orientationEventListener is disabled before the rotation is performed, and is kept that way.

The same fix is applied to imageAnalysis locked rotation test.

Bug: 161689522
Test: ImageCaptureLockedOrientationTest.kt
Change-Id: Ie61fa3b8722471fd299065d01f6087e2e3241c4d