Merge multiple operations on the same fragment
When a SpecialEffectsController already has a
pending Operation for a given Fragment, it now
intelligently merges the new Operation into the
previous one rather than creating an entirely new
Operation.

This ensures that at any time, there's at most one
pending Operation for a fragment that captures two
important pieces of information:
- What the final state of the Fragment should be
after the Operation. This can one of the normal view
visibilities (VISIBLE, GONE, or INVISIBLE) or REMOVED
for the cases where the view should be removed entirely
from the container
- What impact that Operation has on the lifecycle
of the fragment which is used by FragmentStateManager's
computeExpectedState()

By moving to this model rather than explicit types,
we ensure that pairs of actions (such as a postponed
add + remove due to popping the back stack) do not result
in any special effects.

Test: many postponed tests that pop postponed ops now pass
BUG: 147749580
Change-Id: I7a0bdef0e0ba25d0eeca4b3c12222b6ccab776ba