Set CameraX.shutdown timeout for tests
The CameraX.shutdown() might stuck if the camera cannot be closed
successfully, and the test result TIMEOUT is not clear to analysis
the fail reason for the test. It would be more clear if the "camera
cannot be closed" issue can be showed in result FAIL.

Modify the CameraX.shutdown().get() to
CameraX.shutdown().get(10000, TimeUnit.MILLISECONDS),
and the test result will show in FAIL when the shutdown timeout.

Bug: 162048564
Test: ./gradlew bOS &&
./gradlew camera:camera-core:connectedAndroidTest &&
./gradlew camera:camera-camera2:connectedAndroidTest &&
./gradlew camera:camera-extensions:connectedAndroidTest &&
./gradlew camera:camera-core:test &&
./gradlew camera:camera-camera2:test &&
./gradlew camera:camera-extensions:test &&
./gradlew camera:camera-extensions:connectedAndroidTest &&
./gradlew camera:integration-tests:camera-testapp-extensions:connectedAndroidTest

Change-Id: Iaedc14e8edcd9321495f91d2a2d77c8bee351583