Cache type elements per round
This CL adds an XTypeElement caching layer to JavacProcessingEnv so that
we can avoid wrapping type elements repeatedly.

Instead of using a proper cache like Guava's cache, I opted for a simple
hashmap of weak references. This is to avoid keeping items in memory
longer than necessary, which would happen with an LRU cache.
The downside is that if we re-visit element after losing references, we'll
potentially re-wrap the element. Given that these elements might have long
children trees, i still think it is a better trade-off.

Bug: 160323720
Test: XprocessingStepTest
Change-Id: Iebb0a564e04a3a7b46460fa79b49ef0d6c5e409d