Refactor: remove the SurfaceProvider wrapper in Preview
Minor refactor that removes the SurfaceProvider wrapper in Preview. The wrapper was put in place so that the pipeline can be created without a SurfaceProvider. The same can be achieved by saving the SurfaceRequest and send it when SurfaceProvider is later provided.

The CL should also made the change so that the pipeline is not recreated for setting the SurfaceProvider for the first time.

Bug: 159659392
Test: manual test and ./gradlew bOS
Change-Id: Icf04a2d1e40e55fd7c0b3bc9ed51a0ac472cc549