Change DisplayOrientedMeteringPointFactory to take CameraInfo
The DisplayOrientedMeteringPointFactory takes in a CameraInfo so that it
does not need to use the CameraX singleton to query for the the
CameraInfo using the CameraSelector. This makes it so the factory does
not need to rely upon some global state and makes it much easier to
reason about the relationship between which Camera the factory is
generating points for so that ambiguity will be removed.

PreviewViewMeteringPointFactory methods that take in a CameraInfo
instead a CameraSelector instead.

The PreviewView will no longer require taking a CameraSelector for
retrieving the MeteringPointFactory, but will instead retrieve it based
on the currently attached camera.

Removed CameraX singleton dependency in
DisplayOrientedMeterinPointFactoryTest and
PreviewViewMeteringPointFactoryTest

Relnote: "DisplayOrientedMeteringPointFactory takes in a CameraInfo
instance instead of a CameraSelector so there is a direct mapping to
which Camera the factory will be generating points for. All classes
which use DisplayOrientedMeteringPointFactory also take in a CameraInfo
instance instead of CameraSelector."

Bug: b/162265893
Test: test class of modified files
Change-Id: I400c16899037e29fd37d6cf6faaa87ed70188de1