Ensure that cancelled operations have no lifecycle impact
When using forceCompleteAllOperations(), operations
are cancelled prior to calling complete(). The complete()
call that calls through to FragmentStateManager's
moveToExpectedState() synchronously calls through to
getAwaitingCompletionLifecycleImpact() to determine the
final state of the fragment. Since the operation is
already cancelled, it should not be considered in
terms of lifecycle impact.

This fixes an issue where changes in the FragmentManager's
state (specifically the destruction of the view due to
a parent fragment going on the back stack) would not be
properly reflected in the fragment's state, effectively
never moving below AWAITING_EXIT_EFFECTS.

Test: updated SpecialEffectsControllerTest tests pass
BUG: 147749580
Change-Id: I5a13e8674f9537aaef8c773b49749feb89492eb3