PlacementScope.place stops automatically mirror the position
PlacementScope.place() is renamed to PlacementScope.placeRelative().
PlacementScope.placeAbsolute is renamed to PlacementScope.place().

Relnote: "PlacementScope.placeAbsolute() was renamed to PlacementScope.place(), and the previous PlacementScope.place() was renamed to PlacementScope.placeRelative(). As a result, the PlacementScope.place() method will not automatically mirror the position in right-to-left contexts anymore. If this is desired, use PlacementScope.placeRelative() instead."
Test: foundation, foundation-layout, ui, animation tests passed
Fixes: 162916675
Change-Id: I873ac827e6c4d4bf6c85a80b7128174c61602945