Fix NonNullableMutableLiveDataDetector casting issue
Currently if you use an apply block to assign the value when
initializing a MutableLiveData object while within a class, on any
subsequent assignments to the value, the lint check will attempt to cast
a method to a variable and fail.

This fix ensures that if the given caller is not a variable that the
lint check does not crash, but instead just ignores it.

This does not handle the fact that the lint check does not work within
the apply and that may be fixed in a follow-up CL, but users should be
using the constructor for MutableLiveData that takes a value anyway.

Also adds a test for creating a MutableLiveData in a companion object.

Test: Added MutableLiveDataTest
Bug: 162829610
Change-Id: I6243293db30914d0e701ee9f4e3f38b88f493dbc