Add KotlinType information to JavacType classes
This CL adds kotlin metadata type information into JavacType classes
to better handle kotlin nullability information.

Previously, this information was carried over via the nullability
field in JavacType but it is not enough in cases where nullability
cannot be discovered from annotations and rather needs to be read from
KotlinMetadata.

For instance, for a value defined as follows:

`val foo: Map<String, Int?>?`

We can discover the nullability of `foo` from kotlin stub generated
nullability annotations but we cannot for the type parameters of
Map. This surfaces itself as a bug in cases like Query methods
that return `Flow<T?>`, where we don't know if T is nullable or
not as it cannot be annotated and we need the information in Kotlin
Metadata (see b/146755796).

Unfortunately, since kotlin metadata is not always available, the
nullability field is still present in JavacType. I've tried removing
it with a synthetic KmType generation but it becomes very misleading
since Kotlin's nullability is true|false vs Java's nullability is
true|false|unknown. To keep it somewhat manageable, all subclasses
of JavacType has 2 constructors that either receives the nullability
information or kotlinType, but never both.

Bug: 146755796
Test: XNullabilityTest
Change-Id: I0eb2c75090b745e54f3f3bef9f05e49343afb208