Introduce timestamp policy into jetifier.
Previously we have introduced now() as modified time to remedy issues
with some implementations not being able to handle missing time record
on zip entries. However that introduced issues for build tools that rely
on determinism of the output. This change allows developers to change
the strategy via provided policies - now, epoch, keepPrevious. And it
also changes jetifier to use keepPrevious policy as the default. Keep
previous policy means that we now store the original modified time for
each archive and set it again when creating a new archive. This should
provide determinism but also avoid regressing previous cases that were
prone to missing the modified time. There is also a last remedy which
is the epoch that forces time=0 for all the zip entries.

Bug: 159360145
Test: Added tests for times and determinism and manually verified the
commands line interface.
Relnote: "Deterministic timestamps in jetifier. Including configurable
strategies - now, epoch, keepPrevious (default)".

Change-Id: Ib2dfdea882799ad0391390b2480ef1e02e62571d