Wear/Watchface Style: Introduce boolean & ranged style categories
In addition to supporting the new style categories, we allow a user
specified default ID. Note the in watch face UI for the the new
category types is just a placeholder. The UI will move into the system
(watches can still have their own UIs) so this seems OK.

Test: The sample watchface has been modified to test this, with options
to turn on & off the hour pips and to scale the watch hands.

Change-Id: I40aa2fd04570565ee3bbcb45ac6b2392bb449981