Create a 'backend' interface and refactor the current code to use it.
A 'backend' is a system that can store documents and return query
results. Currently LocalBackned is the only implementation. In the
future, we expect at least PlatformBackend to exist, and possibly
others.

This code will shortly be placed into its own gradle project so that
users can use AppSearch in the platform without incurring the cost of
protos and libicing.so.

Bug: 162450769
Test: AppSearchManagerTest
Change-Id: Ia44c00326c7049b2e8d88a4faa48b7662cd326c2