BackdropScaffold material component
This introduces an experimental BackdropScaffold component, implementing
the basic functionality of a backdrop and combining it with an app bar
and snackbars, similar to the existing Scaffold. A backdrop is composed
of two surfaces: a back layer and a front layer. The back layer displays
actions and context, which control and inform the front layer's content.

Bug: 113857957
Bug: 153636126
Test: Added new tests in BackdropScaffoldTest
Relnote: "Added an experimental BackdropScaffold component."
Change-Id: Iad908db4a93687c6dee6cc163721febfe4c814fc