Support RenderNode on M-O devices.
Bug: 162061312

To avoid reflection, a new stub library was created to
support RenderNode on older devices. This should also improve
allocation performance as RenderNodes should be much smaller
and easier to allocate than Views.

This does not currently work on P devices because ART
does not allow access to APIs in the unsupported list on P.

TextInColumnBenchmark#first_draw is around 30% faster in
my local tests.

This CL will make benchmark performance slower on M-O devices,
but it is more representative of actual performance. It turns
out that PictureRecordingCanvas has faster access than
DisplayListCanvas and was giving artificially fast results
on devices that don't support hardware accelerated canvas.

Test: demo on API 23, 26, 29
Test: TextInColumnBenchmark#first_draw
Test: New test for fallback for M RenderNode to View

Change-Id: Ia10caea550f328368c4f0f4384267058d1dbc541