Add LazyListState
Relnote: """Add a new LazyListState class.
This allows for observation and control of the scroll position of
LazyRow and LazyColumn components. Instances can be created using
rememberLazyListState() and passed into the state parameter of the
component. Currently, the first visible item and offsets can be
observed in this initial version."""

Fixes: 159307669
Test: Added new tests
Change-Id: Ic7cb72444d41ccb6add635ba0873bb0e9222cf15