Allow new state manager to reorder ops within a transaction
When using the new state manager without reordering, the replace case
will cause both the exiting and entering fragment views to be created
and added to the container at the same time. Allowing this would be
inconsistent with the behavior of the old state manager.

When the new state manager is active, we should allow the reordering of
ops within a single transaction. This means we can be consistent with
the old state manager and provide improved behavior for the fragment
lifecycle.

RelNote: "The new state manager now consistently removes the exiting
fragment view before adding the entering one."
Test: Fixed workaround in PrimaryNavFragmentTest

Change-Id: I41a6ee1a0bc135c0443d8152a8106498241d3862