Save view focus state before executeOperations
Because the new state manager changes the view visibility when handling
animations, we save the focus state in onViewCreated and we restore it
when the fragment is RESUMED.

There are cases when the view focus changes even later than onViewCreated
(i.e. setting it in onStart()). This means we need to handle uses of the
<requestFocus> tag, as well as late calls of the request focus method.

We need to continue to check the view focus after onViewCreated, because
that is where we need to make the view alpha 0 and the <requestFocus>
tag does not work if the view alpha is 0. we should also save it before
executeOperations() since this is the last possible moment before we start any
special effects. This means we will have the latest state when we restore
it once the Fragment is RESUMED.

Relnote: "Fragments now properly restore their view focus state from
right before they start their animation."
Test: all exisiting test pass
Bug: 167510681

Change-Id: Icc2561fb3dc413250153e5e0fcb2d10846791d53