Add an abstraction on top of ComplicationData.
ComplicationData is a legacy API that we cannot change since it is
part of the protocol between watch faces, complications, and the Wear
OS system. However, it is possible to create a developer abstraction
that hides that implementation.

This commit attempts to demonstrate what this may look like and how we
would be using it.

Test: ./gradlew \
          :wear:wear-complication-data:lint \
          :wear:wear-complication-data:ktlint \
          :wear:wear-complication-data:assembleDebug \
          :wear:wear-complication-data:test
Change-Id: Ic012975828992c8a7316b7291556a2f2c77b74e4