Add ResourcesCompat#getFont with cacheOnly option
There are two APIs in ResourcesCompat for gettting font, sync and async.
Not to block UI thread, async font fetch is recommended but becasue the
callback is called one loop behind from the first draw even if the font
is already fetched and stored in the cache, the flicker happens.
To prevent this flickering, add new API getFont with cacheOnly which
sees the cache entry and returns immediately.

Relnote: Add new API ResourcesCompat#getFont with cacheOnly option

Bug: 145207014
Test: ./gradlew core:core:connectedAndroidTest
Change-Id: Ic38cf93a5c3155eb2bacfbbc093365c31720481a