Clean-up KDoc: replace {@link} with [].
KDoc uses a different way to refer to other classes and members than
Javadoc does and the {@link} currently in the documentation would not
be linked correct and are not checked by the compiler.

This commit removes all the {@link} in Kotlin files for Wear
libraries. It does not however fix all the reference that are either
out of date or using the wrong syntax.

Test: ./gradlew wear:wear:assemble wear:wear-complications-data:assemble wear:wear-complications-rendering:assemble wear:wear-input:assemble wear:wear-watchface:assemble wear:wear-watchface-data:assemble wear:wear-watchface-style:assemble
Bug: 169639460
Change-Id: I24a8b9171ecec986915c5bdb7137bd9ab7a1091d