Fix a few bugs with WatchData<>
We now clear toBeRemoved.

We no longer throw an exception if an observer adds a new observer.

We also rename the observe() method to addObserver() and remove the
dependency on lifecycle's Obseverer in favor of our own fork. NB
this has to be a java interface or we can't use the lambda style.

Test: Presubmit
Change-Id: I4123b15e911285fa1be37acfd1d93a8c78e7c07a