Implement XArrayType in KSP
This CL implements XArrayType in KSP.

In KSP, arrays are just generics so we detect when a type
is wrapped into KspType, similar to what we do in java
processor.

For TypeName, I opted in to java style array rather than
a Parameterized type. What we will do for TypeNames is
up for debate (e.g. at what level we'll swap them with java
representations). It is inconsistent with the rest of KSP
but for arrays, conversion is always so made sense to me
to do it here.

Bug: 160322705
Test: XArrayTypeTest
Change-Id: I2d687d5d5a77d460f66914f7167375c05036fe0d