Re-call onApplyWindowInsets after the update of visibleInsets
On API < 30, the visibleInsets, used to built WindowInsetsCompat, are updated
after the insets dispatch so we don't have the updated visible insets at
the first time onApplyWinwoInsets is called.

As a workaround, we rebuilt the WindowInsetsCompat again in
the next frame so we know that we'll have the right value.

Bug: 167429196
Test: WindowInsetsCompatActivityTest#ime_insets_cleared_on_back
Test: WindowInsetsCompatTest#test_not_equals_root_visible_insets
Test: WindowInsetsCompatTest#test_equals
Change-Id: Icede56e124e14f7d15797c72f570b8d7b626ad9a