Introduce XFiler
The initial abstraction for XProcessing ignored the Filer
API since it is an interface. This CL changes it to have
an abstraction in XProcessing instead to avoid re-implementing
full Filer API in KSP.

Also added a test for code generation. Unfortunately, KSP
does not yet support access to generated code from java sources.
https://github.com/google/ksp/issues/119

Fortunately, it won't affect Room as we don't require access
to generated code.

Bug: 160322705
Test: XProcessingEnvTest.generateCode
Change-Id: I953a17cd857b7360d12269ccda1fe3c6b0489906