Set scrollable for accessibility
AccessibilityNodeInfo scrollable is set to true if we have scroll
semantics actions and states. This is not needed for google3 head
version of talkback, but talkback 9.0 still needs it for auto scroll. So
I think it is better to set it.

Test: tested manually on talkback 9.0. Since google3 head version
talkback doesn't need it, no unit test is added.

Change-Id: Iea099cc90e6110d21548e3dc929ff2057060971e