Replace AppSearchManager with a SearchSession.
SearchSession becomes the way to interact with AppSearch APIs.

In this CL, AppSearchManager is kept to avoid changing the public API,
but it will be removed in the next CL with API approval.

In this CL, we add a SearchResultsHack interface as a shim between the
SearchResults public API and the new implementation. The existing
SearchResults class will be deleted and this interface promoted to
SearchResults in the next CL with API approval.

Bug: 169377054
Test: AppSearchManagerTest
Relnote: N/A
Change-Id: I89e5b577d8628e8bdf5c67e2d537280be48b82d1