Add queries intent in AndroidManifest for vendor library to interact with customized service in Android R
If the device is running Android API level 30 (R) and an app is built with targetSdkVersion 30, an app needs to use <queries> element in AndroidManifest so that it can interact with other apps. For extensions' vendor library implementation, OEMs may need to bind their customized service to enable the effect mode. Adding this generic queries intent in camera-extensions' AndroidManifest, then, it will be included in the 3P app. So that, the extensions vendor library code executed by the 3P app can interact with the customized service on the device and the the effect can be enabled successfully.

Bug: 167704226
Test: manual test
Change-Id: Ic3df1f199fcc9b477ef5e09e3d60624dc0e0e12a