Remove the image.close in default implementation of OnImageCapturedCallback and improve javadoc
1. If developers add the override method in Android Studio, it will
contain the super.onCaptureSuccess which if it is not deleted, it
will result in image to be closed. So we better remove the default
implementation.

2. Improve javadoc of OnImageCapturedCallback to clarify how to get
the rotationDegrees.

Relnote:"OnImageCapturedCallback#onCaptureSuccess base class
implementation no longer closes the image. This is for preventing
unexpected behavior to developers. Developers should not rely on
super.onCaptureSuccess to close the image."
Bug: NA
Test: ImageCaptureTest

Change-Id: Ifbf9c98da68fcbba0c177a1a0579d2d65c120346