Introduce KspPrimitive Type
This CL introduces a new KspPrimitiveType : KspType which is used
to represent java primitives in KSP as primitives do not exist in
Kotlin.

Unfortunately, we cannot decide whether a kotlin type will be a
primitive or not just by looking at the type itself. For instance,
in most cases, a `kotlin.Int!!` is a java `int` but it might also
be mapped to `java.lang.Integer` if the type was resolved from a
type parameter (e.g. List<kotlin.Int>.get returns a boxed Integer).

To support this, I've removed the `wrap(KSType)` method from
KspProcessingEnv and replaced it with a wrap method that also
expects the KSTypeReference. KspProcessingEnv uses that type
reference to see if the references was resolved from a type
parameter. For cases where we always want the boxed type (e.g.
type of a TypeElement, there is a wrapDeclared method).

This is still not fully complete as it will miss overrides but
for overrides, kotlin generates two methods (see aosp/1458703).
I'll tackle that when we can run integration tests.

This CL makes TypeNames a bit inconsistent as the primitive types
are the only TypeNames in java. I'll followup with another CL
where
* TypeNames are always generated in java
* Ksp processing always uses kotlin types (e.g. java.util.List will
be mapped to kotlin.collections.MutableList).

After that CL, consistency will come back and then we'll start
running integration tests.

Bug: 160322705
Test: Existing tests + XProcessingEnvTest#getPrimitives
Change-Id: I271171cf7b35cd8f4249350fbdfadf81f55e6a8a