Optional: Move produceFile parameter to be the last parameter in the list.
I'm not sure whether we should do this. The syntax is a little bit nicer but it also moves the required file parameter to be the last parameter which is a bit weird.

Feature request here: b/169425442

Test: Existing datastore tests
Relnote: The produceFile parameter on DataStoreFactory and PreferenceDataStoreFactory is now the last parameter in the list so it lets you use the kotlin trailing lambda syntax.
Bug: 169425442
Change-Id: Ibe7f1ef0d244dd22ad1513a66e79f228fd088469