Do not wrap SurfaceProvider
This CL is walking back from a premature optimization in a previous CL aosp/1457478. The optimization made core exceptions more catchable by binding use cases before the controller is attached to a PreviewView.

However, it breaks the behavior that CameraController only works when attached to Preview. This will not work, if there are 2 controllers created, as pointed out by API review. The second controller will fail to construct because the camera is already attached to use cases.

Bug: 161377155
Test: manual test and ./gradlew bOS
Change-Id: I197e039cbb4bdbe3d6278bf3737ef6f1ff560e9b