Rename apks for suite testing
* We were observing collisions between apks
with identical names, causing them to be overwritten
during the copy to the zip file. This resulted in multiple
module config files referencing the same (incorrect) file.
To fix, we rename the apks upon copy. Also added a strategy
to the copy spec to fail the build if it would copy duplicates
as a failsafe.

Test: "./gradlew bOS" then unzip the resulting androidTest.xml zip
Change-Id: Ia8c3a77765d36c7ec543bd9cd6ab624fb35d127d