Fix possible memory leak due to inner classes of InteractiveWatchFaceImpl
Some tooling complained about this, the worry is retaining references to
the stubs may lead to leaks.  The lint error suggested these classes
should not be inner classes. I've gone further and removed the
references InteractiveWatchFaceImpl held for the stubs, I don't think
it's necessary for it to retain them.

Test: Presubmit
Change-Id: I5b892f7f7d72d3fe2a599dfc8e8fef2c90a50a6d