Deprecate place(Offset) and placeRelative(Offset)
We shouldn't pretend that we can layout children in floats as it is not true and we just do rounding internally. Rounding should be done explicitly instead if needed, but in most cases we can just use ints.

Test: ./gradlew bOS
Relnote: Deprecate place(Offset) and placeRelative(Offset). Use overloads with int offsets instead
Change-Id: I4c5e75e6ba7382735acccd44324bb96a59d82490