Move semantics props away from foundation
Moves semantics properties out of foundation into ui. This also
required to move ToggleableState. This is based on our conclusion that
the properties are essential which justifies for them to be moved to ui
for general usage. This also breaks dependency of ui-test on foundation.

Also removed supportTextInput and isInMutuallyExclusiveGroup as these
seem redunant from now on.

Bug: 171218089
Test: Compiles
Relnote: "Moved foundation semantics properties to ui"
Change-Id: I6f05cc7c0bdf1e5344440cd6e492fbc0545011e7