Remove deprecated APIs in SafeArgs
In Gradle 4.1 the getApplicationIdTextResource method was deprecated. We
still need this method in safe args, specifically for feature modules
which cannot use getApplicationId to find the id dynamically. The
solution in 4.1 does not work for feature modules, so we will use
reflection to avoid the deprecation method, until 4.2 is stable and we
can use that directly.

Test: ./gradlew --rerun-tasks navigation:navigation-safe-args-gradle-plugin:test
Bug: 159542337

Change-Id: I5e25c19fd9ada5186cf470d988e222941d97bf64