Add InteractionState to foundation TextFields.
This CL adds InteractionState to Basic (and, as an impl, to Core) TextFields, allowing people to react on the Interactions that are happening on the TextField
All 3 current types are supported. Interaction.Pressed when TextField is pressed. Interaction.Focused, when focused, and Interaction.Dragged when TextField is being scrolled.

Change-Id: I61d91b15858fc3914ab9f15409857f0bccf67f34
Relnote: InteractionState support for TextFields has been added.
Fixes: 171484305
Test: new tests added