Stops repeating ahead of still capture when flash is always/auto on Samsung SM-A716 devices.
On Samsung SM-A716 series devices, when taking photos in dark environement with auto/always
flash, it will not notify the results to the app.

Added a workaround to fix the problem by stoping repeating before triggering still image
capture and resume repeating after the capture is done or abort.

Relnote:"Fixed the issue that taking pictures doesn't receive results when flash is always/on
and environment is dark on Samsung SM-A716 devices."
Test: StillCaptureFlowTest,  all camera2 test
Bug: b/172036589

Change-Id: If98b2b45cbf1a7fb7228d4db1885846621644341