Reset input when composition or text changes
When the text or composition changes by the application
IME needs to be informed about this fact. Otherwise
IME thinks that their data is correct and in synch. An easy
culprit for this issue is: if onValueChange always rejects the
given input, keyboard still shows/updates suggestions thinking
that the data that was previously sent is visible to user.

When the IME is not informed what is happening, also the duplicate
text issues occurs since IME still thinks that the composition
range is correct, and that is what the user sees.

This CL changes the EditProcessor behavior to compare the previously
suggested state to the new state it gets. If
- the text is different
- the composition is different

It resets the input.

Test: Added new tests
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:connectedAndroidTest
Test: ./gradlew compose:ui:ui:test
Test: ./gradlew compose:ui:ui:connectedAndroidTest
Test: ./gradlew compose:foundation:foundation:test
Test: ./gradlew compose:foundation:foundation:connectedAndroidTest

RelNote: Added resetInput parameter to TextInputService#onStateUpdated

Bug: 172239032
Bug: 171860947
Change-Id: I3e8f5404553921bd94ae424d2840ca5595b6f90b