Prioritize given args over defaults
When there are multiple deeplinks on a NavDestinations, currently we
choose which deep link to match with based on the number of matching
arguments. This creates a problem when considering default values since
2 default arguments and two given arguments would be considered the same
match.

We should instead choose the deep link with the most matching,
explicitly given, arguments.

RelNote: "When matching deep links, navigation now prioritizes
explicitly given arguments over default arguments."

Test: NavGraphAndroidTest.matchDeepLinkBestMatchPathAndQuery
Bug: 162574940

Change-Id: Ieb84eb7781ff8a5cec635a58b2b92294bc1e3a2a