Only require Activities for inflated FragmentContainerViews
We currently throw an error when attempting to inflate a
FragmentContainerView outside of a FragmentActivity. This requirement is
actually only for views that are being inflated.

This change makes it so we only throw the error for inflated
FragmentContainerViews.

RelNote: "FragmentContainerViews that are not inflated using a class or android:name attribute can now be used outside of a FragmentActivity"
Test: all tests pass
Bug: 172266337
Change-Id: Id4397edf815e6603016e0a7798232c8e9f9f065f