Only requestFocus on child views of Fragments
When resuming a Fragment, before we call requestFocus(), we should check
that the saved focus view is still within the fragment view hierarchy.
If the saved view is not present we should avoid making the call to
requestFocus().

RelNote: "Fragments that remove their focused view during an animation
will no longer attempt to restore the focus on the detached view once
they reach RESUMED"
Test: focusViewRemoved
Bug: 172925703

Change-Id: I38c65f834e9b85c365952fdb1e84e03a82863063