Desktop: add desktop-samples and desktop tests (only building, not running) to buildOnServer
There are often changes in Compose that breaks compilation of desktop samples and tests (approximately once in a week).
For example, we can move/delete classes in androidMain/commonMain sourceSets and forget to change code in desktop samples/tests.
To avoid this we added them into buildOnServer task, so developers could see if their changes breaks desktop samples/tests.

Change-Id: I24a157fea9a4718acfde5ccf99e626c223e056ac
Test: ./gradlew buildOnServer -Pandroidx.compose.multiplatformEnabled=true