Run TypeAdapterStore tests with KSP
This CL has the first integration of KSP tests into room-compiler tests.
The first coversion happens in TypeAdapterStore test.

To make them pass with KSP, we need to start setting nullability properly
in adapters. I've only updated the ones that are necessary to pass the
tests, we'll need to do more but I want to add tests for them before
making the changes (e.g. enum adapter).
This should not have any impact on java processing pipeline as it
ignores nullability in equality etc.

Bug: 160322705
Test: TypeAdapterStoreTest
Change-Id: Ic3a572e422fc59489d41daf90178bb50280f1af6