Improve Ksp isSameType & error types
This CL fixes two issues:
* isSameType in KSP was not working as Room expects for platform types.
We still want it to consider nullability for kotlin types but for types
coming from Java (or .class w/ nullability), we use flexible equality
the same way kotlin does.
We may eventually change how isSameType works as it is not always clear
whether it expects exact equality with nullability or without
nullability.

* We were losing error types by eagerly calling `asMemberOf` which loses
that information. https://github.com/google/ksp/issues/107
To workaround it, we don't call asMemberOf if the type resolution
returns an error type.

I've also updated KspTypeTests to use the XProcessing classes to ensure
they go through the same expected path (which revelaed the asMemberOf
bug). They are still not running w/ KAPT because some things like
wildcard handling changes. I'll do that in a followup.

Bug: 160322705
Bug: 175246617
Test: XTypeTest, KspTypeTest
Change-Id: Ibd241456e4f402e5bc2b9002511ff59b7fab08e2