Add Executor param to PagingData operators for Java
Instead of forcing Java users down the sync path without much control
for execution context, add an Executor argument and remove the -Sync
suffix of the method name.

In cases for Guava, while it's possible to build the executor
into the Future itself, it's still desirable to be able to
control which execution context it receives the result on.

Relnote: "Allow Java callers to use PagingData transform operations in
  an async way, by accepting an Executor into transform operator
  arguments. All of the -Sync transform operators have the -Sync suffix
  removed now, and Kotlin Coroutine users will need to disambiguate by
  calling the extension function which accepts a suspending block instead.
  All PagingData transformation operators have been moved to extensions
  under the static PagingDataTransforms class. Java users will need to
  call them via static helpers e.g.,
  `PagingDataTransforms.map(pagingData, transform)`
  For Kotlin users, the syntax is the same but you'll need to import the
  function."
Fixes: 172895919
Test: ./gradlew paging:paging-common:test
Change-Id: If688562d07d96e943b1abfa3690042132db1c4d0