Remove XElement.asDeclaredType
This CL removes XElement.asDeclaredType as it requires an unsafe cast.
Instead, caller should check the type (isTypeElement) and then use
XTypeElement.type property.

Note that there will be a followup CL that merges XDeclaredType to XType
but before that, we need to cleanup such unsafe usages to make that
migration safe.

Bug: 175417726
Test: existing tests
Change-Id: If5d5af593a1d031db9c4b14c1fb2f02659fe762d