Remove XType.asTypeElement & XType.isType
This CL removes XType.asTypeElement, which was unsafe. Instead, XType
has a new `typeElement:XTypeElement?` property. This property can be
used to get the type element, if it exists. It is nullable, hence the
calling code has to handle the null case. Appearantly, we were not
handling this which would result in compiler exceptions. I've added a
bunch of new error messages to detect them and fail more gracefully.

I've also removed XType.isType. Looks like we were using it just to
prevent Auto-Common from throwing when `isTypeOf` is called. Instead,
we're now safe-guarding it in the java abstraction and simply returning
false.

Bug: 175417726
Test: CustomConverterProcessorTest, DatabaseProcessorTest,
PojoProcessorTest, ShortcutMethodProcessorTest

Change-Id: I4ec706d6fad3e345c7b7fdf01fc90b2436d345e2