Returned the passed insets instead of a copy
Previously, we were saving the last insets and returning the saved
reference to save convesion to platform insets when the same insets are
applied twice. But since we are keeping the same reference that we
passed as the return value, it can be consumed later which would then
stop the dispatch on subsequent called.

We also need to call the listener anyway to avoid some bugs.

Bug: 175391720
Bug: 175713604
Test: WindowInsetsAnimationCompatActivityTest#child_callback_called_when_consumed
Change-Id: Ie4bb29458b66c3634984266af207121b1acd712d