Run KspTypeTests with javac
This CL runs KspTypeTests w/ kapt to ensure they work the same.

This required 3 changes:

* KAPT replaces undefined types with `error.NonExistentClass` which
means the `isError` check on TypeMirror is not reliable. I've updated
Javac implementation to consider `error.NonExistentClass` as error as
well. (Enabling correctErrorTypes should've prevented it but it does not
seem to be happening in kotlin compile testing. I've not digged deeper
because we would like to detect this even when developer did not enable
correctErrorTypes option in kapt)
 -> In KSP, we had a similar fallback class in Room package. I've renamed
    it to use the same class as KAPT for consistency.

* When types are resolved, KSP resolves bounded type variables to their
extends bound instead of a non-type TypeMirror. Initially, we
simply implemented extendsBound to return null for all KSP types which was
inconsistent with Javac. This CL updates KspTypeArgumentType to implement
extends bound properly (basically return the ksType).
 -> We no longer create KspTypeArgumentType for invariant types as they
    iare declared.
 -> If we receive a ksType whose declaration is a type parameter, we
    create KspTypeArgumentType. This might happen after calling
    asMemberOf.

* When resolving nullability from KmType (kotlin-metadata type class),
we read the nullability flag to make a decision, which is not enough to
cover cases like `T : Foo?`. Now, if the KmType has an extendsBound, we
also look at it.

Bug: 160322705
Test: KspTypeTest, also extended existing tests to check for bounds

Change-Id: Idc944519d0300804ea29cb88a0372978c1ffd072