Add API to specify a custom focus traversal order
This Cl adds a Modifier.onKeyEvent to the FocusModifier so that it can receive hardware key events.
Then in response to these events, it attempts to move focus left/right/up/down or next/previous.
Focus moving is not supported yet, This CL adds API to specify a custom focus traversal order, and only moves focus if a custom focus traversal is set.

Bug: 175899543
Bug: 170155556
Bug: 170155429
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.focus.CustomFocusTraversalTest
Relnote: Added Modifier.focusOrder() that can be used to specify a custom focus traversal order
Change-Id: I90cf58d6b1f19341d6cdbc7846fd95934938e9f4