Prevent RuntimeException when calling setDataSource
When retrieving EXIF data from HEIF files, ExifInterface calls
MediaMetadataRetriever.setDataSource, but it throws a
RuntimeException for API 26 and below because support for parsing
HEIF files was added in API 27. Also, support for retrieving EXIF
data from HEIF files was added in API 28 (ag/2799160), so trying
to get it from API 27 is useless as well.

This CL changes the implementation to call getHeifAttributes only
for API 28 and above and print a UnsupportedOperationException for
API 27 and below to provide more information to developers.

Bug: 172025296
Test: Run ./gradlew :exifinterface:exifinterface:connectedCheck

Change-Id: I1c6af5f38cafa56806ddfe8487083cdb59432087