Use link tags to reference Dagger types from the Hilt extension jdocs.
Using `{@link }` to reference symbols outside of androidx or the Android APIs will causes issues when generating docs and for those that slip they'll end up as a 404 link in the reference docs. This moves the various Dagger API reference in the Hilt extension jdocs to be <a> tags that link to dagger.dev.

Test: ./gradlew projects :docs-tip-of-tree:doclavaDocs
Relnote: n/a
Change-Id: Idb732c880d423b7613738d78ad48d6c4a4e96331