Save dialog state as not showing
Before we save the state of DialogFragments we always call hide()
before. The call to hide does not synchronously change the state show of
state of the dialog. This means we save and restore the dialog state as
showing when we really want to make it false.

This change saves ensures the showing state is actually saved as false
like we intend.

RelNote: "Child DialogFragments will now always show in the proper order
across configuration changes"
Test: tested in sample app from b/177439520
Bug: 177439520

Change-Id: I3080648340b7f05eac35f5ac77f8a7a2c289a104