Track whether an Op is associated with the topmost fragment
Multiple back stacks involve saving and later restoring the
set set of FragmentTransactions that make up the back stack
(via addToBackStack()). The replace() operation is a
unique case where you don't need to know what fragment you
are replacing - it just operates on the topmost fragment
for that container.

Due to how expandOps() works, that initial OP_REPLACE
is lost, replaced by the more primitive operations
(namely, OP_UNSET_PRIMARY_NAV, OP_REMOVE, and OP_ADD).
However, in order to save that FragmentTransaction for
later restoration in the multiple back stack case, we
need to know that the OP_REMOVE doesn't apply to a
specific fragment, but instead just the topmost
fragment of the container.

By tracking this information, we ensure that
you can safely restore a set of FragmentTransactions
(i.e., a back stack) onto the FragmentManager even
if changes have been made between saving and restoring
that back stack.

As an example, let's say the you have an initial fragment
in a container of A and two FragmentTransactions that
use addToBackStack() that called replace() with B and C,
respectively. In the multiple back stack world, you could
save both of those FragmentTransactions and return to only
A. You could then create new FragmentTransactions with
addToBackStack() to replace A with D and E, and then later
restore the back stack previously saved. By doing so,
the FragmentTransaction that previously replaced A with B
would now replace E (the topmost fragment of the container)
with B, followed by a replacement of B with C. Thus, you
would be on a stack that had A -> D -> E -> B -> C.

There's still more work needed to build up the rest of the
Fragment APIs for multiple back stacks - this is just
the first step on the underlying implementation.

Test: existing tests pass
BUG: 80029773
Change-Id: I8432e9043eeb14c05b1a121d94e00722cff31688