Update CAR_APP's version to 1.0.0-beta01
As previously discussed, this prepares us to go directliy to beta
given we have gone through EAP and beta with the original library
already.

Also updated app-aaos module to point to version 1.0.0-alpha01, as
that is not intended to be part of the library group release for
the first version.

Relnote: N/A
Bug: 178532393
Test: N/A
Change-Id: Ic179c4981985715ca4c18e46911640dc6bcfcaac