Unify Composer observer map with Layout's observer map
Removes the custom observer map that was used in the Composer
in favor of the observer map produced for use by Layout. The
Layout observer map allocates more on initial use but has
fewer allocations during recomposition and has a performance
advantage especially for a large number of invalidations.

Adds a result to IdentityArraySet.remove() to indicate when an
item was removed from the set.

Adds `remove()` to IdentityScopeMap.

Test: :compose:runtime:runtime:tDUT
Change-Id: If88b740d9a91729a65bdc71eaa6d85dbd651044d