Move SlotTree.kt to separate module ui-tooling-data
Currently ui-tooling combines inspection and preview stuff.
Going forward compose layout inspector is going to be shipped
within compose:compose-ui:ui aar and won't require devs
to manually include tooling in project. In order to do so:
1. inspection specific code will be moved to ui-inspection
   module;
2 common piece between layout inspector and preview
   is moved to tooling-data.

As a drive-by all APIs in tooling-data
are marked as UiToolingDataApi to show that
they aren't stable

Relnote: "CompositionData.asTree() and related APIs moved
to separate ui-tooling-data module and marked as experimental"

Test: moved tests work
Change-Id: Ic95b89456b04b25ef41b8157a72511d10aca09cc