API change: add camera info getter to CameraController
As per feature request from developers, we are adding a CameraInfo getter to CameraController. The functionality of this method overlaps with #getZoomState() and #getTorchState(), but those methods also provide additional wrapping to always return the latest state regardless of the camera being selected.

Relnote: add a CameraInfo getter to CameraController.

Bug: 178251727
Test: manual test and ./gradlew bOS
Change-Id: Ib8138c44ebc17164d3bf91d4dbc684a0fe7a2728