[GH] Navigation: Allow parsing a `NavType#ReferenceType` from String
If Bundle contents can be extracted as a Reference, allow parsing the Bundle contents as a Reference.

When navigating via deeplink URI, the user is required to string-ify navigation arguments in order to append them to the URI as query parameters. This PR supports retrieving a Reference type from a Bundle

## Proposed Changes
  - replace `NavType#ReferenceType::parseValue`'s method functionality with `Integer.parseInt`

## Testing
Updated `NavTypeTest` to validate that reference types can be inserted and retrieved from a Bundle.
Test: /gradlew test connectedCheck

## Issues Fixed

Fixes: b/179166693

This is an imported pull request from https://github.com/androidx/androidx/pull/127.

Resolves #127
Github-Pr-Head-Sha: 78179a9e7dfe156d647a7e30992fd82aaa40a7ad
GitOrigin-RevId: edf5410d3277740bdb4033bce59655e9c8df960a
Change-Id: I80f4d03a07b04756cd7ea1f3cf43688a115116e4