Set ViewTree owners for dialogs in DialogFragment
While Fragments with a view set their ViewTree owners automatically,
since DialogFragments do not always have a view (as in the case of only
using onCreateDialog), it is possible to have a DialogFragment but no
ViewTree owners set.

We should set the ViewTree owners on the DialogFragment dialog's decor
view. This way it can be accessed.

RelNote: "Dialogs within a DialogFragment can now get access to ViewTree
owners through their DecorView"
Test: testDialogFragmentViewTree
Bug: 180691023

Change-Id: Ib92905306e16dcaee63c985f22b5e36b6718dd39