Detecting all "complex" schema changes.
Complex changes are arbitrary changes on the database schema that will require going through the 12-step generalized ALTER TABLE procedure in order to migrate the database. The current implementation does not handle FTS tables.

In addition, a "callback" field has been added to the AutoMigration annotation in preparation for switching to declaring AutoMigrations only by annotation (with an option to implement a callback for the user in case of column/table renames/deletes and/or usage of the onPostMigrate() function).

For testing, JUnit tests have been added to trigger processor errors to be thrown for "complex" schema changes at the Entity level (instead of Column level).

The follow-up CL will replace the processor errors thrown when complex schema changes are encountered at the Entity level, and implement the 12-step migration for these changes.

Test: SchemaDifferTest.kt
Bug: 74441348
Change-Id: I0b3a982b396a7541ad6279d3921b5372de6d9518