Not wait for the CameraCaptureSession close
Currently, CameraX waits for the CameraCaptureSession to be closed before
closing the Camera, this makes higher latency for the camera switching.
Also, we noticed some devices have issues in the camera HAL layer that
might cause an error while waiting for the CameraCaptureSession to close.

This CL makes the following change to reduce the happen rate of the HAL
error and speed up the camera switching:
(1) Not Wait for the CameraCaptureSession is closed before
switching camera or create a new session.
(2) Decrease the DeferrableSurface use count when the camera device is
closed or the new session is created.
(3) Remove unused workaround for DeferrableSurface

Bug: 175679772
Test: manually run CaptureSessionTest on lab devices
Relnote: Speed up Camera switching and reduce the error happen rate of the camera device

Change-Id: I34c99dd6bf6dcf95ee174f28bacc8bbc9c0caa59