Sets a default empty value when converting DocumentProtos to GenericDocuments
This changes the mSchemaMap in AppSearchImpl from maintaining a map of
Prefix -> Set of Schema Types, to Prefix -> Schema Type -> SchemaTypeConfigProto.
This map is used for query and get operations when converting a DocumentProto
to a GenericDocument to provide the converter a fall back look-up of the
SchemaTypeConfigProto for a DocumentProto that has all empty values.
This change also factors out common prefix utilities into its own class to be shared
outside AppSearchImpl.

Bug: 177611968
Test: ./gradlew appsearch:appsearch-local-storage:connectedCheck --info --daemon
Test: GenericDocumentToProtoConverterTest#testConvertDocument_whenPropertyHasEmptyList
Change-Id: I4bf77637970cececaa3e081fe7a1d60c29c27124