Use a placeholder PagingSource to prevent running DataSource.Factory on main thread
Paging3 introduced InitialPagedList which requires a PagingSource which
was previous provided by synchronously calling PagingSourceFactory() on
the main thread. This breaks the back-compat requirement from paging2's
DataSource.Factory#create() to be called on fetchDispatcher, so instead
of using a real PagingSource, we can simply substitute a placeholder
PagingSource implementation which only ever loads empty pages and provides
null keys.

Fixes: 182798948
Test: ./gradlew paging:paging-runtime:cC
Change-Id: I638d2149a9e8b78884b4f82315d0b1d3d5066d44