Fix drawing positioning for Views inside Compose
The bug was happening when the position of a View inside Compose was
changing but the View was not redrawn (due to a layer). This was
happening because the drawing of the View was done by negatively
translating the Canvas to compensate double offsetting the drawing
positition with the View's position otherwise - once by the Compose
canvas translate logic and once by the View's system drawChild canvas
translate logic. Therefore, when the layer containing the View was moved
in Compose, the canvas kept being translated with the original position
of the View (not being updated since the View did not have to redraw),
causing the drawing to appear at a wrong position.  The bug was fixed by
avoiding View's system drawChild canvas translate logic, therefore not
needing the canvas translate compensation we were doing before anymore.

Relnote: "A bug (beta02 regression) causing wrong drawing positioning for AndroidView and AndroidViewBinding was fixed."
Fixes: 182387688
Fixes: 182490311
Fixes: 182450383
Test: AndroidViewCompatTest
Change-Id: I28dbdccdbdb6c4741011be12471be53d59fd8c44