Repeatable annotation support in XProcessing
Repeatable annotation in Java AP work via a container class which we can
use to get all of the annotations.

Unfortunately, KSP does not yet support it where:
* Each annotation shows up as an element in the annotations list of
KSAnntated if it's origin is source and the wrapper annotation shows up if
origin is .class.

To avoid inconsistencies in the XProcessing APIs, I've added a
`getAnnotations(KClass<T:Annotation>)` API which will return all
repeated annotations. It automatically checks if the given annotation is
repeated and if so, uses the container class to query.

I've also renamed `toAnnotationBox` to `getAnnotation` but kept the old
one deprecated for easy migration for XProcessing clients.

This is still not perfect due to two issues in KSP and 1 issue in
Kotlin:
https://youtrack.jetbrains.com/issue/KT-12794
https://github.com/google/ksp/issues/356
https://github.com/google/ksp/issues/358

tl;dr; Kotlin does not support Repeatable annotations unless it has
SOURCE retention.
Moreover, due to #356, we cannot support these in .class files. It
should automatically be fixed in XProcessing when KSP side is fixed
(though it seems to be a kotlin compiler issue hence may take a while to
fix in KSP).

Bug: 160322705
Test: XAnnotationBoxTest

Change-Id: Ib1fb97861ab3a01060ca8245107698a4d46d67cf