Add scroller to Text benchmarks to prevent clipping
TextField benchmarks are based on Text benchmarks. During
b/170013443 work I realized when a Column does not have
space left it will send restrictions/constraints on the
measurement and for TextField it would cause scrollers
to display.

In order to prevent that for TextFields aosp/1652688 adds
a scroller to the main Column.

Applied the same rule to the Text benchmarks.

Test: Ran TextToggleTextBenchmark on device
Test: Ran TextBasicBenchmark  on device

Bug: 170013443
Bug: 170960653
Change-Id: I050b1f4e98fa554b1b172602e2bc67777de4961d