Reduce license notifications in emoji2 artifact
This change removes the dependencies that added OFL and UNICODE files
to the emoji2 library.

As a result, this module does not have access to the emojicompat font
for testing. All tests for emojicompat that depend upon the font are
moved into emoji2-bundled.

Some methods were promoted to a visibily higher than LIBRARY to allow this refactoring:

- EmojiSpan.getMetadata // LIBRARY_GROUP
- EmojiSpan.getHeight // TESTS
- EmojiCompat.getMetadataRepoLoader() // LIBRARY_GROUP

In addition, MetedataRepo's test factory was refactored to allow it
to be called from outside of the emoji2 module:

- MetadataRepo.create(Typeface) replaces MetadataRepo.create(Typeface,
MetadatList)

As the MetadataList class is not constructable outside of the emoji2
module due to the embedded library file structure from bundleInside.

None of these functions should have higher visibility, as they are only
useful for writing tests of emoji2, and are not useful for writing tests
of an app that uses emoji2.

Test: Instrumentation tests
Bug: b/178035684
Change-Id: I3233ccacfff9a4b77e05f24f86e36b714720b757