Make ComponentActivity provided CompositionLocals nullable
While in mosts cases the CompositionLocals provided
by androidx.activity are always available (as Compose
~always suggests the usage of ComponentActivity or its
subclasses), making the CompositionLocals nullable
allows developers to customize their fallback behavior
if the APIs are not available.

Relnote: "`LocalOnBackPressedDispatcherOwner.current` and
`LocalActivityResultRegistryOwner.current` now return a
nullable value to better determine whether it is available
in the current composition. APIs that requires those APIs,
such as `BackHandler` and `registerForActivityResult`,
respectively, will now throw a more descriptive error if
the underlying owner is not found. `NavHost` now works
even when an `OnBackPressedDispatcherOwner` is not found,
such is the case when previewing the `NavHost`."
Test: updated tests still pass

Change-Id: I7d8b4662b2d30515a4536e212bf6631357a5357f