Add Box container and modifiers to NavHost
NavHost currently does not have a container composable destinations and
just takes the shape of whatever compose it is placed in. Really, you
always want to wrap a NavHost in a Box since that ensures that the
destinations share the same window. This will become even more important
once we introduce transitions, where wrapping a NavHost with something
like a column would result in the entering and exiting destinations
being stacked horizontally in drawing order instead of layered on top
of each other.

This change wraps the destination composable in a Box, and NavHost now
takes modifiers that are passed to that Box.

RelNote: "`NavHost` now accepts a `Modifier`, which is passed down to
the composable container that wraps the destination composables."
Test: All existing tests pass
Bug: 172112072

Change-Id: I85aca06d44b0a1f55b7a8ef5ca7dae34577d28fa