Run AutoValuePojoProcessorDelegate tests with KSP
This CL runs all auto value tests with KSP.
Unfortunately, we've hit a known bug in KSP where we
cannot detect static modifier in dependencies.
https://github.com/google/ksp/issues/231

I've also added support for passing annotation / symbol
processors to the compileFiles method, it was necessary
to compile some source with AutoValue.

I've also removed the classpath filter in XProcessing compiler. It
matters a bit for performance but filtering for Room classes is not
correct for a general purpose library (it also broke the auto value
case). Removed that filtering as the main performance issues is already
fixed in Kotlin Compile Testing.

Bug: 160322705
Test: AutoValuePojoProcessorDelegateTest
Change-Id: Ieb7ebb6ae827d077df5fb3e0685dd0d764f56a8e