Run PojoProcessorTest with KSP
This CL moves PojoProcessorTest to xprocessing.

There are two differences with JavaAP.
Error detection for primitives do not work in KSP since all
types in KSP have a type element (a.k.a. no primitives).
Technically, we could check if their typename is primitive
but they already fail for lack of constructor so I changed
their errors instead in the test.

Also, there is an existing bug in KSP where it loses KSType.isError
information in some cases. Changed the expected error for that test
for now.

https://github.com/google/ksp/issues/371

Bug: 160322705
Test: PojoProcessorTest
Change-Id: I4a102bf9a6e5e749711783ac55955284d4d0b5fb