Run ElementExtTest with KSP
Also includes a minor test cleanup where previously, it provided an
assertThat method which autmatically maps the input to method names. It
was confusing to read so I moved it to an explicit names() function.

With this CL, we can remove the simple run helper method as it won't
be used anymore.

I also tried to make this test run with pre-compiled sources but this
test hits the KSP static bug very frequently, and there is no easy
workaround to ignore it in tests without one-off if statements.
Instead, I created a bug to update it: b/184588339
https://github.com/google/ksp/issues/231

Bug: 184588339
Bug: 160322705
Test: ElementExtTest
Change-Id: Ic9823f7ccdd307f69260533841509fb5d8878560