[GH] [Room][Compiler Processing] Originating element improvements
## Proposed Changes
Right now the compiler processing library only exposes a way to set originating elements when generating a file with JavaPoet, and KotlinPoet should be supported for the case of creating kotlin files. Additionally when writing a file with KSP the incremental mode defaults to isolating with no way to specify that it should be aggregating.

This change:
  - Adds a new boolean parameter to XFiler `write` functions to optionally specify isolating vs aggregating behavior (defaults to isolating)
  - Adds a kotlinpoet extension function to add originating elements.
  - Exposes a public function to convert XElement to an originating element in case users need to access it directly

## Testing

Test: OriginatingElementsTest

Unit tests that check originating elements are created from XElement and added to kotlin and java poet type builders.

## Issues Fixed

Fixes: https://issuetracker.google.com/issues/184568922

This is an imported pull request from https://github.com/androidx/androidx/pull/155.

Resolves #155
Github-Pr-Head-Sha: 6fe69f742bff7599dfb6e3dd1c2e3eedda426062
GitOrigin-RevId: fe90e5a0210d0f0b2159f8f249c11ec8264778f4
Change-Id: Ie5510e7043f7a2b23c01d68491781bda60d7ae97