Desktop: dispatch all snapshot observers in the right thread.
Snapshot applying can be performed in any thread.
So if this is not the thread in which we run DesktopOwner (Swing thread for ComposeWindow, main thread for TestComposeWindow),
we will encounter a race condition and a state inconsistency.

After this CL all commands that should be performed on any snapshot change
will be dispatched using the right dispatcher.
They are guaranteed to run before the next frame because of dispatcher.flush()

Fixes https://github.com/JetBrains/compose-jb/issues/486

Change-Id: I065374afc93b8da8e7cccaf7a1e662da4ea0020a
Test: ./gradlew jvmTest desktopTest -Pandroidx.compose.multiplatformEnabled=true