Return a stable ActivityResultLauncher instance
Use rememberUpdatedState to keep a reference
to the current ActivityResultContract instead
of returning a brand new ActivityResultLauncher
whenever the contract changes. This ensures
that developers who capture the launcher
(i.e., in a LaunchedEffect block) will always
have a reference to the most up to date
launcher even if they don't explicitly add
the launcher as a key.

Relnote: "`rememberLauncherForActivityResult` now
returns a stable reference to the same
`ActivityResultLauncher` instance even if the
contract changes due to recompositions."
Test: new tests pass
BUG: 184856855

Change-Id: Id2d6d602149aee89a7c0b7b1f2c7b8226cb8ed48