Handle a null ComplicationProviderInfo.
The `IProviderInfoService.getProviderInfos()` method returns an array
of `ComplicationProviderInfo` elements that may be nullable.
Unfortunately, the generated Java class for the AIDL does not provide
the right information about nullability, so our code was incorrectly
assuming that it would not be `null`.  In fact, `null` is the value
used to represent the empty complication provider, and therefore we
need to handle that accordingly.

This changelist takes care of the `null` value, which is mapped to a
`null` value in the API `ComplicationProviderInfo`.

Bug: 186458999
Test: manually tests with sample watch face
Change-Id: Ie3349a035460366f47f4c1777e84d3b39955935c