Move Service Connection to View Model
Changing night mode destroys and recreates the activity.
This means bind/unbind at create/destory is very expensive and would reset
the app state which is not desired.
We also cannot avoid unbinding when onDestroy is called as it will leak
a connection.

This change updates the lifecycle of the service connection by moving it
into ServiceConnectionManger and then a view model object.

We also cannot report the destroyed state to the host since it will mark the
app host as destroyed which is not a reversable act. Instead view model
will call terminate on the host when needed.

Test: Added unit tests

Change-Id: If885809bc34a683bab361c9da3e614452b06d14c