Convert ComposeNavigator+NavHost to new Navigator APIs
Utilize the new Navigator APIs and specifically
the NavigatorState within ComposeNavigator and
NavHost. This ensures a single source of truth from
the NavController to the NavigatorState to the
NavHost.

Due to the use of DisposableEffect, the NavHost
composes at least once before the graph is set
(and therefore before the NavigatorState is available).
This logic is encapsulated within ComposeNavigator
such that the NavHost can always assume the
back stack is available.

The ComposeNavigatorTest suite was redone to
focus on the logic that ComposeNavigator now contains
around the back stack, thus avoiding duplicating the
logic already tested in the base NavController and
Navigator tests.

Relnote: N/A
BUG: 80029773
Test: updated ComposeNavigatorTest tests pass
Change-Id: I0698d2a5e0e7bf8379bab8636f54882d33a464b3