Add support for restoring state
Hook up the restoreState to calls to navigate(),
allowing the NavBackStackEntry state to be restored
back to the state it was originally saved in.

This means that even if the underlying Navigator
does not participate in saving state, the
NavController will still navigate() to the same set
of destinations with the same NavBackStackEntry state
when the saveState/restoreState APIs are used.

BUG: 80029773
Test: New NavControllerTest tests pass
Change-Id: I4b69f189314e6edc998391c29c35ab1c4d35fa2c