Update KSP to ksp-1.4.32-1.0.0-alpha08
In this version, KSP changed how symbol processors are registered, which
broken Kotlin Compile Testing's KSP integration.

To be able to apply the update (which is necessary for kotlin 1.5
update), I've replicated what KCT KSP integration does in our repo and
removed dependency on that module.

I've kept it in KCT's package with the same changes from
https://github.com/tschuchortdev/kotlin-compile-testing/pull/140 such
that once KCT has a new version, we can delete the file and depend back
on to the KCT-KSP module.

I've also updated Room to use the new provider API for registering its
processor.

Bug: 160322705
Test: room tests

Change-Id: I03b73303d0559dd1034e467971843c4c85accc53