Handle type aliases in ksp
Type aliases in KSP resolve to a KSTypeAlias declaration which
requires special treatment while producing the TypeName not to create a
false type name.

We had two options on how XProcessing can handle type aliases:

a) Resolve them to the aliased type immediately
b) Resolve them to the aliased type only for TypeName

I opted for option a in this implementaion to mimic what KAPT does.
Moreover, due to nullable type aliases, doing it lazily creates a
difficult nullability decision.
e.g. you can declare:
typealias Foo = MyClass?
Then the nullability of:
val f : Foo is actually nullable even though the declaration is
non-null.

In addition to that, if we keep wrapped type aliases in the codebase,
we risk more place where we might find the wrong declaration (e.g. via
ksType.declaration).

Last but not least, getClassDeclarationByName does not find KSTypeAlias
declarations hence we would need to synthesize them if we wanted to
support individual KSType (or do the unwrapping when we find the
declaration which could cause more inconsistencies)

Bug: 186782654
Test: TypeAliasTest, TableEntityProcessorTest, TypeAdapterStoreTest
Change-Id: I6c3923ac2136882d58ba66742bca8d2194d7daa2