Re-enabled emoji2 default configuration
- Added a 500ms delay prior to thread creation

Total cost in app startup code:
 - Create a handler
 - Call EmojiCompat.init with a delayed configuration
 - postDelayed to wake up and load the font

After a fixed delay:
 - Create background thread to run default configuration loading and
load the font on the same thread.

This is accomplished by adding a new EmojiCompat.Config that attempts to
load the real DefaultEmojiCompatConfig during the font load path. If the
phone does not have the compat font, the config fails out, otherwise it
delegates to the real config.

In doing this, it became clear that many developers will need to set the
thread on DefaultEmojiCompatConfig, so changed the return to
FontRequestEmojiCompatConfig.

Test: Used new benchmarks for profiling
Relnote: "Add short delay to EmojiCompat initialization to avoid
creating work during startup."
Bug: "b/187328685"

Change-Id: I46a50d4ff11036329b9a892531876bc43f6682e7