Fix BroadcastObsever usage
BroadcastObservers are tied to a context and we were incorrectly
assuming that context was always the same. Its not for an APK that
contains multiple watchfaces.  Also there's no need for Headless
instances to register observers, this is important because the
headless instance is destructed on the UI thread but the
WatchFaceControlService (and associated context) goes away first
on a background thread leading to problems.

Test: Manual testing
Bug: 187425827
Bug: 186875155
Change-Id: Iaefeb4e71d2b4b8c386cf93e90b73ec1b341f3c8