Add workflow to run integration tests
This CL adds a new workflow to run integration tests via the new
androidx-ci-action.

I've not removed the FTL step from the main build file yet and will do
so once we fully migrate to androidx-ci-action. Unfortunately, post
workflow trigger works only on the main branch hence we need this
workflow file to run end-to-end test on production.

Bug: n/a
Test: https://github.com/androidx/androidx/actions/runs/851468271
Change-Id: I1f9fa96ef36dab900c2242527bb47983622cb5ee