Add @Discouraged annotation
This annotation can be used to mark elements that are allowed and are not
deprecated but are inadvisable for some reason, such as performance. In
the IDE, lint will flag elements with @Discouraged and return the
annotation's message metadata as a warning.

Relnote: "Adding @Discouraged annotation for marking elements that
         are discouraged."
Bug: 178477913
Test: Build
Change-Id: Ib2549fc20843a13384fa74befc2f733e61549253