Support self referencing types in KSP
This CL fixes (a.k.a hacks) an issue in TypeName generation logic in the
KSP pipeline where we would go into an infinite loop if we are trying to
create TypeName for a self referencing type, e.g.:
class SelfReferencing<T : SelfReferencing<T>>

JavaPoet solves this using an internal constructor in TypeVariableName
and changing bounds after it is created. For this reason, we are also
finding that constructor now and using it to construct a
TypeVariableName reflectively.
https://github.com/square/javapoet/issues/842

Unfortunately, this fix hits a bug in KSP.
https://github.com/google/ksp/issues/476
I kept the test where it expects the Error class so we know when it is
fixed.

Also, there is still saddle differences between KAPT and KSP. More
specifically, KAPT might generate wildcard type parameters due to
variance. We do that at code generation phase and doing it in typename
creation is fairly complicated (a.k.a. i couldn't find an easy way).
This does not really matter much so I'm keeping them separate for now.

I've also changed compilation test capabilities to only consider kotlin
major-minor versions as KSP is more stable now and we should try to
avoid disabling it.

Bug: 187572913
Test: XTypeTest
Change-Id: I9d85f999cd9ec976149dad961b6e94154dde8c16