MediaRouter: set connection state when MRB is created
A media route button starts animation when the connection state
of the selected route is changed.

Since we initialize the connection state as "disconnected", if the button
is created after connecting a remote route, it plays the "connected" animation.
This CL prevents it by setting connection state at the <init> of MediaRouteButton.

Bug: 189407920
Test: ./gradlew mediarouter:mediarouter:connectedCheck &&
manually check by calling supportInvalidateOptionsMenu()

Change-Id: I857f7288bbea58ecba1fd6b14f7e265fbf7a4535