Add TileProviderClient interface and default implementation.
This will be extended shortly with a -testing lib, that contains testing
stubs to communicate with a known TileProviderService. Currently though,
TileConnection has been updated to conform to the TileProviderClient
interface, and all usages updated. This also then exposes
TileProviderClient to allow developer implementations to communicate
with a TileProviderService.

All this class really does is to hide the AIDL interface, and do all the
proto marshalling/unmarshalling.

This has another change. Now that we have the TileProviderClient
interface, TileClient (which used to handle the binding and invoking the
renderer) has been renamed to TileUiClient to make it clear that it's
not related to the TileClient interface.

Bug: 186543504
Test: Current tests updated and extended.
Relnote: "Add TileProviderClient interface, and expose
DefaultTileProviderClient, allowing implementations to bind to a
TileProviderService interface."

Change-Id: I691659d10e9cc011cb612fc2df0f7f2a8d509a73