Replace custom sharing logic with SharedFlow
This CL replaces custom multicast logic with built in shared flow APIs.

Shared flows do not end so i needed to change the CachedPageEventFlow to
account for it and have custom "ending" logic.

We could also consider taking advantage of additional functionality in
shareIn, like timeout for upstream. Unfortunately, it is not that
trivial because the actual value saved is from caching PagingData's flow
events which cannot be restarted so it needs to be kept active. But if
we keep it active, than caching Flow<PagingData> is not really any cost
since it is mostly a wrapper.

We can evaluate those options later, this CL keeps the behavior the same
with what we had before.

UpdateAPI complained about some restricted APIs but these are not
public, not sure why it complained but i just followed instructions to
regenarate api file.

Relnote: N/A
Fixes: 189349369
Test: CachingTest, CachedPageEventFlowTest
Change-Id: Ic998a199660f66a9eb44b9696efd626393ead1eb