Add support for line_height in Spannables, and remove line_spacing.
The line_spacing API was nasty to use; to meaningfully use line_spacing
to emulate line_height, you need the exact font metrics (just the font
size is **not** sufficient; see the implementation of
TextViewCompat#setLineSpacing). Instead, this CL updates this to
line_height (which has the side effect of unifying the API between Text
and Spannable).

Note that this is on the Spannable, not the font. This is nasty to
support on the font (especially on Android), as line_height is
implemented at the paragraph level, not the span level (or rather, if
you add it in a Span, then the last one wins). Because of that, it's
cleaner to put it on the top-level Spannable, rather than individual
spans.

Relnote: Add support for line_height in Spannables, remove line_spacing.
Test: Golden tests will be incoming soon.
Change-Id: Ibeb54523224d4514ba5c94d83a0a34d7d4abfb21