Fix bug when using proportional dimensions.
This could cause an error when the image, for whatever reason, was
measured with a constraint of 0x0. This can happen when a proportionally
sized image, with one side being expand(), was in a box with dimensions
wrap().

We don't allow this case normally; a container having dimension = wrap()
should only be allowed if at least one of its children has a known,
measurable size. This of course isn't the case where the image is both
expand, and proportional.

This CL updates the logic in TileRenderer to catch this case; a
proportional dimension is only measurable if the other axis is
measurable, or is expand and the parent axis is measurable.

Relnote: Fix bug when using proportional dimensions.
Test: Incoming with additional tests/golden tests.
Change-Id: I37ace9ae4619027467b4593aa798f2ce54611907