Handle replace() transaction with saveBackStack()
When using commit() with a replace() operation, the
FragmentTransaction is added to the back stack
before expandOps() is called on it (which only happens
when executing the transaction).

saveBackStack() needs to take into account both the
normal case (where the back stack only contains
expanded transactions) and the case where both the
non-expanded transaction and saveBackStack() are both
pending operations.

Since a OP_REPLACE is expanded into a OP_ADD of that
fragment, we can treat both the same for the purposes
of saveBackStack().

Due to the use or reordered transactions, the
fragment that is used in the replace() and then
immediately saved never goes through even onAttach()
so we add specific logic to ensure that some state
is saved (and a better error message if we missed
any other edge case).

Relnote: "Fixed an issue where the call to
`saveBackStack()` as part of supporting multiple back
stacks would fail when done at the same time as
running a `FragmentTransaction` that used `replace()`."
Test: new SaveRestoreBackStackTest test passes
BUG: 188896069

Change-Id: I731372db582f00dc3da4fb05a4f2892378945618