Handle properties without backing fields
This CL fixes a bug KSP abstraction where it would consider any property
as a field even when the property does not have a backing field.

Now instead, we properly exclude them from fields list.

We also had another small bug where we didn't create fields for
companion properties. KAPT does so now we do as well. For that, I've
changed the KSDeclaration isStatic helper to check for companion
objects. It is necessary as `isStatic` is used for overrides, asMemberOf
etc. (asMemberOf won't work between a class and its companion).

We've hit a bug in KSP where backing field value is wrong for static
companion fields in the classpath so that test path is excluded for now:
https://github.com/google/ksp/issues/491

Also fixed the declared fields implementation in KSP to properly return
based on backing field.

Bug: 177660733
Test: XTypeElementTest
Change-Id: I978c285e5b720bd8cbabb4df563a85153cbb5918