Centralize method / field collection
This CL replaces MethodCollector with a generic DeclarationCollector
that can be re-used for collecting declarations from a type element
and its parents. This new class is also used to collect fields instead
of the custom implementations in each XTypeElement (javac, ksp).

As part of it, I also fixed the JavacTypeElement implementation not to
include enum constant in declared fields list.

In a followup CL, i'll change those methods to return sequence instead
of list in XTypeElement to avoid collecting too many methods.

Bug: n/a
Test: existing tests
Change-Id: Ia805c862a0a841edc8f50a84d1ad7d17399fb660