Report accessibility class name ComposeView instead of ViewGroup
ComposeView is comparable to WebView from an accessibility
service point of view in that it's the root of a different UI framework
that communicates its content via virtual views. Some accessibility
services are currently treating WebViews specially for the parts of the
Android framework API which cannot support virtual views, for example
focusSearch.  Reporting ourselves with a different classname will allow
such services to more easily adapt to Compose.

Bug: 192274274
Test: composeViewComposedContent
Change-Id: I059c6a89a334e270cd4690e809491273a17c46ec