Throw when launching with unregistered requestCode
If a user attempts to launch an ActivityResultLauncher when there is no
corresponding requestCode in the ActivityResultRegistry, we should
throw. If the ActivityResultLauncher has no requestCode, it is not
currently registered and therefore it could never return a result.
Instead of allowing this unexpected behavior, we should always throw.

RelNote: "The `ActivityResultRegistry` will now throw an
`IllegalStateException` when attempting to call `launch()` on an
`ActivityResultLauncher` that has not be registered or that has been
unregistered."
Test: testLaunchUnregistered
Bug: 192567522

Change-Id: Ida75dca1c2947d74376c453f6e3b65c1adf1684d