Add support for radial alignment to CurvedRow
Radial alignment specifies where to lay down children that are thiner than the
CurvedRow, either closer to the center (INNER), apart from the center (OUTER) or in the middle
point (CENTER). It's similar to the vertical alignment in a row.

* Introduce a scope for components on a CurvedRow, so we can introduce a
modifier to specify the radial alignment of CurvedRow's children.
* Add radial alignment parameter to CurvedRow, as a default for children
that don't specify one.

Relnote: "Add support for radial alignment to CurvedRow (similar to the
vertical alignment in a row)"
Test: Added to the demo.
Bug: 190807301

Change-Id: Id9de5375161076290a6ec3935724381f8fa8e403