Add getPhysicalCameraIds to CameraCharacteristicsCompat and getCameraCharacteristicsMap to Camera2CameraInfo
To support Advanced Vendor interface, we need to provide the
CameraCharacteristicsMap when querying extension availability and
initialize extensions. This CameraCharacteristicsMap consists of
CameraId and its CameraCharacteristics.

In addition to the main camera id and its cameraCharacteristics,
it will also contain all physical camera ids and their
cameraCharacteristics if it is a logical camera.

The following methods are added to support this:

1) getPhysicalCameraIds() in CameraCharacteristicsCompat
2) getCameraCharacteristicsMap() in Camera2CameraInfo

Test: Camera2CameraInfoTest, CameraCharacteristicsCompatTest
Bug: 183683118
Change-Id: Ibd6fb7fc7d751daa11c5930aa4bf277c2a8a186c