Updating OAuthRequest to automatically choose redirect_url
redirect_url for sending the request should be different based
on whether the device is RoW or China.

Chaning OAuthRequest.Builder to have context that is needed for
determing which device sends the request.

Bug: 193413932
Test: Existing ones + additional for China case
Relnote: "We have updated RemoteAuthClient library to automatically pick redirect_uri based on device type (RoW/China)."
Change-Id: I38866d48ffdeaa961edd5e67360b509d7ccffd6c