Fix restoreBackStack() on pending added fragment
When a Fragment has been added (via
add() or replace()), it is not added to the
FragmentManager until the operation is
executed. This means that in the case where you
add a fragment, then immediately save its state,
then restore its state all in the same set
of pending transactions, we cannot rely on the
FragmentManager holding the instance of the
Fragment that needs to be saved+restored.

By going through all pending transactions to
find those that have the mBeingSaved flag set,
we can build up the set of pending fragments
which can be used to supplement the set of
fragments that have already had their state
saved.

Relnote: "Fixed an issue with multiple back
stacks when rapidly swapping betweeen back stacks
that would appear as an `IllegalStateException`
while restoring a `FragmentTransaction` or as
a second copy of a fragment appearing."
BUG: 193685189
Test: newly added tests now pass

Change-Id: I9039f01490dae53cb66ef435161b153b4a0568ff