diagnose-build-failure now automatically focusing the search to a single error
It now autodetects the message that identifies a failure based on running build_log_simplifier to identify unusual messages and selecting the longest such message to search for.

Now, once diagnose-build-failure.sh encounters a failure, it will focus the binary search to look for an explanation for specifically what caused that failure, rather than just looking for the cause of any failure.

This can be important when it is in the initial phase of attempting to automatically simplify the list of tasks to run. In some cases, removing a task can cause the build to still fail but to fail with a different error. For example, even though the task `:createDiffArchiveForAndroidxArchCore` runs as part of the build, attempting to run simply `./gradlew :createDiffArchiveForAndroidxArchCore` fails because the task doesn't exist without also configuring the required arch-core project

Bug: 187822154
Test: ./gradlew :help && echo | tee `find ~/.gradle -name "*.jar"` && ./development/diagnose-build-failure/diagnose-build-failure.sh ":help"
      See that it observes this error:

        Exception in thread "main" java.lang.ClassNotFoundException: org.gradle.launcher.GradleMain
        	at java.base/java.net.URLClassLoader.findClass(URLClassLoader.java:471)
        	at java.base/java.lang.ClassLoader.loadClass(ClassLoader.java:588)
        	at java.base/java.lang.ClassLoader.loadClass(ClassLoader.java:521)
        	at org.gradle.wrapper.BootstrapMainStarter.start(BootstrapMainStarter.java:33)
        	at org.gradle.wrapper.WrapperExecutor.execute(WrapperExecutor.java:108)
        	at org.gradle.wrapper.GradleWrapperMain.main(GradleWrapperMain.java:63)

      See that it automatically decides to check for the presence of this line to recognize the error:

          Exception in thread "main" java.lang.ClassNotFoundException: org.gradle.launcher.GradleMain

Change-Id: I945a32ae68d495059eb0c2cd6c22bfad67b9f8a4