Fix popWithTransition in TestNavigatorState
When we popWithTransition the entry being popped should remain in a
CREATED state before being the transition is completed and the entry
coming to the top of the back stack should remain in the STARTED state
until its transition is completed.

This change adds the logic to ensure we hold the entries in their proper
state while the transition is happening.

RelNote: "When calling `popWithTransition()` from `TestNavigatorState`
the entries will now have the proper state while the transition is
running before it receives the `onTransitionComplete()` callback."
Test: modified TestNavigatorStateTest
Bug: 172112072

Change-Id: I66cddae843f476c19ac3a6c40a7cd5031355768a