Excluding disabled tasks from duplicate-outputfile check
It seems that the multiplatform plugin disables :collection:collection2:jar and replaces it with :collection:collection2:jvmJar .

This change also changes the way that exemptions are combined.
Previously if multiple tasks contained overlapping outputs, it was required for all of them to be exempt in order for the build to pass.
Now if multiple tasks contain overlapping outputs, it is only required for all of those tasks except one to be exempt in order for the build to pass, in case some of those tasks are disabled or otherwise known to be not problematic. This avoids having to also explicitly mention which tasks they overlap with.

Bug: 149103692
Bug: 193558867
Test: ./gradlew listTaskOutputs
Change-Id: I62c2c6554fe9895ac80d028e6fe6426d75d9d6a4