Hide media route button when no mr2 routes
When media route button is expected to show the output switcher,
the button should be hidden when there are no routes to transfer
but it showed the button if there is routes obtained from legacy route provider.

This CL fixes this issue.

The app may override this by MediaRouteButton#setAlwaysVisible.

Bug: 170912258

Test: Manually check the button is hidden when
no mr2 routes exist

Change-Id: Icda98d60579f230ccca6a57e378a408cd3135a7e