Default converter for storing UUID
A default TypeAdapter for UUID that stores it as a byte array in the
database.

Used as a fallback default adapater if no custom TypeConverter for UUID is supplied.

If UUID field is annotated with TEXT TypeAffinity, it is still binded
and read as BLOB unless user provides two-way UUID/String converters.

Bug: 195413406
Fixes: 195413406
Test: ./gradlew :room:integration-tests:room-testapp-kotlin:cC
Test: ./gradlew :room:room-compiler:test
Change-Id: I40d526186711948521e534e0989319a628c8b82f