[RevokePermissions] Remove the use of lambdas in PackageManagerCompat
This CL removes the use of lambdas in PackageManagerCompat as there is
currently a bug with LambdaDesugaring when trying to update the
androidx.core library in google3.

Also, this CL fixes the name of a class referenced by the Jetifier
migration.config file.

Relnote: Remove the use of lambdas in PackageManagerCompat to avoid LambdaDesugaring bug
FIXED: 196257508
Test: PackageManagerCompatTest
Change-Id: I36c875108090468d2e66cc4d8a54402eb0d905f9