Initialize complications with NoDataComplicationData
This means CanvasComplicationDrawable always renders something for
a complication slot even if the system hasn't sent us data yet.

Complications are sent one at a time asynchronously and the net
result of this patch is a reduction in complication flickering on
start up. Note its not elicited because it's typically not
possible for the system to send us complication data immediately
due to latency.

Bug: 196958807
Test: Manual testing
Relnote: ComplicationSlots are now initialized with NoDataComplicationData, ComplicationSlot.complicationData now always has a value and CanvasComplicationDrawable.complicationData is no longer nullable.
Change-Id: I4dfd6e996b84f3d6d664fce98bc8e6601d9298cd