[GH] [Room][XProcessing] Fix NPE in getTypeElementsFromPackage
## Proposed Changes
The javac implementation of XProcessingEnv.getTypeElementsFromPackage contains a bug, where an NPE is thrown if the requested package is not found. This is because `elementUtils.getPackageElement` returns null for an unknown package, and this was not previously handled in the code because it is a platform type and this case was overlooked (by me, when I originally contributed it).

## Testing

Test: Added the unit test getElementsFromPackageReturnsEmptyListForUnknownPackage

## Issues Fixed

Fixes: N/A

This is an imported pull request from https://github.com/androidx/androidx/pull/222.

Resolves #222
Github-Pr-Head-Sha: c4618595dc3d043343acc5533f2f8b32258098da
GitOrigin-RevId: 7d58671b84d236490eff6ba58b6d4ba6e27028f8
Change-Id: Ic285b1818adf5d9f5095d00d301dfb6f12e9b638