Separate layout size from touch target size in SemanticsNode
Fixes: 197751214

SemanticsNode was changed to reflect the touch target size.
This caused many tests to have to change to be able to read
the real layout size. A new property was added to SemanticsNode
to have the size of the touch target area.

Relnote: "Added touchBoundsInRoot to SemanticsNode that
includes the minimum touch target size so that developers
can ensure that touch targets meet accessibility minimums."

Test: ran existing tests. AccessibilityScanner for checkboxes
Change-Id: I2e14bf1bab7a745aa2421353f44c734540d2489c