Fix SCROLL_STATE_IDLE not sent
Fixes: 197802324
Fixes: 197100023

The RecyclerView was incorrectly returning true for a fling
during over scroll stretch. That caused the system to think
that it was flinging the view scroll state when it wasn't.
Now flinging during over scroll returns false, indicating
that there will not be a scroll animation.

This CL also fixes the broken StretchEdgeEffect tests.

Test: manual testing. New regression test.
Change-Id: I84c532e0eb12125f2edfac1da14b946dd53e9013