Prevent GC from removing LoadStateCapture in PagedList tests
We currently track load state output in tests using a WeakReference
which gets GC'd in the middle of the test since nothing references it.
This CL changes LoadStateCapture to accept a lambda for the body of the
test so we can assure that listener does not get GC'd. This fixes
existing failures on ToT.

Test: ./gradlew paging:paging-common:test
Change-Id: I1d8a07966592deb9bef940ca519659436c4e857c