Fixes TabRow and ScrollableTabRow not measuring divider with a minimum height of 0
Previously the divider was measured with the same constraints for the entire tab row, so setting Modifier.height(100.dp) on the tab row would cause the divider to also be measured with 100.dp height. Now the divider is measured with a min height of 0, so by default the divider will still be the correct size.

Fixes: b/197254738
Test: TabTest
Change-Id: I795836de40c8dfe023ae203ecea32132f6a43837