Update SwipeDismissableNavHost to support saveable state.
Also, add backgroundKey and contentKey to public API
for SwipeToDismissBox as part of SwipeToDismissDefaults,
so that they can be passed in calling code that may not
always have a non-trivial value available. And also,
set backgroundKey and contentKey in Wear Compose
Integration demo so that state is preserved.

Test: androidx.wear.compose.navigation and
androidx.wear.compose.integration.demos.test

RelNote: "Updated SwipeDismissableNavHost to support
rememberSaveable by setting the key identity for background and content"

Bug: 199054308
Change-Id: I746fd6214b2f58bee625a86fd3f422122a0528ed