Remove initial CombinedLoadStates value from presenter APIs
Any presenter class, such as PagingDataAdapter, currently is initialized
with a CombinedLoadStates that has null for mediator states and
NotLoading.Incomplete for source states. This makes the API harder to
use for those using NotLoading to track idle state past the initial
state, and many users have to add .drop(1) to loadStateFlow without any
API contract that guarantees this behavior. Since the initial idle state
is easily hooked into on init, removing the initial state allows
loadStateFlow to be used more fluently, and also guarantees that
mediators states will always be populated if RemoteMediator is used.

RelNote: "`.loadStateFlow` and `.addLoadStateListener` on presenter APIs
  no longer redundantly send an initial `CombinedLoadStates` that always
  has mediator states set to `null` and source states set to
  `NotLoading(endOfPaginationReached = false)`. This means that:
    1. mediator states will always be populated if you use RemoteMediator.
    2. Registering a new loadState listener or a new collector on
      `.loadStateFlow` will no longer immediately emit the current value
      if it hasn't received a real `CombinedLoadStates` from
      `PagingData`. This can happen if a collector or listener starts
      before a `PagingData` has been submitted."
Test: ./gradlew paging:paging-common:test
Test: ./gradlew paging:paging-runtime:cC
Change-Id: I1a7485edd077f1ae08f7d9656cbcdfb6bf9ec99f