Disable swipe gesture in SwipeToDismissBox API if content has no background.
When there is no background content, the swipe-to-dismiss gesture should
be disabled - otherwise, there is no content to which to return after
swiping. For example, at the top of the navigation hierarchy, if swipe
was enabled, a blank screen would have been shown.

Test: androidx.wear.compose.material tests

RelNote: "Added hasBackground parameter to the SwipeToDismissBox
 API so that the swipe gesture can be disabled when there
 is no background content to display."

Bug: 199381364
Change-Id: I313d86797efcef05eaf9d5ef4d86ae1b0c761a17