Handle kotlin default impls in generics
This CL fixes a bug where `hasDefaultImpl` returned false for methods
that receive generic parameters.

It also expands the test for cases where the method is inherited in an
interface. To handle that case properly, I needed to change
JavacMethodElement to use the original VariableElement when doing type
checking, not the `XVariableElement.type` because the
XVariableElement.type is resolved based on the containing type, hence
might have a resolved type for a generic instead of the TypeVariable.

I also needed to chage KspClassFileUtility to not require full binary
match for interfaces. This extra check is only used in tests and cannot
handle private methods in interfaces. Moreover, KSP already has an API
for this work so we'll probably get rid of this.

Bug: 199888180
Test: XExecutableElementTest
Change-Id: I0fe20efae76d297f295d6dca9245f67f2ff76290