Initial version of two-pane preference fragment.
An intial version of two-pane preference. This fragment shows one or more headers of preferences,
each of which is associated with a fragment or an intent to display the detail preferences of that header.
This fragment leverage the exisiting Preference XML definition to build a header hierarchy, and
SlidingPaneLayout as two-pane layout can dynamically adapt to large screens.

Bug: 202078370
Test: Sample app.
Relnote: "Added PreferenceHeaderFragmentCompat for two-pane preference"

Change-Id: I9a2d87178737bb0dd2df8e4f6173f8cc690eb20c