Fix SwipeToDismissBox sometimes not intercepting swipes.
Implements `canScrollVertically` and `canScrollHorizontally` in
`AndroidComposeView` to support the `SwipeDismissLayout` at the top of
WearOS windows when it asks if any views in the tree are scrollable to
determine whether to intercept touch events for swipe detection. Because
the `canScroll*` methods don't respond to a particular offset and are
meant to query the view as a whole, the initial down event is recorded
as the position to check for scrollability. This works because
`ScrollDismissLayout` only asks its children if they're scrollable after
touch slop has been reached.

On the Compose side, the semantics tree is used to determine whether
there is a scrollable node at a given offset in the compose view. The
Wear-specific implementation of `Swipeable` is given fake scrollable
semantics to support this.

This entire solution is admittedly pretty hacky, but it should unblock
basic swipe navigation in WearOS for now.

Test: Tests added for Wear `Swipeable`, `AndroidComposeView`, and
`AndroidComposeViewAccessibilityDeletgateCompat`.
Fixes: b/199908428
Relnote: "Fix WearOS SwipeToDismissBox sometimes not handling swipes."

Change-Id: I9387e6584e9d04d7ffda3a4319e3da2da13788b5