Update TypeConverters API to be more explicit
Previous TypeConverters API did a shortest path from
N types to N types. In practice, it is always called
with 1 -> N (into statement), N -> 1 (from cursor) or
1 -> 1 (when we know both types).

This CL updates the API to match the usage patter.
I've not changed the implementation to avoid destabilization
but the null aware converter will have different heuristic based
on the call reason.

Bug: 160322705
Test: Existing tests
Change-Id: Ice3cfbd305a2329e8380486dbb0281d7bb799447