Process hints syncronously
This CL fixes a bug in paging where we would drop rapid hints without
possibly processing them since hints come from UI thread but collection
runs on another thread (and we don't have buffer).

This especially happens if the UI processes the items in an unprediable
manner. For instance, while going in the append direction, if the UI
lays out items in the order 90..100; we would end up only processing 90
instead of 100. Similarly, on the prepend direction, if it lays them out
20..10, we would only consider 20 hint. This is not necessarily always
the case as the collector coroutine might be running but it is likely to
be the case based on scheduling.

Given that the prioritizatoin order for hints is clear (e.g. we always
want to expand), this CL does that work syncronously as hints arrive and
the output flow is an always incrementing(append) / decrementing
(prepend). Similarly, the last access hint is also moved into the
HintHandler to centralize this.

Each direction can be reset (for retries) and that is the only time these
flows are not monolithic.

Relnote: n/a
Fixes: 200740064
Test: PageFetcherSnapshotTest, HintHandlerTest
Change-Id: I09d4ee16746946513ce38274b6552ed22ebfa293