Move implementation of image capturing to camera-camera2
(1) Modify the CameraControlInternal#submitStillCaptureRequests(),
it will receives CaptureConfigs, CaptureType & flashType parameter input,
all the implementation details would be in the CameraXCamera2 module.
(2) Add a new file Camera2CapturePipeline to put all the
implementation of #submitStillCaptureRequests() in it.
(3) Add Camera2CapturePipelineTest to test the logic of the
Camera2CapturePipeline.
(4) Remove Camera2CameraControlImplTest since the related test was
rewritten or moved to Camera2CapturePipelineTest.
(5) The Camera2CameraControlImplDeviceTest will test on the maxQuality
and miniLatency capturing on real devices

Bug: 205648057
Test: run Camera2CapturePipelineTest, run Camera2CameraControlImplDeviceTest on lab
Change-Id: Ibc17ebb57ec480037595594400e4d6c93cebf41c