[GH] Use relative paths for ktlint arguments
The latest ktlint update removed the support for absolute paths as
arguments.
https://github.com/pinterest/ktlint/issues/1131

Github workflow used to prepend the checkout source to make paths
absolute and now it does not work anymore. This PR
replaces it with `..` since we run ktlint inside `activity` project.

There is probably a better shell magic to get this but I think .. is good
enough for now. (there does not seem to be a consistent solution between
mac and linux for relative paths).

Bug: n/a
Test: locally verified that relative paths with .. works by manually
breaking a file's style in room and invoked ktlint in activity as the
workflow does.

This is an imported pull request from https://github.com/androidx/androidx/pull/247.

Resolves #247
Github-Pr-Head-Sha: da899158ad056ba3b40fe19c6f6750163bc34808
GitOrigin-RevId: 80adfcd76e26d9454aaddb15e6914a16eab5f9e8
Change-Id: Ie6ec09c66a57f89970aff9da581135769b43ca9f