Modify NavBackStackEntry equality
Including the Lifecycle and SavedStateRegistry in the NavBackStackEntry
equals and hashcode functions. This will fix our issues in compose where
different instances of NavBackStackEntry were considered the same and we
were not recomposing properly.

This change always required the following:

Ensuring that ViewModels are not cleared during singleTop operations
Making sure markTransitionComplete clears all entries with the same ID
Ensure when NavController calls onSingleTop we update the links between
child and parents

We also made onLaunchSingleTop in NavigatorState a single atomic
operation to remove the old entry and add the new one.

RelNote: "NavBackStackEntries now with different Lifecycles are not
longer considered equal. This means NavHost will properly recompose all
destinations when doing navigation with singleTop and when reselecting
bottom menu items."
Bug: 196997433
Test: Added tests

Change-Id: I1b3510b238d3c4e6ed5cc86a6fddf795bd9acb02