Ensure DialogFragmentNavigator state stays in sync
When manually calling dismiss() on a DialogFragment,
we need to make sure that the dialog is popped from
the DialogFragmentNavigator's back stack. As the
back stack is a _stack_, all dialogs stacked on top
of that now dismissed dialog should also be
popped.

If developers are using popUpTo and the NavController
APIs, this should never happen, but for the cases
where it does, we'll log a message and pop all
of the dialogs to ensure the visual state and the
navigator state stay in sync.

Relnote: "The visual state of dialog destinations now
correctly stays in sync with the `DialogFragmentNavigator`'s
state. This means that manually calling the asynchronous
`dismiss()` API for `DialogFragment` now properly clears all
dialog destinations above the now dismissed dialog. This
does not affect cases where you use `popUpTo` or
`popBackStack()` to dismiss your dialog."
BUG: 191073055
Test: new tests pass

Change-Id: I77ba23fb7a816fc555d0c383213cac01cbafee43