Initial support for switches
The backport uses two stacked ImageViews, which will allow us to apply different tints to the track and thumb as well as simplifying motion animations.
Currently, the backport doesn't animate between the two states. This will be added in a future CL with either a custom AVD or using an animated-selector with the translationX attribute.

Bug: 201783875
Test: locally
Relnote: n/a
Change-Id: I68fcc898458b449fc31466d412acac41b830d232