Set tileResources to default instance if none provided.
We don't try and fetch tile resources if the resource version is empty,
but we did a non-null assertion on the resource bundle before passing to
the renderer.

The renderer can't cope with null resources, so if we don't fetch a
resource bundle, set tileResources to just be an empty resource bundle.

Test: Manually tested.
Relnote: "Fixed NullPointerException in TileUiClient when an empty
resource version was provided."
Bug: 184852037

Change-Id: I0586ea13311ffe1e8fba6b0304c508f344b1c25a