Generate column index for queries whose column result order is known.
For queries without a top-level star projection the column result set is deterministic and therefore Room can generated the column index for a field by using the column result set of a JDBC analyzed query.

For optimizes queries via '@RewriteQueriesToDropUnusedColumns' the column index is also determined at compile-time even if the original query has a start projection, since the optimization rewrites the query with a known column order.

Bug: 160714163
Test: :room:integration-tests:room-testapp:cAT
Change-Id: I22a130ef7fa4979191c7f6cba878e34f4ebb1c17