Add a fake implementation of WindowMetricsCalculator.
Add a fake implementation of WindowMetricsCalculator that returns an
approximate but inaccurate result. This is done to support clients that
rely on Robolectric testing or need the supplied values of the
calculator to be consistent across android versions for tests.
Robolectric does not implement all APIs to match the actual platform
implementation leading to crashes at P and below. This test rule can be
used when a test does not need to be completely correct. One example
would be if are doing a transformation on WindowMetrics into a local
abstraction. If a developer decides to create their own size
abstraction. The important part is the transformation that can be
approximated by the display size and trust the actual implementation to
be correct.

Relnote: Add a test rule for a simple WindowMetricsCalculator.
Bug: 201382530
Bug: 202456663
Test: ./gradlew window:window-testing:cAT
Change-Id: Ibacdb52705d86c127cd1021a36069602440a5aaa