Treat the ICU tokenizer as custom.
When creating FTS tables don't add the tokenizer and args if user defined it to be ICU since the JDBC library used in Room is not compiled with the -DSQLITE_ENABLE_ICU flag.
See https://github.com/xerial/sqlite-jdbc/blob/3.36.0/Makefile

Bug: 201753224
Test: ./gradlew :room:integration-tests:room-testapp-kotlin:cAT
Relnote: Fix a bug where Room would fail to compile with a SQL error when an FTS entity declared to use the ICU tokenizer.
Change-Id: I00db9e36315f69a872a1498b4f5360ec6b576645