Lower min api tested for StartupTimingQuery to API 24
Bug: 183129298
Bug: 190219056
Test: StartupTimingQueryTest
Relnote: N/A

Adds support for hot launches pre-api 29, which weren't traced. We do
this with a very simple approximation for now, measuring from
activityresume to end of the next DrawFrame.

Additionally, adds tracing wrapper with fallback reflection for the
single startup-critical trace section, as it factors into
StartupTimingQuery.

This accounts for the only known platform trace section content
differences going back to intended min support level of 21.

This does not solve other compatibility issues on these old devices,
such as missing trace content, which needs to be investigated
separately. For creating the sample traces, `Thread.sleep()` and
retrying when missing content were both used.

Change-Id: Ief6dd7eff6ab90607df1088adf456c89921a55bb