Add a sample that does not handle Config changes.
Add a DisplayFeature sample that does not handle configuration changes.
We have a report of a leak in the library. One way to reproduce this is
to launch the DisplayFeature Activity that does not handle
configuration changes and rotate the device multiple times.
Add a dependency on leak canary so that leaks in the samples are
reported in the future.

Bug: 202250276
Bug: 202989046
Test: manual - launch display feature activity without config change
Test: rotate the device a few times and see Leak Canary report a leak.
Change-Id: Iae7a99bb689fff69366f560053ecfecd9e87a2f0