Fix insets handling when using OnExitAnimationListener
On API 31, the SplashScreenView sets window.setDecorFitsSystemWindows(false) when an
OnExitAnimationListener is used. This also affects the application content that will
be pushed up under the status bar even though it didn't requested it.
And once the SplashScreenView is removed, the the whole layout jumps back below the
status bar.
Fortunately, this happens only after the view is attached, so we have time to record
the value of window.setDecorFitsSystemWindows() before the splash screen
modifies it and reapply the correct value to the window.

Test: SplashscreenParametrizedTest#decorFitSystemStableContentView#decorFitSystemStableContentView
Bug: 197870763
Change-Id: I59850c8e66c38ca9a7188863735489547f7efe9b