Fix camera2interop options is erased after unbind/bind
The request options will be kept for the lifeCycle camera
in the main thread, but the camera control will clear the
options after the camera control is inactive in the camera thread.
The operations in different thread may cause a timing issue
that may clear the option before lifeCycle camera store it.

Solution: Don't clear the capture request options when control
is inactive, the lifeCycle camera  will clear the options after
is stopped.

bug: 202934891
Test: run Camera2InteropIntegrationTest
Change-Id: I4aba4c7ae0e3596d4162f1a00690117e430ac01a