Stepper implementation for WearCompose
Stepper allows users make a selection from a range of values.
It's a full screen control with increase button on the top, decrease button on the bottom and a slot in the middle
Value can be increased and decreased by clicking on the increase and decrease buttons.
Buttons can have custom icons - decreaseIcon and increaseIcon.
Step value is calculated as a difference between min and max values divided by steps-1

Bug: 200919118
Test: Run tests in androidx.wear.compose.integration.demos.test
Relnote: "Added Stepper for wear compose. Stepper allows users make a selection from a range of values. It's a full-screen control with increase button on the top, decrease button on the bottom and a slot in the middle"
Change-Id: I625fe06a1b1e90d1e59cb20be9fc5b8b19f387bf