Add memory leak tests for Paging
This CL adds new test that tracks allocated objects and ensure we are
not leaking.

This was actually reproducing the attached bug before but it is fixed
after the Change-Id: Ib682ea334e061bc5fa850663cc1ebea0c7a7ebbf.

Regardless, it is good to have these tests.
There is still a relatively smaller leak where we keep the previous
paging state, it is not fixed yet so this CL simply asserts the bug :)
This test should be updated when b/204125064 is fixed.

Bug: 203594733
Bug: 204125064
Test: CachedPageEventFlowLeakTest
Change-Id: I09ad1f1d7a5d11d5c94bc029626c94165e306dee