Fix NavBackStackEntry initial Lifecycle state
NavBackStackEntry currently relies on its host lifecycle owner to
determine the initial host lifecycle state of the entry. This can cause
issues cause the host state might not reflect the dispatched state, so
we can end up with a scenario where the entry moves to a state before
the NavController's lifecycleObserver has dispatched that state.

We should instead make NavBackStackEntries host lifecycle state rely
directly on NavController's lifecycleObserver's dispatched state. This
sure that the entries lifecycle is always properly in sync.

RelNote: "NavBackStackEntries will not longer be pushed down to
`Lifecycle.State.CREATED` after moving to `Lifecycle.State.STARTED` when
the `NavHost` is added directly to the activity's `setContent()`."
Test: NavHostTest
Bug: 203536683

Change-Id: Ia5ac1c2c7e83f90780dd85187519c08debaa6eca