Quirk for UHD video recording on huawei hma-l29
Root Cause: From the CamcorderProfile, UHD (3840x2160) is supported on Huawei Mate 20 and Mate 20 Pro. However, there is not any video frame output from Camera. The CaptureSession is opened and configured, but something error in the HAL. See b/202080832#comment8.

Use the quirk to exclude the UHD quality on these devices.

Bug: b/202080832
Test: manual test VideoRecordingTest on the remote problematic devices &&  ./gradlew :camera:camera-video:connectedAndroidTest
Change-Id: I5327afd754fb9ae461f66ffc8456b6e822015ecf