Fixed VideoEncoderTest flaky test failure
VideoRecordingTest uses CameraUtil#createCameraUseCaseAdapter() util function but not calls CameraXUtil#shutdown() in its teardown function. This caused that the created CameraX instance in CameraXUtil class was not released and reset as null. Therefore, the afterward function to call the CameraXUtil#initialize() util function will encounter the IllegalStateException.

To avoid multiple CameraX instances to exist in the VideoRecordingTest and cause unexpected error, removing the usage of CameraUtil#createCameraUseCaseAdapter(). A hidden isUseCasesCombinationSupported() API is added in the Camera interface to allow the tests to check whether a use cases combination can be supported by the returned Camera instance after calling ProcessCameraProvider#bindToLifecycle().

Bug: 204322933
Test: ./gradlew camera:camera-video:connectedAndroidTest
Change-Id: Ifd69b0745e370941b503bf3f0e7cbcba1cc93734