Improve watch face permission request flow
Test: Manual testing
Bug: 188298101
Relnote: We've added WatchFace.setComplicationDeniedDialogIntent and setComplicationRationaleDialogIntent. These intents are launched to to show a rationale dialog before requesting complication permissions, and another dialog explaining that complication permission is needed when trying to edit a complication when permissions have been denied (the provider chooser will fail to open so the dialog is needed).
Change-Id: I3a29c8eec8c80e741176596dc2fd3a320688f902