Remove resolution parameter from the ExtensionsManager#getEstimatedCaptureLatencyRange API
By using CameraX, CameraX library supports to find the available sizes for the bound use cases. According to the vendor library implementation, the devices might use the YUV_420_888 stream for ImageCapture when extension modes are enabled and maybe only a subset of the supported sizes can be used. It is too complex and difficult for users to understand and obtain this information to query the image capture latency information. In most cases, the ImageCapture can use the maximum supported size unless the users specify a smaller target resolution. Removing the resolution parameter from the ExtensionsManager#getEstimatedCaptureLatencyRange API will simplify the usage and it can fulfill the requirements of most apps.

Relnote: "Remove resolution parameter from the ExtensionsManager#getEstimatedCaptureLatencyRange API since users can't know which sizes are supported for the ImageCapture use case and can't distinguish the returned latency information is for the maximum capture output size or the input resolution parameter."

Bug: 198396873
Test: ./gradlew camera:camera-extensions:build && ./gradlew camera:camera-extensions:connectedAndroidTest
Change-Id: I74bb23e0bcb340772f19a1d620282f4f3607576f