Update SidecarAdapter to use SpecificationComputer.
Update SidecarAdapter to use SpecificationComputer when validating some
objects from Sidecar. I noticed that our tests are sometimes swallowing
errors because it is hard to write a good sidecar test. We have to know
the device to know if we should expect a feature to be present. With a
strict validation we can have the translation throw an exception if a
check fails in tests and either log or stay silent for productnion
versions of the library.

Bug: 205617231
Bug: 205617208
Test: run `./gradlew window:window:cAT` on a device with flaky sidecar
Test: see better error messages.
Change-Id: I05488bc90d41c67c1b137847c399aad1f838e19b