Use restored state instead of default state
When restoring a SavedStateHandle after
process death and recreation, the set of
default values were being merged with the
restored state. This has the effect of re-adding
any default values from keys specifically
removed from the SavedStateHandle.

Instead, we should only use the default values
when there is no restored state, relying 100%
on the restored state as the true state of the world
that needs to be restored.

Relnote: "Fixed an issue where the default value provided
to a `SavedStateHandle` would re-appear after process
death and recreation, even if it was specifically
removed from the `SavedStateHandle`. As a consequence of
this, `SavedStateHandle` will no longer merge default
values and restored values together, instead only using
the restored values as the source of truth."
Test: new test passes
BUG: 204506306

Change-Id: I53a4bbffc6990f8dea902584daf47e780ae9abdb