Rename XMethodElement.name to jvmName
This is in preparation for separating method names from their jvm names.
This CL is a no-op simple rename. The followup will introduce name back
which will refer to the name in source code. Moreover, that CL will also
change Room to use name whenever appropriate. For instance, when
matching fields to their setters/getters, we actually want the name, not
jvm name because jvm name of methods might differenciate based on the
visibility of the method.

Separating name vs jvmName has two benefits:

a) Performance. Resolving jvm name is expensive in KSP because it
requires resolving types (to see if they are inline). Note that gaining
this performance for classpath inputs will require changing XProcessing
to not always return fields / methods in their declaration order (make
it an additional API instead).

b) Logically, we always want the name except for java code generation.

Bug: 205289020
Test: existing tests

Change-Id: I7fd61193db833922ee3b7af64f6cf20eb3d244c0