Add onCheckedChange actions for glance-appwidget compound buttons
This change adds an onCheckedChange parameter for glance-appwidget
compound buttons (e.g. Switch, CheckBox). This new parameter takes
an action that is run when the button is clicked. The action is passed
the current state of the button. CompoundButtonAppWidget is updated to
be interactable using onCheckedChange.

Test: ./gradlew :glance:glance-appwidget:integration-tests:installDebug
 then use CompoundButtonAppWidget
Test: ./gradlew :glance:glance-appwidget:test
Relnote: "Added onCheckedChange actions for compound buttons"
Change-Id: Ia60b4e66fcca0d6517b887f226822b68b7926dc6