Keep dialog back stack entries until the dialog is disposed
When a Dialog is dismissed the content within is recomposed again before
the dialog is disposed. This causes an issue in Navigation where the
dialog content gets a DESTROYED back stack entry.

To get around this we can pop dialogs with transitions and wait until we
get an onDispose callback to move the back stack entry state.

RelNote: "Fixed an issue where attempting to retrieve a viewModel from
a dialogs `NavBackStackEntry` would fail when the dialog was dismissed."
Test: NavHostTest
Bug: 206465487

Change-Id: I6b96d1ce4bcd703340c1d20968d2416c56b9fe46