Fix ActivityNavigator.Destination's equality check
android.content.Intent doesn't override equals()
or hashCode() which meant that two identically
constructed instances of ActivityNavigator.Destination
would not be considered the same by equals() or
hashCode().

By using filterEquals() and filterHashCode(), we
can ensure that the logic in NavController
onGraphCreated() in testing graph equality would
correctly handle activity destinations.

Because many of the fields of ActivityNavigator.Destination
are part of the Intent (and part of the filter equality
check), they do not need to also be present in
equals() and hashCode() and have been removed as part
of this fix.

Relnote: "Fixed an issue when using `activity` destinations
with Navigation Compose's `NavHost` that would result
in infinite recompositions."
Test: new ActivityNavigatorTest tests
BUG: 206677589

Change-Id: I8f64c2edbd87f77662f8b92a6101c179a1eee585