Enable drawing during captureRegionToImage()
Using the new HardwareRendererCompat from androidx.test:monitor, we
enable drawing during the invocation to captureRegionToImage() if it was
disabled. This should enable us to capture images on headless devices.

Bug: 203781965
Test: ./gradlew bOS
Relnote: "`SemanticsNodeInteraction.captureToImage()` will now also work
if `HardwareRenderer.isDrawingEnabled()` is false, by enabling it for
the duration of the call"

Change-Id: Idf3d09c3a63458914d30b66e86bdea17101f7791