Verify more capture requests in Camera2CameraControlDeviceTest
(1) The Camera2CameraControlDeviceTest only verify the first capture request
after the interop interface add the new parameter to the camera,
it could fail if there are any other update comes during the
verification.
The update will verify the request value on the next 60 updates to
ensure it can detect the updated signal.
(2) Use the shared VerifyResultListener in the test

Bug: 208827848
Test: Camera2CameraControlDeviceTest
Change-Id: Ica5a619ae4ce675c09b5612412027785c8549680