Make PreferenceInflater constructor public
Android Studio's layoutlib seems to be using
reflection to access the package private
PreferenceInflater class, but doesn't call
setAccessible(true) when getting access to
the constructor which it just assumes is public.

By changing the constructor back to public,
Android Studio should be able to access the
PreferenceInflater via reflection again.

Relnote: "Fixed an issue where Android Studio
would fail to load the preview for a
`PreferenceScreen` built with Preferences
`1.2.0-alpha02` or higher."
Test: tested in Android Studio preview
BUG: 209572848

Change-Id: I533aa2e0627df3e6f9fcecb3272c69e3f2cf68c8