[Multi-Window] Enable the active resuming of the camera for resuming the camera when interrupted
To makes app seamlessly resume the camera if the camera is
interrupted by other client in the multi-window scenario.

When in active resuming mode, it will actively retry opening the
 camera periodically to resume regardless of the camera
availability if the camera is interrupted in OPEN/OPENING/
PENDING_OPEN state.

When not in active resuming mode, it will retry opening camera
only when camerabecomes available. (Existing behavior)

The CL does not deal with onCameraAccessPrioritiesChange yet and
there would be another CL to handle that after the
onCameraAccessPrioritiiesChange issue is fixed.

doc: go/CameraX-multi-window-support

Relnote: "Enabled applications to restore the camera when camera
is interrupted by other higher priority application in the
multi-window mode and the focus changes back to the application.
Please note that there could be some latency(1 second to 10
seconds) between focus changes and camera reopened due to some
framework issue."
Bug: 124224342
Test: Camera2CAmeraImplCameraReopenTest

Change-Id: I4d0929f80b644387b8ba1ec3ec8b8c792a8a2481