Webkit: Block test thread not main thread in RenderProcessClientTest
Blocking the main thread may block other async initialization calls that
happen on the main thread such as initializing the GPU service channels.
This leads to blocking the initialization of the renderer and not
evaluating any javascripts and thus test failures, see
https://crbug.com/1269552.

The initialization process moved from the IO thread to the UI thread in
https://crbug.com/904556 starting M95 which caused the regression in
WebViewRenderProcessClientTest tests.

Bug: 204197604
Test: Run the tests 100 times using a repeat rule on sdk=30 emulator and Webview M98.
Change-Id: Ie0f61192413804f8d39c768dcbfdb08299bc2764