Avoid finding false conversion paths between user types
This CL changes the algorithm in NullAwareTypeConverter to ensure it
doesn't find conversion paths between two user types (non db types) by
going through db type converters.

For instance, if both TypeA and TypeB can be saved as String in the
database, it doesn't mean that TypeA can be converted into TypeB.

Normally, this wouldn't be an issue in practice as that path would be
more expensive. Unfortunately, due to converters that are written
without thinking about nullability, such paths might become cheaper than
adding a not null check.

The solution tracks whether the current type converter node converts
between a known column type and a user type (or non-column type). If
that is the case, any new converter created by appending into the
current one must preserve the same semantics.

There is also a possible followup to add non-null checks when finding
statement binders. Created a separate bug for it: b/211521846

Bug: 206961709
Bug: 211521846
Test: NullabilityAwareTypeConverterStoreTest

Change-Id: I840a61efbbe2bf82c92e6a64ee88145bf2d03db5