Introduce LookaheadLayout - part 1
This CL introduces a new layout pass for pre-calculation. This
pass allows us to look ahead and calculate a new layout while allowing
the actual measurement & placement every frame to be different than
the pre-calculation. This therefore enables a gradual change over time
towards the lookahead results.

When this is used with animation, lookahead results can be used as
animation targets, whereas the measure/layout will be consuming these
animation values in the `Modifier.intermediateLayout`. When animations
end, the measure/layout is expected to be the same as the lookahead.

The lookahead pass:
- Is optimized so that layout nodes can be re-measured and re-laid-out
  independently, thus narrowing the scope of invalidation
- Supports alignment lines & instrinsics
- Contains a measure/layout pass for the purpose of calculation
- Accommodates nested LookaheadLayout and ensure only one lookahead
  pass for all of the LookaheadLayouts
- Makes the lookahead result observable in the post-lookahead measure
  and layout pass.
- Does not yet support SubComposeLayouts.

Relnote: "New LookaheadLayout that supports a lookahead pass before
the actual measure/layout. This allows a pre-calculation of the
layout when it changes, while permitting the post-lookahead
measure/layout to use the pre-calculated size/position to animate
the size and positions towards the target. SubcomposeLayouts
are not yet supported, but will be in an upcoming release."

Test: Included in the CL

Change-Id: I477f57d1f9efeb0aafd9fb509a2cac0ad8edfaf8