Add VisibilityDocument to Jetpack, Step 2.
This CL split the check visibility function from VisibilityStore and
create VisibilityChecker for this.
Therefore, the VisibilityStore exists in both AppSearch-framework and
AppSearch-jetpack. And VisibilityChecker only exists in
AppSearch-framework.

This CL is also a part of sync upstream from framework.
Classes are reviewed in ag/16183145 and ag/16184850:
VisibilityStore (The check visibility methods are removed)
VisibilityStoreMigrationHelperFromV0

Bug: 202194495
Test: AppSearchImplTest
Test: VIsibilityStoreMigrationHelperTest
Change-Id: I797657519cd80313e9b0fd230ec123f63c86e281