Add a new API to configure XProcessingEnv behavior
This CL adds a new API that lets XProcessing clients to configure its
behavior.

Currently, it is not used for any functionality but I've added one
argument that we'll need to be able to set the class hierarchy
properly.

For KSP processors, they need to pass it in the constructor.
For Javac processors, they need to pass a lambda to configure it. This
is done to allow changing configuration based on options, which are not
available at the construction time of the processor for javac.

Also added a new API for testing to load a configuration provider.
If the configuration for an annotation processor does not match the
default, it would be very inconvenient to provide it into every single
runProcessorTest call. To make it easier, we also support loading a
provider implementation via the service locator
(XProcessingEnvironmentTestConfigurationProvider).

Bug: 160322705
Test: XProcessingEnvConfigTest
Change-Id: I32815e6589ca3b41fd2b4dba75646d3a94599704