Fix continuous backspace on CJK keyboards
Delete operations in TextFieldKeyInput.kt used
onValueChanged(TextFieldValue).
This causes the input connection to be reset; which in turns
prevent continuous delete from keyboard.

This CL changes TextFieldKeyInput.kt to use EditCommands; which applies
the changes in EditProcessor so that the processor can know to reset the
input connection or not.

The CL also fills in performContextMenuAction in a hacky way; but with
various keyboard tested, it works right now (previously they wouldn't
work). Added TODO's to improve this part later.

Test: manual test on emulator / gboard
Test: manual test on samsung s8 / samsung korean keyboard
Test: manual test on samsung s8 / hardware keyboard
Test: manual test on samsung s8 / swift keyboard
Test: manual test on desktop build
Test: ANDROIDX_PROJECTS=COMPOSE ./gradlew jvmTest desktopTest \
      -Pandroidx.compose.multiplatformEnabled=true --no-daemon
Test: ./gradlew compose:foundation:foundation:test
Test: ./gradlew compose:foundation:foundation:cAT

Bug: 194817435
Change-Id: I9fe0144f2f4804802a745dded441de28f73270a7