Add support for sticky placeholders
Normally placeholders are finished when there is not enough space
to show them by the side of the primary activity. However, in some
cases it makes sense to continue showing the placeholder stacked
on top when the window is small after it first appeared in a split
with a larger window size.

This CL introduces an option to make placeholders sticky and behave
similar to regular split rules in stacked case for consistency.

Bug: 202810486
Relnote: Added a new configuration option for activity split rules.
Test: New option in the demo app
Change-Id: Iec6af1c9c7c7209f9d38d30aaa32e1961cd78ddd