Fix ParcelUtilsTest for T
Update testBundlingExpectedCrash to explicitly attempt to load the
test VersionedParcelable from the serialized bundle.

Previously, the test called Bundle.getString to load the control
string. Prior to T, this would greedily deserialize the bundle,
throwing a BadParcelableException when the bundle encountered the
VersionedParcelable and was unable to deserialize it due to the
bundle's null class loader. Bundles in T derserialize parcelables
lazily, so we must attempt to deserialize it explicitly.

Test: ParcelUtilsTest
Fixes: 214053897
Change-Id: I3db9cc8daa6189ea5dfb9fa42150cb20d4489f87