Un-deprecate SparseArrayCompat.removeAtRange
removeAtRange is not a deprecated method from the original collection
source, but was deprecated during the migration to kmp due to odd
behavior noted in b/214092493. Instead of deprecating for removal, this
CL moves us towards just fixing the behavior and keeping the method to
maintain existing parity with android.util.SparseArray.

This unfortunately causes metalava to mark the method as final, but will
fix this in a separate CL, since there are other methods that need to be
fixed too.

RelNote: "Un-deprecate SparseArrayCompat.removeAtRange"
Bug: 214092493
Test: ./gradlew collection2:collection2:checkApi
Change-Id: If445fafb7de9c7baf7229a71e051bf12a0e89b89