Make SubcomposeLayout slots reuse customizable
New SubcomposeSlotReusePolicy interface was added and a SubcomposeLayoutState constructor accepting it.
This policy allows SubcomposeLayout to retain some of slots which we were used but not used anymore instead of disposing them. Next time when you try to compose a new slot instead of creating a completely new slot the layout would reuse the kept slot. This allows to do less work especially if the slot contents are similar.

Bug: 193664975
Relnote: SubcomposeLayoutState constructor accepting maxSlotsToRetainForReuse is now deprecated. Instead there is a new constructor accepting SubcomposeSlotReusePolicy - a new interface allowing more granular control on what slots should be retained for the future reuse.
Test: SubcomposeLayoutTest.kt
Change-Id: I52c4d0360cd987ce03504807312beabdbe410ab0