Add upstream to TextLayout horizontal
This CL connects the upstream position calculation
with previously implemented code.

When text is RTL, and text direction is LTR the code
would throw an exception. Fixes the issue and added
tests for the case.

Finally, android always calculates downstream position,
therefore added a shortcut to LayoutHelper for downstream
calculation.

Test: ./gradlew :text:text:test
Test: ./gradlew :text:text:cAT

Bug: 143556460
Change-Id: Ieb4585ffda9238bf70a0c1c251e58939c0a3055e