Fix getViewModelStore() prior to ON_CREATE
The intention is that the ViewModels associated
with a NavBackStackEntry cannot be accessed prior
to the NavBackStackEntry being added to the
NavController. We were previously using the ON_CREATE
event to signify when this happens, but in fact there are
cases, such as in a Fragment's onCreate(), where a
NavBackStackEntry can be retrieved from the NavController
*before* the Lifecycle reaches CREATED (due to how
fragment lifecycle callbacks relate to Lifecycle changes).

By performing the restore exactly once the first time
updateState() is called, we ensure that ViewModels are
accessible immediately after being added to the NavController.

Test: updated test suites
BUG: 213504272
Relnote: "Fixed an issue where accessing a ViewModel created
via `by navGraphViewModels()` from a Fragment's `onCreate()`
would fail with an `IllegalStateException`."

Change-Id: I8a14dd596195d4ddfa8199c8023a7aedcd286113