De-flake ScalingLazyColumn Tests
Add a temporary extra check on ScalingLazyListState.initialized == true to the ScalingLazyColumn tests that will ensure that all of the initialization has completed before making assertions. This extra check can be removed once the 2 stage initialization of the ScalingLazyColumn is no longer needed - this needs a fix to the underlying LazyList which is being tracked - TODOs added to remove the extra check once the underlying is fixed.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 216138159
Change-Id: Ib70c94077336720c2988a88a80f5112c5d82c950