Improve custom AndroidViewModel parameter error message
When implementing a custom AndroidViewModel, the order of the parameters
must match the order in the super class or there will be an error.

We should improve the error message so that developers know that they
need to reorder their parameters.

RelNote: "The error message when custom `AndroidViewModel` classes have
parameters in the wrong order has been improved."
Test: SavedStateFactoryTest
Bug: 177667711

Change-Id: I340f7696bbf90e331e0b727e9195210cf1dadd0a