Fix ClipDrawTest
Because Outline clip requirements
have been relaxed, path outlines
can now be used for outline clips.
As a result, compose's graphicsLayer
no longer falls back into the manual
clipping behavior which does a full
invalidation + redraw of the layer. Instead
the layer is updated without re-recording
drawing commands. This means that
draw modifiers will not be re-invoked
if a change in graphicsLayer shape is
made. To take this into account, we
move the CountdownLatch#countDown call from the
draw callback to the graphics layer block
to ensure that we no longer run into timeouts
waiting to verify the screenshot result.

Additionally introduce some variance in pixel
sampling coordinates to account for non-deterministic
anti-aliasing behavior.

Fixes: 213947480
Test: Updated ClipDrawTest
Change-Id: I8fe25274e5a022564e8b8893399ea0db4e12a6f9