ScalingLazyColumn: change size semantics to match LazyColumn
ScalingLazyColumn was always taking all of the available space offered to it (unlike LazyColumn) regardless of the size of its contents. This was happening because the layout in verticalNegativePadding was using fixed constraints based on the max size of the parent when adding additional space in order to allow for additional listItems to be composed in the underlying LazyColumn in case they are needed as a result of scaling/shrinking of list list items due to the fish-eye nature of the component. A secondary issue was that the constraints used in the layout where not the constraints from the containing BoxWithConstraints - resulting in incorrect handling of min/requiredSize constraints.

RelNote: "We have modified ScalingLazyColumn so that it no longer greedily fills all of the space in its parent. Instead it will take its size from the size of its contents. This makes it consistent with the behaviour of LazyColumn. If you want to reinstate the old behaviour then pass Modifier.fillMaxWidth()/width()/widthIn() to the ScalingLazyColumn"
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 214316016
Change-Id: I51bf811c2e31e8dd797634e3ce42f7ce9805d79c