Use ScalingLazyColumn for Dialog.
Switched to using ScalingLazyColumn instead of Column. Also added PositionIndicator and Vignette to the Dialog. Added Dialog demos for Alert and Confirmation dialog content.

Bug: 217151157
Test: androidx.wear.compose.integration.demos.test and DialogTests in androidx.wear.compose.material (updated to verify spacing between items in ScalingLazyColumn rather than absolute position).
Relnote: "For consistency with Scaffold, our full-screen Dialog component now displays a PositionIndicator and a Vignette. We have also changing to using ScalingLazyColumn instead of Column, which means that the Dialog contents are now in ScalingLazyListScope (and typically need to be enclosed by item { /* content */ }). Dialog supports the verticalArrangement parameter accordingly."
Change-Id: Idea13159860bce107a0bf2bd492fabfc575038fb