Allow injecting closeable objects into ViewModel
Rather than requiring developers manually override
onCleared() and close any open resources, provide
APIs to allow developers to add one or more
Closeable objects to the ViewModel that will be
closed when the ViewModel is cleared.

As a convenience, provide an additional constructor
for ViewModel that can directly takes one or more
Closeable objects, thus ensuring that injected
parameters can be directly added to the ViewModel.

BUG: 206789617
Test: new ViewModelTest tests
Relnote: "Added a `addCloseable()` API and a new
constructor overload that allow you to add one or more
`Closeable` objects to the `ViewModel` that will be
closed when the ViewModel is cleared without requiring
any manual work in `onCleared()`."

Change-Id: I55ea006f3ddcfbf358c488b30d65a21f4fc85998