Correct trace skipping behavior to fix cuttlefish presubmit
Test: Presubmit

In presubmit, all tests run, including benchmarks, which causes us to
capture traces in tests that don't check for abi support.

Avoid this by skipping tracing specifically for cuttlefish in this config.

Also remove now incorrect check for unbundled tracing support -
unbundled tracebox is contained in benchmark-common, so it's always
available on API 21+.

Change-Id: I3b250633cb2e24fe0a5c3976bbd2901f625ddfac