Add suppress for collections' Kotlin migration
Moving SparseArrayCompat to Kotlin allows NavController to no longer
need a safe call since .valueAt(index) returns a non-null result and
the generic's nullity is now able to be correctly carried through to the
use-site.

This is needed since internal CI is built against the latest version of
AndroidX dependencies, although it is a bit too early to have navigation
depend on colletions ToT directly.

Test: ./gradlew navigation:navigation-runtime:compileDebugKotlin
Change-Id: I332318997ae893908cb74c8f7968439d1197b47c