Fix the List length bug in the CL 1990530.
In the cl 1990530, there is a bug that we should use the length of the
screenArray instead of the itemLimit to control the number of items to
be displayed. I didn't catch the bug while testing because I was testing against different screen size but not different number of items. This CL will fix the bug.

Test: Manual testing with Polestar2 emulator and Automotive emulator,
also tested with different number of items from 1 -> 8 -> 16.
Bug: 215910755
Relnote: Fix a list length bug in the MiscTemplateDemoScreen.

Change-Id: Ibcafcf4969141651bfebb2ee56b35bd223eb4783