ScalingLazyListItemInfo - add unadjusted size
Because of the vagaries of float maths it is not safe to get the unanjusted size of a list item from its scaled size and scaling factor. So we are making it available in the ScalingLazyListItemInfo for developers to use if they need it. It can be used along with unadjustedOffset (already in the interface) to determine the location of an item before it was scaled.

Bug: 221079441
RelNote: "We have added unadjustedSize to the ScalingLazyListItemInfo as it is not safe to calculate the original item size using the scaled size and scaling factor due to float maths precision."
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Change-Id: I54657254014989469bed08b6ef3859cab814b826