Prioritize CreationExtras for ViewModelFactories
We currently have a hard separation between the new stateless
ViewModelFactory that supports CreationExtras and the old stateful
factory that does not. While this was intended, this can cause problems
for compat with older versions, specifically Hilt, which makes a direct
call to the factory create function.

For a compat approach, we can instead allow using CreationExtras when
using the stateful constructor, preferring CreationExtras over any state
that was passed in. This way if there are components that have yet to
integrate CreationExtras, they can still work with newer components that
have.

RelNote: "`SavedStateViewFactory` now supports using `CreationExtras`
even when it was initialized with a `SavedStateRegistryOwner`. If extras
are provided, the initialized arguments are ignored."
Test: SavedStateFactoryTest
Bug: 224844583

Change-Id: I6c43bfd75888cb4b8bdd610cd07d4962aaba37ea