Use local density when asserting layout values
The density in a LayoutNode on which we make a bounds assertion can be
different from the density of the root LayoutNode. It is therefor
incorrect to evaluate and make assertions on the bounds of a node, using
the root's density to convert between px and dp.

Instead, use the node's local density for such conversions.

Bug: 224429422
Test: ./gradlew compose:ui:ui-test:cC \
        -Pandroid.testInstrumentationRunnerArguments.class=androidx.compose.ui.test.assertions.BoundsAssertionsTest
Change-Id: If07601871485b8d40448b27b8ef1e7f995002c9f