Update test app for video write fail with MediaStore
It seems to be a device file system issue, the
output stream of MediaStore.Video cannot successfully
be opened on twist-2-pro and itel-w6004.

The _data columns in the MediaStore.Video are unavailable
on the above devices, it is difficult to fallback
to the file path mechanism.

Store the file to a different position or use FileOutputOption
can workaround this issue.

Bug: 223576109
Test: RecorderTest, manually test recording with CoreTestApp
on twist-2-pro and itel-w6004
Change-Id: I49b25dfef4c56ee6b9164902a5d70d3a478d1c50