Fix selection & getLineBottom with top/bottom padding
When includeFontPadding is false TextLayout adds additional padding
around the original Android layout to prevent clipping (aosp/1992050).

In such cases getLineBottom returns the line bottom from the
previously clipped case. This CL adds the bottomPadding to the last
line bottom, and a test for it.

Moreover, when topPadding is set to a non-zero value, selection is
drawn higher than it should be. This CL fixes the selection path
by offsetting it by topPadding to draw in its original location.
However offsetting does not correct the height of the selection for
each line.

Note: SelectionPathHelper is added in the first version of this CL
to illustrate a patch to draw all of selection with the correct
height. However it is not clear if the visual fix is worth the
any performance cost.

Test: New test
Test: ./gradlew :text:text:test
Test: ./gradlew :text:text:cAT
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT

Bug: 171394808
Change-Id: I22673819a80c9b6ba4d8ad10a2bb020b348c9e27