Expose beforeContentPadding and afterContentPadding on lazy lists and grids layout info
This is useful in situations like when you implement a snapping fling behavior and need to know the real size available for the content after all the paddings. Currently we have to ask for a user to manually pass endContentPadding in order to make it work.

Test: new tests for lists and grids
Fixes: 200920410
Relnote: LazyListLayoutInfo and LazyGridLayoutInfo now have beforeContentPadding and afterContentPadding fields
Change-Id: I3b628e4f4fdfc27b7629c3ac1cd2b47375ff31f2