Fix camera is opened at the INITIALIZED state
The current state machine may move from CLOSING state to
the INITIALIZED state (means fully camera is closed)
accidentally even when it is opening.
It could cause the camera to stay in the opened status
and we didn't close it after the lifecycle is closed.

Not to force open camera in the REOPENING state would fix this issue

Bug: 224686053
Test: PreviewViewFragmentTest
Change-Id: Ifd5bdaf01160c839abae4cbc0cc75184c72ea035