Converted SavedStateRegistryOwner to Kotlin
Part of the kotlinification process of SavedState library.

SavedStateRegistryOwner is converted into Kotlin with source-incompatible changes. Dependencies in navigation, lifecycle-viewmodel-savedstate, and compose ui is updated to use latest SavedStateRegistryOwner.

Test: ./gradlew savedstate:savedState:cC
Test: ./gradlew savedstate:savedState-ktx:cC
Test: ./gradlew navigation:navigation-common:test
Test: ./gradlew navigation:navigation-common:cC
Bug: 220191285

Relnote: "`SavedStateRegistryOwner` is now in Kotlin. `SavedStateRegistry`
is now a property of `SaveStateRegistryOwner` for kotlin users which needs to be overridden via
overriding the property's default get() accessor."

Change-Id: I91fc991e6fdb11ea558a7158f319cfea4aae1e64