Call Processor.startWork() in `DelayMetCommandHandler` from serial executor
After this CL `Processor.startWork()` is always called from serial executor,
so there is now a guarantee for other work in the serial executor, that now new
work will be started in parallel. This is important for work updating feature because
we want to guarantee that if update was successful, then next worker execution will have
updated spec.

bug: 219446409
Test: ./gradlew :work:work-runtime:cC
Change-Id: Ic9ff9e0fdcc6524b12b3761b9976f328e834e95a