Allow using SafeArgs with no manifest file
Gradle now supports namespaces and with namespaces it is possible to
completely remove the manifest if it is not being used. SafeArgs
supports using namespace now and should not crash if the manifest is not
present.

RelNote: "`Navigation Safe Args` will no longer crash when using
namespaces and no `AndroidManifest.xml` is present."
Test: tested in sample app
Bug: 227229815

Change-Id: I17ccf95b5f4137762d7c3db0503ec394812e89e8