Prevent test from leaking state
The ServiceWorkerclientCompatTest was setting a
ServiceWorkerClientCompat but was not cleaning it up, causing failures
in ServiceWorkerWebSettingsCompatTest.

ServiceWorkerControllerImpl was passing a @Nullable reference into a
@NonNull method without checking, which led to NullPointerExceptions
that manifested once the test set the client to null after testing.

This change also merges in the abandoned test first implemented in
https://android-review.googlesource.com/c/platform/frameworks/support/+/1710634

Update to lint-baseline.xml for now, as the underlying issue will be
addressed as part of b/204918082

Fixes: 228071896
Fixes: 143102352
Bug: 204918082
Test: install_webview.py -V 102.0.4973.2 # From chromium/clank/bin
Test: ./gradlew :webkit:webkit:connectedAndroidTest
Change-Id: I92ac5618f0a133dc0b4785754702fd444fc8c314