Prevent LineHeight Clip
aosp/2058653 adds LineHeightBehavior. When lineHeight is smaller than
the font size, and there is a maxLines defined, Compose clips
according to the height of the Paragraph.

This CL updates additional top and bottom paddings for Paragraph
based on the information collected in LineHeightBehaviorSpan.

The handling of Top and Bottom paddings is the same as what was
needed for preventing clipping with includeFontPadding=false
configuration.

Test: Tested on Demo app
Test: Added tests
Test: ./gradlew text:text:test
Test: ./gradlew text:text:cAT
Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT
Test: Tested on API 24, 29, 33

Bug: 181155707
Change-Id: I33029953006acde28c60a8133eecb2d5a0dad614