PopWithTransition on dismiss
DialogNavigator should be using popWithTransition when calling dismiss
on a dialog as well. That way we don't dismiss too early when the dialog
is being destroyed.

RelNote: "Fixed a race condition when using a `ViewModel` within `dialog` destination that would cause an `IllegalStateException` when dismissing the dialog by tapping outside the dialog."
Test: existing tests pass
Bug: 226552301

Change-Id: Id7376c0b8db5be869d8ff53185e15b0603bf8582