Use LineHeightStyle as default for includeFontPadding=true
LineHeightStyle(Alignment.Proportional, Trim.Both) is the current default for
includeFontPadding=false.

This CL makes it the default for includeFontPadding=true as well.

The previous behavior includeFontPadding=true should not change since
it was the same or similar as LineHeightStyle(proportional, trim
both).

The change required no update in regular tests. Treehugger will
validate screenshot tests.

Test: ./gradlew compose:ui:ui-text:test
Test: ./gradlew compose:ui:ui-text:cAT
Test: ./gradlew text:text:test
Test: ./gradlew text:text:cAT

Bug: 181155707
Change-Id: Ib22aaeef41d2b2b6888be7a295524a2baa43c615