Partially migrate preview selection logic from CameraX
Partially migrate preview selection logic from CameraX, and replace the
use of WindowManager with DisplayManager.

CameraX retrieves the application context, stores it and passes it to
its dependencies. This causes crashes in the integration layer of
CameraPipe since it uses WindowManager which cannot be retrieved with
an application context. This CL takes the opportunity to partially
migrate the logic from CameraX (camera2) internals.

Note that there is an existing "extra cropping quirk" that isn't
migrated in this CL (see TODO in CameraUseCaseAdapter.kt).

Bug: None
Test: Run CameraX core test app with CameraPipe
Change-Id: If90f7f734230ffd9ef650e4eb7e1e724f4e8de67