Fix audio/video out of sync in the beginning of the recording
Video data takes more time to process, which means there is a
difference in timestamps between the video and audio data arriving at
the muxer. Because we only cache the latest data, when the first video
I-frame is ready, the "simultaneous" audio data has been discarded.

A ring buffer is used to cache the audio data. The saved audio data will
be filtered by the timestamp and written to the muxer when it is ready.

Bug: 227125710
Test: ./gradlew camera:camera-video:connectedAndroidTest
Change-Id: I321e265436b275effd6c8b0a80165fbf5ecdd777