Deprecate XTypeElement.superType in favor of superClass.
The function name superType was misleading and the behaviour more precisely reflected that of javac's getSuperClass(). Specifically since XTypeElement also represent interfaces, the super types of those when no super interface is present is Object / Any. Meanwhile the super class of interfaces is null, they have none since they are not classes.

Test: XTypeElementTest
Change-Id: Ie0cd5418c05cdadebf2f2c55cd883ff9fd949a66