Draggable fix for dropping events
Added a fix to draggable gesture detection. Because of the nature of how
trackpad to touch event chain worked on ChromeOS, some of its interaction with
our gesture detection system caused the forEachGesture loop to miss some events
and this made the trackpad to behave incorrectly. This fix fixes this by looping
the drag cycle inside the awaitPointerScope and relying on the coroutine scope
to eventually cancel/leave the cycle.

Test: Additional test and all previous tests should pass.
Fixes:205373134

Change-Id: I30a88afcc520055dc00ab20fa236209bede5d9db