[ZSL] Expose CameraX Zero-Shutter Lag API as experimental
Relnote: "Add new capture mode CAPTURE_MODE_ZERO_SHUTTER_LAG in ImageCapture and add isZslSupported in CameraInfo to query the device capability. CAPTURE_MODE_ZERO_SHUTTER_LAG mode is aiming to provide the minimum latency for instant capture. It is implemented based on a ring buffer, which caches intermediate capture results for later reprocessing when user presses buttons to take picture. If {@link VideoCapture} is bound or flash mode is not OFF or OEM Extension is ON, this mode will be disabled automatically."

Bug: b/232558859
Test: Run test app and test the new capture mode on real devices
Change-Id: I9ae74de91ce38e607af4d6b7b65bc4f7e88217c9