Change WakeLockStrategy
With the previous strategy if the WF crashed in draw() then the
wake lock could be held for a long time which is undesirable for
battery usage.

The new strategy should avoid holding the wake lock for extended
periods, however if draw() is excessivly slow then the watch may
go to sleep midway with unpredictable consequenes.

Test: The WF seems to work as expected in ambient, but this is hard to test properly.
Bug: 232718606
Change-Id: I419ea9800d0798300c2db1b9a34ed2a60e1e643d