Refactor DemoApp so that swipe-to-dismiss does not cause recomposition.
We have eliminated the flash and redraw seen in wear.compose.integration.demo when swiping-to-dismiss, by adopting the same pattern as SwipeDismissableNavHost for boxed content from the navigation stack. That is, the SwipeToDismissBox should be given a single line of composable content so that a composable is deemed equivalent whether it is drawn in the background (during the swipe) or the foreground (after the swipe). When the content is seen as equivalent, no recomposition occurs, so we do not see the flash.

Bug: 223358884
Test: androidx.wear.compose.integration.demo
Relnote: N/A
Change-Id: I6a58c839bb84c0f9c574a73bd6c28daea0d79f5e