Initial setup of the post-processing architecture.
Based on the discussion in go/post-processing-discussion, in this CL, we layout the architecture of CameraX post-processing. Please see the JavaDoc of Node.java for an overview. Future implementations are marked as TODO comments.

Please note that the design in this CL are NOT final. Espeically for the public API part, which are subject to change due to API/design review. Due to our limited understanding of the problem, we might not cover everything in this design. If new problem arises which cannot be handled by the architecture, we should discuss and make changes.

Bug: 233280347
Test: manual test and ./gradlew bOS
Change-Id: Ib5fd1be5116bae1953d3901f0aa22ab0f80770e0