Introduce UrlAnnotation to support link a11y.
This change is basically just a copy of `VerbatimTtsAnnotation`.
Includes the new annotation type, methods on `AnnotatedString` and
`Builder`, a new `Saver`, support for converting to `URLSpan`s when creating
`SpannableString`s, a demo, and tests including a missing test for
`VerbatimTtsAnnotation`. All new APIs are marked experimental.

Fixes: b/231495122
Test: ./gradlew :compose:ui:ui-text:test
Test: Tested TalkBack integration manually
Relnote: "Introduced the `UrlAnnotation` annotation type and associated
    methods to support TalkBack link support in `AnnotatedString`s."
Change-Id: I1c754dfa0ce88a36989888b43816333dfc94b0aa