Exclude Compose Compiler from being auto-updated
The default behavior of the script is to update
any RC or stable versions to the next alpha version.
We want to customize this behavior so as a first step
I'm going to exclude the compiler from an update.
Then we should be able to provide some more custom
logic.

Test: ./test_update_versions_for_release.py
Change-Id: I2c75b553e78a5f2badb25546c9957f7c4d15c212