Auto fill-in packageName for ${applicationId} pattern
Currently when using an Activity destination in navigation XML, if you have a placeholder that reflects the "${applicationId}" pattern as part of the "targetPackage" attribute it will automatically be filled in with the packageName of the context. For the "data" and "dataPattern" attributes this is not the case and those need to be filled in manually.

We should add the same support for the "data" and "dataPattern" attributes that is currently in place for "targetPackage"

Test: ./gradlew navigation:navigation-runtime:cC
Fixes: 234223561

RelNote: "When using the exact string `${applicationId}` as the placeholder in the `app:data` and `app:dataPattern` attributes in the activity element of a navigation XML file, the placeholder will be automatically filled in with the packageName of the context upon inflation."

Change-Id: Iaabde79a03b4d9b54bf1633d4abf4c297dd30de8