Avoid overwriting uiMode on activity or application context
Fixes an issue where our context wrapper reused the application
context's backing resource implementation.

We now eagerly use the base context's resources when we don't have
any configuration override, which avoids reusing the application's
resources.

As a side-effect, we now get configuration updates that are (correctly)
missing our overrides. To counteract this, we let the delegate re-apply
overrides and then we replace the configuration update contents with the
desired overridden configuration contents.

Bug: 227502879
Fixes: 236394773
Test: FilternatorTest
Test: NavDrawerActivityTest
Test: entire AppCompat test suite
Change-Id: Idf9d51987fdf953bcc892970f625f451a07b0d3d