Integrate NavHost into ComposeCameraActivity
Extract ComposeCameraApp from ComposeCameraActivity to shift the Composable away from the Activity into its own file.

Create a Scaffold in ComposeCameraApp to hold the navigation component
and content of the screen.

The navigation component for our application is a row of Tabs.
The TabRow code is based on the Navigation Codelab by Android.

The content is rendered is the NavHost, which manages navigation and the
screens that we want to render.

Bug: 234080273
Test: n/a
Change-Id: Iffd0082c733ac609163dd2ead7e72a16bb866927