Restrict devices for ComponentActivity menu tests
Prior to SDK 26, it was possible to pass null menus to Activity menu
APIs. b/142843126 added nullability annotations to framework Activity
that ensured that any menu passed to the menu APIs were non-null.
LeakCanary was implemented in Kotlin with the expectations that the new
framework Activity nullability contracts would hold true. In reality
that is not the case, and on API levels 25 and lower, you can get a
non-null menu passed to a framework API.

There is nothing we can do to address this in the framework. The only
way to run into this issue is by using ActionBar with ComponentActivity
and attempting to use a null menu. Any devs interacting with ActionBar
should be using AppCompantActivity which addresses this issue.
Otherwise, they should be using ToolBar, which also addresses this.

We are going to move these particular test to only run on API 26 and
above to avoid these failures.

Test: updated tests
Fixes: 236868265
Fixes: 229013665
Change-Id: Ib4f606d4a1533bf9b56bd3346c0b5772d5e3cf47