Clean up DynamicFeatures LiveData observer
The test installation failure has special code to ensure that we don't
end the test synchronously. That code uses a liveData to get the proper
callback and DynamicInstallMonitor requires all liveData's to be removed
when they get a failure callback. So even though the particular liveData
in the test uses observe() and would eventually clean itself up, it
needs to do so immediately to avoid the error from
DynamicInstallMonitor.

Test: this fixes the test
Fixes: 236976534
Change-Id: I9c192872995e823526769ff683f3fb30032597b2