Add support for weights to ArcLayout
Test: Added a new unit test
Bug: 230730731
Relnote: We've added weight to ArcLayout.LayoutParams which allows a widget to expand to fill the available space, if there's more then one widget then their share of the available space is proportional to their weight. In addition we've added ArcLayout.setMaxAngleDegrees so you can for example cap the expansion at 90 degrees (NB this does not affect layout of any fixed sized child widgets). Finally ArcLayout.Widget now has setSweepAngleDegrees which allows the ArcLayout to inform a Widget with a non-zero weight of its size.
Change-Id: I75f2449c6bed88e2f6f8a443dd8951b2ce275d76