Prevent direct access to NavController's backQueue
Despite backQueue never being part of the public API surface,
RestrictTo is not enough to prevent some developers from
using it in their own projects (despite it not ever being
the right tool). Directly modifying the backQueue will
cause the NavController to go out of sync with
reality, particularly in cases like Navigation with Fragments.

Instead, we'll provide a new, still restricted, API
that provides the back stack as a
StateFlow<List<NavBackStackEntry>> that provides snapshots
of what the current back stack is without being directly
editable in a way that would adversly affect the
NavController.

The TestNavHostController API, which *does* provide a
public backStack property, has been rewritten on top of
the new restricted API to maintain behavior compatibility.

Test: navigation-testing tests still pass
Change-Id: I94db868dc29efab4799f4f72987c56dc42c554e8