Create CameraBackend interfaces for CameraPipe
This begins the process of formalizing the concept of a "CameraBackend"
in CameraPipe. Backends are responsible for managing the properties,
lifecycle, and interactions across one or more camera devices.

CameraPipe will always have at least one "default backend" that may be
overriden at configuration time for testing and development purposes.
Additional backends are defined by providing a map of factory functions
that will be used to load these backends on demand if requested. This
ensures that expensive resources/connections are not initialized if an
app is not going to use or interact with a specific camera backend.

CameraGraph(s) interact with a CameraBackend by creating a
CameraController. This object is used to start / stop / restart the
underlying connection to a camera. While a CameraBackend may only
support one active camera at a time, it must support multiple clients
that may have overlaping interactions, even if only one is allowed to
submit requests at a time.

Test: ./gradlew\
 :camera:camera-camera2-pipe:testDebugUnitTest\
 :camera:camera-camera2-pipe-testing:testDebugUnitTest\
 :camera:camera-camera2-pipe-integration:testDebugUnitTest

Change-Id: Iea28a396dad5baed54dd685b6b0a7ff3dbf47c8c