Avoid NaN being passed to CubicBezierEasing
An assertion was added in aosp/2137944 that is triggered by some swiping tests in Picker. These showed that a divide-by-zero occurs when constructing a CubicBezierCurve when the distance to animate is zero. Instead, we should omit the animation and return.

Test: androidx.wear.compose.material tests
Bug: 238889897
Relnote: "We have fixed a bug where a divide-by-zero could occur with some fling behaviour when scrolling."
Change-Id: I86cb623e42ad97bfda3382f58f1b525b73a6a595