Add deprecation suppressions to generated fromBundle
The getParcelableArray() and get() functions on Bundle have been
deprecated in SDK 33. This causes warning in the generated Safe Args
classes that use the fromBundle function with custom types.

We need to continue to use the old functions since we need to support
older platform versions so we should suppress the warnings on these
calls.

RelNote: "Navigation Safe Args will no longer cause deprecation warnings
in generated classes when using custom types."
Test: ./gradlew :navigation:navigation-safe-args-generator:test
--rerun-tasks
Bug: 237725966

Change-Id: Id86edf231b90176b5a0f03239bff628171a0284c