Make ActivityFilter more testable.
Add a local class ClassInfo to substitute for ComponentName.
This make it easier to write unit tests since ComponentName is part of
the Android SDK and we can not reference Android SDK  classes through
JVM tests.
Update unit tests to use local class instead of mocks.  Concrete classes
are easier to maintain.
Move comparison methods to MatcherUtil to make MatcherUtil and
ActivityFilter more focused classes.
Add basic tests for MatcherUtils to avoid regressions.

Relnote: Add APIs to check if an Activity is part of the ActivityFilter.
Bug: 238767712
Test: ./gradlew window:window:test
Change-Id: Ia43cfc0fe4d7108a4a2d84d204bba51913309c8f