Move all Fragment animations in the same direction
Say you have 3 fragments, A, B, and C. If you add A and B with one set
of entering and exiting animations. And then pop B, before adding C with
an different set of entering and exiting animations while using the
recommended setReorderingAllowed=true flag on your fragment
transactions, it is possible to get in a scenario where the visible
exiting fragment, B, uses the animations from the popping opertion to A
instead of the adding operation with C.

The reason for this is that the animations for B are set during the
first operation with the pop from B to A. The system is then
techinically doing a replace operation from A to C, since B has already
been popped, so it sets the proper animations on A and leaves B
untouched. This causes B to run the stale animations from the pop
instead of the current animations from the replace.

We should ensure that we always run the animations from the operation
that determines what will be visible to user on the screen.

RelNote: "Fragments will now run the proper animations when mixing pop
and replace operations."
Test: added FragmentAnimationTest
Bug: 214835303

Change-Id: Ib1c07bd0e05c0c1a3d785e29456bad9afb183e2d