Roll forward "Create a FakeSurfaceEffect for testing."
This reverts commit 0325096840d8f337482bc11d72c909e419e026f2.

The original CL was rolled back because of flaky tests: https://android-build.googleplex.com/builds/tests/view?invocationId=I97500010072254785&testResultId=TR97628150100029629&tab=results

The crash is due to the closing of the SurfaceTexture in FakeSurfaceEffect#finalize() on API level 33. Not sure why it would happen though, maybe because SurfaceTexutre has new behaviors on API 33.

To fix this, we remove the finalize() override and add a autoClose flag:
- if the flag is true which is for the most of cases, we automatically closes the SurfaceOutput
- if the flag is false, it's the calling test's responsibility to close the SurfaceOutput.

Bug: 239409308
Test: manual test and ./gradlew bOS
Change-Id: If66816b6cfc3b0ba24e3c906bf0d955485d30762