Prevent text cut for singleLine TextField
When maxLines is defined on TextField, the height is
calculated based on a constant height.

When maxLines is 1 (and softwrap is false), and a
character with greater height is entered, this causes
the taller glyph to be cut.

When maxLines is not 1, and softwrap is true, this
does not cause a problem since it is possible to scroll
the TextField.

This CL updates the height constraints based on text
entered for maxLines modifier when maxLines = 1.

RelNote: N/A

Test: TBD
Test: ./gradlew compose:foundation:foundation:test
Test: ./gradlew compose:foundation:foundation:cAT
Bug: 228512135
Change-Id: I7c2df57d26e9f7bd1a01a4c37883627b8ef7c86a