Update XAnnotationValue API to include the value type.
This information is already available via InternalXAnnotationValue, so
this CL simply exposes it to the public XAnnotationValue. The value type
can normally be retrieved in java source code by looking at the method
return type or in kotlin source code by looking at the constructor
parameter type. However, providing the type explicitely from the
annotation value provides a convienent, backend agnostic way of getting
the type of the annotation value.

This CL also changes JvmDescriptorTypeVisitor to throw a
TypeNotPresentException when visiting an ErrorType rather than a generic
error. This helps Dagger determine the difference between an illegal
state versus an invalid type (e.g. due to the type not being on the
classpath).

Test: XAnnotationValueTest
Change-Id: I3dbc8929598c69bb642f96fa042e6931229f424d