Fix errorprone StringSplitter warnings in :core:core-i18n
Relnote: "Fix errorprone StringSplitter warnings in :core:core-i18n"

This is used to "parse" plural rule samples, and it is expected to skip
the empty fragments. So the behavior was correct, only unclear.

Also used the opportunity to bring this closer to the official ICU4J code
so that we can diff easier, and completely bring them in sync at some point.
(only inconsequential changes, like comments, diamond operator, etc.)

Bug: 239828124

Test: ./gradlew test connectedAndroidTest on all (emulator) API [19-33] images
Change-Id: I789f5705874e0cf1b978e96f2835303e601d418f