Add support for Chips and Buttons with an outlined Style
Update Chip and Button allow drawing of an optional Border. Extend ChipDefaults and ButtonDefaults to add outline styles which have transparent backgrounds and a thin border.

Bug: 236099544
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Relnote: "We have added a new outlined style for Chips and Buttons and new OutlinedChip and OutlinedButton composables that provide a transparent Chip/Button with a thin border."
Change-Id: Id597231272756493c4261f0aeb2d27357f64bb28