Fix Toolbar onPrepareMenu timing with MenuHost
Instead of unconditionally calling onPrepareMenu() when Toolbar executes
onCreateMenu(), it should wait until the menu is about to be shown to
make the call.

This will ensure that Toolbar properly fulfills the MenuHost contract.

RelNote: "Toolbar will now call `onPrepareMenu()` before the menu is shown
instead of unconditionally when the menu is created."
Test: modified the ToolbarMenuHostTest
Bug: 232206677

Change-Id: I2a58df0064c08b9ae87962a41d87e7bc60073df0