Removes pointer position from Overscroll APIs
Pointer position is not fundamentally part of overscroll - it's an implementation detail for certain overscroll effects that shouldn't be part of the common API surface. This makes it easier to programatically control overscroll and makes it more aligned with other APIs like NestedScrollConnection.

Test: Existing tests and OverscrollBenchmark
Fixes: b/241239306
Relnote: Removed pointerPosition from OverscrollEffect - effects that care about pointer position can use Modifier.pointerInput { } in the effectModifier to get the current pointer position instead.
Change-Id: I9f606c640cc23d7483b32ed98b7caf790c9d2e36