Introduce SplitAttributes
SplitAttributes extract splitRatio and layoutDirection from SplitRule,
which is necessary because we are going to provide the ability to
change the current layout with the current device state.
We may also support to change layout at runtime in the future.

Bug: 240912390
Bug: 207494880
Test: ./gradlew window:window:test
Test: ./gradlew window:extensions:test
Relnote: " Add SplitAttributes class to support split the task bounds
horizontally"

Change-Id: Ib0418f19642f58d167ee8867e54fdb7713c4759f