Add IntDef return types for APIs returning ints
The AppAuthenticator APIs support verifying an app's signing identity
based on a permission model for an app accepting connections as well
as expected signing identities for apps installed on the device. This
commit adds new IntDef return types for these APIs to limit the
results to a set of known values.

Fixes: 241247906
Test: gradlew :security:security-app-authenticator:check
Change-Id: Ifb254e2bdc26e2a62761706c5610cda386a85836