Encode SessionInfo as Bundle instead of Bundleable
As the client code doesn't set the classloader for intent extras upon
receipt of an intent, older versions of the client code cannot be passed
custom objects within the extras, otherwise they'll crash due to a
ClassNotFoundException. Because of this, we need to encode objects via
Bundle as the client can already understand Android objects. This
prevents a crash on older clients, while still allowing the host to send
the data it needs to BEFORE it knows the client version.

Test: ./gradlew :car:app:app:test
Fixes: 242066376
Change-Id: I9ad6a6baf0fee278b35a43d249dfcc3a502cbcfc