Call super in OnPictureInPictureModeChanged above API 26
In Activity 1.5 ComponentActivity was changed to use the new
onPictureInPictureModeChangedProvider interface to handle
onPictureInPictureModeChanged callbacks. This allowed for separating
that behavior our from framework Activity.

The onPictureInPictureModeChange() callback has two versions, one param
for < API 26, and two param for 26+. If a component that relies on
ComponentActivity does not handle both versions, it is possible to get
into a case where Activity does not dispatch to the expected callback.

To ensure the callback is always dispatched, we should call super in the
two param version. This will ensure framework Activity has the chance to
dispatch to the proper version.

RelNote: "Classes that extend `ComponentActivity` will now always have
both of their `onPictureInPictureModeChanged()` callbacks dispatched."
Test: modified onPictureInPictureModeChanged test
Bug: 241802137

Change-Id: Ib7fdbb7cbbfe12347df81b43e8e8f3fe841555fd