Call super in onMultiWindowModeChanged above API 26
In Activity 1.5 ComponentActivity was changed to use the new
onMultiWindowModeChangedProvider interface to handle
onMultiWindowModeChanged callbacks. This allowed for separating
that behavior our from framework Activity.

The onMultiWindowModeChanged() callback has two versions, one param
for < API 26, and two param for 26+. If a component that relies on
ComponentActivity does not handle both versions, it is possible to get
into a case where Activity does not dispatch to the expected callback.

To ensure the callback is always dispatched, we should call super in the
two param version. This will ensure framework Activity has the chance to
dispatch to the proper version.

RelNote: "Classes that extend `ComponentActivity` will now always have
both of their `onMultiWindowModeChanged()` callbacks dispatched."
Test: modified onMultiWindowModeChanged test

Change-Id: Ic4d85882270447666ed7b15019c39c04aa4d4602