Fix Navigator singleTop
When doing onLaunchSingleTop the Navigator currently calls the navigate
function that takes a destination, which only the Navigator actually
implements. We should instead be using the navigate call that takes
NavBackStackEntries which all the extended Navigators also implement,
which will ensure they are called even if they don't already explicitly
handle single top.

RelNote: "Navigation will now properly replace the Fragment instance
when navigating to the same destination and setting the
`launchSingleTop` flag to true."
Test: Added NavControllerWithFragmentTest
Bug: 237374580

Change-Id: I5a2f15e1c2a9a4bc5bae6f5d016a9b06b76aa448