Preserve initial value of StateFlow in FlowLiveData's asLiveData
Currently, when FlowLiveData creates a LiveData object
from a given Flow using its asLiveData function, it does
not contain any initial value. Meaning that if the Flow
is a StateFlow, which always contains an initial value,
that initial value will not be initially accessible in
the LiveData. We will now preserve this initial value.

RelNote: "`FlowLiveData`'s `asLiveData` function will now
preserve the initial value of a `StateFlow` when creating
the new `LiveData` object."
Test: asLiveData_preserveStateFlowInitialValue
Fixes: 157380488

Change-Id: I3f53068b1a071e2f334f2cbab2819e7f7a4c60b2