Fix LazyColumn position maintaining when the constraints changes and fillParentSize() is used
The issue was reproducible when an item had Modifier.fillParentSize() and the parent constraints changes. The size known to Modifier.fillParentSize() was only changing on the next frame which is enough for the scroll position to be adjusted as the list assumes we have empty space left after putting all the items(they had the size smaller than needed in this frame).
The fix is to use the new size during the measure pass, not during the composition. When the child will be remeasured because the constraints changed such modifier will know the correct new size already.
Same fix is ported to the TV implementation.

Fixes: 236391559
Test: LazyListTest for both regular and TV LazyColumn
Change-Id: Ic3da187eb10333bbe7308867d363707a5abab966