Update UiWidget tests to reset the cameraImpl
For Parameterized test with camera-camera2/camera-pipe,
UiWidget tests will shut down camerax in @After block,
and it will require to force config the cameraImpl
every time when start the test Activity.

Add an extra Intent key to ignore the configured
implementation type, so it will be able to configure
the camera implement again when next time to start the
Activity.

Bug: 243758509
Test: All tests under UiWidget test app
Change-Id: Ifcda3dd5930cea63b91887a40941eef03b2c106e