Fix cpuDuration calculation and expose totalDuration
A missing metric (the time taken to queue the data sync to RenderThread)
in FrameMetrics made the cpuDuration data in JankStats less accurate than
it could be, and made it harder to determine the equivalent of
"totalDuration" in comparison to data one could get from FrameMetrics.
Further complicating the problem was the overlap of SWAP duration
in the calculation of both GPU and CPU durations on API31, making it
even trickier to align all of these similar-yet-different metrics.

There was a request (see the bug) to expose totalDuration for JankStats
overall, but this is problematic since it is not possible to expose it
consistently across different releases (on 31 it includes GPU, on 24 it
does not, and on 16 it has even less data behind it). Instead of
exposing it overall, we now expose it just in API31, in the
FrameDataApi31 structure, where iut has all of the data it needs to
provide an accurate measure of total duration.

Bug: 243694893
Test: Ran JankStats on API24 and 31
RelNote: cpuDuration now more accurate, also new totalDuration on API31
Change-Id: I59ce8c67f06a168f96893375c8aeca5516a55d81