[CustomTab] Add a new extra flag
Adding the androidx.browser.customtabs.extra.ACTIVITY_RESIZE_BEHAVIOR
extra flag for CustomTabsIntent. It will define the resize behavior for
a custom tab activity when acting as a bottom sheet.

Adding convenience methods for the recently added extra flags that are
enabeling a custom tab to behave as a bottom sheet.

Test: ./gradlew browser:browser:assemble and run tests
Relnote: "Adding a new extra flag for defining the resizing behavior of
    a custom tab when behaving as a bottom sheet.
    Adding convenience methods for the recently added extra flags that
    are enabeling a custom tab to behave as a bottom sheet."

Change-Id: I48bd3f5ceebfac5b9a479c0da555d50f1c3e1725