Enabled instrinsic rmember by default
This change corrects the detection of when it is valid to use `composer.cached()`
instead of `remember` ane enables it by default.

Using `composer.cached` instead of `remember` uses the parent group's
slot space instead of creating a new group. Using the parent group
reduces the overhead of using `remember`. Also if a parameter has already
been compared for skipping the result of the skipping comparision is used
further saving a redundant slot.

Intrnsic remember can be disabled by passing:

  `-Pplugin:androidx.compose.compiler.plugins.kotlin:intrinsicRemember=false`

as a parameter to the Kotlin compiler.

Test: ./gradlew :compose:compiler:c-h:i-t:tDUT
Fixed: b/187523740

Change-Id: Iba66bedca48802f5875e30743e706f55c3641081