Request pin GlanceAppWidget
Adapt the requestPinAppWidget method from AppWidgetManager
for GlanceAppWidgets.

The main difference is that this method renders the preview
when provided.

For just generating the RemoteViews tree we do not need recomposition. This will avoid making requestPin method suspend.

Test: added integration test that check if request pin is displayed
Fix: 207748795
Relnote: "New requestPinGlanceAppWidget to request the launcher to pin the given GlanceAppWidget"
Change-Id: Ic6e47bd3316560f000a4bccffadec31066b62b3c