Fix BackHandler inside a Compose Dialog
System back is delivered to the topmost window which,
in the case of a Compose managed Dialog, is the
dialog itself and not the underlying Activity. This means
that any BackHandler needs to register with the dialog
and not with the activity in order to be triggered.

The underlying work of hooking up a BackHandler to a
dialog was done in Activity 1.5's ComponentDialog class, so
by switching Compose UI to depend on Activity 1.5 and using
ComponentDialog instead of android.app.Dialog as the
base class for Compose's Dialog, we get BackHandler support
for free.

Note that this CL also moves off of the deprecated
onBackPressed() API on Dialog to instead of use AndroidX's
OnBackPressedDispatcher, ensuring that Compose's Dialog
is compatible with enabling Predictive Back. However, due
to how the onDismissRequest API works, Compose's Dialog
still won't be able to support any future enhancements to
Predictive Back such as a predictive closing animation of
the dialog. That work can be handled as a separate CL.

BUG: 201909853
Relnote: "AndroidX Activity's `BackHandler` API now works
within a `Dialog` composable."
Test: new test dialogTest_backHandler_isCalled_backButtonPressed

Change-Id: I3534219bb887fdc6c46c659a1b1f1465d2c8aa7d