Correct the $changed flags for restarted functions
When creating the restart lambda for a restartable function,
the $changed flags are captured along with the previous values
used to call the function. The $changed flags are calculated
by the caller of the function and are used to pass the
comparison information of each parameter.

When a caller passes a parameter that it has already compared
with its previous value either a Same(01) or Different(10)
bit pattern given as part of the changed flags for the parameter.
The callee can then use this information to avoid having to
compare the values itself, saving both time and space.

When the restart lambda calls the function to restart it
because some data it read changed, the same $changed flags
are passed (with a forced execution bit set). However, all
flags that are Different(0b10) are now potentially  wrong as
the values are the same as the previous time the function was
invoked, not different, and therefore, should be Same(01)
instead.

These flags cannot be ignored or changed as they control
how many comparisons a function performs and slot offsets
used to the values to compare.

The fix is to change all Different(0b10) values to
Same(0b01) before restarting the function. This is done
for 10 parameters at a time using a the new utility functions
`updateChangedFlags0`, for changed which includes the forced
execution bit, and `updateChangedFlagsN` for all other
changed flags parameters.

Relnote: N/A
Fixes: 245806803,241426925
Test: ./gradlew :compose:r:r:tDUT :compose:compiler:c-h:i-t:tDUT

Change-Id: I7a7b1453622706e1e24c2f9ccfd652c4ca697b65