Add ComplicationPersistencePolicy and ComplicationDisplayPolicy
Note we've added a base builder to the ComplicationData builders
to reduce repetition, this doesn't change the API from the
caller's point of view.

Test: Added a new test to check a complication with DO_NOT_PERSIST is not persisted, and another to check a complication with DO_NOT_SHOW_WHEN_DEVICE_LOCKED is not shown when locked.
Bug: 246532732
Bug: 236743673
Bug: 160567102
Relnote: We've added ComplicationPersistencePolicy and setCachePolicy to ComplicationData which currently allows a provider to control whether a complication is persisted or not (i.e. whether it's cached past reboot).  Most complications won't need to set cache control, but doing so can fix corner cases with stale data for some complications that update frequently (e.g. health data complications).  We've also added ComplicationDisplayPolicy where DO_NOT_SHOW_WHEN_DEVICE_LOCKED instructs a compatiable watch face to not display the complication when the device is locked.
Change-Id: Ic9574b2c2055b0f4b8436efc913fe467a8a2ebde