Fix missing property accessors for Companion objects
For properties defined in a companion object, we only
expose getters and setters in the enclosing class
when @JvmStatic is present(could be on the property or on
accessors with @get:JvmStatic or @set:JvmStatic). When
@JvmStatic is not present, there's no way to find these
accessors. This CL fixes that.

Now when calling getDeclaredMethods() on the companion object
related property accessors will always be included.

Test: XElementTest::enclosingElementKotlinCompanion()
Bug: b/245792321
Change-Id: I624a1c2423914eb0c41a62a77d1887b6c0b51090