[CameraPipe] Temporarily implement camera release
The CameraInternal#release should be able to wait
for all the output pipeline is finished.
Currently, we didn't have a state of the from
CameraPipe to identify if the camera is closed or
not; so this CL monitors the usage of the output
surface instead of the camera output status.
And complete the release future when all the
Surfaces are inactive.

Bug: 247419208
Test: Parameterized tests for CameraPipe
Change-Id: If6bf154b5d96105d5f001e0ffa06ffeb2bf88ae7