Mitigate incorrect fold data after config changes.
Update logic to use OnConfigurationChangedProvider instead of
ComponentCallbacks.  ComponentCallbacks are called at the Application
level, but happen before the Activity receives configuration changes.
The window bounds do not match the orientation causing empty values.
A mitigation is to add a listener through
OnConfigurationChangedProvider. OnConfigurationChangedProvider provides
a signal when the Activity gets the onConfigurationChanged value.

Note, this only affects R since it is using an early version of Sidecar.
This issue was fixed for SC-V2+ (12L) in b/205342008 and should be easily
backported to S if necessary.

Bug: 230779309
Test: manual - run the samples that handle configuration changes on R.
Test: expect fold to match orientation after rotations.
Change-Id: I66f6ccb24f649d7db044ba6c7a1b6294ef7a8ea2