Add experimental placeholder support to Compose for Wear OS.
Add support for the three different parts of the Compose for Wear OS Placeholder motion design.
1. Add support for a brush to be used to draw and wipe-off a effect that is applied on top of container backgrounds (such as chip and card).
2. Add a Modifier.placeholder() to draw and wipe-off a stadium shaped placeholder shape over the top of content that is waiting to load
3. A Modifier.placeholderShimmer() that draws and wipes off an orchestrated shimmer/gradient effect over the top of components with contents that are waiting to load.

Bug: 226311070
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have added experimental placeholder support. This has three distinct visual effects designed to work together. Firstly a placeholder background brush effect used in containers such as Chip and Cards to drawn over the normal background when waiting for content to load. Secondly a modifier (Modifier.placeholder()) to draw a stadium shaped placeholder widget over the top of content that is being loaded. Thirdly a modifier gradient/shimmer effect (Modifier.placeholderShimmer())that is drawn over the top of the other effects to indicate to users that we are waiting for data to load. All of these effects are designed to be coordinated and shimmer and wipe-off in an orchestrated fashion."
Change-Id: I3c33961fa8bdda098b9a992896d49550185e2b86