Rename SurfaceEffect to SurfaceProcessor
No logic change. Just renaming.

In the public API, we want to use the word "Effect" for a different meaning. Effect should contain two functionalities: 1) driving the camera, and 2) processing the output. A processor, e.g. SurfaceProcessor is for 2) only. We will add the "driving the camera" part later.

In the next CL, we will remove EffectBundle and replace it with a new class.

Bug: 242538531
Test: manual test and ./gradlew bOS
Change-Id: I4fa3ac5e179b0b124582fcdef600b9840426d03e