Fix isFinal() for properties and visibility for lateinit properties
This CL changes the meaning of final properties from
"not open/not overridable" to "immutable" to match kapt and Java final
fields when the backend is KSP.

This CL also changes the visibility of lateinit properties for KSP backend
from private to the setter's visibility because 1) "Late-Initialized
properties are also exposed as fields" and 2) "The visibility of the
field will be the same as the visibility of lateinit property setter"[1].

With these changes, a lateinit property like

```
lateinit var x: String
```

is now considered not final and public.

Note that the visibility fix only works for Kotlin source. We have no way
to tell if a property is lateinit for Kotlin libraries. We'll need a new
KSP release for that: b/250567151

[1] https://kotlinlang.org/docs/java-to-kotlin-interop.html#instance-fields

Bug: b/245934092
Test: XTypeElementTest::lateinitFields(), KspFieldElementTest
Change-Id: Iebebb2d9383972fc5b27bcf18be8481990f32ffe