Add onSessionFinalized to CaptureSessionWrapper
This change allows a newly configured session to "finalize" a
previously created session, and for a CameraDevice instance to
also "finalize" a session that was created as part of close().
This is intended to ensure that all objects are eventually
released, even if the underlying framework call is not invoked.

Bug: 249258992

Test: ./gradlew\
 :camera:camera-camera2-pipe:testDebugUnitTest\
 :camera:camera-camera2-pipe-testing:testDebugUnitTest\
 :camera:camera-camera2-pipe-integration:testDebugUnitTest

Change-Id: I6eb1e9fc9e8c81be26d680d56507b2d1582682ca