Support property reading and writing in Kotlin code generation.
When generating Kotlin code don't use JVM getter and setters for a Kotlin property, instead use the property syntax to get and set the 'field'. This CL only identifies when a setter / getter is a synthetic method for a property and doesn't do any extra work to recognize the backing field is for a property, it also relies on XProcessing and certain Kotlin language aspects.

* XProcessing API for inspecting fields from Kotlin properties only considers properties with backing-fields.
* XProcessing in KSP exposes synthetic getter and setter for properties as part of method inspection.
* Kotlin enforces getter to have same visibility as property, but setter can be different
* Kotlin disallows getter-like and setter-like functions that would conflict with the property's JVM generated ones.

Thus this CL changes the getter and setter codegen logic to be based on the language being emitted and the new CallType that is identified by exposing a `isKotlinPropertyMethod()` in `XMethodElement`.

Bug: 127483380
Test: KotlinCodeGenTest
Change-Id: I64786d7a2026ab1788ad403db62da2e2fcf5cc6a