ObserverNode interface
Modifier.Node implementations that need to notified when a value
that they read changed, can use this interface. It provides
a onObservedReadsChanged function that is called whenever a
value read within an observeReads block changes.

Bug: 247716483
Test: ./gradlew compose:ui:ui:cC -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.node.ObserverNodeTest
Relnote: "Added ObserverNode interface that can be used by Modifier.Node implementations that need to be notified when a value that they had previously read has changed"
Change-Id: I5728be00aad7c3fb9e44240c257f4d45a081da6f