Migrate DaoWriter raw query codegen to XPoet
A small but noticeable change is that we no longer assign the SupportSQLiteQuery param to a `final` local variable since for a while Room has been generating overrides with `final` parameters.

Also fixes the usage of a MemberName in EntityCursorConverterWriter, it was converted in aosp/2234531, but no Kotlin codegen test that used it was written until now.

Bug: 127483380
Test: KotlinCodeGenTest + existing.
Change-Id: Id3b5aea4a095da6fa9996abd82c4cc43bfa1aea3