Reject goals tracking if data types aren't tracked
Rejecting a passive monitoring request for goals the goal data type are
not tracked. Added a validation in PassiveListenerConfig and validated the config before registering the goal. This validation checks if the goal datatype is already set for tracking or not.

Relnote: "Allow passive monitoring tracking for goals only if the same data types are also tracked"
Bug: 254298217
Test: ./gradlew :health:health-services-client:test

Change-Id: Ibed8da24bafb1786c47e72c069e9cec6ac9ee564