Reply on Image dimension and Exif rotation to calculate HAL rotation.
When there is rotation in the HAL, the output image should respect 2 values:

- the width/height of the media.Image, and
- the rotation in Exif.

The former is framework code and the latter is guarded by CTS. Anything else, e.g. the Exif size used by the old pipeline, cannot be trusted.

This CL uses those informations to calculate the crop rect and transformation.

Bug: 249593716
Test: manual test and ./gradlew bOS
Change-Id: I6d6869bd90b5dced12050b15fabae69626aaac5f