Backported setBufferTransform
Updated SurfaceControlV29 to support
setBufferTransform for API Q and above.
Added logic to lazily call setGeometry
in commit calls if there are pending
setBuffer calls in order to infer the
size of the buffer as well as the
destination.

Relnote: "Fixed missing RequiresApi annotation
for addTransactionCommitListener which was
introduced in Android S"

Fixes: 254387178
Test: Added Test to SurfaceControlWrapperTests +
re-ran graphics-core tests

Change-Id: I0a0358dd907c313ba0aa5d284fb360df5fa40d31