ScalingLazyListState.layoutInfo.visibleItemsInfo - return emptyList() until initialization is complete
ScalingLazyColumn has a multiple pass initialization. At the moment interim visibleItemInfo values are visible (and misleading) during the initialization of SLC. There is no way through the public API for the developer to know that the initialization is still ongoing. To correct this we will return an emptyList for visibleItemsInfo until initialization is complete with all  items being visible and having the correct offsets.

Bug: 254440356
Bug: 254455955

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Relnote: "We have corrected a bug in the ScalingLazyListState.layoutInfo.visibleItemsInfo which was reporting incorrect offsets during ScalingLazyColumn initialization. Now an empty list will be returned until all list items are visible and have the correct offsets. Check for ScalingLazyListState.layoutInfo.visibleItemsInfo.isNotEmpty() will confirm that the ScalingLazyColumn initialization is complete and items are visible."
Change-Id: I3a3b883d3347784a013b655a0a1e8312d0914ba2