Adds canScrollForward and canScrollBackward to ScrollableState
Bug: b/255557085
Test: existing tests
Relnote: "Adds ScrollableState#canScrollForward and ScrollableState#canScrollBackward to query whether a ScrollableState has room to scroll in either direction (whether it is at the minimum / maximum of its range). This defaults to true for backwards compatibility with existing implementations of ScrollableState. Consumers can use this to show indication to the user that there is still room to scroll, and this could also be used to avoid dispatching delta to ScrollableStates that have no room to scroll in a given direction, to reduce unnecessary work."
Change-Id: Idf1a066bdd86d2205d4bd5bab121c73059ada054