Update documentation for Modifier.edgeSwipeToDismiss
Modifier.edgeSwipeToDismiss must exist within a SwipeToDismissBox that uses the same SwipeToDismissBoxState. Otherwise, edgeSwipeToDismiss waits indefinitely for the anchors to be provided that would tell it the swipeable limits.

Bug: 256543970
Test: N/A
Relnote: "We have clarified the usage of Modifier.edgeSwipeToDismiss in the documentation - it is a requirement to use edgeSwipeToDismiss within a SwipeToDismissBox, sharing the same SwipeToDismissBoxState."
Change-Id: I4abae19c36cedd8f307365b609dfb1d6e596c953