Temporarily generate RoomOpenHelper.Delegate in its own Java file.
Generating the RoomOpenHelper.Delegate in its own file helps us migrate DatabaseWriter to XPoet so it can generate Kotlin code without having to also migrate SQLiteOpenHelperWriter.

Since the delegate is no longer an inner class two API had to be opened up in RoomDatabase to be called from the delegate. These are temporary and will be removed once the delegate is moved back to an inner class.

Bug: 127483380
Test: Existing
Relnote: Add restricted temporary API in RoomDatabase to be accessed by generated OpenHelper Delegate.
Change-Id: I2f206155ba48a235c580596dcad8bd3179a3ca24