Cache result of getEnabledAccessibilityServiceList
Instead of calling getEnabledAccessibilityServiceList on every call to
the isEnabled property, we cache the result of the
getEnabledAccessibilityServiceList call.
We add listeneris fired whenever AccessibilityManager#isEnabled or
AccessibilityManager#isTouchExplorationEnabled change to recalculate the list of the enabled services. Ideally we'd
want to observe the changes to the list of the enabled services but the
listener is only avaiable in API 33+

One of the deteriorated benchmark results:

* Before aosp/2283214:
p50 3.4, p90 4.2, p95 4.4, p99 5.0

* With aosp/2283214:
p50 4.6, p90 5.3, p95 5.5, p99 6.5

* With this CL improvement:
p50 3.4, p90 4.2, p95 4.4, p99 4.9

Test: ran benchmark to verify that this's an improvement
Test: added some new tests to AnroidAccessibilityTest
Bug: 257995472

Change-Id: Ie2f0d5477e32c49baa609c933858cb1645c4ee1a