Add a #hasCameraTransform API to TransformationInfo
When the Surface is connected to the camera, camera2 set the camera transform on the Surface which will be used by TextureView to transform the result. When the Surface is connected to a OpenGL renderer, the value is absense. PreviewView needs to transform the output differently based on this value.

- Add a #hasCameraTransform API TransformationInfo. The return value indicates that whether the associated Surface has the camera transform embedded in it. It's hidden for now, but we will make it public later.
- In SettableSurface#notifyTransformationInfoUpdate, set this value to match the value of the SettableSurface.
- In Preview & VideoCapture, set this value to true if there is no processing.
- In Preview, do not update the TransformationInfo if the processor exists. Set the rotation on the SettableSurface instead.

Bug: 241910577
Test: manual test and ./gradlew bOS
Change-Id: I4231917d39ef418cbe87de07c212bb4a762eae75