LazyListItemPlacementAnimator improvements
In order to simplify code and make the logic more reliable we changed how we calculate the target offsets for items moving out of the bounds. Not we just use an offset right after the lazy layout bounds instead of trying to calculate a real offset where this item is suppose to be located if we were to compose all the items. As we do not really compose items in between we can't know the offset for sure and there is not much of the benefits in trying to be so precise - from the user's perspective items are moving out of the bounds and that is enough to indicate the intention. We still make sure that two items do not animate to the same target if there are multiple parallel animations - we add such items one by one after the screen bound so they have different target offsets.

Fixes: 235259330
Test: existing and new tests in LazyListAnimateItemPlacementTest
Change-Id: Id3028028bff70a01d8ca0e2a9fc4e54715830fee