Allow modifiers to consume CompositionLocals
Modifier.Nodes can implement the `CompositionLocalConsumerModifierNode`
interface to get access to CompositionLocals at the position of the
attached layout node.

To allow this, the Composer now exposes an opaque list of
CompositionLocals in the currentCompositionLocalMap property.
LayoutNodes store this list and invalidate modifiers that read
CompositionLocals whenever the resolved locals are reassigned.

Reading a CompositionLocal in Modifier.Node happens once, imperatively
at the time of read. We discourage reads of CompositionLocals in the
onAttach and onDetach lifecycle events, both in documentation and via a
lint inspection.

Relnote: """
    Allow Modifier.Nodes to read CompositionLocals by implementing the
    CompositionLocalConsumerModifierNode interface.
    """"
Test: ./gradlew compose:ui:ui:connectedAndroidTest
Change-Id: Ib44df147ceaad520c9102c416440d20fadadc403