Update Placeholder to WipeOff immediately that data become ready.
Update the experimental placeholder UI to initial wipeoff of placeholder effects as soon as the content is ready to display.
Also make some other adjustments to the shimmer and wipe off to match interim specification including applying an alpha curve and changing the speed and coordinates of the wipe off and shimmer effects.

Bug: 253667329
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Change-Id: I5a7f45c0ce6f2262716de3c2c8131de663910d7b