Convert DialogFragmentNavigator to Navigation v2 APIs
Changing DialogFragmentNavigator to use push/popWithTransition to adjust
the Lifecycle state of the NavBackStackEntry in relation to the
DialogFragment.

Now the entry will be moved to RESUMED with the DialogFragment moves to
RESUMED, and it will not be DESTROYED until the DialogFragment is
DESTROYED.

This also adds an onLaunchSingleTopWithTransitions API that adds both the
entering and exiting NavBackStackEntries to the list of
transitionsInProgress.

RelNote: "`DialogFragment` destination `NavBackStackEntry`s will no
longer be `DESTROYED` before the fragment itself is `DESTROYED`."
Bug: 261213893
Test: DialogFragmentNavigatorTest

Change-Id: I53ee5ff20ea8dd04dfffbce10015504a9bdca135