Remove minSdkVersion restriction from test
This test was previously flawed as it relied on screen density information when deciding the amount of content padding to be applied as it applied padding in DP which could actually be larger than the actual SLC viewport size resulting in no actual list items being visible. This was fixed for PositionIndicatorTest.kt as a by product of aosp/2318789.

Bug: 260235088
Test: /gradlew :wear:compose:compose-material:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.wear.compose.material.PositionIndicatorTest
Change-Id: I0d6542ce5de42c4a6f10a83fa134e6766d6a55f2