Introduce SplitRule#maxAspectRatioInPortrait/Landscape
The new properties are to check whether the aspect ratio of the parent
bounds satisfy the requirments. When the aspect ratio is greater than
the maxAspectRatio, the split pair of activities will be stacked.

RelNote: "
SplitRule now takes maxAspectRatioInPortrait/Landscape. It only allows
activities split when the aspect ratio of the parent bounds is smaller
or equal to the requested maxAspectRatio.
"
Bug: 242970511
Test: ./gradlew window:window:cAT

Change-Id: Ia5990b64c87987a9328c228753e792bd293e139a