docs: Adding proper summary sentence fragment, and other minor changes
Per bug, the "e.g." was confusing the doc engine, which thought it was
a sentence-end. In any event, we really want a *short* sentence fragment
for the various summary tables. So, added that fragment, and also made
very minor punctuation changes to (hopefully) make that long sentence
clearer.

Staged to:

* go/dac-stage/reference/androidx/wear/watchface/ListenableCanvasRenderer2

Change-Id: I5159150d78d70cc3abd5051afcae6b68f9ee8aac
BUG: 231768798
TEST: [docs build]