[CameraPipe] Port FocusMeteringControlTest from camera-camera2
These tests were originally designed to test camera-camera2 FocusMeteringControl logics. Porting them for camera-pipe will help in ensuring all the logics. Not all of the tests will be applicable and some new tests might need to be added due to the coding differences in the two modules, this CL is for porting as much as possible for now and getting the general idea on what else may be remaining.

Bug: b/254790453
Test: camera-camera2-pipe-integration / FocusMeteringControlTest
Change-Id: I25b5fdee4ca41c7e490dd51df88b512a42fc6535