[GH] Remove mismatched androidTest dependency already provided by implementation
Fixes a "resolution ambiguity" error with lifecycle-runtime-ktx which  is broken on playground as androidTest resolves a different version of the artifact than the snapshot, but on AOSP both resolve to the project successfully, which works around the problem.

Test: cd appcompat && ./gradlew appcompat:appcompat:compileDebugAndroidTestKotlin
Change-Id: I3c4b8cc388d8a6c38c38f30ae5bbd955f613a755

This is an imported pull request from https://github.com/androidx/androidx/pull/479.

Resolves #479
Github-Pr-Head-Sha: 8f3b4e49cb69a00fb727f78a48b7734c893444df
GitOrigin-RevId: 32ca91d3311bf17af6d14a4b3487fd6914b4786b
Change-Id: Ic714c7704b486488e84823e27c95faa30df83f39