Allow resetting state of Modifier.Node when Layout is reused
Added a new `Modifier.Node.onReset()` callback to be called when the node is about to be moved to a pool of layouts ready to be reused. For example it happens when the node is part of the item of `LazyColumn` after this item is scrolled out of the viewport. This means this node could be in future reused for a `Layout` displaying a semantically different content when the list will be populating a new item.
Use this callback to reset some local item specific state. As part of this change updated `FocusTargetModifierNode` to properly reset its focused state.

Relnote: Added a new `Modifier.Node.onReset()` callback allowing you to reset some local state to properly handle the case when the `Layout` will be reused (for example as an item of `LazyColumn`). Fixed `FocusTargetModifierNode` to properly reset the focused state.
Test: ModifierStateResetTest
Fixes: 265201972
Change-Id: I65495712cdeafeb3fedc3e4272627c024f952ddb