Add a "hasCameraTransform" property to UseCase
Camera2 writes camera transform(orientation, mirroring) to the Surface. The transform is used by UI or OpenGL to correct the output. If a UseCase is connected to camera2, it always has the transform.

However, when the UseCase is connected to StreamSharing, the transform matrix is lost during the buffer copy. We need to pass that information to the UseCase so that it can handle the transformation differently.

Bug: 265818567
Test: manual test and ./gradlew bOS
Change-Id: I9cf21adb8e330d3e98a6937701457524a7075574