Fix error with 19 letter long args not wrapping
When an argument is exactly 19 letters long, it hits the end of the line
and safeargs improperly wraps the map function causing it to fail.

We should add `·` to emit a space that never wraps.

RelNote: "Navigation SafeArgs will no longer cause a compilation error
when using arguments that are exactly 19 characters long."
Test: Modified KotlinNavWriterTest
Bug: 257110095

Change-Id: Id60bcbd017753261a97b002d76960e1e2b1d18ed