Add animateScrollToOption method to Picker
This method allows Picker to be scrolled to specific option with animation. A smooth scroll always happens to the closest item if picker is repeatable.

A usage of this method was added into DatePicker - so that the date will be animated from 31 to 30 or 28 when month is changing

Bug: 249957779
Test: wear.compose.material.PickerTest.kt
Relnote: "Add animateScrollToOption method to Picker"

Change-Id: I6fe679e1c9d168f1da542467ee8041a6ac0a7586