Support lookahead in LazyList
This change creates an optimized behavior for lookahead pass and
post-lookahead pass in LazyList. More specifically:
- Items are retained until they are no longer needed for
 either pass
- Lookahead pass will maintain the source of truth for scroll
position.  The scroll position in post-lookahead pass will only
be temporarily different than lookahead pass when there's
a scrollback needed to not reveal beyond the bottom of the list.
In that case, we'll apply an animation to animate the delta
back to 0.
- In the rare cases where lookahead pass failed to subcompose
all slots needed for postLookahead, they will be composed
in post-lookahead. These nodes will have the same lookahead
constraints and placement as post-lookahead.
- Item placement animations are also updated to ensure lookahead
pass skips the animations.

Test: New tests added

Change-Id: I7cd715c3d71d498f76b8abbe5be9ddad48717419