Improve test flakiness.
The UI doesn't always know when the RemoteViews have fully settled when
running the Integration Tests. This change adds some timeouts and retries
to make this better, but doesn't fully fix it.

Adds more verbose logging for tracking down the failures when run on the
test runners.

Renames the loading views in the tests so we can see when they are being
loaded in the logging.

Bug: 267494219
Test: GlanceAppWidgetReceiverTest
Change-Id: Id87a662306b42ac35ea617ea6dce35b393eee504