Add Lint check for margins in Chains
Adds a check for when users have chains and are trying to apply margins
to its element by creating equivalent constraints with a margin in it.

The check lets the user know the appropriate way to apply margins to
chains and proposes a fix.

Relnote: "Added Lint check for when chains are used in conjunction with
equivalent constraints that try to apply a margin. It will suggest to
instead use `LayoutReference.withChainParams()`."

Bug: 268213648
Test: manual
Change-Id: I62eb07e51c36d762aaeff3bfa0031fcfbf6e3923