Supporting value classes in Room for KSP.
Value classes are now supported for KSP with Kotlin codegen in Room.

In KAPT stubs and in bytecode the places where the value class is used is replaced by the underlying type. This is the reason why we can't tell in KAPT or javac if the type is a value class without looking at the metadata. We have filed b/273592453 to further investigate this and figure out a solution for handling this case.

We are planning to follow-up and add a check such that, if user is on KAPT, straight out error, if user is on KSP gen Java, hint to @JvmInline and @JvmName requirements (if we can conclusively say they work this way) to use value classes with Java codegen and point to full-support with Kotlin codegen.

Bug: 124624218
Bug: 272820290
Test: TypeAdapterStoreTest.kt
Change-Id: I0a380ada951df776ed7bcb96757b86103743b0c5