JS evaluation sends data via pipe and not Binder
This CL uses pipes to send data to the sandboxed process instead of
relying on Binder. The JS scripts can be large and as a result, there
was a risk of overflowing the Binder transaction buffer leading to
TransactionTooLargeException. Even though the buffer has a maximum size
of 1MB, this is shared with the entire process and the actual limit is
dependent on runtime parameters.

A new flag has been added for clients to check if this new behaviour is
available in the underlying WebView. Currently, the flag has a ":DEV"
suffix. This will be removed once development is complete.

Relnote: "Script evaluation is no longer bound by Binder limits"
Test: ./gradlew :javascriptengine:javascriptengine:connectedAndroidTest
Bug: 261440543
Change-Id: I28d1ad66d60cda2facdebdab109aac638df3f43c