Preview distorted when using View display
In PreviewView, we recalculate the transformation when 1) layout is changed or 2) display is changed. However it seems that on certain devices, Display#getRotation returns the previous value at the time DisplayRotationListener#onDisplayChanged is invoked, which leads to the incorrect result. A mitigation would be always setting the current display rotation before calculating the transformation.

Change:
- always set display rotation before redrawing. So if DisplayRotationListener#onDisplayChanged returns the wrong display, OnLayoutChangeListener#onLayoutChange can fix it.

Relnote: Fixed a PreviewView bug that if the app handles screen orientation itself. the preview might become distorted upon rotation. This only happens on certain devices. It can be reproduced on Pixel a4, Pixel 5, and Pixel 6a.

Bug: 223488673
Test: manual test and ./gradlew bOS
Change-Id: I14340cc9790f22c70717217f8f1d9fdfe338dfeb