Support custom reverse duration
For repeated animation with repeat mode as reverse, a custom reverse
duration can be specified in Repeatable, so that the duration for the reverse part of the
animation can be  different from its forward part.

This change also moves duration, delay and easing into
AnimationParameters class; for repeared aniamtion, the
AnimationParameters can be override for its repeated forward and reverse
passes separately.

Current implementation ignores the duration and easing in the
forwardParametersOverride, will do in followup CLs, filed bug
b/274452709 to track.

Test: test case added
Bug: 265791537
Relnote: Support custom duration for reverse animation
Change-Id: I3251f683678b064261d06b15c9f878f70b472657