Rename cameraSessionListener_receivesClose_afterUnbindAll test
(1) The SessionStateCallback of CameraX may redirect the call from Camera2
framework, however, there are known issues that
CameraCaptureSession.StateCallback#onClosed may not always be called.

(2) The original test `cameraSessionListener_receivesClose_afterUnbindAll`
doesn't verify the onClosed state, modify the name of the test.

(3) The verification to the `unbindAllCalled` flag may not be set
before the onReady callback is invoked, which may cause the test to be flaky.

Modify the name of the test, and not verify the `unbindAllCalled` flag.

Bug: 269523374
Test: Camera2InteropIntegrationTest on Nexus 4 API 21
Change-Id: I7f030b02776484f3cec88249b30815e8ecdbf075