Replace GridItem.Builder#setBadge() with overloaded setImage() methods.
The method setBadge() does not make it clear that the
badge is associated with the GridItem image. To make this relationship
more obvious, add the badge as a parameter to overloaded setImage()
methods.

This change has the benefit of guaranteeing that a badge will only be set when an image is also set. The build() method will still check that the badge isn't set without an image to protect against future changes to the API.

Test: unit tests
Bug: 272320122
Relnote: "Replaced GridItem.Builder#setBadge() with overloaded setImage()
methods"

Change-Id: Id200086be38764876c358e3d1eb0bea42fb1c328