Don't remove a group from insertTable when copying it to the new group during applyChanges
The insertTable will be disposed and recreated after changes applied.

Motivation: removing a group requires moving the gap twice that implies copying the parts of the array. It seems that it's not necessary since the insert table can be cleaned up later all at once.

This change helps to run applyChanges faster in cases when we insert a lot of new items (~2-3 times improvement).

Test: ./gradlew :compose:runtime:runtime:testDebugUnitTest
Change-Id: Ia8e850d98a777822fc34f0159fbdcf0879c4ae22