[TimePicker] Change zIndex of checked item toggle
Because of the way talkback sees a clickable element, if two elements use minimumInteractiveSize modifier but they overlap, talkback only sees the one on top as respecting the minimum touch target size. Since only the one that's not selected is clickable, changing the zIndex makes it so that the clickable one is on top therefore with the right bounding box.

Test: Tested with a11y scanner with PM and AM checked/unchecked
Fixes: 267505053

Change-Id: I9e6a0f2399a808db79b3d51d742b7e018f01543d