Smooth Rotary Encoder Scrolling for Low Resolution Devices
Low resolution rotary encoders produce larger motion events. With the current mechanism of doing a regular scroll in reaction to each motion event, these encoders produce a choppy experience, as each motion event causes a large jump in pixels on the screen.

To mitigate this, we are now using smooth scrolling for such low resolution encoders. This causes the scrolls to be smooth/animated, instead of being an instant jump.

Note that, we are not changing the rotary encoder scroll mechanism to smooth scrolling for all rotary encoders. This is because smooth scrolling does not suit encoders with high resolution (i.e produce more frequent, smaller motion events). This is because the accumulated smooth scrolls rapidly increase in those situations, causing the visual scroll to be very laggy and inconsistent.

Bug: 275621219
Test: unit tests added, manually verified behavior
Change-Id: I5ecfbdb85c51192a09593cf73227023e6049bdc3