disable dependency constraints for KMP projects
This CL disables dependency constraints for KMP projects (including
android targets).

KMP plugin has a bug where it tries to resolve constraints as if
they are regular dependencies, triggering a stack-overflow.

It will be fixed in 1.9, hence we need to disable constraints in
KMP projects until then.

Note that this change DOES NOT change the Project.shouldAddGroupConstraints
property since it is used to validate we don't ship projects w/o
constraints. Instead, it just skips adding them to the KMP projects.

We should do a followup to limit the impact into just native targets.
I couldn't figure out how to do it.

Bug: 274786186
Bug: https://youtrack.jetbrains.com/issue/KT-57531/
Test: local
Change-Id: Ief77146c133fe203bcdf58e065de84cefe6dadf6