Update expandable items and demos.
Remove "Show Less" buttons, as they are not part of UX spec, replace
custom chevron with material icon.
Create new expandableButton extension to ScalingLazyListScope to
simplify creating a button that collapses when the content is expanded.

Test: N/A
RelNote: "Update expandables examples. Added expandableButton extension
to ScalingLazyListScope to simplify creating a button that collapses
when the content is expanded."

Change-Id: Iae3091470e3f1b531757172e838406fe3e075150