Update UnrememberedStateDetector lint inspection
Currently, we have a lint inspection to ensure that mutableStateOf()
calls are `remember`-ed when invoked from a Composable function, but
this won't work outside of the compose Runtime package. To open the
inspection more broadly, this CL updates the existing
`UnrememberedStateDetector` inspection to be annotation-based. Users
can now annotate functions with a new `@StateFactory` annotation, which
mandates the invocation be `remember`-ed when inside a composable.

Relnote: "Added a `@StateFactoryMarker` annotation, to mark functions
that create states and must be wrapped in a `remember` call."

Test: Existing and new tests in UnrememberedStateDetectorTest
Change-Id: I64550a0cb1647096260f00b75e7d35897878c617