Move filter from property on TextFieldState to param on BasicTextField2.
Filtering is not really associated with anything else in state.
Programmatic state changes do not go through the filter. Filtering is
only a property of processing user input (keyboard, pasting/dropping, a11y
services, tests), which is done by the text field composable.

Furthermore, it is valid text fields consuming the same TextFieldState to be
configured with different filters, since filters are only ran on
_input_, not _output_ (rendering).

Test: EditProcessorTest
Test: BasicTextField2Test
Relnote: n/a
Change-Id: Ib6b348d19119b456d853dfe67364f5b6201b5c1c