Fix AnimationHandler so that AnimatorTestRule can isolate tested animations.
At some point the AnimationCallbackData was changed to static, which meant that when AnimationHandler.setTestHandler() was called, it didn't actually use an isolated list of AnimationFrameCallbacks.  This meant that animations registered but not completed in one test would end up executing code in future tests. If test cleanup led to state the animation callbacks didn't expect (e.g. a null), this could cause the later test to crash.

Fixes: 275602127
Test: AnimatorTestRuleIsolationTest
Change-Id: I1c5a9bef21cd36db53fad6a4276941b307be095e