Remove open val usage in APIs.
Relying on overridden field during construction time in the base class leads to unexpected behavior. The fields may act like they have not yet been initialized. The kotlin documentation also discourage the usage of `open` members in the constructor / init blocks:
https://kotlinlang.org/docs/inheritance.html#derived-class-initialization-order

The reason we used `open val` was because we want to hide it everywhere
and only have one specific subcalss (custom type) to override the
visibility by overriding the property. In this change, we adopt a new
approach where we just hide the property everywhere and on that custom
typed subclass, we expose seperate getters for these values.

Bug: 275950646
Test: unit tested
Relnote: "Remove Credential Manager API open val usages"

Change-Id: I0c8c5da07b568a112b49adda1dbafbc12b9721f3