Run all jvmTests of datastore-core on android
They were only being run on the jvm target of android
tests (android stubs) but we would really want them
to also run on actual device. Note that i did not remove
the android jvm target of them since it is a common
case for users to use for testing, hence it is good to
keep the support for it.

Bug: n/a
Test: CI
Change-Id: I77791dcddeed209478e9cfdebd1974b61f74d032