Allow capping the size of bound data types
This allows controlling the total number of bound data types in the pipeline.

Control granularity is on expression level. If an entire expression does not fit in the remaining quota, the evaluator will throw an exception.

This also wraps all the different arguments to bind methods into a DynamicTypeBindingRequest type. It allows the called to retry binding the request later if it fails.

RelNote: Allow capping the size of bound data types.
Bug: b/204318149
Test: Unit tests added.
Change-Id: Ie29669c5e3fd935de8cad63b3a61d79db582d68f