Fixed current profile can write check to work on Android U
Due to a behavior change on Android U the current profile file is not
created anymore, therefore `File#canWrite` returns always false. To
understand if the file is writeable, we need to try to create an empty
file.

Test: ./gradlew :profileinstaller:integration-tests:profile-verification:test
Bug: 277638667
Change-Id: Ie3899ddf7fd38be4d91fb0281ea4ead9f9f8288e
Relnote: Fixed profile installer on Android U failing due to current profile not created empty when process starts.