Small fixes to transitions
Bug: 259288155

Listeners collections were shared between the transition and the
one being run. Transitions often add listeners to themselves or
to the root. When this happens, it must remove the transition when
it is finished, which developers often forget to do. It is also
impossible to know for sure when a transition is finished with
a listener when seeking. Also, because the listener collection
is shared, if the same transition is used multiple times, the
listeners will be called by an irrelevant instance, so the
transition must compare the parameter with the expected one.

Instead, the collection is copied when the transition starts so
that a listener added will no longer be added to the one that
it was cloned from.

Also, pause and resume listeners are called prior to the
transition starting. However, seeking was causing the pause
and resume to stop being called in some cases. This fixes
that problem.

Test: new tests
Change-Id: I48fa106c4dd1f0b6ff4a2e76797a1dafe3db7822