Add an experimental extension layout element
This can be used by any library that integrates with ProtoLayout to
add support for their use case specific layout elements. They need to
provide a renderer extension which ProtoLayout renderer will use
whenever it tries to inflate an `ExtensionLayoutElement`.

Relnote: "We've added an experimental extension layout element. Note that this can't be used by default and requires a renderer extension capable of understanding the layout element."
Bug:278503614
Change-Id: I6581dae6a6570c761f53ddd7f5634aa59beb64ef