Add proguard keep rules to keep Quirk classes
"quirk.getClass() == quirkClass" is used to get or check whether a Quirk class exists. If those classes become the same class due to merging, the result of the reflective getClass() call will be the same for them. This falls into the same category as more general reflection needing keep rules. For more information, please see b/278818703.

Bug: 276483197
Test: ./gradlew bOS
Change-Id: I7e3985a45a9a21020e5f538d01bba9c1d12bb4ea