Fix Transition tests for API 18-
Fixes: 279455368

A few problems were discovered by tests on API 18 and below.
First, one of the AnimatedListenerAdapters was added that
implemented AnimatorPauseListenerCompat, but not onAnimationPause()
or onAnimationResume(). Second, some AnimatedListenerAdapters that
did implement onAnimationPause()/onAnimationResume() did NOT
implement AnimatorPauseListenerCompat. Third, the prior
alpha was not being saved during onDisappear() for the Fade
transition, so the alpha was being lost.

Test: ran existing tests on KitKat
Change-Id: I0de325708044c0d708514a2c2138970bae1a3055