Expose isImageAnalysisSupported API to ExtensionsManager
ImageAnalysis support is not mandatory when Extensions are
enabled. Developer needs a way to query if ImageAnalysis
is supported or not before binding the ImageAnalysis.

Binding ImageAnalysis when ImageAnalysis is not supported
will throw an exception.

Relnote: "Added an isImageAnalysisSupported API to the
ExtensionsManager which apps can determine if an
ImageAnalysis use case can be bound along with Preview
and ImageCapture when Extensions are enabled."
Bug: 267728014
Test: NA

Change-Id: I1bc6361c6fcb32f4ece0b616ebac1059b1ecfe76