Expand asPagingSourceFactory API to singe list
In addition to Flow<List<Value>>.asPagingSourceFactory API, we added a List<Value>.asPagingSourceFactory. All PagingSoruces produced by this new API will load from the exact same static list of data. It supports multi-generational operations such as refresh, but it does not support any updates to the source data. Because it operates on a static list, collection is not required which removes the requirement for a coroutineScope.

Test: ./gradlew paging:paging-testing:test
Bug: 280655188
Relnote: "Added new API List<Value>.asPagingSourceFactory to get a factory that creates PagingSource instances that loads from the immutable list of data."
Change-Id: Id34d1361119413d0c19a054a2e23dd26241faa6f