Fix a newly surfaced flake in TextFieldTest
In a previous CL aosp/2493878, we have attempted to fix another flake in BasicTextField tests and added an extra checkpoint to catch any other flakes that could surface. This actually happened and told us that probably text delete action isn't working as we thought it would. Now, we are replacing all the previous semantics actions with the new ones introduced recently.

Fix: b/279809276
Test: TextFieldTest
Change-Id: I5e04e4f51052180673e47f957040358380019e2a