Fix Java TypeName for @JvmInline types
This CL fixes an issue with our [TypeName] calculation for types that
are annotated with `@JvmInline`. These types have interesting rules
for when they're actually replaced with the actual type or the
`@JvmInline` type, which I've updated [KSTypeJavaPoetExt] to handle.

In addition, some of the test cases I added to [KspTypeNamesGoldenTest]
caused a breakage in [KspClassFileUtility] when ordering the
methods/fields. However, KSP has since fixed the initial issue
(https://github.com/google/ksp/issues/250), so I went ahead and updated
the code to use the new `Resolver#getDeclarationsInSourceOrder()`
method, which fixed the breakages.

Test: KspTypeNamesGoldenTest.kt
Change-Id: I9ad48edda0fbb518b89be116df2999545d68f33b