Do not rerun placement block of layout modifiers when parent moves children
Currently we run all the user provided layout modifiers even when the node is not marked dirty. For example it happens on every scroll of lazy lists. We can skip this work in most of the conditions.
A few other small optimizations are added as part of this cl.

This cl improves LazyListScrollingBenchmark.scrollProgrammatically_noNewItems() by ≈ 11 percent from 1,643,646 ns to 1,477,123 ns.

Test: DrawReorderingTest, GraphicsLayerTest and PlacementLayoutCoordinatesTest
Change-Id: I6b849ea0ddc00110446998ee68023a9eeab0bdeb