Improved error messaging of painterResource
Updated painterResource implementation to provide
more descriptive error messages and include the
name of the asset that is attempted to be loaded.

Added explicit messaging of webp support.

Relnote: "Add ResourceResolutionException type
to wrap throwables thrown when attempting to load
bitmap assets with a description of the asset path
that failed to load."

Fixes: 230166331
Fixes: 278424788
Test: Added test to verify webp image loading
Change-Id: I19f445605ae8b171532b73b7ae4697cec0453767