Do not call state::confirmValueChange on custom drags in AnchoredDraggable.
Since primary idea behind confirmValuechange is to veto the changes influcted by the settling, asking for veto during programmatic drags via state::anchoredDrag is unnecessary. In fact, some usages rely on the precise timing of the confirmValueChange - that is, being called right after the drag stopped and the components start to settle.

This change brings back the desirable behaviour.

Test: new regression tests added
Bug: 282175220
Change-Id: I46c60515e699fc434a87fb21a719c7ab151dca19