Clear FragmentStore savedState on resume
Currently savedState survives resume and the stale state gets stored. When resumed, the stale state is restored.

Now we ensure stale states are cleared upon resume.

Test: ./graldew fragment:fragment:cC
Bug: 246289075
Relnote: "Fixed an issue where the saved state stored when the activity was stopped but not destroyed would be incorrectly cached even after the fragment instance was moved back to the RESUMED state. This would cause that cached state to be reused if that fragment instance was on the back stack when using the multiple back stacks API to save and restore that fragment."
Change-Id: I712884633ae73e0784ba0a7fe50bb5a1046e2275