Add experimental API for enabling Preditive BAck
Provide an experimental API for controller whether FragmentManager
should use Predictive Back.

Currently the feature defaults to on, and in the event there is a
regression can be switched off to avoid being blocked while the
regression is addressed. The API will only exists to verify a regression
is caused by the new feature.

Test: tested in sample app
Bug: 258863498
RelNote: "Added a new *experimental* API for controller whether the
FragmentManager uses the new predictive back feature for handling
gesture back."

Change-Id: Iabe27ec91e8e48553dbd3a51bbfb4b6dc30a11c2