Extended build type will now copy base build type sources
This PR ensures that the extended build types created in the app target
module for baseline profile gradle plugin, copy also the sources of the
base build type. For example, `nonMinifiedRelease` should use all the
sources of `release`. In order to test this, Hilt has been added to
baselineprofile-consumer integration test.

Bug: 282787832
Test: ./gradlew :benchmark:benchmark-baseline-profile-gradle-plugin:test :benchmark:integration-tests:baselineprofile-library-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-flavor-consumer:testBaselineProfileGeneration :benchmark:integration-tests:baselineprofile-consumer:testBaselineProfileGeneration
Change-Id: Ib4d56a85daae71f2330acc839a04f90a3dd57a38