Set Window#setDecorFitsSystemWindows on Android R+ and View#setFitsSystemWindows on Android Q- instead of using WindowCompat#setDecorFitsSystemWindows
On Android Q, WindowCompat#setDecorFitsSystemWindows sets the flag SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION which causes the app to behave as if the system bars are hidden (when they aren't). This stops system bars from responding to touch events. This CL fixes this issue by calling setDecorFitsSystemWindows for Android R+ and setFitsSystemWindows on Android Q- instead of executing the WindowCompat implementation.

Test: Unit and manual
Bug: 283985939
Change-Id: I8517c4044306af424abf9e799e695db3450e07fc