Add truncation to ListTemplate
Impose a 100 item limit on ListTemplate as a stop-gap solution for
preventing the client from overflowing the 1MB binder limitation. This
is a "good enough" solution for the time being while a longer term
solution is planned and implemented.

Fixes: 279995622
Test: ./gradlew :car:app:app:test
Relnote: "Introduces an automatic 100 item limit to ListTemplate to
avoid overflowing the 1MB binder limitation"

Change-Id: I017f69cf3449cf774ca40ea0670a17bdd2f9da01