Fix issue with ClassName when simple name contains '$'
As part of the logic to convert a type to a `ClassName` we first get the
jvm descriptor and convert it to a `ClassName`. However, this doesn't
work when the simple name contains a '$' since the jvm descriptor will
use '$' as the marker for separating simple names and we'll end up
splitting the simple name in incorrect places.

This CL fixes this issue by only applying the transform when it's
absolutely necessary.

Test: XTypeElementTest.kt
Change-Id: I7bbbae1881046823d07955f825ac241e02bbbe00