Force withJava for KMP projects without Android targets
API files are only generated if the project applies android or java plugins.
It is possible to have a KMP project that needs neither of those, which
subsequently disables API / lint checks in the project.

This CL adds a quick-fix to force java plugin to KMP projects if they don't
declare an Android target. Ideally, the real fix would be to handle the case
properly in a followup CL.

As the API / lint were not running, i also updated those files for KMP projects.

Relnote: n/a
Bug: 286852059
Test: ./gradlew updateApi
Change-Id: I3fb03fca186a5341fc3dee0c5d0d0e0b2a8c5b96