Support for micro metrics to report multiple values
Test: MetricsContainerTest
Test: TrivialKotlinBenchmark
Bug: 286306579

Counters should start/stop together, but report multiple values, so
changed MetricsCapture/MetricsContainer to support this, by writing to
an output array (and transposing the storage in MetricsContainer).

Initially, this CL would have used a single LongArray(size =
measureCount * metricCount), but this caused warmup measruments of an
empty measureRepeated block to go from 75ns to 400+ns on a Bramble
running API 30 in BenchmarkStateConfigTest.

Trivial benchmarks seem to show no noticable increase, but it's
possible this causes a small regression in extremely fast/trivial
benchmarks. If it does turn out to be impactful, we could have a
special optimized path for time capture (capture at end of start,
don't pass to MetricCapture, and handle pausing directly within
MetricsContainer).

Change-Id: Ifdf52098ccb8405a012166648c5bbd9ba3cf1f21