Add documentation for creating icons with resource
Tried adding active validation but decided not to due to
1. The type of icon is only available >= api 28 so cannot
be validated before that anyway
2. GPM is setting calling app icon in create flows for which they
don't have resource. They have scaled the icon down. May amplify this
documentation with a recommeneded size but there doesn;t seem to be
a way to get the size of the icon programmetically either

Bug: 269501289
Test: N/A - documentation change only

Change-Id: Idae7903bf57b21090a5f7cc1ccc6bfb4719ad559