Fix flicker in AutoTransition
Fixes: 284196136

When seeking an AutoTransition where a View is faded
in and another is faded out, if the seeking position
is at the start, neither View has had any fading done.
If a new transition is started in which the Views
are restored to their original fade value, the wrong
value for the transition alpha was being used, causing
the views to think they were already faded out/in by
the seek. This CL fixes the problem.

Test: new tests, manual tests

Change-Id: I16dc71c336a4553ceadda4991d2f4e2f529f9f4a