Make origin private and have a separate getter that takes in allowlsit
this is to make sure that providers are consciously trusting select
calling packages for populating the origin, before returning credentials
for this origin (which is supposed to be different than the calling
package)..

Providers must pass in a list of accepatable apps and the custom getter
will validate that the calling package (and signatures) from the
incoming request are present in the allowlist, and onky then return the
origin..

This version does not have the matching logic in yet as the JSON
schema has to be defined first. Providers will then need to pass in a
JSON that follows that schema.

Bug:289402940
Test: Unit tests
Relnote: "Expose a custom origin getter that takes in allowlist"

Change-Id: I0c1b48411a1b0ec8480e8a87e4ae16507d5b89bd