diagnose-build-failure now looking for the cleanest state that reproduces the failure message
rather than the dirtiest state that doesn't reproduce it.

Gradle saves the task-up-to-datedness metadata of most tasks in one file, and if previously if diagnose-build-failure were unlucky, it could report that that the dirtiest state that doesn't reproduce the failure could use all of the dirty files except for the task state file which needed to be clean. This is a true statement but not very helpful: it just says that the build failure can be avoided by deleting Gradle's memory of having previously run the tasks, which is equivalent to rerunning all tasks.

It's more helpful if diagnose-build-failure reports all of the dirty files that if made clean will prevent the failure from reproducing. This list of files should be essentially all of the interesting files rather than just one interesting file chosen arbitrarily.

Previously we were looking for successful builds because we didn't have a good way to distinguish builds that fail with the expected error vs builds that fail with an unrelated error. Now that build_log_simplifier.py can identify which lines of output are unexpected in a passing build, diagnose-build-failure.sh uses that to identify the specific error to and focus the search to states that reproduce a similar message.

Bug: 187822154

Test: Create an incremental build failure, for example: \

  ./busytown/impl/build.sh --diagnose --no-daemon :compose:ui:ui:compileTestKotlinDesktop \
  sed 's/\( *\)\(private var compositionToken.*\)/\1\2\n\1private var sample1: Int = 0/' -i ./compose/runtime/runtime/src/commonMain/kotlin/androidx/compose/runtime/Composer.kt \

 and ask build.sh to build and diagnose it:

  ./busytown/impl/build.sh --diagnose --no-daemon :compose:ui:ui:compileTestKotlinDesktop

 and find that the diagnosis lists lots of generated files that contribute to the repro rather than just out/gradle-project-cache/8.0/executionHistory/executionHistory.bin

Change-Id: Id47f5a3c58c52bdbe57d7bb93cbfcb1cf2e65086