Tests for Framework2Jetpack DomException Converts
This adds tests that validates the framework 2 jetpack createFrom
methods generated in the last change list. This test validates that the
flow works for one type of DomException. Further parameterized tests can
be added, but considering that these objects are type gaurded, the main
flow can be confirmed with these tests.

Bug: 290269134
Test: These are JUnitTests
Change-Id: Iaa2afb47068e11e0fa636960a0d0545c2a58c549