Create DifferentialMotionFlingHelper
This class will control fling algorithm for differential motion. It is
designed to be independent of the View it may be used on. Currently,
it's situated under recyclerview/, and is being used by RecyclerView
only. Once we fully implement its details and make RecyclerView use it
to generate real flings, we can finalize the API by pulling it out of
recyclerview/ module (and putting it in some common module like view/),
and making it public, so that other Jetpack widgets can use it. Once the
API has been finalized/made public, we can also enable developers to
inject their own implementation of DifferentialMotionFlingHelper to
widgets to allow custom fling behaviors.

This implementation is currently empty - it just defines the structure
of the class, and makes RecyclerView use it. The algorithm to actually
control fling will be implemented in a follow up, along with unit tests.

Bug: 290680625
Test: N/A
Change-Id: I2c0473af5618fd5a489ead341269021566ce59bf