Update how key is recreated.
Since pageCount may change dynamically without a recomposition, we'll need to update how the key parameter is used and updated in Pager's lazy layout configuration. The dataset that is used in pageCount may change without recomposition which may not be reflected in the key lambda if we use the same dataset there, we need to make sure to update it as well so we always have the latest version of the key lambda.

Relnote: N/A
Fixes: 289715050
Test: Added test to verify behavior.
Change-Id: Ifaab7b71f0f321cd2cbda2799a1885fbfb66d46a