Split test for ranged value drawing.
This commit splits the test for ranged value and drawing of an arc
into multiple tests with different inputs instead of using a for-loop
to iterate over the values.

This means that each test is testing a single scenario and it is more
obvious if one of the tests fails.

Moreover, this removed a call to the setUp() method within the test
body, which causes issues when trying to move away from openMocks(),
which is deprecated.

Test: ./gradlew :wear:watchface:watchface-complications-rendering:testDebug
Bug: 251211092
Change-Id: I02d0aabbb62ab126ee9bc6ff2583e457dc95e159