Add more mime types for webkit
WebKit relies on the OS provided MIME types
but this is pretty lacking so adding all the other
mime types that Chromium has.

Matching the behaviour we are doing in Chromium
which is to fallback to our hardcoded mime list
if we didn't get anything from the operating
system.

We will unfortunately have to manually maintain
this list. The good news is this only puts us
in a better position than it was before.

Bug: crbug/1019528
Test: gradlew :webkit:integration-tests:instrumentation:connectedAndroidTest
Change-Id: Ie5a878b46d868011feadde437ae7ca66ccd64169