Wait stream to idle after Preview unbind in UseCaseCombinationTest
Wait 1 sec before verifying the preview output.
The test flow is unbind Preview -> wait for preview output to stop
It originally only allowed less than 10 frames preview output.

The Preview may take time to post process the detach flow; it's difficult
to estimate the number of frame outputs before it is done.
As an alternative way, this CL sets 1 sec time to wait for the unbind to
be done, and then verify there's no more output after it.

Bug: 290862773
Test: UseCaseCombinationTest
Change-Id: I3aded57dc83ed931758248c12f4c4bf1b42890ac