Add benchmark test for MotionScene
MotionScene is an object that represents the animations of a
MotionLayout.

This object can be failry complex and building it may be more costly
than it should.

As such, this change adds a couple of benchmarking test for MotionScene.
With the purpose to track its performance as we work to improve on it.

Bug: 291331242
Test: n/a
Change-Id: I8ee4dd0a91504b12a6491bd4914930a65ba17a7d