Fix nested drag test flakiness
These tests were expecting a drag to result in fling settling, which
shouldn't occur.  Due to the way the drag method was implemented, it
would usually, but not always, result in settling.

This fixes the drag method to dwell at the end of the drag so that it
*never* results in settling, and changes the tests to assert that no
corresponding settling events occur.

Fixes: 291637658
Test: Ran RecyclerViewLayoutTest
Change-Id: I95bef0262f69e2aa97f2ea3adc11f06f6d9ba22a