Introduce NoOpEffects in Fragments
When there are cases where we won't run any effect, instead of completing
immediately, we should create a NoOpEffect that holds the operation and
only completes it when we run all of our other Effects.

This will allow us to determine when we want to execute operations that
may not have a real effect associated, but do not yet need to be
changed, i.e. the predictive back with no effect case.

Test: all tests pass and tested in sample app
Bug: 287137044
Change-Id: Ibbb964eee8bf15fb34578473d8f456467d47c740