Fix ChildHelper removeView re-enter removeViewIfHidden
If ItemAnimation is running, a layout pass removeAndRecycleView
the animating view, the following may happen:

- Initially the ChildHelper binary bits are 1000, 4th view
  is hidden to LM.
- LM tries to remove the 3rd view.
- ChildHelper.removeView modifies the binary bits (
  1000 -> 100 since it's going to remove 3rd View)
- ChildHelper.removeView calls Callback.removeViewAt(2)
- Callback.removeViewAt calls dispatchViewDetached
- custom adapter dispatchViewDettached clears ViewPropertyAnimator
- ItemAnimation stops
- ChildHelper.removeViewIfHidden is called
- Everything falls apart since the ChildHelper is already
  in removeView and state was changed half way. Since the
  bits was changed to 100, ChildHelper incorrectly thinks
  the 3rd View is hidden and calls removeViewAt(2) second
  time.

It appears TV apps using Presenter are all clearing the
ViewPropertyAnimator during dispatchViewAttached due to it
wants to clear the "transient state" which prevents a view
being recycled in scroll pass.  It not only stops
the custom animations, but also stops the ItemAnimation.
See the original CL that added the animate().cancel() call:
https://android.googlesource.com/platform/frameworks/support/+/5358b0ca6ed795892bd097fdf15d41fb6b1a03d2

Interstingly, there was a simlar bug fix and unit test for
removing a "fade-in" view in scroll pass back in 2015:
https://android.googlesource.com/platform/frameworks/support/+/95ff60ba852e21922aa606b5117f6a0c4f4f33d3
In 2015, predictive layout wasn't supported in leanback.
Reproduce step of this bug needs predictive layout (slide
an item into the RV). So no wonder we didn't find this crash
in 2015. This bug was first reported in 2017, after predictive
animation support was added in leanback.

Since we haven't document how the app should behave in
dispatchViewDetached, even we documented, it's very likely
developer may ignore. The better fix will be making the
recyclerview more defensive.
So the CL adds a re-entering check within ChildHelper by
tracking the calling of removeView/removeViewAt/removeViewIfHidden.

Bug: 241014260
Bug: 292114537
Bug: 157105352

Test: Created recyclerview tests based on the fake LayoutManager:
removeSlideInViewLeftToSlideOutViewAndCancelAnimationInOnDetach
removeSlideInViewLeftToSlideOutView

Test: A leanback regression test running on the real layoutmanager
was already merged (but disabled pending on the fix)
https://android-review.googlesource.com/#/q/Icf8fd1373f16d5544bc14db5721a670ba7e5b7ec

Change-Id: I42f22b0e805db12ee4cd478b3beb6b9d1a86cdaa