ScatterMap/MutableScatterMap, a fast, allocation-free hash table
ScatterMap is a container with a Map-like interface based on a flat
hash table implementation. The underlying implementation is designed to
avoid all allocations on insertion, removal, retrieval, and iteration.
Allocations may still happen on insertion when the underlying storage needs
to grow to accommodate newly added entries to the table. In addition, this
implementation minimizes memory usage by avoiding the use of separate
objects to hold key/value pairs.

This implementation makes no guarantee as to the order of the keys and
values stored, nor does it make guarantees that the order remains constant
over time.

This implementation is not thread-safe: if multiple threads access this
container concurrently, and one or more threads modify the structure of
the map (insertion or removal for instance), the calling code must provide
the appropriate synchronization. Concurrent reads are however safe.

Because of the implementation based on linear arrays, this container can
offer performance on par with LinkedHashMap/mutableMapOf().

The comparison benchmarks below show the performance of basic operations
on tables with 10, 100, 1,000, and 16,000 entries. The `read` operation
means that every entry added to the map is then read using the original
keys. The `forEach` operation visits every key/value pair. The `remove`
operation removes every entry by its original keyafter adding them all.

The benchmarks were run on a Pixel 6 running Android 13 (user).

ScatterMap

316     ns        3 allocs    insert[size=10]
72.4    ns        0 allocs    remove[size=10]
38.6    ns        0 allocs    forEach[size=10]
110     ns        0 allocs    read[size=10]

2,451   ns        4 allocs    insert[size=100]
1,105   ns        0 allocs    remove[size=100]
355     ns        0 allocs    forEach[size=100]
1,111   ns        0 allocs    read[size=100]

24,654  ns        4 allocs    insert[size=1,000]
7,840   ns        0 allocs    remove[size=1,000]
4,044   ns        0 allocs    forEach[size=1,000]
10,678  ns        0 allocs    read[size=1,000]

372,545   ns      4 allocs    insert[size=16,000]
131,530   ns      0 allocs    remove[size=16,000]
139,169   ns      0 allocs    forEach[size=16,000]
214,885   ns      0 allocs    read[size=16,000]

LinkedHashMap

366     ns       12 allocs    insert[size=10]
74.5    ns        0 allocs    remove[size=10]
132     ns        1 allocs    forEach[size=10]
77.0    ns        0 allocs    read[size=10]

4,505   ns      102 allocs    insert[size=100]
653     ns        0 allocs    remove[size=100]
1,072   ns        1 allocs    forEach[size=100]
865     ns        0 allocs    read[size=100]

43,073  ns     1003 allocs    insert[size=1,000]
6,642   ns        0 allocs    remove[size=1,000]
10,308  ns        1 allocs    forEach[size=1,000]
9,832   ns        0 allocs    read[size=1,000]

779,423 ns    16001 allocs    insert[size=16,000]
107,690 ns        0 allocs    remove[size=16,000]
168,991 ns        1 allocs    forEach[size=16,000]
191,582 ns        0 allocs    read[size=16,000]

RelNote: ScatterMap is a new allocation-free container
Test: ScatterMapTest

Change-Id: I4de4b6683ffb750925671dc4188490deeb5656c2