Fix flaking RecyclerViewLayoutTest.flingFrozen test
horizontalCounter starts at 3 and decrements once for each frame of
fling. Sometimes, it's getting down to 1 (callback called twice), but
not to zero.

While it's probably reasonable to expect 3 frames most of the time,
that's not an actual technical requirement, and a sufficiently laggy
device could likely run a scroll in 2 frames. This lowers the starting
value to 2 to reduce flakiness.  If it continues flaking, we should
lower it to 1 (AKA just swap in a boolean, probably).

Fixes: 291661857
Test: run flingFrozen and dragFrozen tests; verify that they still pass
Change-Id: I490e33a154f25b57441c680cf41181f7c7aa29c2