Resolve BasicCallControlsTest flaky tests.
In the backwards compat tests, we're seeing a lot of flakiness in the
test runs as a result of test isolation issues. Upon investigation,
the call state reported in CallSessionLegacy was not in par with the
state reflected by the platform. One such instance of dissimilarity
resulted in a call being disconnected and then set back to active
shortly after. Consequently, there were no calls being tracked in
MockInCallService even though the call hadn't been cleaned up properly,
which ends up putting Telecom into a bad state, causing subsequent tests
to fail out.

The fix is to ensure that we wait for the state changes to occur instead
of sending out successive requests to ensure the platform has enough
time to receive the request and process the necessary updates.

Bug: 290699346
Test: BasicCallControlsTest on Pixel 2 API 28, 30, 33
Change-Id: Id7057f8614fa04018e8460104413e7e1e0c64d99