Fix for not running a placement block on not placed subhierarchy
The bug was a side effect of recent aosp/2675675. First, inside of `remeasureAndRelayoutIfNeeded` we started checking for layoutNode.isPlacedByParent only in order to decide if we need to rerun the placement logic. But the parent should be not placed as well, meaning we don't need to do the work. And second issue which was unveiled by this logic is that we don't want to run the relayouts during forceMeasureTheSubtree() pass as it is too early to decide if this node will be considered placed in the end or not.

Fixes: 294073813
Test: PlacedChildTest
Change-Id: I236298836b5db9cf73d3ab704a62fa945675595b