Ensure state reads in CodepointTransformation always invalidate layout.
See comments on new test for info. This change will run the
CodepointTransformation every time the layout cache is read, which for
simple transformations is fine but might prove to be too expensive since
transformations can contain third-party code. If that's the case, see
aosp/2697358 for a proposal to cache the result of transformations.

Test: TextFieldLayoutStateCacheTest
Change-Id: Icccc98fea6a9a1bc41f905db916fc972a16d59f9