Delay alpha & scale animation to layout stage
This is to accommodate the special timing for shared element where
we won't know whether shared element animations should run until
after the composition when we find a match or fail to find a match.
By delaying alpha & scale animation to the measure/layout stage,
we are able to disable the animations if we don't find a match.

Test: All tests pass

Change-Id: I1aa098acdd61ec6a9fe26073b7932156b39cdcbf