Roll forward: fix SurfaceView crash.
This CL rolls forward a fix for a crash in StreamSharing/SurfaceView. For details, see the description of the original CL: aosp/2699899

The original CL was rolled back because it causes crash on LEGACY devices. The issue is twofold:

- SurfaceRequest created by SurfaceEdge depends on the termination of DeferrableSurface to close it. The termination can happen on any thread. In this CL, we depend on the closure to disconnect on main thread, which throws exception. The fix is that run the onProvideClosed on main thread, since closure can happen on any thead.
- Also when provider is closed, we don't care if the current edge is closed or not. We only need the producer to stop producing. So replace ::disconnect with ::disconnectWithoutCheckingClosed

There is another issue that causes black screen on LEGACY devices. Since it's unrelated to this fix, I will address it in a separate CL: aosp/2701993

Bug: 295114036
Test: manual test and ./gradlew bOS
Change-Id: I6487d5ba5df2fe7060f75d7a074d62e98d7df279