Bugfix: allow recreating ImageCapture pipeline when stream sharing is enabled.
On LEGACY devices, when the app is backgrounded, it will trigger StreamSharing's SessionConfig error callback and recreate children pipeline. In this CL, we updated the code to only check for non-virtual camera.

Also there is a quirk on certain devices that affects GL transformation(b/295352116) we will investigate it separately. This bug is not a RC blocker.

Bug: 295114036
Test: manual test and ./gradlew bOS
Change-Id: I4229c0d87f554517045eea8e40d35d46849a7083