Disabled OnApplyWindowInsetsListenerTest on API 34
Follow-up to I2878d0bceba41441099aca05c64b67d85dd9991c, which didn't
work well in tradefed.

The functionality is broken on UDC (API 34). Fixed on UDC-QPR, which is
also API 34. No reliable way to differentiate between them in a test, so
forced to disable the test for the whole API 34.

Fixes: 284406283
Test: OnApplyWindowInsetsListenerTest
Change-Id: I04c12b065168a25fa56c1bfb02b8dc414d51acaf