Reduce CaptureSessionTest#openSecondSession flake
Reduce false alerts. This CL doesn't change the major items
to be verified.

There's a small timing issue when the CaptureSessionTest is going to
verify the Session state but the state update is not completed yet.
Since the RELEASED state will occur after the onClosed is called,
the verification should be able to check the state after the
onClosed callback.

Bug: 294223171
Test: CaptureSessionTest
Change-Id: Ie42abd9a9f5dcaa17090498112a95e8a2a7c7bdf