New API CustomTabSession#setSecondaryToolbarSwipeUp
This CL introduceds a new API to CustomTabSession for the host app
to receive an Intent when the CustomTab it launched gets a swipe-up
gesture on its bottom toolbar view.

This API is migrated from Chrome codebase to androidx to allow adoption
among 3P apps.

Test: ./gradlew browser:browser:test
Relnote: Added a new API for custom tabs.
Bug: crbug/1473390
Change-Id: Id42a2a75706c6bd966f417c470ec8a710e383fe7