Optimize MeasureAndLayoutDelegate.measureOnly()
Previously we had a separate logic doing a recursive traversal through the tree to find nodes which can affect a root size. But we already have this logic in the system as there is a contract that if we call remeasure() on some LayoutNode the system will make sure all the nodes down the tree are remeasured and the size is final. This is implemented via MeasureAndLayoutDelegate.forceMeasureTheSubtree().
I also optimized this logic to do less checks on each rerun and not allocate a lambda. Plus we skip some lookahead specific logic when we know there were no lookahead invalidations scheduled.

Test: existing MeasureOnlyTest and PlacedChildTest
Fixes: 209980990
Change-Id: Ie186f67a2ec4ea245e06633ec6b860cf89b49413