Fix datastore stress tests
These tests were a bit too optimistic about how long it would take to
run 50K updates with high congestion.

I've updated the test to have smaller numbers and cleaned them to be
better with closing their scopes.

Also fixed a bug in the assertions where it assumed we'll always read
a newer value but that is not always the case with dirty reads, hence
changed the assertions to do `isAtMost` for consecutive values instead
of `isLessThan`.

To validate, i've locally made the test parameterized and run 100 times,
all runs completed in ~ 1s on an api 21 emulator.

Fixes: 270197519
Bug: 178641154
Test: SingleProcessDataStoreStressTest
Change-Id: Iba93e300a90c3c0f4d4941a514e238c4e9160dce