Add benchmarks for calls to rememberSaveable()
Added benchmarks for rememberSaveable() in preparation for adding
optimization that target it and composable functions that are
similar to it.

Added a more "focused" benchmark system that pauses the clock when
composing the overhead associated with composition and added
composition benchmarks using this system. rememberSaveable()
benchmarks only use this.

Disables source information for benchmarks that are counting
groups and slots to better match the release version of
compose applications which do not have this information.

Test: N/A - benchmark
Change-Id: I5be546a802407fbed574bf3495575af637e7a0c8