disabling unused sourcejar tasks
There are several tasks that claim to create the same source jar filepaths

We don't actually want to be able to create our source jars in multiple different ways.

This change disables some of these tasks so that it's easier to verify the appropriate way to build a source jar.

Bug: 296250264

Test: ./development/validateRefactor.sh HEAD^

Change-Id: I9f6d7ec3f167d3b4320f63c5405d4818cfa580a2