Prevent running Animations with Transitions
You should never be running Animations if there are any Transitions that
are going to be started.

We were populating the startedTransitions lists too late and we actually
no longer need it, we can just looks at whether we already have effects
on the operation.

RelNote: "Fixed a regressions where Animations were allowed to run with
Transitions."
Test: fixed tests so they actually run
Bug: 296988065

Change-Id: I59f36ad415c506f20a5cadbd82e209d6ff98dc37