KMP-ify the Shapes library
Compose needs to use the Shapes library, including adding API
and functionality that depends on it. Therefore, Shapes itself must
be a KMP library (Compose cannot use a library with Android
dependencies).

This CL makes Shapes a KMP library. It has the same Android artifact
as before (:graphics:graphics-shapes), but provides it in a way that
makes it accessible from Compose code.

Bug: 292289543
Test: Unit tests and integration test apps work

Change-Id: Ic84b4bb8bdba26879f94f74908b289f123e4e577