Introduce FallbackPolicy for backwards compat solution.
This new policy will allow developers to inform the library how to
treat the data transfer work on Android 14- devices. The default value
allows developers to handle the backwards compat solution with their own
logic. The FGS option will allow them to provide the library an FGS
which will be used to perform the work on Android 14- devices.

Bug: 280127935
Test: ./gradlew :work:work-datatransfer:assemble
Test: UserInitiatedTaskRequestTest
Change-Id: I4a8811e38c76abdab5f8f44ddaa2477f69a48812