Make TextFieldBuffer.setTextIfChanged only replace the changed region of the buffer.
This uses a really simple diffing algorithm to find the common prefix
and suffix of two CharSequences, and then only replaces the region
between them in the buffer with the region between them from the new
text.

The previous implementation did a full contentEquals and then a full
replace, which means if only a single character in a long text changed,
the entire string was copied unnecessarily, and would be recorded as a
complete replacement.

With this implementation, only that one character would be replaced, and
the change list would reflect that.

Test: updated existing tests for setTextIfChanged in TextFieldBufferTest
Test: added tests for diffing algorithm in TextFieldBufferTest
Test: added tests for new replace params in TextFieldBufferTest
Relnote: n/a
Change-Id: I0c55abbfbb66deeff52e713ba1d1f5c4c083ab86