Kill JSE sandbox on isolate heap exhaustion
The out-of-memory behavior for JavaScriptEngine is currently not well
handled by the service side and the long-term strategy for OOM handling
is undecided. This CL changes the behavior to unconditionally kill the
sandbox (from the client-side) if an isolate exceeds its memory limit,
which provides the greatest flexibility for API refinement going forward
and avoids inconsistent behaviors across different versions of the
service.

Bug: 291294904
Test: ./gradlew :javascriptengine:javascriptengine:connectedAndroidTest
Change-Id: I0038592b5a699a537154bd3d75fb8c549d825931