move datastore preferences proguard file to the core artifact
Previously, proguard file was only kept in the datastore android
preferences module which broke it for clients that want to depend
on the core artifact (android free). Given that the proguard
classes are in core, not datastore-preferences, this CL moves
them to the right artifact.

Unfortunately, AndroidX / AGP doesn't support testing proguarded
builds so I could add a test but validated with a local build.

I've also left a note to the build file of core just in case
someone adds an android target in the future.

Fixes: 253554198
Test: local test
Change-Id: Ib09ab300716ef89f125688d875cee35fccd3c5b9