Fix visible NavBackStackEntry that should be marked STARTED
When there is a floating window (i.e. dialog) on top of backStack, all visible entires under the floating window should be marked STARTED.
Previously, visible entries only included the first visible fragment and its parents. However this breaks when there are other floating windows in between the top window and first visible fragment.

Now we ensure that these floating windows (and parents that are not also the parents of resumed destinations) are also marked STARTED as well.

Test: ./gradlew navigation:navigation-runtime:cC
Bug: 289257213
Change-Id: I843df573276d8532b2ea6a04226b4b12c44ca85b