Fix testAppSearchSchema_toString
The string comparison is too restrictive and older tests will break if
we change the toString method at all. The testGetSchema* tests in
AppSearchSessionCtsTest are sufficient to ensure that every field set in
the schema builder makes it to the resulting schema.

Just do a loose check to make the test pass without changing the actual
literal string. Now we just make sure all the previous expected lines
are present in the new output.

Test: testAppSearchSchema_toString
Bug: 287720061
Change-Id: I98b443615bf286a8491f9e3f5d9f4da186275247