Support Camera2Interop callback in the Surface processing pipeline.
Camera2Interop allows the app to configure each stream with camera2 API. Stream sharing share 1 parent stream with 2 children UseCase. That's why the Camera2Interop configured on the children are not effective on the parent.

To solve this issue, we add the children's configu to the parent. In this CL we also reenabled the stress tests that are previously disabled.

Note:
- Camera2Interop itself is an experimental API and not offically supported by the CameraX team.

Bug: 297132112
Bug: 297311194
Test: manual test and ./gradlew bOS
Change-Id: Iad5a400c2bbe413439853a5b9035f799461a7461