Catch errors in the composition and send the errorUiLayout
SessionWorker wraps call to 'Composition.setContent' and
'Recomposer.runRecomposeAndApplyChanges' in try/catch to catch errors in
the composition. Also adds exception handler for the effect coroutine
context.

SessionWorker reports these errors to the session through
`Session.onCompositionError`. AppWidgetSession sends the errorUiLayout
when there is a composition error.

Add unit and device tests to make sure errors are caught correctly when
thrown from provideGlance, from directly in the composition (on initial
run and recomposition), and from SideEffect and LaunchedEffect.

Bug: none
Test: manual demo with ErrorUiAppWidget
Test: SessionWorkerTest, AppWidgetSessionTest,
 GlanceAppWidgetReceiverTest
Change-Id: I5d666ddb35e8ac5837e1f97ca8ba69a1d957e86b