Fix slide interrupted position
Fixes: 297427333

When a Slide transition was used to animate a View
into the scene, and a new Slide transition animates
the same view to be removed, the position of the
View was jumping to the wrong starting location.

This CL changes this by properly recording the
position when the animation is interrupted and
keeping it around until the animation continues.

Relnote: "Fixed Slide Transition interruption bug.
When a Slide Transition interrupted an entering
transition to remove a View, it would jump to an
incorrect position."

Test: new test

Change-Id: I946f826a338304d07e3bd7527e371ee655464e77