Use AndroidUiDispatcher.Main in Paging Compose
Ensure that PagingDataDiffer and Compose use the
same concept of what the main dispatcher is. This may
prevent cases where an extra dispatch is required
as data is delivered vs when it is available as
part of snapshot state.

Relnote: "Paging Compose now internally uses
`AndroidUiDispatcher.Main` to ensure that new
data is available in the same frame as loading
completes."
BUG: 258744755
Test: verified in sample app

Change-Id: Ia55af9ad4aeb366d7a280922563d29424619abf9