Fix issue with bouncing pages
If we're scrolling at the edge of the pager we may generate deltas that will make the pager bounce to the next page. Introduce the use of positionalThresholdFraction, the same one we use to decide if we moved enough to change targetPage. In here, it will be used to judge if the page moved enough to consider going to the next page or not.

Test: Added test to verify behavior.
Fixes: 297088699
Relnote: N/A
Change-Id: I81bcffc06dfc0ede8228ffa3e807d8a7cd0097c7