Apply snapshot before drawing
Fixes: 291647821

When a state changes during measurement/layout that affects drawing,
it must be applied in the drawing stage. This CL calls
Snapshot.sendApplyNotifications() immediately prior to drawing
so that any state modifications will properly invalidate
the drawing state and will be applied in the same frame as
drawing. It also means that the state read won't cause
multiple drawings for the same value.

Test: new test
Change-Id: Ic3043e36bf2a3ac37b1d51864be0451e9e72fa2f