Allow parent type information in GenericDocument to be stored as a synthetic property
This CL creates a new AppSearchConfig class that wraps all AppSearch configurations, and added a new flag to choose whether to store the parent types list as a meta field or as a synthetic property. The synthetic property is a workaround for platform storage to pass information to Jetpack with only existing public APIs.

Bug: 290389974
Test: Presubmit
Change-Id: I3d008c4f497d6e434b3777841893a46277e432d6