Disable YUV related image capture tests for Legacy/Limited HW level
The following core-test-app ImageCaptureTest tests are failing since aosp/2658659 has been merged, which basically adds MeteringRepeating only if surface combination is supported.
- canSaveJpegFileWithRotation_whenSoftwareJpegIsEnabled
- returnYuvImage_whenSoftwareJpegIsEnabledWithYuvBufferFormat
- returnJpegImage_whenSoftwareJpegIsEnabled
- returnYuvImage_withYuvBufferFormat

Resolution selector in camera-core suggests YUV with MAXIMUM resolution for ImageCapture only case, because it doesn't know the existence of MeteringRepeating which is added later in camera-camera2. In LEGACY/LIMITED devices, YUV with MAXIMUM size is not supported with any other surface so MeteringRepeating is no longer added and image capture does not work. But this won't happen in real devices at the moment, because YUV is added either when enabled by extension (but will have Preview added as well then) or when enabled by JpegHalCorruptImageQuirk. Only 3 devices are included in the quirk and none of them satisfy all the conditions for this issue.

Bug: 298138582
Test: core-test-app ImageCaptureTest
Change-Id: I9596eb4aa9e86529da89ab193fdba5f9b74d81bb