Fix flakiness in GraphicsSurfaceTest
Make fullscreen screenshots more robust by making sure we run
on the UI thread and perform retries when the screenshot fails
(based on heuristics from DeviceCapture.takeScreenshot()).

These tests cannot use DeviceCapture directly: DeviceCapture
will throw when a screenshot cannot be taken (returns null),
but that is expected on older API levels when a secure surface
is showing. In addition, this change provides extra
modifications to crop the screenshot to the screen bounds of
the target node.

Fixes: 299366648
Test: GraphicsSurfaceTest
RelNote: N/A
Change-Id: I1a188badff8b54986954704f7869ce1fd6f2026a