Rename of ic parameter
When using proguard-android.txt for minification, having
Resource.getIdentifier() used in the code base will cause R8 to
retain resources that match anything in the string pool. Since many
drawables in Android are prefixed with `ic_`, having a parameter named
ic can cause a substantial number of drawables to be retained despite
not being used in the resulting APK.

Although this is ideally fixed by setting shrinkMode to strict and using
the keep attribute to identify assets that may be accessed via
getIdentifier, that presents a fairly substantial burden/risk for apps
with a large number of resources and dependencies. Given the prevalence
of `ic_` assets, it seems worth doing the parameter rename on our side
for this specific case.

Bug: 258598164
Test: Manually valided no more matching of string pool constant ic
Change-Id: I428085eaadd81fd93f21d79724ac7e3669f09ceb