Update motionFromChange for drag function
Change on AwaitPointerEventScope.drag send the check for motion change to the function instead of performing it inside ot it. The name of the passed lambda was also changed to express it better. This is pre-work for Draggable2d where the check would be performed to an offset delta when implementing draggable2d.

Test: Previous tests should pass
Change-Id: If17de897d9be2d7d3b521e2145911f6cb8844fe5