Removing background padding in Modal Navigation Drawer
Modal Navigation Drawer's background content is padded to ensure
content is not accidentally lost under the drawer. EAP feedback suggests
that this is more of a hindrance as getting the background images to extend
to the start under the drawer feels hacky.
Based on this feedback, removing the additional padding and requiring the
developers to pad their content themselves.

Updated the kdocs and samples to reflect the same.

This also avoids the side-effect of the background content, occasionally
jumping as the calculated padding was animated back.

fixes: 299491364
fixes: 299998247
fixes: 276890666

Test: Added tests to ensure that background always extends to the edge under
the drawer.

Change-Id: Ied625de5fca8a190835b0aa8297e6ec60cd63855