Remove PlacementScope companion object
Fixes: 279891775

The PlacementScope companion object was used for
a singleton scope that can be used throughout Compose.
This avoids allocations, but opens Compose up to
threading issues. System UI can use multiple UI
threads, and if Compose is used in them, then
the PlacementScope singleton's state could be
mutated and accessed by multiple UIs simultaneously.

This CL removes the PlacementScope companion object
and creates a scope for each place it can be called.

Test: ran existing tests, new test
Change-Id: Iccc86b0ddad5cc72f167c328c051f24752457fe1