Fix pop signal in FragmentNavigator
We used to rely on pop signal from FragmentManager to determine whether an entry should have been popped or not. However this pop signal only represents the last operation within a transaciton where multiple operations (pop and navigate alike) are batched together. This leads us to misidentify an entry
that should be popped as one that should not be, and throw when this entry was correctly popped.

Now we track entry states based on the signal from FragmentManager as well as tracking pop operations that were initiated through NavController.

Test: ./gradlew navigation:navigation-fragment:cC
Bug: 301887045
Relnote: "Fix issue where interrupting a Fragment navigation with another navigate with popUpTo will cause FragmentNavigaor to crash."
Change-Id: I3c8483110b22e6e7d4b31cde88f2391784d9007d