Changed SchemaType and Doc id to de-flaky test testAddObserver
The observer in test maybe polluted by other tests in this class. We
are using same type(Email) and doc id(id1) everywhere in this class.
Although we re-create the search session for each test, all sessions
connected to one AppSearchImpl, and we don't have a lock in local-storage searchSessionImpl. The database maybe polluted between the searchSessionImpl puts a document and dispatches notifications. This is not a bug in product. Just make our test flaky.

Bug: 292743317
Test: GlobalSearchSessionCtsTest
Change-Id: I8693efee882e1b9298e6e5c0e282eb52e5a552f2