Update delta change collection and post slop calculation
For `awaitPointerSlopOrCancellation`, there were two ways that decide if we are looking for the delta change in one axis or not: `triggerOnMainAxisSlop` and `PointerDirectionConfig`. Here, we make changes to perform getting the delta change and calculating the `postSlopOffset` inside PointerDirectionConfig. This allows us to simplify the code and allow us to add `BidirectionalPointerDirectionConfig` that will be used in Draggable2d.

Test: Previous tests should pass
Change-Id: Id7b6c18ddb72e2b02f32f90601dcfa971343b958