Add logs for attempting to seek with only sharedElementTransition
If you set a sharedElementTransition for a FragmentTransaction and do
not also set a transition on the entering or exiting fragment,
DefaultSpecialEffectsController does not run any transitions.

This change is adding logs to indicate to devs exactly why the
transition is not running.

RelNote: "Added logs to indicate why setting a sharedElement without any
other transitions will fail to run."
Test: add exiting tests pass
Fixes: 300657793

Change-Id: Iec48ef375eb1c498bc13da82d5d8355be58ba0ff