Add a lint check for setting current state from coroutine
Since TestLifecycleOwner uses runBlocking it is not safe to set the currentState property, for instance, from the context of a coroutine, since it is thread-blocking. Instead, users should use the new setCurrentContext suspending function if using TestLifecycleOwner from something like runTest.

RelNote: "Providing a new Lint check to avoid setting the Lifecycle.State of the TestLifecycleOwner by using the currentState field when inside of a coroutine."
Test: TestLifecycleOwnerInCoroutineDetectorTest
Bug: 297880630
Change-Id: Icf728aa93a9aeca7c3cd89311289aae970d1f651