Fix popped Dialogs not marked complete
When navigating and popping Dialogs in quick succession, the popped dialogs may not have gone through composition. This means the listeners that are usually attached upon composition are not attached to these popped dialogs. As such, the popped dialogs remain in transition and get leaked into NavController.visibleEntries.

Now we make sure to mark all popped dialogs as complete when the incoming Dialog is composed.

Test: ./gradlew navigation:navigation-compose:cC
Bug: 287969970
Relnote: "Dialogs that were navigated to and dismissed in quick succession will no longer leak into the list of NavController.visibleEntries"
Change-Id: I67586735d33659e524ca7b2e4ae44b2df3494f3e