Simplify live edit slot table invalidation
This change simplifies the way the slot table is invalidated
during live edit changes. This ensures that both the children
and the slots of a group are discarded instead of just the children
as just discarding the children will not sufficiently invalidate
the slot table if remember does not create its own group which
is planned b/254480106

Includes removal of unrelated dead-code.

Bug: 254480106
Test: ./gradlew :composer:r:r:tDUT
Change-Id: I99f21182cdc59d32f562c34043c87e6b643e3be5