Ensure solo shared element transitions do not cause crashes
There is a case where if you do either an add transaction with only a
shared element transition or a replace transaction with only a shared
element transition and manipulate the visiliblity of your fragment
views, there is a NPE crash caused by having a null transition.

While it is not correct to have a shared element transition without
either an enter or exit transition with it, we should avoid crashing in
that case.

RelNote: "Fixed a `NullPointerException` caused by setting a shared
element transition and failing to set an enter/exitTransition as well."
Test: added FragmentTransitionTest
Bug: 304304379
Bug: 301332995

Change-Id: I8472b865cf93d1d3463f73942c460263f746fef5