Delegate Shader caching to a DerivedState in ShaderBrushSpan
When a custom `ShaderBrush` is used to create a Shader and a state read happens in its `createShader` function, the previous eager caching strategy prevented from redrawing the text with the new values. This change makes it more efficient to animate Brushes in spans without re-layouting the text due to Span changes.

Test: AndroidParagraphTest
Test: SpannableExtensionsTest
Change-Id: I235f0fa466d78afbddfb5db6ea7bf7bf658a4f6e