Fix flaky test testAddObserver.
The test set a schema and put a document to db. We should have 1 schema
notification 1 document notification. However we keep receiving 2 document
notifications.

Both SetSchema() and Put() will trigger dispatch notification jobs. The
first dispatch notification job triggered by SetSchema will be scheduled
in the executor. And before it get executed, the document change is made
and got picked up by the first dispatch notification job.

This shouldn't happened since we copy and clean the pending change
before schedule the dispatch notification job in the executor.
However, we only copy and clean non-empty changes, but new change could be
added to the empty-change between we schedule the job and the execution.

This change also clean the observer in the end of tests.

Bug: 292743317
Test: https://atp.googleplex.com/test_runs/348014056
Change-Id: I63a649ebe8acae82f7e3c34a3beeea63eb4108f4