Fix flaky invalidationForDrawWithCache test
Bug: 306598104

DrawModifierTest#invalidationForDrawWithCache has started flaking
on nexus4 API 21. I have a hammerhead running API 22 and was
able to reproduce the flaky test and a couple others. I suspect
that there is some timing issues, and have changed the tests
to use CountDownLatch and it appears to have cleared up the
flakiness.

Test: ran many times on hammerhead API 22
Change-Id: Ib7a2ddbe765916a79c065ba0fc81f142db97e4f0