Improve CanvasFrontBufferedRenderer rotation tests
Updated rotation tests for CanvasFrontBufferedRenderer to
attempt to rotate the device under test 3 times and settle
successfully before proceeding. If rotation fails then skip
executing the test.

Updated release logic to avoid attempting to release the
renderer if it was not created which can occur if the test
is skipped due to rotation preconditions failing.

Fixes: 307799190
Test: this
Change-Id: I74e03b7a679bdc94848c21da75bff4d5556c3445