Remove ScrollPositionUpdater logic from lazy lists
It was required earlier when we were managing key to index relation in a different manner, so we were able to update indexes for the items currently being composed. Right now it is done differently inside of LazyLayoutItemContentFactory based on  itemProvider.getKey(index) which will continue to work even without ScrollPositionUpdater.

Test: tests in lazy package
Fixes: 308610905
Change-Id: Id23fd972df3e3cbe4e0c667194c7bc173179befd