Fix off by one error when calculating selected region
When computing where the text toolbar should be placed, we read the bounding boxes of some characters at the first and last positions of selected text. The end character being used was one past the last selected character. This change corrects it to using the last character of the selection.

Additionally, the code around the bug was refactored to be more easily tested.

Test: SelectionManagerGetSelectedRegionRectTest
Change-Id: I1bb5b42bc9d19657187485b0356f894b84e06360