Pass in minApkVersion to GMS availability check
androidx.credentials_play_Services_auth module does an availability
check on GMS availability that is not passing in a minApkVersion
currently. When not passed in, it falls back to JAR build version
code determined at compile time. For 3P apps, this works as it is
derived from public released SDK. For GMS however, it is derived from
the client library at HEAD which is always higher than the one on the
device. Hence, to make it work across 3P and 1P, we need to add a
minApkVersion when checking availability to avoid the fallback.

Bug: 308719692
Test: Unit tests & sample app testing

Change-Id: Ib9f6579ee04e5100a586eae8a13bfdc42214ffea