Ensure prediction does not goes backwards while lifting
At the moment the prediction will be aborted if we detect that
the stylus is likely to be lifted (pressure < 0.1); however, this
causes the prediction to disappear and make it feel like it's
going backwards due to the pressure prediction not matching the
real result. This change will force the prediction to be generated
up to the time of the last one, since the previous prediction did
not predict the lift.

Bug: 302300930
Test: `gradlew :input:input-motionprediction:test`
Change-Id: I3c254e2e3b559f88482aee7f2e339ee81bae724b