Allow CameraEffect to configure prefered transformation.
Previously, the effects pipeline is built based on the assumption that CameraX will make all the decisions and send transformation to the image processing library via the SurfaceProcessor API. During the integration with other image processing libraries, we realized that it's hard for the image processing lib to handle arbitrary transformation. For example, MediaPipe has built in mechanism to handle the texture transformation and the Surface transformation. It's hard to disable those and apply the value provided by the effects pipeline.

To fix this, we will allow the implementer of the CameraEffect to configure the desired transformation in this CL.

- Update CameraEffect and add a new option called transformation. TRANSFORMATION_ARBITRARY is the current behavior, and TRANSFORMATION_CAMERA_AND_TARGET_ROTATION is the behavior that handles texture and surface rotation (but not cropping).
- Update StreamSharing to apply the rotation in the effect node if the option is selected. Update VirtualCamera to handle this accordingly.
- Update the test app to use TRANSFORMATION_CAMERA_AND_TARGET_ROTATION as the default option for testing.

The API is hidden for now. After it stabilizes, we will consider making it an official API.

Bug: 310260193
Test: manual test and ./gradlew bOS
Change-Id: I8f3264a68706fda82d96aef13febe76de2179af3