Unbundled SQLite driver implementation
Adds a :sqliteMultiplatform implementation that uses a compiled SQLite and targets Android (i.e. unbundled SQLite) and JVM via JNI along with iOS, Mac and Linux via statically included.

The minimal JNI binding are located in sqlite_bindings.cpp and is compiled with androidx's clang infra and included as a native lib for both Android and JVM targets.

For native targets, the actual UnbundledSQLite* classes simply type alias to NativeSQLite* but will statically include the compiled SQLite instead of dynamically linking to it. To expose NativeSQLite* classes between the libraries without opening up the API, `@RestrictTo` is used which required updating androidx.annotation to target `androidNative()`

Lastly identified an inconsistency between 0-indexed columns and 1-indexed binding args for which it is now fixed and aligns with SQLite's C API indices.

Test: UnbundledSQLiteDriverTest
Change-Id: I74a042b97f522cbc013bfe8957d3e9753614d6a6