Rotate buffer in StreamSharing.
Previously, we let the virtual camera in StreamSharing inheriting the real camera (except for Preview). We did this to be consistent with the existing behavior. Now we realized that some image processing libraries handles the image transformation, such as MediaPipe. This means that CameraX needs to handles the transformation in the sharing step too. This will also make Preview consistent with other UseCases when acting as a child of StreamSharing.

In this CL, we update StreamSharing to always rotate the buffer for its children:
- In VirtualCameraAdapter, apply the rotation in the sharing node. This include the remaining rotation degrees from upstream node, plus the delta between the parent rotation and the child rotation if there is any.

Also fixed in this CL:
- In DefaultSurfaceProcessor, fixed a bug where we apply the GL flip and the texture transformation in the wrong order.
- Fixed a bug in EffectsFragment where the video cannot be recorded twice.
- Cleanup Surface in tests

Bug: 310260193
Test: manual test and ./gradlew bOS
Change-Id: I73565589b880261cb744224de32e4bdeceed60ce