Adds ripple API to material library
The new ripple API replaces the deprecated rememberRipple API - it creates a ripple instance that will use the default theme values. ripple
does not need to be remembered, similar to modifiers - if the parameters compare equally, then the same node can be reused internally. Also
adds RippleDefaults that contain the default values used by material.

This CL also migrates all material components to use the new ripple. There is a temporary CompositionLocal,
LocalUseFallbackRippleImplementation, which can be set to true to fallback to using the old RippleTheme / rememberRipple implementation, but
this is strongly discouraged as it is much less performant than the new implementation. This fallback will be removed in the next stable
release, and exists here to aid migration.

Bug: b/298048146
Bug: b/304985887
Test: RippleTest
Relnote: "Adds new ripple API in material which replaces the deprecated rememberRipple. Also adds a temporary CompositionLocal, LocalUseFallbackRippleImplementation, to revert Material components to using the deprecated rememberRipple / RippleTheme APIs. This will be removed in the next stable release, and is only intended to be a temporary migration aid for cases where you are providing a custom RippleTheme. See developer.android.com for migration information and more background information behind this change."
Change-Id: I31629bf341de9ee401489b608135dcfdeeeb8fb6