Fix NavGraph equals to consider other graph nodes
Fixing a bug in the NavGraph equals() functions that compared the nodes
of `this` graph to itself instead of the other.

RelNote: "The `NavGraph` `equals()` function now correctly considers the
nodes of the other graph instead of just the calling one. This will
ensure that graphs that have nodes with different ids will no longer be
considered equal"
Test: added NavGraphTest
Bug: 311414915

Change-Id: I401cb2f6f71d0bc914afeea6057e0d8c34dfc840