Fix FragmentNavigator crash on pop up to start destination
FragmentNavigator crashes when navigating with pop up to initial entry (first entry in stack) because we normally don't add initial entries to pendingOps when it is getting popped, so FragmentNavigator does not expect a pop callback for it from FragmentManager.

But in the case of popping up to initial entry (inclusive), we do get a pop callback from FragmentManager. We can differentiate an isolated pop intialEntry (which should not be added to pendingOps) from a popUpTo initialEntry based on the fact that if it is popUpTo, initialEntry would have been added to pendingOps already as an incomingEntry when the entry on top of it was getting popped. So if initialEntry was already in pendingOps, we will re-add it to pendingOps with isPop = true so that FragmentNavigator knows it is getting popped.

Test: ./gradlew navigation:navigation-fragment:cC
Bug: 314652481
Change-Id: I241a93d79cf7cbc7a41e0ff92e016dd7c786cf2e