Fix for a crash in LayoutNode deactivation
We don't yet know how to reproduce it, but we have reports that sometimes the code might crash when is trying to invalidate the semantics from LayoutNode#onDeactivate. invalidateSemantics() was added here recently in beta01. We are not yet sure when the deactivation can happen on a non-attached node, but I don't see any risks in fixing it with a simple if statement, as if the node is detached, then the semantics were already updated to exclude this node.

Bug: 310732023
Test: existing tests
Change-Id: If1e48224265db6b6af6614eb2b1fdcea42fe531b