Add helpful logging for shape params in ShapeEditor
There's a button in the testapp-compose integration test app which
says "Dump to Logcat", but it only does so if an internal DEBUG flag
is true (which it is not by default). This change enables dumping to
logcat always for these params, so now you can click on the button
(renamed "Output Details") to see how the shape is configured.

Also, the output is issued to both logcat and a share intent (so it can be
copied and pasted elsewhere).

Also, the output includes a snippet of code with the given parameters
which should be pasteable to recreate the shape in an app.

Also, the output includes a simple SVG "d" string containing the initial move
(M) and subsequent cubic (C) operations.

Bug: 315531982
Test: ran testapp-compose, saw the logs, verified the SVG string is correct
Change-Id: Idc85393bb9ea987de49065b055319300f6487a59