Move coroutineScope, density and childConstraints into LazyListMeasureResult
So we don't have any public/internal setters on LazyListState and the only way to set some variable on a state is to apply a new LazyListMeasureResult. It makes it easier to reason about as we never set some temporary variables on a state from composition or measurement.
Will do the same for other lazy layouts.

Test: existing tests
Change-Id: Id33ed5b2d12078da15869f8608f903d4198c5536