Convert Consumer to Kotlin
By taking advantage of Kotlin's generics,
we ensure that the type that the Consumer is
created with (e.g., Consumer<String>) matches the
nullability of what the accept() method receives.

This means that a Consumer<String> is guaranteed
a non-null String while a Consumer<String?> may
receive a nullable String and this is now enforced
by the Kotlin compiler.

Rather than the generic name 't' for the parameter
to accept, the name 'value' was chosen to mirror
APIs in the Kotlin stdlib.

Test: ./gradlew checkApi
Relnote: "The `androidx.core.util.Consumer` class has
been rewritten in Kotlin and now enforce that the generic
type `T` now matches the nullability of what the
`accept()` method receives."

Change-Id: Ie21e0fdf7843c799a5cbcbcccde31e0770e6b3c1
Signed-off-by: tribalfs <tribalfs@gmail.com>