Only wait to finish fragments in a container
If your fragment does not have a container, it will never have an
operation enqueued into the SpecialEffectController and therefore it
will never be correctly moved to its final state.

We should only for a exit transition for fragment that are actually in a
container and will be added to the SpecialEffectController.

RelNote: "Fixed an issue where when commiting a Predictive Back gesture
on a fragment that is not in a container, that fragment would never be
destroyed."
Test: Added FragmentAniamtionTest
Fixed: 325541321

Change-Id: Ida0d901038b60ae42e42942fb16342e443247573
Signed-off-by: tribalfs <tribalfs@gmail.com>