Execute OnBackStackChangeCancelled sooner
We currently fire the
onBackStackChangedListener.onBackStackChangeCancelled callback after
the call to executePendingTransactions() instead of executing during
the executeOps like all of the other callbacks.

We now fire these callbacks in a runOnCommit Runnable, ensuring they are
called as part of executing the ops.

RelNote: "The onBackStackChangeCancelled callback on the
FragmentManagers OnBackStackChangedListener interface now fires as part
of executing operations in FragmentManager, moving it closer in line
with the timing of the onBackStackChangeCommitted callback."
Test: Modified cancel test
Bug: 332916112

Change-Id: I5ebfb12840173c4e239ff239856c2a05464648f5
Signed-off-by: tribalfs <tribalfs@gmail.com>